+++ framework/src/bcs/ADFunctionPenaltyDirichletBC.C
@@ -19,6 +19,10 @@ InputParameters
 ADFunctionPenaltyDirichletBC<compute_stage>::validParams()
 {
   InputParameters params = ADIntegratedBC<compute_stage>::validParams();
+  params.addClassDescription(
+      "Enforces a (possibly) time and space-dependent MOOSE Function Dirichlet boundary condition "
+      "in a weak sense by penalizing differences between the current "
+      "solution and the Dirichlet data.");
   params.addRequiredParam<Real>("penalty", "Penalty scalar");
   params.addRequiredParam<FunctionName>("function", "Forcing function");
 
