+++ framework/doc/content/source/bcs/ADFunctionPenaltyDirichletBC.md
@@ -1,34 +1,46 @@
-# PenaltyDirichletBC
+# ADFunctionPenaltyDirichletBC
 
-!syntax description /BCs/PenaltyDirichletBC
+!syntax description /BCs/ADFunctionPenaltyDirichletBC
 
 ## Description
 
-`PenaltyDirichletBC` is a `NodalBC` used for enforcing Dirichlet boundary conditions
-which differs from the [`DirichletBC`](/DirichletBC.md) class in the way in which it handles the enforcement.
-It is appropriate for partial differential equations (PDEs) in the form
+`ADFunctionPenaltyDirichletBC` is a generalization of [`ADPenaltyDirichletBC`](/ADPenaltyDirichletBC.md) which
+imposes a possibly temporally- and spatially-dependent value defined
+by a MOOSE [`Function`](/Functions/index.md) object on a particular set of degrees of freedom
+(DOFs) defined by the `boundary` parameter. That is, for a
+PDE of the form
 
 \begin{equation}
 \begin{aligned}
   -\nabla^2 u &= f && \quad \in \Omega \\
-  u &= g && \quad \in \partial \Omega_D \\
-  \frac{\partial u}{\partial n} &= h && \quad \in \partial \Omega_N
+  u &= g(t,\vec{x}) && \quad \in \partial \Omega_D \\
+  \frac{\partial u}{\partial n} &= h(t,\vec{x}) && \quad \in \partial \Omega_N
 \end{aligned}
 \end{equation}
 
+where $\Omega \subset \mathbb{R}^n$ is the domain, and $\partial
+\Omega = \partial \Omega_D \cup \partial \Omega_N$ is its boundary,
+a `ADFunctionPenaltyDirichletBC` object can be used to impose the
+condition (2) if the function is well-defined for $\vec{x} \in
+\partial \Omega_D$. In this case, the `function` parameter corresponds to a
+MOOSE `Function` object which represents the mathematical function
+$g(t,\vec{x})$, and the user must define one or more sidesets
+corresponding to the boundary subset $\partial \Omega_D$ via the
+`boundary` parameter.
+
 Instead of imposing the Dirichlet condition directly on the basis by replacing the
 equations associated with those degrees of freedom (DOFs) by the auxiliary equation
-$u-g=0$, the `PenaltyDirichletBC` is based on the variational statement:
+$u-g(t,\vec{x})=0$, the `ADFunctionPenaltyDirichletBC` is based on the variational statement:
 find $u \in H^1(\Omega)$ such that
 \begin{equation}
   \label{weakform}
   \int_{\Omega} \left( \nabla u \cdot \nabla v - fv \right) \,\text{d}x
   -\int_{\partial \Omega_N} hv \,\text{d}s
-  +\int_{\partial \Omega_D} \frac{1}{\epsilon} (u-g)v \,\text{d}s = 0
+  +\int_{\partial \Omega_D} \frac{1}{\epsilon} (u-g(t,\vec{x}))v \,\text{d}s = 0
 \end{equation}
 holds for every $v \in H^1(\Omega)$. In [weakform], $\epsilon
 \ll 1$ is a user-selected parameter which must be taken small enough
-to ensure that $u \approx g$ on $\partial \Omega_D$. The
+to ensure that $u \approx g(t,\vec{x})$ on $\partial \Omega_D$. The
 user-selectable class parameter `penalty` corresponds to
 $\frac{1}{\epsilon}$, and must be chosen large enough to ensure
[…]
