+++ framework/src/outputs/XMLOutput.C
@@ -56,7 +56,7 @@ XMLOutput::outputVectorPostprocessors()
   // is not required.
   vec_node.append_attribute("time") = _problem_ptr->time();
   vec_node.append_attribute("timestep") = _problem_ptr->timeStep();
-  if (_execute_enum.contains(EXEC_LINEAR))
+  if (_execute_enum.contains(EXEC_LINEAR) && !_on_nonlinear_residual)
     vec_node.append_attribute("linear_iteration") = _linear_iter;
   if (_execute_enum.contains(EXEC_NONLINEAR))
     vec_node.append_attribute("nonlinear_iteration") = _nonlinear_iter;
