+++ framework/src/meshgenerators/DistributedRectilinearMeshGenerator.C
@@ -7,7 +7,7 @@
 //* Licensed under LGPL 2.1, please see LICENSE for details
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
-#include "DistributedGeneratedMesh.h"
+#include "DistributedRectilinearMeshGenerator.h"
 
 #include "SerializerGuard.h"
 
@@ -16,8 +16,6 @@
 #include "libmesh/periodic_boundaries.h"
 #include "libmesh/periodic_boundary_base.h"
 #include "libmesh/unstructured_mesh.h"
-#include "libmesh/partitioner.h"
-#include "libmesh/metis_csr_graph.h"
 #include "libmesh/edge_edge2.h"
 #include "libmesh/edge_edge3.h"
 #include "libmesh/edge_edge4.h"
@@ -25,41 +23,32 @@
 #include "libmesh/remote_elem.h"
 #include "libmesh/face_quad4.h"
 #include "libmesh/cell_hex8.h"
-#include "libmesh/parmetis_helper.h"
+#include "libmesh/partitioner.h"
+
+// TIMPI includes
+#include "timpi/communicator.h"
+#include "timpi/parallel_sync.h"
 
 // C++ includes
 #include <cmath> // provides round, not std::round (see http://www.cplusplus.com/reference/cmath/round/)
 
-#ifdef LIBMESH_HAVE_METIS
-// MIPSPro 7.4.2 gets confused about these nested namespaces
-#ifdef __sgi
-#include <cstdarg>
-#endif
-namespace Metis
-{
-extern "C" {
-#include "libmesh/ignore_warnings.h"
-#include "metis.h"
-#include "libmesh/restore_warnings.h"
-}
-}
-#else
-#include "libmesh/sfc_partitioner.h"
-#endif
+registerMooseObject("MooseApp", DistributedRectilinearMeshGenerator);
 
-registerMooseObject("MooseApp", DistributedGeneratedMesh);
 
-template <>
 InputParameters
-validParams<DistributedGeneratedMesh>()
+DistributedRectilinearMeshGenerator::validParams()
 {
-  InputParameters params = validParams<MooseMesh>();
+  InputParameters params = MeshGenerator::validParams();
 
   params.addParam<bool>("verbose", false, "Turn on verbose printing for the mesh generation");
 
-  MooseEnum elem_types(
-      "EDGE EDGE2 EDGE3 EDGE4 QUAD QUAD4 QUAD8 QUAD9 TRI3 TRI6 HEX HEX8 HEX20 HEX27 TET4 TET10 "
-      "PRISM6 PRISM15 PRISM18 PYRAMID5 PYRAMID13 PYRAMID14"); // no default
+  MooseEnum partPackage("parmetis ptscotch hierarch", "parmetis", false);
+
+  params.addParam<MooseEnum>("part_package",
+                              partPackage,
+                             "The external package is used for partitioning the mesh via PETSc");
+
+  params.addParam<dof_id_type>("num_cores_per_compute_node", 1, "Number of cores per compute node for hierarchical partitioning");
 
   MooseEnum dims("1=1 2 3");
   params.addRequiredParam<MooseEnum>(
@@ -74,11 +63,16 @@ validParams<DistributedGeneratedMesh>()
   params.addParam<Real>("xmax", 1.0, "Upper X Coordinate of the generated mesh");
   params.addParam<Real>("ymax", 1.0, "Upper Y Coordinate of the generated mesh");
   params.addParam<Real>("zmax", 1.0, "Upper Z Coordinate of the generated mesh");
+
+  MooseEnum elem_types(
+      "EDGE EDGE2 EDGE3 EDGE4 QUAD QUAD4 QUAD8 QUAD9 TRI3 TRI6 HEX HEX8 HEX20 HEX27 TET4 TET10 "
[…]
