+++ framework/include/meshgenerators/DistributedRectilinearMeshGenerator.h
@@ -7,45 +7,266 @@
 //* Licensed under LGPL 2.1, please see LICENSE for details
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
-#ifndef DISTRIBUTEDGENERATEDMESH_H
-#define DISTRIBUTEDGENERATEDMESH_H
+#pragma once
 
-#include "MooseMesh.h"
-
-class DistributedGeneratedMesh;
-
-template <>
-InputParameters validParams<DistributedGeneratedMesh>();
+#include "MeshGenerator.h"
 
 /**
  * Mesh generated from parameters
  */
-class DistributedGeneratedMesh : public MooseMesh
+class DistributedRectilinearMeshGenerator : public MeshGenerator
 {
 public:
-  DistributedGeneratedMesh(const InputParameters & parameters);
-  DistributedGeneratedMesh(const DistributedGeneratedMesh & /* other_mesh */) = default;
+  static InputParameters validParams();
 
-  // No copy
-  DistributedGeneratedMesh & operator=(const DistributedGeneratedMesh & other_mesh) = delete;
+  DistributedRectilinearMeshGenerator(const InputParameters & parameters);
 
-  virtual MooseMesh & clone() const override;
-  virtual void buildMesh() override;
-  virtual Real getMinInDimension(unsigned int component) const override;
-  virtual Real getMaxInDimension(unsigned int component) const override;
+  std::unique_ptr<MeshBase> generate() override;
+
+  template <typename T>
+  void build_cube(UnstructuredMesh & mesh, const unsigned int nx, unsigned int ny,
+             unsigned int nz, const Real xmin, const Real xmax, const Real ymin, const Real ymax, const Real zmin, const Real zmax,
+             const ElemType type, bool verbose);
+
+  /**
+    * Get the element ID for a given hex
+    *
+    * @param nx The number of elements in the x direction
+    * @param ny The number of elements in the y direction
+    * @param i The x index of this element
+    * @param j The y index of this element
+    * @param k The z index of this element
+    * @return The ID of the i,j element
+    */
+
+    template <typename T>
+    inline dof_id_type
+    elem_id(const dof_id_type /*nx*/,
+                     const dof_id_type /*ny*/,
+                     const dof_id_type /*i*/,
+                     const dof_id_type /*j*/,
+                     const dof_id_type /*k*/)
+    {
+      mooseError("elem_id not implemented for this element type in DistributedGeneratedMesh");
+    }
+
+    /**
+      * Get the number of neighbors this element will have
+      *
+      * @param nx The number of elements in the x direction
+      * @param ny The number of elements in the y direction
+      * @param nz The number of elements in the z direction
+      * @param i The x index of this element
+      * @param j The y index of this element
+      * @param k The z index of this element
+      * @return The number of neighboring elements
+      */
+     template <typename T>
+     inline dof_id_type
+     num_neighbors(const dof_id_type /*nx*/,
+                   const dof_id_type /*ny*/,
+                   const dof_id_type /*nz*/,
+                   const dof_id_type /*i*/,
+                   const dof_id_type /*j*/,
+                   const dof_id_type /*k*/)
+     {
+       mooseError("num_neighbors not implemented for this element type in DistributedGeneratedMesh");
+     }
+
+     /**
+      * Get the IDs of the neighbors of a given element
+      *
+      * @param nx The number of elements in the x direction
+      * @param nx The number of elements in the y direction
+      * @param nz The number of elements in the z direction
+      * @param i The x index of this element
+      * @param j The y index of this element
+      * @param k The z index of this element
+      * @param neighbors This will be filled with the IDs of the two neighbors or invalid_dof_id if there
+      * is no neighbor.  THIS MUST be of size 6 BEFORE calling this function
+      */
[…]
