+++ modules/geochemistry/unit/Makefile
@@ -2,67 +2,41 @@
 ################### MOOSE Application Standard Makefile #######################
 ###############################################################################
 #
-# Required Environment variables (one of the following)
-# PACKAGES_DIR  - Location of the MOOSE redistributable package
-#
 # Optional Environment variables
-# MOOSE_DIR     - Root directory of the MOOSE project
-# FRAMEWORK_DIR - Location of the MOOSE framework
+# MOOSE_DIR        - Root directory of the MOOSE project
+# MODULE_DIR       - Location of the MOOSE modules directory
+# FRAMEWORK_DIR    - Location of the MOOSE framework
 #
 ###############################################################################
-# Use the MOOSE submodule if it exists and MOOSE_DIR is not set
-MOOSE_SUBMODULE    := $(CURDIR)/../moose
-ifneq ($(wildcard $(MOOSE_SUBMODULE)/framework/Makefile),)
-  MOOSE_DIR        ?= $(MOOSE_SUBMODULE)
-else
-  MOOSE_DIR        ?= $(shell dirname `pwd`)/../moose
-endif
+THIS_MODULE_DIR    ?= $(shell dirname `pwd`)
+MODULE_DIR         ?= $(shell dirname $(THIS_MODULE_DIR))
+MOOSE_DIR          ?= $(shell dirname $(MODULE_DIR))
 FRAMEWORK_DIR      ?= $(MOOSE_DIR)/framework
+ADDITIONAL_CPPFLAGS += -Wall -Wextra
 ###############################################################################
+
+# Extra stuff for GTEST
+ADDITIONAL_INCLUDES	:= -isystem $(FRAMEWORK_DIR)/contrib/gtest
+ADDITIONAL_LIBS 	:= $(FRAMEWORK_DIR)/contrib/gtest/libgtest.la
 
 # framework
 include $(FRAMEWORK_DIR)/build.mk
 include $(FRAMEWORK_DIR)/moose.mk
 
 ################################## MODULES ####################################
-# set desired physics modules equal to 'yes' to enable them
-CHEMICAL_REACTIONS        := no
-CONTACT                   := no
-FLUID_PROPERTIES          := no
-HEAT_CONDUCTION           := no
-MISC                      := no
-NAVIER_STOKES             := no
-PHASE_FIELD               := no
-RDG                       := no
-RICHARDS                  := no
-SOLID_MECHANICS           := no
-STOCHASTIC_TOOLS          := no
-TENSOR_MECHANICS          := no
-XFEM                      := no
-POROUS_FLOW               := no
-LEVEL_SET                 := no
+GEOCHEMISTRY := yes
 include           $(MOOSE_DIR)/modules/modules.mk
 ###############################################################################
 
-# Extra stuff for GTEST
-ADDITIONAL_INCLUDES := -I$(FRAMEWORK_DIR)/contrib/gtest
-ADDITIONAL_LIBS     := $(FRAMEWORK_DIR)/contrib/gtest/libgtest.la
-
-# dep apps
-CURRENT_DIR        := $(shell pwd)
-APPLICATION_DIR    := $(CURRENT_DIR)/..
-APPLICATION_NAME   := geochemistry
-include            $(FRAMEWORK_DIR)/app.mk
-
-APPLICATION_DIR    := $(CURRENT_DIR)
-APPLICATION_NAME   := geochemistry-unit
-BUILD_EXEC         := yes
-
+APPLICATION_DIR  := $(shell pwd)
+APPLICATION_NAME := geochemistry-unit
+BUILD_EXEC       := yes
+app_BASE_DIR     :=
[…]
