### 2020-02-26 f33abe2
PetscExternalPartitioner: create an graph partitioning interface
So that we can use it in other places

Closes #11485
