+++ framework/src/partitioner/PetscExternalPartitioner.C
@@ -34,6 +34,8 @@ PetscExternalPartitioner::validParams()
                              partPackage,
                              "The external package is used for partitioning the mesh via PETSc");
 
+  params.addParam<dof_id_type>("num_cores_per_compute_node", 1, "Number of cores per compute node for hierarchical partitioning");
+
   params.addParam<bool>("apply_element_weight",
                         false,
                         "Indicate if we are going to apply element weights to partitioners");
@@ -51,7 +53,8 @@ PetscExternalPartitioner::PetscExternalPartitioner(const InputParameters & param
   : MoosePartitioner(params),
     _part_package(params.get<MooseEnum>("part_package")),
     _apply_element_weight(params.get<bool>("apply_element_weight")),
-    _apply_side_weight(params.get<bool>("apply_side_weight"))
+    _apply_side_weight(params.get<bool>("apply_side_weight")),
+    _num_parts_per_compute_node(params.get<dof_id_type>("num_cores_per_compute_node"))
 {
   if ((_apply_element_weight || _apply_side_weight) &&
       (_part_package == "chaco" || _part_package == "party"))
@@ -95,38 +98,23 @@ PetscExternalPartitioner::partition(MeshBase & mesh, const unsigned int n_parts)
 void
 PetscExternalPartitioner::_do_partition(MeshBase & mesh, const unsigned int n_parts)
 {
-#ifdef LIBMESH_HAVE_PETSC
-  // construct a dual graph
-  Mat dual;
-  PetscInt *i, *j, *values, *elem_weights, nrows, nj, ncols, local_elem_id;
-  const PetscInt * parts;
-  unsigned int side;
-  MatPartitioning part;
-  IS is;
-  // Let us check PETSc return code
-  PetscErrorCode ierr;
+  dof_id_type num_edges, num_local_elems, local_elem_id, nj, side;
+  std::vector<dof_id_type> side_weights;
+  std::vector<dof_id_type> elem_weights;
 
-  i = 0;
-  j = 0;
-  values = 0;
-  elem_weights = 0;
-
+  // Call libmesh to build the dual graph of mesh
   build_graph(mesh);
-  nrows = _dual_graph.size();
-  ierr = PetscCalloc1(nrows + 1, &i);
-  CHKERRABORT(mesh.comm().get(), ierr);
+  num_local_elems = _dual_graph.size();
+
+  elem_weights.clear();
   if (_apply_element_weight)
-  {
-    ierr = PetscCalloc1(nrows + 1, &elem_weights);
-    CHKERRABORT(mesh.comm().get(), ierr);
-  }
+    elem_weights.resize(num_local_elems);
 
-  // Set graph offsets and
+  num_edges = 0;
   // compute element weight
-  for (PetscInt k = 0; k < nrows; k++)
+  for (dof_id_type k = 0; k < num_local_elems; k++)
   {
-    i[k + 1] = i[k] + _dual_graph[k].size();
-
+    num_edges += _dual_graph[k].size();
     if (_apply_element_weight)
     {
       // Get the original element
@@ -138,16 +126,10 @@ PetscExternalPartitioner::_do_partition(MeshBase & mesh, const unsigned int n_pa
     }
   }
 
-  // Graph adjacency
-  ierr = PetscCalloc1(i[nrows], &j);
-  CHKERRABORT(mesh.comm().get(), ierr);
-
+   side_weights.clear();
   // Edge weights represent the communication
   if (_apply_side_weight)
-  {
-    ierr = PetscCalloc1(i[nrows], &values);
[…]
