### 2020-02-26 #2118 @WilkAndy
Why doesn't the paper compile?  It's clearly in the repository:

https://github.com/idaholab/moose/blob/next/modules/porous_flow/paper.md

and @whedon can compile it from my own repo:

https://github.com/WilkAndy/PorousFlowDescription

I submitted an issue at  openjournals/whedon#63  just in case it's not my fault.
