### 2020-02-26 3104810
Added check functionality for optional mutually exclusive parameters
Closes #19804
