### 2020-02-26 52011dc
Don't enable extra unit test build warnings by default
