+++ modules/geochemistry/unit/Makefile
@@ -12,7 +12,6 @@ THIS_MODULE_DIR    ?= $(shell dirname `pwd`)
 MODULE_DIR         ?= $(shell dirname $(THIS_MODULE_DIR))
 MOOSE_DIR          ?= $(shell dirname $(MODULE_DIR))
 FRAMEWORK_DIR      ?= $(MOOSE_DIR)/framework
-ADDITIONAL_CPPFLAGS += -Wall -Wextra
 ###############################################################################
 
 # Extra stuff for GTEST
