### 2020-02-26 8042cc4
Update MooseDocs build script and config.yml file
