+++ modules/geochemistry/doc/config.yml
@@ -1,21 +1,23 @@
 Content:
-    geochemistry:
-        root_dir: ${ROOT_DIR}/doc/content
-    moose:
-        root_dir: ${MOOSE_DIR}/framework/doc/content
-        content:
-            - js/*
-            - css/*
-            - contrib/**
-            - media/**
+    - framework/doc/content
+    - modules/geochemistry/doc/content
 
 Renderer:
     type: MooseDocs.base.MaterializeRenderer
-
 Extensions:
     MooseDocs.extensions.navigation:
-        name: geochemistry
+        name: MOOSE Geochemistry Module
+        repo: https://github.com/idaholab/moose
+        home: /modules/geochemistry
     MooseDocs.extensions.appsyntax:
-        executable: ${ROOT_DIR}
+        executable: ${MOOSE_DIR}/modules/geochemistry
+        hide: !include framework/doc/hidden.yml
+        remove: !include framework/doc/remove.yml
         includes:
-            - include
+            - framework/include
+            - modules/geochemistry/include
+    MooseDocs.extensions.common:
+        shortcuts: !include framework/doc/globals.yml
+    MooseDocs.extensions.acronym:
+        acronyms: !include framework/doc/acronyms.yml
+    MooseDocs.extensions.sqa: disable
