+++ framework/src/interfacekernels/InterfaceKernelBase.C
@@ -99,6 +99,7 @@ InterfaceKernelBase::InterfaceKernelBase(const InputParameters & parameters)
     _current_elem(_assembly.elem()),
     _current_elem_volume(_assembly.elemVolume()),
     _neighbor_elem(_assembly.neighbor()),
+    _neighbor_elem_volume(_assembly.neighborVolume()),
     _current_side(_assembly.side()),
     _current_side_elem(_assembly.sideElem()),
     _current_side_volume(_assembly.sideElemVolume()),
