+++ framework/src/dgkernels/ADDGDiffusion.C
@@ -48,8 +48,8 @@ ADDGDiffusion<compute_stage>::computeQpResidual(Moose::DGResidualType type)
   ADReal r = 0;
 
   const unsigned int elem_b_order = _var.order();
-  const double h_elem =
-      _current_elem->volume() / _current_side_elem->volume() * 1. / Utility::pow<2>(elem_b_order);
+  const Real h_elem =
+      this->_current_elem_volume / this->_current_side_volume * 1. / Utility::pow<2>(elem_b_order);
 
   switch (type)
   {
