+++ modules/thm/doc/hidden.yml
@@ -264,6 +264,7 @@
 - /Materials/MultiComponentFluidPropertiesMaterialPT
 - /Materials/PressureRelaxationRateMaterial
 - /Materials/RDG7EqnNCGMaterial
+- /Materials/ReynoldsNumberMaterial
 - /Materials/SaturationTemperatureMaterial
 - /Materials/SpecificInterfacialArea7EqnMaterial
 - /Materials/SpecificInterfacialAreaSimpleMaterial
