### 2020-02-26 771e8de
Adding convective heat transfer coefficient material
Refs #19682
