### 2020-02-26 0630da3
Restore ADDITIONAL_CPPFLAGS to unit test Makefile for consistency with other modules
