### 2020-02-26 9b9cd08
Modify SampleTest.C to avoid unused value warning
