+++ modules/geochemistry/unit/src/SampleTest.C
@@ -18,8 +18,8 @@ TEST(MySampleTests, descriptiveTestName)
   // compare equality and immediately terminate this test if it fails
   // ASSERT_EQ(2, 1);
 
-  // this won't run if you uncomment the above test because above assert will fail
-  ASSERT_NO_THROW(1 + 1);
+  // if you uncomment this test as well as the above test, this test will not run because above assert will fail
+  // ASSERT_NO_THROW(1 + 1);
 
   // for a complete list of assertions and for more unit testing documentation see:
   // https://github.com/google/googletest/blob/master/googletest/docs/Primer.md
