+++ framework/include/loops/ComputeFVFaceResidualsThread.h
@@ -42,48 +42,23 @@ struct OnScopeExit
   }
 };
 
-/**
- * Base class for assembly-like calculations.
- */
 template <typename RangeType>
-class ComputeFVFaceResidualsThread
+class ThreadedFaceLoop
 {
 public:
-  ComputeFVFaceResidualsThread(FEProblemBase & fe_problem, const std::set<TagID> & tags);
+  ThreadedFaceLoop(FEProblemBase & fe_problem, const std::set<TagID> & tags);
 
-  ComputeFVFaceResidualsThread(ComputeFVFaceResidualsThread & x, Threads::split split);
+  ThreadedFaceLoop(ThreadedFaceLoop & x, Threads::split split);
 
-  virtual ~ComputeFVFaceResidualsThread();
+  virtual ~ThreadedFaceLoop();
 
   virtual void operator()(const RangeType & range, bool bypass_threading = false);
 
-  void join(const ComputeFVFaceResidualsThread & /*y*/){};
+  void join(const ThreadedFaceLoop & /*y*/){};
 
-  /**
-   * Called before the element range loop
-   */
-  virtual void pre();
+  virtual void onFace(const FaceInfo & fi) = 0;
 
-  /**
-   * Called after the element range loop
-   */
-  virtual void post();
-
-  /**
-   * Called before the boundary assembly
-   *
-   * @param elem - The element we are checking is on the boundary.
-   * @param side - The side of the element in question.
-   * @param bnd_id - ID of the boundary we are at
-   */
-  virtual void onFace(const FaceInfo & fi);
-
-  /**
-   * Called before the boundary assembly
-   *
-   * @param bnd_id - ID of the boundary we are at
-   */
-  virtual void onBoundary(const FaceInfo & fi, BoundaryID boundary);
+  virtual void onBoundary(const FaceInfo & fi, BoundaryID boundary) = 0;
 
   /**
    * Called every time the current subdomain changes (i.e. the subdomain of _this_ element
@@ -91,7 +66,7 @@ public:
    * You might think that you can do some expensive stuff in here and get away with it...
    * but there are applications that have TONS of subdomains....
    */
-  virtual void subdomainChanged();
+  virtual void subdomainChanged(){};
 
   /**
    * Called every time the neighbor subdomain changes (i.e. the subdomain of _this_ neighbor
@@ -107,7 +82,7 @@ public:
    */
   virtual void caughtMooseException(MooseException &){};
 
-private:
+protected:
   FEProblemBase & _fe_problem;
   MooseMesh & _mesh;
   const std::set<TagID> & _tags;
@@ -124,45 +99,44 @@ private:
 
   /// The subdomain for the last neighbor
   SubdomainID _old_neighbor_subdomain;
-
-  OnScopeExit reinitVariables(const FaceInfo & fi);
-
-  std::set<MooseVariableFVBase *> _needed_fv_vars;
 };
 
 template <typename RangeType>
-ComputeFVFaceResidualsThread<RangeType>::ComputeFVFaceResidualsThread(FEProblemBase & fe_problem,
-                                                                      const std::set<TagID> & tags)
+ThreadedFaceLoop<RangeType>::ThreadedFaceLoop(FEProblemBase & fe_problem,
+                                              const std::set<TagID> & tags)
   : _fe_problem(fe_problem), _mesh(fe_problem.mesh()), _tags(tags)
 {
 }
 
 template <typename RangeType>
-ComputeFVFaceResidualsThread<RangeType>::ComputeFVFaceResidualsThread(
-    ComputeFVFaceResidualsThread & x, Threads::split /*split*/)
[…]
