### 2020-02-26 2998710
Fixup stork module MooseDocs
- Renames previous `moosedocs.py` to `moosedocs.py.app`
- Renames previous `config.yml` to `config.yml.app`
- Adds `moosedocs.py.module`
- Adds `config.yml.module`
- Modifies `stork.sh` to properly select between new files

Closes #14786
