+++ modules/thm/doc/content/source/materials/PrandtlNumberMaterial.md
@@ -1,0 +1,18 @@
+<!-- MOOSE Documentation Stub: Remove this when content is added. -->
+
+# PrandtlNumberMaterial
+
+!alert construction title=Undocumented Class
+The PrandtlNumberMaterial has not been documented. The content contained on this page includes the
+typical automatic documentation associated with a MooseObject; however, what is contained is
+ultimately determined by what is necessary to make the documentation clear for users.
+
+!syntax description /Materials/PrandtlNumberMaterial
+
+!syntax parameters /Materials/PrandtlNumberMaterial
+
+!syntax inputs /Materials/PrandtlNumberMaterial
+
+!syntax children /Materials/PrandtlNumberMaterial
+
+!bibtex bibliography
