### 2020-02-26 39ee76d
Set proper permissions for moosedocs.py in stork script
