+++ scripts/stork.sh
@@ -97,6 +97,7 @@ mv "$dir/include/base/${srcname}App.h" "$dir/include/base/${dstname}App.h"
 mv "$dir/test/include/base/${srcname}TestApp.h" "$dir/test/include/base/${dstname}TestApp.h"
 mv "$dir/doc/config.yml.${kind}" "$dir/doc/config.yml"
 mv "$dir/doc/moosedocs.py.${kind}" "$dir/doc/moosedocs.py"
+chmod a+x "$dir/doc/moosedocs.py"
 chmod a+x "$dir/run_tests"
 
 # remove unnecessary files
