### 2020-02-26 60ac65c
Use new partition interface
