+++ framework/src/meshgenerators/DistributedRectilinearMeshGenerator.C
@@ -8,6 +8,7 @@
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
 #include "DistributedRectilinearMeshGenerator.h"
+#include "PetscExternalPartitioner.h"
 
 #include "SerializerGuard.h"
 
@@ -38,17 +39,10 @@ registerMooseObject("MooseApp", DistributedRectilinearMeshGenerator);
 InputParameters
 DistributedRectilinearMeshGenerator::validParams()
 {
-  InputParameters params = MeshGenerator::validParams();
+  InputParameters params = PetscExternalPartitioner::validParams();
+  params += MeshGenerator::validParams();
 
   params.addParam<bool>("verbose", false, "Turn on verbose printing for the mesh generation");
-
-  MooseEnum partPackage("parmetis ptscotch hierarch", "parmetis", false);
-
-  params.addParam<MooseEnum>("part_package",
-                              partPackage,
-                             "The external package is used for partitioning the mesh via PETSc");
-
-  params.addParam<dof_id_type>("num_cores_per_compute_node", 1, "Number of cores per compute node for hierarchical partitioning");
 
   MooseEnum dims("1=1 2 3");
   params.addRequiredParam<MooseEnum>(
@@ -113,7 +107,8 @@ DistributedRectilinearMeshGenerator::DistributedRectilinearMeshGenerator(const I
     _bias_x(getParam<Real>("bias_x")),
     _bias_y(getParam<Real>("bias_y")),
     _bias_z(getParam<Real>("bias_z")),
-    _part_package(getParam<MooseEnum>("part_package"))
+    _part_package(getParam<MooseEnum>("part_package")),
+    _num_parts_per_compute_node(getParam<dof_id_type>("num_cores_per_compute_node"))
 {
 }
 
@@ -233,10 +228,12 @@ DistributedRectilinearMeshGenerator::add_element<Edge2>(const dof_id_type nx,
 
   auto node0_ptr = mesh.add_point(Point(static_cast<Real>(node_offset) / nx, 0, 0), node_offset);
   node0_ptr->set_unique_id() = node_offset;
+  node0_ptr->processor_id()  = pid;
 
   auto node1_ptr =
       mesh.add_point(Point(static_cast<Real>(node_offset + 1) / nx, 0, 0), node_offset + 1);
   node1_ptr->set_unique_id() = node_offset + 1;
+  node1_ptr->processor_id()  = pid;
 
   if (verbose)
     Moose::out << "Adding elem: " << elem_id << std::endl;
@@ -433,24 +430,28 @@ DistributedRectilinearMeshGenerator::add_element<Quad4>(const dof_id_type nx,
   auto node0_ptr = mesh.add_point(Point(static_cast<Real>(i) / nx, static_cast<Real>(j) / ny, 0),
                                   node_id<Quad4>(type, nx, 0, i, j, 0));
   node0_ptr->set_unique_id() = node_id<Quad4>(type, nx, 0, i, j, 0);
+  node0_ptr->processor_id() = pid;
 
   // Bottom Right
   auto node1_ptr =
       mesh.add_point(Point(static_cast<Real>(i + 1) / nx, static_cast<Real>(j) / ny, 0),
                      node_id<Quad4>(type, nx, 0, i + 1, j, 0));
   node1_ptr->set_unique_id() = node_id<Quad4>(type, nx, 0, i + 1, j, 0);
+  node1_ptr->processor_id() = pid;
 
   // Top Right
   auto node2_ptr =
[…]
