### 2020-02-26 52887b7
Use new partition interface
