### 2020-02-26 6c50413
geochemistry theory doco added.  Refs #14789
