+++ modules/chemical_reactions/doc/content/modules/chemical_reactions/equilibrium.md
@@ -1,0 +1,277 @@
+# Equilibrium equations and solution method
+
+Author: Andy Wilkins
+
+Notation and definitions are described in [nomenclature.md].
+
+This page follows [!cite](bethke_2007).
+
+## Notation
+
+Introduce the following:
+
+- $A_{w}$: water.
+- $A_{i}$: aqueous [basis species](basis.md).  This is a label, not a quantity.  for example, one of the $A_{i}$ might be Na$^{+}$.
+- $A_{j}$: other aqueous species, which are called the "secondary" species.
+- $A_{k}$: minerals in the equilibrium system of reactions that appear in the [basis](basis.md).
+- $A_{l}$: all minerals, even those that do not exist in the equilibrium system.  This might include minerals whose reactions are kinetically controlled.
+- $A_{m}$: gases of known fugacity that appear in the [basis](basis.md).
+- $A_{n}$: all gases.
+- $A_{q}$: sorbed species.  These do not get transported --- they are surface complexes.
+- $A_{p}$: sorbing sites.  In the Langmuir approch and ion-exchange approach there is just one of these, and all sorbing species compete to sorb.  In the "surface complexation" approach, there are an equal number of $A_{p}$ as $A_{q}$.
+
+## Basis
+
+Note that the index $i$, $j$, etc, gives meaning to the symbol that it is indexing.  Although unusual in mathematics literature, this is the common convention in geochemistry.  There are $1+N_{i}+N_{k}+N_{m}$ unknowns and the basis is
+\begin{equation}
+\mathrm{basis} = (A_{w}, A_{i}, A_{k}, A_{m}, A_{p}) \ .
+\end{equation}
+Being a basis, all other other components can be expressed in terms of it.
+
+## Secondary species
+
+All secondary species are in equilibrium with the basis.  They are not sorbing, so
+\begin{equation}
+A_{j} \rightleftharpoons \nu_{wj}A_{w} + \sum_{i}\nu_{ij}A_{i} + \sum_{k}\nu_{kj}A_{k} + \sum_{m}\nu_{mj}A_{m} \ .
+\end{equation}
+This equation is specified in the [database](database.md) with the $\nu$ being the stoichiometric reaction coefficients.  The database specifies the equilibrium constant, $K_{j}$, and [mass-action equilibrium](equilibrium_reactions.md) reads
+\begin{equation}
+K_{j} = \frac{a_{w}^{\nu_{wj}}\cdot \prod_{i}(\gamma_{i}m_{i})^{\nu_{ij}} \cdot \prod_{k}a_{k}^{\nu_{kj}} \cdot \prod_{m}f_{m}^{\nu_{mj}}}{\gamma_{j}m_{j}} \ .
+\end{equation}
+Here:
+
+- $a_{w}$ is the [activity of water](activity_coefficients.md)
+- $\gamma_{i}m_{i}$ is the [activity](activity_coefficients.md) of the primary species $i$
+- $a_{k}$ is the activity of mineral $k$, which is [assumed to be unity](activity_coefficients.md) but is written here for clarity
+- $f_{m}$ is the known gas [fugacity](fugacity.md)
+- $m_{j}$ is the molality of the secondary species
+- $\gamma_{j}$ is the [activity coefficient](activity_coefficients.md) of the secondary species.
+- $m$ denotes [molality](nomenclature.md)
+
[…]
