+++ modules/chemical_reactions/doc/content/modules/chemical_reactions/activity_coefficients.md
@@ -1,0 +1,116 @@
+# Activity coefficients
+
+Author: Andy Wilkins
+
+!alert note
+Only the Debye-Huckel B-dot model along with the formulae below for neutral species and water are coded into the `geochemistry` module
+
+Notation and definitions are described in [nomenclature.md].
+
+The material below is taken largely from Chapter 8 of [!cite](bethke_2007).
+
+Recall that the activity, $a$, of a species is defined in terms of the chemical potential, $\mu$
+\begin{equation}
+\mu = mu^{0} + RT \log a \ .
+\end{equation}
+For minerals in mineralisation reactions, $a=1$, but for species in equilibrium reactions, the activity is a dimensionless measure of effective concentration of a constituent in a solution
+\begin{equation}
+a = \gamma m/m_{\mathrm{ref}} \ ,
+\end{equation}
+where $\gamma$ is the activity coefficient, $m$ is the molality, and $m_{\mathrm{ref}}=1\,$mol.kg$^{-1}$.  It is the purpose of this page to describe models for computing the activity coefficient $\gamma$.
+
+!alert warning
+Explain how to compute these!!
+
+## Ionic strength and stoichiometric ionic strength
+
+Denote the "ionic strength" of the solution by $I$.  Its units are mol.kg$^{-1}$:
+\begin{equation}
+I = \frac{1}{2}\sum_{i}m_{i}z_{i}^{2} \ .
+\end{equation}
+Here, the sum runs over the free ions in the solution, so neutral complexes are not counted.
+
+Denote the "stoichiometric ionic strength" of the solution by $I_{s}$.  Its units are mol.kg$^{-1}$:
+\begin{equation}
+I_{s} = \frac{1}{2}\sum_{i}m_{i}z_{i}^{2} \ .
+\end{equation}
+Here, complete dissociation of complexes is assumed, and then the sum runs over all ions in the hypothetical solution.
+
+## Debye-Huckel B-dot model
+
+### Electrically-charged species
+
+Here (see Eqn(8.5) of [!cite](bethke_2007))
+\begin{equation}
+\log_{10}\gamma = -\frac{Az^{2}\sqrt{I}}{1 + \mathring{a}B\sqrt{I}} + \dot{B}I \ .
+\end{equation}
+In this equation
+
+- $z$ is the charge number of the species (if $z=0$ then see below).
+- $I$ is the ionic strength of the solution
+- $\mathring{a}$ is the ion size parameter, measured in Angstrom ($\mathring{A}$) and given for all primary basis species, redox couples and secondary species in the [chemical database](database.md)
+- Coefficients $A$, $B$ and $\dot{B}$ depend on temperature, and are given in the [chemical database](database.md).  At at 25$^{\circ}$C: $A=0.5092\,$mol$^{-1/2}$.kg$^{1/2}$, $B=0.3283\,$mol$^{-1/2}$.kg$^{1/2}$.$\mathring{A}^{-1}$ and $\dot{B}=0.0410\,$mol$^{-1}$.kg.
+
+### Neutral species
+
+When $z=0$, the activity coefficients may be computed using (Eqn(8.6) of [!cite](bethke_2007))
+\begin{equation}
+\log_{10}\gamma = aI + bI^{2} + cI^{3} + dI^{4} \ .
+\end{equation}
[…]
