+++ modules/chemical_reactions/doc/content/modules/chemical_reactions/transport.md
@@ -1,0 +1,165 @@
+# Transport
+
+Author: Andy Wilkins
+
+Only rudimentary transport is available as part of the `geochemistry` module.  Coupling with the `PorousFlow` module allows users to access more advanced features such as:
+
+- pressure and temperature that tightly couple with fluid flows, instead of being specified by uncoupled dynamics
+- densities, viscosities, etc, that depend on solute concentrations, temperature and pressure, instead of being constant
+- porosity and permeability that change with precipitation and dissolution
+- multiphase flow
+- coupling with geomechanics
+- sophisticated numerical stabilization
+
+Notation and definitions are described in [nomenclature.md].
+
+## Volumes, concentrations and mass conservation.
+
+Traditional formulations of geochemistry, such as presented in the pages on [equilibrium](equilibrium.md) and [kinetics](kinetics.md), consider just the aqueous solution and a vague notion of immobile minerals and surfaces for surface complexation.  When considering transport through a porous medium, these vague notions become concrete.
+
+Define the concentration, $C$ of a component (water, basis aqueous species, secondary aqueous species, etc) as
+\begin{equation}
+C = \frac{M}{V_{\mathrm{solution}}} \ ,
+\end{equation}
+where $V_{\mathrm{solution}}$ \[m$^{3}$\] is the total volume of the aqueous solution.  Hence the units of $C$ are mol.m$^{-3}$.  Now introduce the porous skeleton.  The aqueous solution inhabits its void space.  Introduce the porosity,
+\begin{equation}
+\phi = \frac{V_{\mathrm{void}}}{V_{\mathrm{solid}} + V_{\mathrm{void}}} = \frac{V_{\mathrm{void}}}{V} \ .
+\end{equation}
+Here, $V_{\mathrm{solid}}$ is the volume of the rock grains of the porous skeleton: it is the volume of the porous skeleton if it were crushed so that it contained no voids.  $V$ is a spatial volume containing the porous skeleton.  Hence, the quantity
+\begin{equation}
+\int_{V} \phi C \ ,
+\end{equation}
+is the total number of moles of aqueous solution within a volume $V$ of porous material.
+
+The dynamics of $\phi$, if any, are not provided by the `geochemistry` module.
+
+The continuity equation (mass conservation) reads
+\begin{equation}
+\frac{\mathrm{d}}{\mathrm{d} t} \int_{V}\phi C = \int_{\partial V}n.F + \int_{V} \phi Q
+\end{equation}
+or
+\begin{equation}
+\frac{\partial}{\partial t}(\phi C) + \nabla\cdot F = \phi Q \ .
+\end{equation}
+Here:
+
+- $t$ \[s\] is time
+- $\nabla$ \[m$^{-1}$\] is the vector spatial derivatives
+- $F$ \[mol.s$^{-1}$.m$^{-2}$\] is the fluid flux vector, so $\int_{\partial V}n.F$ is the flux \[mol.s$^{-1}$\] into the volume $V$ through its boundary.
[…]

[…]
