### 2020-02-26 c8ff5b3
loop: call feproblem.prepare - add todos about assembly prepare/reinit
