+++ modules/phase_field/src/vectorpostprocessors/FeatureVolumeVectorPostprocessor.C
@@ -224,7 +224,7 @@ FeatureVolumeVectorPostprocessor::accumulateVolumes(
 
   // Accumulate the entire element volume into the dominant feature. Do not use the integral value
   if (_single_feature_per_elem && dominant_feature_id != FeatureFloodCount::invalid_id)
-    _feature_volumes[dominant_feature_id] += elem->volume();
+    _feature_volumes[dominant_feature_id] += _assembly.elementVolume(elem);
 }
 
 Real
@@ -274,7 +274,7 @@ FeatureVolumeVectorPostprocessor::accumulateBoundaryFaces(
 
   // Accumulate the boundary area/length into the dominant feature. Do not use the integral value
   if (_single_feature_per_elem && dominant_feature_id != FeatureFloodCount::invalid_id)
-    _feature_volumes[dominant_feature_id] += elem->side_ptr(side)->volume();
+    _feature_volumes[dominant_feature_id] += _assembly.elementVolume(elem->side_ptr(side).get());
 }
 
 Real
