### 2020-02-27 3e88b03
Change `geochemistry` to `Geochemistry` in App and TestApp
