+++ modules/geochemistry/src/main.C
@@ -7,7 +7,7 @@
 //* Licensed under LGPL 2.1, please see LICENSE for details
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
-#include "geochemistryTestApp.h"
+#include "GeochemistryTestApp.h"
 #include "MooseInit.h"
 #include "Moose.h"
 #include "MooseApp.h"
@@ -24,10 +24,10 @@ main(int argc, char * argv[])
   MooseInit init(argc, argv);
 
   // Register this application's MooseApp and any it depends on
-  geochemistryTestApp::registerApps();
+  GeochemistryTestApp::registerApps();
 
   // Create an instance of the application and store it in a smart pointer for easy cleanup
-  std::shared_ptr<MooseApp> app = AppFactory::createAppShared("geochemistryTestApp", argc, argv);
+  std::shared_ptr<MooseApp> app = AppFactory::createAppShared("GeochemistryTestApp", argc, argv);
 
   // Execute the application
   app->run();
