+++ modules/geochemistry/doc/content/index.md
@@ -1,3 +1,3 @@
 !config navigation breadcrumbs=False scrollspy=False
 
-# geochemistryApp
+# GeochemistryApp
