### 2020-02-27 eb96e38
Allow setting more font options
Allow users to set:
- bold type
- italic type
- font family ('Arial', 'Times' and 'Courier')

refs #8771
