+++ python/chigger/utils/FontOptions.py
@@ -24,6 +24,9 @@ def get_options():
     opt.add('text_opacity', 1, "The text opacity.", vtype=float)
     opt.add('text', None, "The text to display.", vtype=str)
     opt.add('font_size', 24, "The text font size.", vtype=int)
+    opt.add('font_family', None, "The font family.", vtype=str)
+    opt.add('bold', False, "Enable/disable text bolding.", vtype=bool)
+    opt.add('italic', False, "Enable/disable text italic.", vtype=bool)
     return opt
 
 
@@ -53,5 +56,14 @@ def set_options(tprop, options):
     if options.isOptionValid('text_opacity'):
         tprop.SetOpacity(options['text_opacity'])
 
+    if options.isOptionValid('font_family'):
+        tprop.SetFontFamilyAsString(options['font_family'])
+
     if options.isOptionValid('font_size'):
         tprop.SetFontSize(options['font_size'])
+
+    if options.isOptionValid('bold'):
+        tprop.SetBold(options['bold'])
+
+    if options.isOptionValid('italic'):
+        tprop.SetItalic(options['italic'])
