### 2020-02-27 6f971b2
Remove Dirac kernel for contact and deprecate the use of 'system' in the Contact action. Refs. #7283
