### 2020-02-27 1732cb2
Added missing ghost elements
Add all elements that connect to at least one of local vertices

Closes #11485
