+++ framework/include/meshgenerators/DistributedRectilinearMeshGenerator.h
@@ -24,236 +24,233 @@ public:
   std::unique_ptr<MeshBase> generate() override;
 
   template <typename T>
-  void build_cube(UnstructuredMesh & mesh, const unsigned int nx, unsigned int ny,
-             unsigned int nz, const Real xmin, const Real xmax, const Real ymin, const Real ymax, const Real zmin, const Real zmax,
-             const ElemType type, bool verbose);
+  void build_cube(UnstructuredMesh & mesh,
+                  const unsigned int nx,
+                  unsigned int ny,
+                  unsigned int nz,
+                  const Real xmin,
+                  const Real xmax,
+                  const Real ymin,
+                  const Real ymax,
+                  const Real zmin,
+                  const Real zmax,
+                  const ElemType type,
+                  bool verbose);
 
   /**
-    * Get the element ID for a given hex
-    *
-    * @param nx The number of elements in the x direction
-    * @param ny The number of elements in the y direction
-    * @param i The x index of this element
-    * @param j The y index of this element
-    * @param k The z index of this element
-    * @return The ID of the i,j element
-    */
+   * Get the element ID for a given hex
+   *
+   * @param nx The number of elements in the x direction
+   * @param ny The number of elements in the y direction
+   * @param i The x index of this element
+   * @param j The y index of this element
+   * @param k The z index of this element
+   * @return The ID of the i,j element
+   */
 
-    template <typename T>
-    inline dof_id_type
-    elem_id(const dof_id_type /*nx*/,
-                     const dof_id_type /*ny*/,
-                     const dof_id_type /*i*/,
-                     const dof_id_type /*j*/,
-                     const dof_id_type /*k*/)
-    {
-      mooseError("elem_id not implemented for this element type in DistributedGeneratedMesh");
-    }
+  template <typename T>
+  inline dof_id_type elem_id(const dof_id_type /*nx*/,
+                             const dof_id_type /*ny*/,
+                             const dof_id_type /*i*/,
+                             const dof_id_type /*j*/,
+                             const dof_id_type /*k*/)
+  {
+    mooseError("elem_id not implemented for this element type in DistributedGeneratedMesh");
+  }
 
-    /**
-      * Get the number of neighbors this element will have
-      *
-      * @param nx The number of elements in the x direction
-      * @param ny The number of elements in the y direction
-      * @param nz The number of elements in the z direction
-      * @param i The x index of this element
-      * @param j The y index of this element
-      * @param k The z index of this element
-      * @return The number of neighboring elements
-      */
-     template <typename T>
-     inline dof_id_type
-     num_neighbors(const dof_id_type /*nx*/,
-                   const dof_id_type /*ny*/,
-                   const dof_id_type /*nz*/,
-                   const dof_id_type /*i*/,
-                   const dof_id_type /*j*/,
-                   const dof_id_type /*k*/)
-     {
-       mooseError("num_neighbors not implemented for this element type in DistributedGeneratedMesh");
-     }
+  /**
+   * Get the number of neighbors this element will have
+   *
+   * @param nx The number of elements in the x direction
+   * @param ny The number of elements in the y direction
+   * @param nz The number of elements in the z direction
+   * @param i The x index of this element
+   * @param j The y index of this element
+   * @param k The z index of this element
+   * @return The number of neighboring elements
+   */
+  template <typename T>
+  inline dof_id_type num_neighbors(const dof_id_type /*nx*/,
+                                   const dof_id_type /*ny*/,
+                                   const dof_id_type /*nz*/,
+                                   const dof_id_type /*i*/,
+                                   const dof_id_type /*j*/,
[…]
