+++ framework/src/partitioner/PetscExternalPartitioner.C
@@ -34,7 +34,9 @@ PetscExternalPartitioner::validParams()
                              partPackage,
                              "The external package is used for partitioning the mesh via PETSc");
 
-  params.addParam<dof_id_type>("num_cores_per_compute_node", 1, "Number of cores per compute node for hierarchical partitioning");
+  params.addParam<dof_id_type>("num_cores_per_compute_node",
+                               1,
+                               "Number of cores per compute node for hierarchical partitioning");
 
   params.addParam<bool>("apply_element_weight",
                         false,
@@ -126,7 +128,7 @@ PetscExternalPartitioner::_do_partition(MeshBase & mesh, const unsigned int n_pa
     }
   }
 
-   side_weights.clear();
+  side_weights.clear();
   // Edge weights represent the communication
   if (_apply_side_weight)
     side_weights.resize(num_edges);
@@ -153,7 +155,6 @@ PetscExternalPartitioner::_do_partition(MeshBase & mesh, const unsigned int n_pa
 
         nj++;
         n_neighbors++;
-
       }
 
       side++;
@@ -167,28 +168,35 @@ PetscExternalPartitioner::_do_partition(MeshBase & mesh, const unsigned int n_pa
 
   std::vector<dof_id_type> partition;
   // Partition graph
-  partitionGraph(comm(), _dual_graph, elem_weights, side_weights, n_parts, _num_parts_per_compute_node, _part_package, partition);
+  partitionGraph(comm(),
+                 _dual_graph,
+                 elem_weights,
+                 side_weights,
+                 n_parts,
+                 _num_parts_per_compute_node,
+                 _part_package,
+                 partition);
   // Assign partition to mesh
   assign_partitioning(mesh, partition);
 }
 
 void
 PetscExternalPartitioner::partitionGraph(const Parallel::Communicator & comm,
-                           const std::vector<std::vector<dof_id_type>> & graph,
-                           const std::vector<dof_id_type> & elem_weights,
-                           const std::vector<dof_id_type> & side_weights,
-                           const dof_id_type num_parts, const dof_id_type num_parts_per_compute_node,
-                           const std::string & part_package,
-                           std::vector<dof_id_type> & partition)
+                                         const std::vector<std::vector<dof_id_type>> & graph,
+                                         const std::vector<dof_id_type> & elem_weights,
+                                         const std::vector<dof_id_type> & side_weights,
+                                         const dof_id_type num_parts,
+                                         const dof_id_type num_parts_per_compute_node,
+                                         const std::string & part_package,
+                                         std::vector<dof_id_type> & partition)
 {
 #ifdef LIBMESH_HAVE_PETSC
 
   PetscErrorCode ierr;
   Mat dual;
-  PetscInt num_local_elems, num_elems,
-           *xadj=nullptr, *adjncy=nullptr,i,
-          *values=nullptr, *petsc_elem_weights=nullptr;
-  const PetscInt *parts;
+  PetscInt num_local_elems, num_elems, *xadj = nullptr, *adjncy = nullptr, i, *values = nullptr,
+                                       *petsc_elem_weights = nullptr;
+  const PetscInt * parts;
   MatPartitioning part;
   IS is;
 
@@ -197,38 +205,41 @@ PetscExternalPartitioner::partitionGraph(const Parallel::Communicator & comm,
   // Figure out the total of elements
[…]
