+++ framework/src/meshgenerators/DistributedRectilinearMeshGenerator.C
@@ -35,7 +35,6 @@
 
 registerMooseObject("MooseApp", DistributedRectilinearMeshGenerator);
 
-
 InputParameters
 DistributedRectilinearMeshGenerator::validParams()
 {
@@ -91,7 +90,8 @@ DistributedRectilinearMeshGenerator::validParams()
   return params;
 }
 
-DistributedRectilinearMeshGenerator::DistributedRectilinearMeshGenerator(const InputParameters & parameters)
+DistributedRectilinearMeshGenerator::DistributedRectilinearMeshGenerator(
+    const InputParameters & parameters)
   : MeshGenerator(parameters),
     _verbose(getParam<bool>("verbose")),
     _dim(getParam<MooseEnum>("dim")),
@@ -115,11 +115,11 @@ DistributedRectilinearMeshGenerator::DistributedRectilinearMeshGenerator(const I
 template <>
 inline dof_id_type
 DistributedRectilinearMeshGenerator::num_neighbors<Edge2>(const dof_id_type nx,
-                     const dof_id_type /*ny*/,
-                     const dof_id_type /*nz*/,
-                     const dof_id_type i,
-                     const dof_id_type /*j*/,
-                     const dof_id_type /*k*/)
+                                                          const dof_id_type /*ny*/,
+                                                          const dof_id_type /*nz*/,
+                                                          const dof_id_type i,
+                                                          const dof_id_type /*j*/,
+                                                          const dof_id_type /*k*/)
 {
   // The ends only have one neighbor
   if (i == 0 || i == nx - 1)
@@ -128,15 +128,30 @@ DistributedRectilinearMeshGenerator::num_neighbors<Edge2>(const dof_id_type nx,
   return 2;
 }
 
+template <typename T>
+inline void
+get_neighbors(const dof_id_type /*nx*/,
+              const dof_id_type /*ny*/,
+              const dof_id_type /*nz*/,
+              const dof_id_type /*i*/,
+              const dof_id_type /*j*/,
+              const dof_id_type /*k*/,
+              std::vector<dof_id_type> & /*neighbors*/,
+              const bool /*corner*/)
+{
+  mooseError("get_neighbors not implemented for this element type in DistributedGeneratedMesh");
+}
+
 template <>
 inline void
 DistributedRectilinearMeshGenerator::get_neighbors<Edge2>(const dof_id_type nx,
-                     const dof_id_type /*ny*/,
-                     const dof_id_type /*nz*/,
-                     const dof_id_type i,
-                     const dof_id_type /*j*/,
-                     const dof_id_type /*k*/,
-                     std::vector<dof_id_type> & neighbors)
+                                                          const dof_id_type /*ny*/,
+                                                          const dof_id_type /*nz*/,
+                                                          const dof_id_type i,
+                                                          const dof_id_type /*j*/,
+                                                          const dof_id_type /*k*/,
+                                                          std::vector<dof_id_type> & neighbors,
+                                                          const bool /*corner*/)
 
 {
   neighbors[0] = i - 1;
@@ -154,11 +169,11 @@ DistributedRectilinearMeshGenerator::get_neighbors<Edge2>(const dof_id_type nx,
 template <>
 inline void
 DistributedRectilinearMeshGenerator::get_indices<Edge2>(const dof_id_type /*nx*/,
-                   const dof_id_type /*ny*/,
-                   const dof_id_type elem_id,
-                   dof_id_type & i,
-                   dof_id_type & /*j*/,
-                   dof_id_type & /*k*/)
+                                                        const dof_id_type /*ny*/,
+                                                        const dof_id_type elem_id,
+                                                        dof_id_type & i,
+                                                        dof_id_type & /*j*/,
+                                                        dof_id_type & /*k*/)
 {
   i = elem_id;
 }
[…]
