+++ framework/src/mesh/DistributedGeneratedMesh.C
@@ -1,1517 +1,0 @@
-//* This file is part of the MOOSE framework
-//* https://www.mooseframework.org
-//*
-//* All rights reserved, see COPYRIGHT for full restrictions
-//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
-//*
-//* Licensed under LGPL 2.1, please see LICENSE for details
-//* https://www.gnu.org/licenses/lgpl-2.1.html
-
-#include "DistributedGeneratedMesh.h"
-
-#include "libmesh/mesh_generation.h"
-#include "libmesh/string_to_enum.h"
-#include "libmesh/periodic_boundaries.h"
-#include "libmesh/periodic_boundary_base.h"
-#include "libmesh/unstructured_mesh.h"
-#include "libmesh/partitioner.h"
-#include "libmesh/metis_csr_graph.h"
-#include "libmesh/edge_edge2.h"
-#include "libmesh/edge_edge3.h"
-#include "libmesh/edge_edge4.h"
-#include "libmesh/mesh_communication.h"
-#include "libmesh/remote_elem.h"
-#include "libmesh/face_quad4.h"
-#include "libmesh/cell_hex8.h"
-
-// C++ includes
-#include <cmath> // provides round, not std::round (see http://www.cplusplus.com/reference/cmath/round/)
-
-#ifdef LIBMESH_HAVE_METIS
-// MIPSPro 7.4.2 gets confused about these nested namespaces
-#ifdef __sgi
-#include <cstdarg>
-#endif
-namespace Metis
-{
-extern "C"
-{
-#include "libmesh/ignore_warnings.h"
-#include "metis.h"
-#include "libmesh/restore_warnings.h"
-}
-}
-#else
-#include "libmesh/sfc_partitioner.h"
-#endif
-
-registerMooseObject("MooseApp", DistributedGeneratedMesh);
-
-defineLegacyParams(DistributedGeneratedMesh);
-
-InputParameters
-DistributedGeneratedMesh::validParams()
-{
-  InputParameters params = MooseMesh::validParams();
-
-  params.addParam<bool>("verbose", false, "Turn on verbose printing for the mesh generation");
-
-  MooseEnum elem_types(
-      "EDGE EDGE2 EDGE3 EDGE4 QUAD QUAD4 QUAD8 QUAD9 TRI3 TRI6 HEX HEX8 HEX20 HEX27 TET4 TET10 "
-      "PRISM6 PRISM15 PRISM18 PYRAMID5 PYRAMID13 PYRAMID14"); // no default
-
-  MooseEnum dims("1=1 2 3");
-  params.addRequiredParam<MooseEnum>(
-      "dim", dims, "The dimension of the mesh to be generated"); // Make this parameter required
-
-  params.addParam<dof_id_type>("nx", 1, "Number of elements in the X direction");
-  params.addParam<dof_id_type>("ny", 1, "Number of elements in the Y direction");
-  params.addParam<dof_id_type>("nz", 1, "Number of elements in the Z direction");
-  params.addParam<Real>("xmin", 0.0, "Lower X Coordinate of the generated mesh");
-  params.addParam<Real>("ymin", 0.0, "Lower Y Coordinate of the generated mesh");
-  params.addParam<Real>("zmin", 0.0, "Lower Z Coordinate of the generated mesh");
-  params.addParam<Real>("xmax", 1.0, "Upper X Coordinate of the generated mesh");
-  params.addParam<Real>("ymax", 1.0, "Upper Y Coordinate of the generated mesh");
-  params.addParam<Real>("zmax", 1.0, "Upper Z Coordinate of the generated mesh");
-  params.addParam<MooseEnum>("elem_type",
-                             elem_types,
-                             "The type of element from libMesh to "
[…]
