+++ modules/doc/content/newsletter/2018_06.md
@@ -1,10 +1,10 @@
 # MOOSE Newsletter (June 2018)
 
-## DistributedGeneratedMesh
+## DistributedRectilinearMeshGenerator
 
 Since the beginning, MOOSE has had the [/GeneratedMesh.md] for generating cartesian meshes in 1D, 2D and 3D.  However, the way this has worked in parallel for distributed-mesh is non-optimal.  In parallel, the complete [/GeneratedMesh.md] is actually built on every processor.  If you're using `parallel_type = distributed` the entire mesh will still be built on every processor, partitioned, and then each processor will delete the portion not assigned to it.
 
-We have [added a new capability](https://github.com/idaholab/moose/pull/11505) called [/DistributedGeneratedMesh.md].  This new capability allows for each processor to uniquely create only the portion of the mesh that belongs on that processor.  In this way, much larger generated meshes can be created in parallel much more quickly and efficiently.
+We have [added a new capability](https://github.com/idaholab/moose/pull/11505) called [/DistributedRectilinearMeshGenerator.md].  This new capability allows for each processor to uniquely create only the portion of the mesh that belongs on that processor.  In this way, much larger generated meshes can be created in parallel much more quickly and efficiently.
 
 ## VectorPostprocessor Parallel Consistency
 
