### 2020-02-27 289fd47
DistributedRectilinearMeshGenerator: added tests
Closes #11485
