+++ framework/include/meshgenerators/DistributedRectilinearMeshGenerator.h
@@ -24,7 +24,7 @@ public:
   std::unique_ptr<MeshBase> generate() override;
 
   template <typename T>
-  void build_cube(UnstructuredMesh & mesh,
+  void buildCube(UnstructuredMesh & mesh,
                   const unsigned int nx,
                   unsigned int ny,
                   unsigned int nz,
@@ -49,14 +49,14 @@ public:
    */
 
   template <typename T>
-  inline dof_id_type elem_id(const dof_id_type /*nx*/,
+  inline dof_id_type elemId(const dof_id_type /*nx*/,
                              const dof_id_type /*ny*/,
                              const dof_id_type /*i*/,
                              const dof_id_type /*j*/,
                              const dof_id_type /*k*/)
   {
     mooseError(
-        "elem_id not implemented for this element type in DistributedRectilinearMeshGenerator");
+        "elemId not implemented for this element type in DistributedRectilinearMeshGenerator");
   }
 
   /**
@@ -71,14 +71,14 @@ public:
    * @return The number of neighboring elements
    */
   template <typename T>
-  inline dof_id_type num_neighbors(const dof_id_type /*nx*/,
+  inline dof_id_type numNeighbors(const dof_id_type /*nx*/,
                                    const dof_id_type /*ny*/,
                                    const dof_id_type /*nz*/,
                                    const dof_id_type /*i*/,
                                    const dof_id_type /*j*/,
                                    const dof_id_type /*k*/)
   {
-    mooseError("num_neighbors not implemented for this element type in "
+    mooseError("numNeighbors not implemented for this element type in "
                "DistributedRectilinearMeshGenerator");
   }
 
@@ -95,14 +95,18 @@ public:
    * there is no neighbor.  THIS MUST be of size 6 BEFORE calling this function
    */
   template <typename T>
-  inline void get_neighbors(const dof_id_type /*nx*/,
+  inline void getNeighbors(const dof_id_type /*nx*/,
                             const dof_id_type /*ny*/,
                             const dof_id_type /*nz*/,
                             const dof_id_type /*i*/,
                             const dof_id_type /*j*/,
                             const dof_id_type /*k*/,
                             std::vector<dof_id_type> & /*neighbors*/,
-                            const bool corner = false);
+                            const bool corner = false)
+{
+  mooseError(
+      "getNeighbors not implemented for this element type in DistributedRectilinearMeshGenerator");
+}
 
   /**
    * The ID of the i,j,k node
@@ -116,7 +120,7 @@ public:
    * @param k The z index of this node
    */
   template <typename T>
-  inline dof_id_type node_id(const ElemType /*type*/,
+  inline dof_id_type nodeId(const ElemType /*type*/,
                              const dof_id_type /*nx*/,
                              const dof_id_type /*ny*/,
                              const dof_id_type /*i*/,
@@ -124,7 +128,7 @@ public:
                              const dof_id_type /*k*/)
   {
     mooseError(
-        "node_id not implemented for this element type in DistributedRectilinearMeshGenerator");
+        "nodeId not implemented for this element type in DistributedRectilinearMeshGenerator");
   }
 
   /**
@@ -140,7 +144,7 @@ public:
    * @param mesh The mesh to add it to
    */
   template <typename T>
-  Node * add_point(const dof_id_type /*nx*/,
+  Node * addPoint(const dof_id_type /*nx*/,
                    const dof_id_type /*ny*/,
                    const dof_id_type /*nz*/,
                    const dof_id_type /*i*/,
@@ -150,7 +154,7 @@ public:
                    MeshBase & /*mesh*/)
   {
     mooseError(
[…]
