+++ framework/src/meshgenerators/DistributedRectilinearMeshGenerator.C
@@ -8,6 +8,7 @@
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
 #include "DistributedRectilinearMeshGenerator.h"
+#include "CastUniquePointer.h"
 #include "PetscExternalPartitioner.h"
 
 #include "SerializerGuard.h"
@@ -95,9 +96,9 @@ DistributedRectilinearMeshGenerator::DistributedRectilinearMeshGenerator(
   : MeshGenerator(parameters),
     _verbose(getParam<bool>("verbose")),
     _dim(getParam<MooseEnum>("dim")),
-    _nx(declareMeshProperty("num_elements_x", getParam<unsigned int>("nx"))),
-    _ny(declareMeshProperty("num_elements_y", getParam<unsigned int>("ny"))),
-    _nz(declareMeshProperty("num_elements_z", getParam<unsigned int>("nz"))),
+    _nx(declareMeshProperty("num_elements_x", getParam<dof_id_type>("nx"))),
+    _ny(declareMeshProperty("num_elements_y", getParam<dof_id_type>("ny"))),
+    _nz(declareMeshProperty("num_elements_z", getParam<dof_id_type>("nz"))),
     _xmin(declareMeshProperty("xmin", getParam<Real>("xmin"))),
     _xmax(declareMeshProperty("xmax", getParam<Real>("xmax"))),
     _ymin(declareMeshProperty("ymin", getParam<Real>("ymin"))),
@@ -126,21 +127,6 @@ DistributedRectilinearMeshGenerator::num_neighbors<Edge2>(const dof_id_type nx,
     return 1;
 
   return 2;
-}
-
-template <typename T>
-inline void
-get_neighbors(const dof_id_type /*nx*/,
-              const dof_id_type /*ny*/,
-              const dof_id_type /*nz*/,
-              const dof_id_type /*i*/,
-              const dof_id_type /*j*/,
-              const dof_id_type /*k*/,
-              std::vector<dof_id_type> & /*neighbors*/,
-              const bool /*corner*/)
-{
-  mooseError(
-      "get_neighbors not implemented for this element type in DistributedRectilinearMeshGenerator");
 }
 
 template <>
@@ -223,7 +209,7 @@ DistributedRectilinearMeshGenerator::elem_id<Edge2>(const dof_id_type /*nx*/,
 
 template <>
 void
-DistributedRectilinearMeshGenerator::add_element<Edge2>(const dof_id_type nx,
+DistributedRectilinearMeshGenerator::addElement<Edge2>(const dof_id_type nx,
                                                         const dof_id_type /*ny*/,
                                                         const dof_id_type /*nz*/,
                                                         const dof_id_type /*i*/,
@@ -346,11 +332,12 @@ DistributedRectilinearMeshGenerator::get_neighbors<Quad4>(const dof_id_type nx,
 
   if (corner)
   {
-    int nnb = 0;
-    for (int ii = -1; ii <= 1; ii++)
-      for (int jj = -1; jj <= 1; jj++)
-        if ((i + ii >= 0) && (i + ii < nx) && (j + jj >= 0) && (j + jj < ny))
-          neighbors[nnb++] = elem_id<Quad4>(nx, 0, i + ii, j + jj, 0);
+    // libMesh dof_id_type looks like unsigned int
+    unsigned int nnb = 0;
+    for (unsigned int ii = 0; ii <= 2; ii++)
+      for (unsigned int jj = 0; jj <= 2; jj++)
+        if ((i + ii >= 1) && (i + ii <= nx) && (j + jj >= 1) && (j + jj <= ny))
+          neighbors[nnb++] = elem_id<Quad4>(nx, 0, i + ii - 1, j + jj - 1, 0);
 
     return;
   }
[…]

[…]
