### 2020-02-28 #2118 @WilkAndy
@Kevin-Mattheus-Moerman , is it possible to change the Repo of this to

https://github.com/WilkAndy/PorousFlowDescription

?  The code is NOT at that location, only the paper.md, paper.bib, LICENSE and README.md that points to the code's location.

We've read the whedon Ruby script and think the first stage is to clone the Repo, which is failing.  This is because https://github.com/idaholab/moose/tree/master/modules/porous_flow isn't directly clonable, only https://github.com/idaholab/moose/ is itself.   There are about 20 physics modules of MOOSE - PorousFlow is only one of them - and we feel it's likely the authors of some other physics modules (eg, solid mechanics) will want to write JOSS papers too, so we don't want to have https://github.com/idaholab/moose/ as our repo.
