### 2020-02-28 #2118 @jedbrown
We could change it, but standard practice is to put it in the parent repository and rename or delete the paper source after publication. There have been a few Landlab examples of this pattern: https://github.com/openjournals/joss-reviews/issues?utf8=%E2%9C%93&q=is%3Aissue+landlab
