### 2020-02-28 idaholab/moose#14796 @WilkAndy
create joss_porous_flow branch in idaholab/moose
## Reason

For our recently submitted article to Journal of Open Source Software (JOSS), the bot that creates the paper needs to have the address of the repo

https://github.com/idaholab/moose

It then searches for a file "paper.md" and "paper.bib".   We can give JOSS this information, but then it might appear like we're claiming that we wrote the entire MOOSE framework and all physics modules!  It also precludes anyone other MOOSE person publishing in JOSS while our paper is under review.   A better idea, suggested in openjournals/joss-reviews#2118 is to create a separate branch, say `joss_porous_flow` , get the bot to build from that , and do all amendments in that, and when the paper is published, merge the branch into `next`.

## Design

Create `joss_porous_flow` branch in idaholab/moose from current `next` or `master`.    Perhaps you want to do it @permcody or other MOOSE devels?

## Impact

The JOSS bot will be able to build our paper, and we'll be separate from the rest of the MOOSE community.
