### 2020-02-28 3d199c7
Cleanup of SodiumProperties
