+++ modules/fluid_properties/include/userobjects/SodiumProperties.h
@@ -10,7 +10,6 @@
 #pragma once
 
 #include "FluidProperties.h"
-#include "Moose.h"
 
 class SodiumProperties;
 
@@ -28,48 +27,53 @@ public:
   SodiumProperties(const InputParameters & parameters);
 
   /**
-   * Thermal conductivity as a function of temperature
-   * @param T temperature
-   * From page 181. Valid from 371 to 1500 K.
+   * Thermal conductivity as a function of temperature. From page 181. Valid from 371 to 1500 K.
+   * @param T temperature in K
+   * @return Thermal conductivity of liquid sodium in W/m/K
    */
   Real k(Real temperature) const;
 
   /**
    * Enthalpy of liquid Na (relative to solid Na at STP) in J/kg as a function of temperature
-   * @param T temperature
    * From page 4. Valid from 371 to 2000 K, and relative to enthalpy of solid Na at 298.15.
+   * @param T temperature in K
+   * @return enthalpy of liquid sodium in J/kg
    */
   Real h(Real temperature) const;
 
   /**
-   * Heat capacity of liquid Na in J/kg-K as a function of temperature
-   * @param T temperature
-   * From page 29 (or by differentiating enthalpy). Valid from 371 to 2000 K.
+   * Heat capacity of liquid Na in J/kg-K as a function of temperature. From page 29 (or by
+   * differentiating enthalpy). Valid from 371 to 2000 K.
+   * @param T temperature in K
+   * @ return Heat capacity of liquid sodium in J/kg/K
    */
   Real heatCapacity(Real temperature) const;
 
   /**
-   * Inverse solve for temperature [K] from enthalpy [J/kg]
-   * @param h enthalpy
+   * Inverse solve for temperature from enthalpy
+   * @param h enthalpy in J/kg
+   * @return temperature of liquid sodium in K
    */
   Real temperature(Real enthalpy) const;
 
   /**
-   * Density as a function of temperature
-   * @param T temperature
+   * Density as a function of temperature. Valid 371 K < T < 2503.7 K
+   * @param T temperature in K
+   * @return density of liquid sodium in kg/m^3
    */
   Real rho(Real temperature) const;
 
   /**
-   * Derivative of density w.r.t temperature
-   * @param T temperature
+   * Derivative of density w.r.t temperature. Valid 371 K < T < 2503.7 K
+   * @param T temperature in K
+   * @return derivative of density of liquid sodium with respect to temperature
    */
   Real drho_dT(Real temperature) const;
 
   /**
-   * Derivative of density w.r.t enthalpy
-   * @param h enthalpy
+   * Derivative of density w.r.t enthalpy. Valid 371 K < T < 2503.7 K
+   * @param T enthalpy in J/kg
+   * @return derivative of density of liquid sodium with respect to enthalpy
    */
   Real drho_dh(Real enthalpy) const;
 };
-
