+++ modules/fluid_properties/doc/content/source/materials/SodiumPropertiesMaterial.md
@@ -1,0 +1,28 @@
+# Sodium Properties Material
+
+!syntax description /Materials/SodiumPropertiesMaterial
+
+`SodiumPropertiesMaterial` samples the functions availble in [SodiumProperties](SodiumProperties.md)
+and provides them as material properties.
+
+The following materials created for liquid sodium are:
+
+- Thermal Conductivity: `k(temperature)`
+- Enthalpy: `h(temperature)`
+- Heat capacity: `heatCapacity(temperature)`
+- Temperature: `temperature(h)`
+- Density: `rho(temperature)`
+- Derivative of Density w.r.t temperature: `drho_dT(temperature)`
+- Derivative of Density w.r.t enthalpy: `drho_dh(h)`
+
+Properties of liquid sodium are obtained from [!cite](Fink:1995bf).
+
+!listing modules/fluid_properties/test/tests/sodium/exact.i block=Materials
+
+!syntax parameters /Materials/SodiumPropertiesMaterial
+
+!syntax inputs /Materials/SodiumPropertiesMaterial
+
+!syntax children /Materials/SodiumPropertiesMaterial
+
+!bibtex bibliography
