### 2020-02-28 34800d7
templated SodiumProperties
