+++ modules/fluid_properties/include/userobjects/SodiumProperties.h
@@ -31,7 +31,13 @@ public:
    * @param T temperature in K
    * @return Thermal conductivity of liquid sodium in W/m/K
    */
-  Real k(Real temperature) const;
+  template <typename T>
+  T k(T temperature) const
+  {
+    const T temperature2 = temperature * temperature;
+    const T temperature3 = temperature2 * temperature;
+    return 124.67 - 0.11381 * temperature + 5.5226e-5 * temperature2 - 1.1842e-8 * temperature3;
+  }
 
   /**
    * Enthalpy of liquid Na (relative to solid Na at STP) in J/kg as a function of temperature
@@ -39,7 +45,16 @@ public:
    * @param T temperature in K
    * @return enthalpy of liquid sodium in J/kg
    */
-  Real h(Real temperature) const;
+  template <typename T>
+  T h(T temperature) const
+  {
+    const T temperature2 = temperature * temperature;
+    const T temperature3 = temperature2 * temperature;
+
+    // Converted from kJ/kg to J/kg.
+    return -365.77e3 + 1.6582e3 * temperature - 4.2395e-1 * temperature2 +
+           1.4847e-4 * temperature3 + 2992.6e3 / temperature;
+  }
 
   /**
    * Heat capacity of liquid Na in J/kg-K as a function of temperature. From page 29 (or by
@@ -47,33 +62,83 @@ public:
    * @param T temperature in K
    * @ return Heat capacity of liquid sodium in J/kg/K
    */
-  Real heatCapacity(Real temperature) const;
+  template <typename T>
+  T heatCapacity(T temperature) const
+  {
+    const T temperature2 = temperature * temperature;
+
+    // Converted from kJ/kg-K to J/kg-K.
+    return 1.6582e3 - 8.4790e-1 * temperature + 4.4541e-4 * temperature2 - 2992.6e3 / temperature2;
+  }
 
   /**
    * Inverse solve for temperature from enthalpy
    * @param h enthalpy in J/kg
    * @return temperature of liquid sodium in K
    */
-  Real temperature(Real enthalpy) const;
+  template <typename T>
+  T temperature(T enthalpy) const
+  {
+    // Estimate initial guess from linear part of enthalpy.
+    T temperature = (enthalpy + 365.77e3) / 1.6582e3;
+
+    // Newton-Raphson for this equation: enthalpy(T) - enthalpy = 0 = residual. This is easy because
+    // dResidual/dT is just dH/dT, which is heat capacity.
+    for (unsigned iteration = 0; iteration < 10; ++iteration)
+    {
+      const T residual = h(temperature) - enthalpy;
+      temperature -= residual / heatCapacity(temperature);
+      if (std::abs(residual / enthalpy) < 1e-6)
+        break;
+    }
+    // If we get here, enthalpy is probably out of bounds. However, due to the nature of the JFNK
+    // calculation, we probably just want to ignore the error and spit out a bogus T so that the
+    // solver keeps rolling.
+    return temperature;
+  }
 
   /**
    * Density as a function of temperature. Valid 371 K < T < 2503.7 K
    * @param T temperature in K
    * @return density of liquid sodium in kg/m^3
    */
-  Real rho(Real temperature) const;
+  template <typename T>
+  T rho(T temperature) const
+  {
+    const T rhoc = 219.0; // kg/m^3
+    const T f = 275.32;
+    const T g = 511.58;
+    const T Tc = 2503.7; // critical temperature, K
+    mooseAssert(temperature < Tc, "Temperature is greater than critical temperature 2503.7 K ");
+
+    return rhoc + f * (1.0 - temperature / Tc) + g * std::sqrt(1.0 - temperature / Tc);
+  }
 
[…]
