### 2020-02-28 a066870
Added ElementAverageFunctionDifference #14800
