+++ framework/doc/content/source/postprocessors/ElementAverageFunctionDifference.md
@@ -1,0 +1,19 @@
+# ElementAverageFunctionDifference
+
+!syntax description /Postprocessors/ElementalVariableValue
+
+# Description and Syntax
+
+The `ElementAverageFunctionDifference` is a `ElementIntegralPostprocessor` that computes the
+difference between the elemental average of a variable with a function. The `absolute_value`
+input parameter optionally returns the absolute value of the difference.
+
+!syntax description /Postprocessors/AreaPostprocessor
+
+!syntax parameters /Postprocessors/AreaPostprocessor
+
+!syntax inputs /Postprocessors/AreaPostprocessor
+
+!syntax children /Postprocessors/AreaPostprocessor
+
+!bibtex bibliography
