### 2020-02-29 38506d5
Protect against non-existing dlopen (#14591)
