### 2020-02-29 22afb08
Finished Simplification Algorithm
