+++ modules/phase_field/include/utils/ExpressionBuilder.h
@@ -54,7 +54,7 @@ public:
   class EBTermNode;
   class EBFunction;
   class EBSubstitutionRule;
-  class Comparer;
+  struct Comparer;
   typedef std::vector<EBTerm> EBTermList;
   typedef std::vector<EBTermNode *> EBTermNodeList;
   typedef std::vector<const EBSubstitutionRule *> EBSubstitutionRuleList;
@@ -74,8 +74,28 @@ public:
       return os << node.stringify();
     }
     virtual std::vector<EBTermNode *> getChildren() { return std::vector<EBTermNode *>(0); }
-    virtual bool compare(EBTermNode *, std::vector<EBTermNode *>) = 0;
-    virtual std::vector<EBTermNode *> getPermutations() { return std::vector<EBTermNode *>(0); }
+    virtual bool compare(EBTermNode *, std::vector<EBTermNode *> &, bool) = 0;
+    virtual std::vector<EBTermNode *> getPermutations(bool, bool)
+    {
+      return std::vector<EBTermNode *>(0);
+    }
+
+    virtual bool compareRule(EBTermNode * rule,
+                             std::vector<EBTermNode *> & stars,
+                             std::vector<int> & current_index,
+                             bool = false);
+
+    virtual EBTermNode * simplify(std::pair<EBTermNode *, EBTermNode *> rule, bool & changed);
+
+    virtual EBTermNode * replaceRule(std::vector<EBTermNode *>) { return this; }
+
+    virtual bool isCommutative() const { return false; }
+    std::vector<std::vector<EBTermNode *>> permuteChildren(std::vector<EBTermNode *> children);
+    std::vector<std::vector<EBTermNode *>>
+    permuteChildren(std::vector<EBTermNode *> children,
+                    std::vector<std::vector<int>> & nnary_index);
+
+    virtual EBTermNode * toNNary() { return this; }
   };
 
   /// Template class for leaf nodes holding numbers in the expression tree
@@ -93,11 +113,11 @@ public:
 
     T getValue() { return _value; }
 
-    virtual bool compare(EBTermNode * compare_to, std::vector<EBTermNode *>)
+    virtual bool compare(EBTermNode * compare_to, std::vector<EBTermNode *> &, bool)
     {
       EBNumberNode<T> * c_to = dynamic_cast<EBNumberNode<T> *>(compare_to);
 
-      if(c_to != NULL && c_to->getValue() - _value < 0.000001)
+      if (c_to != NULL && c_to->getValue() - _value < 0.000001)
         return true;
       return false;
     }
@@ -117,12 +137,12 @@ public:
 
     std::string getSymbol() { return _symbol; }
 
-    virtual bool compare(EBTermNode * compare_to, std::vector<EBTermNode *>)
+    virtual bool compare(EBTermNode * compare_to, std::vector<EBTermNode *> &, bool)
     {
       EBSymbolNode * c_to = dynamic_cast<EBSymbolNode *>(compare_to);
 
       // We will do a hash compare first in the future
-      if(c_to != NULL && c_to->getSymbol() == _symbol)
+      if (c_to != NULL && c_to->getSymbol() == _symbol)
         return true;
       return false;
     }
@@ -143,7 +163,7 @@ public:
     virtual std::string stringify() const; // returns "[idnumber]"
     virtual int precedence() const { return 0; }
 
-    virtual bool compare(EBTermNode *, std::vector<EBTermNode *>)
+    virtual bool compare(EBTermNode *, std::vector<EBTermNode *> &, bool)
     {
       mooseError("Cannot compare with temp id node");
     }
@@ -159,7 +179,28 @@ public:
[…]
