### 2020-03-02 0c08340
Apply clang format
