### 2020-03-02 499e7cb
Fixed inheritance
