+++ framework/doc/content/source/postprocessors/ElementAverageFunctionDifference.md
@@ -1,19 +1,26 @@
 # ElementAverageFunctionDifference
 
-!syntax description /Postprocessors/ElementalVariableValue
+!syntax description /Postprocessors/ElementAverageFunctionDifference
 
 # Description and Syntax
 
 The `ElementAverageFunctionDifference` is a `ElementIntegralPostprocessor` that computes the
-difference between the elemental average of a variable with a function. The `absolute_value`
-input parameter optionally returns the absolute value of the difference.
+difference between the elemental average of a variable with a function evaluated at a given point,
+\begin{equation}
+  u_{avg} - f(t,x,y,z).
+\end{equation}
 
-!syntax description /Postprocessors/AreaPostprocessor
+The `absolute_value` input parameter optionally returns the absolute value of the difference,
+\begin{equation}
+  |(|u_{avg}| - |f(t,x,y,z)|)|.
+\end{equation}
 
-!syntax parameters /Postprocessors/AreaPostprocessor
+!listing test/tests/postprocessors/element_average_function_difference/test.i block=Postprocessors
 
-!syntax inputs /Postprocessors/AreaPostprocessor
+!syntax parameters /Postprocessors/ElementAverageFunctionDifference
 
-!syntax children /Postprocessors/AreaPostprocessor
+!syntax inputs /Postprocessors/ElementAverageFunctionDifference
+
+!syntax children /Postprocessors/ElementAverageFunctionDifference
 
 !bibtex bibliography
