### 2020-03-02 fee0151
Ensure that neighbor is init'd before interface stateful props
Previously we had not reinit'd the neighbor before calling
InterfaceMaterial::initStatefulProperties, leading to a nullptr
in _neighbor_elem. Thanks to @arovinelli for catching this

Closes #14770
