### 2020-03-02 552a3b6
Update param name to something more descriptive
(refs #14457)
