### 2020-03-02 42cc769
Require subcommands in MooseDocs
(refs #13699)
