+++ modules/fluid_properties/unit/include/SodiumPropertiesTest.h
@@ -15,7 +15,7 @@
 class SodiumPropertiesTest : public MooseObjectUnitTest
 {
 public:
-  SodiumPropertiesTest() : MooseObjectUnitTest("MooseUnitApp") { buildObjects(); }
+  SodiumPropertiesTest() : MooseObjectUnitTest("FluidPropertiesApp") { buildObjects(); }
 
 protected:
   void buildObjects()
