+++ modules/fluid_properties/unit/include/MethaneFluidPropertiesTest.h
@@ -15,7 +15,7 @@
 class MethaneFluidPropertiesTest : public MooseObjectUnitTest
 {
 public:
-  MethaneFluidPropertiesTest() : MooseObjectUnitTest("MooseUnitApp") { buildObjects(); }
+  MethaneFluidPropertiesTest() : MooseObjectUnitTest("FluidPropertiesApp") { buildObjects(); }
 
 protected:
   void buildObjects()
