+++ modules/doc/content/newsletter/2020_03.md
@@ -1,1 +1,20 @@
 # MOOSE Newsletter (March 2020)
+
+## Stochastic Tools
+
+The [stochastic tools module](modules/stochastic_tools/index.md) is under heavy development and has
+gained some powerful features, with many more to come in the near future.
+
+1. Support for distributed VectorPostprocessor (VPP) data was added to the framework. This allows
+   massively parallel stochastic simulations to be performed with minimal memory impact. The VPP
+   must be programmed to support data that is distributed or replicated. The mode of operation
+   is controlled by the "parallel_type" parameter. When set to distributed mode an output file
+   for each processor will be created.
+
+2. The VPP statistics calculator was generalized to support replicated and distributed VPP data as
+   well as for multiple vectors, see [vectorpostprocessors/Statistics.md].
+
+3. Support for computing bootstrap confidence level intervals was added to the statistics calculator,
+   see [vectorpostprocessors/Statistics.md].
+
+4. An [XML](XMLOutput.md) output object was created to allow for single file VPP output.
