### 2020-03-03 bb54cc2
Adding simplification, substitution, derivatives, and tensors to modified form of ExpressionBuilder
