+++ modules/phase_field/src/utils/ExpressionBuilder.C
@@ -8,181 +8,43 @@
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
 #include "ExpressionBuilder.h"
-#include "EBSimpleTrees.h"
 
-std::vector<std::pair<ExpressionBuilder::EBTermNode *, ExpressionBuilder::EBTermNode *>>
-    ExpressionBuilder::EBTerm::_prep_simplification_rules = PrepEBSimpleTrees();
-std::vector<std::pair<ExpressionBuilder::EBTermNode *, ExpressionBuilder::EBTermNode *>>
-    ExpressionBuilder::EBTerm::_simplification_rules = EBSimpleTrees();
-
-/******************************************
- * The Beginning of the Node implementations
- * EBTermNode
- *******************************************/
-bool
-ExpressionBuilder::EBTermNode::compareRule(EBTermNode * rule,
-                                           std::vector<EBTermNode *> & stars,
-                                           std::vector<int> & current_index,
-                                           bool leafCompare)
+ExpressionBuilder::EBTermList
+operator, (const ExpressionBuilder::EBTerm & larg, const ExpressionBuilder::EBTerm & rarg)
 {
-  std::stack<EBTermNode *> self_stack;
-  self_stack.emplace(this);
-  std::stack<EBTermNode *> rule_stack;
-  rule_stack.emplace(rule);
-  std::vector<EBTermNode *> star_vec(stars);
-  std::stack<Comparer *> permutations;
-  bool isHead = true;
-  while (rule_stack.size() > 0)
-  {
-    // Get the children of the top of the stack
-    std::vector<EBTermNode *> self_children = self_stack.top()->getChildren();
-    std::vector<EBTermNode *> rule_children = rule_stack.top()->getChildren();
-    // Check the nodes to make sure they match
-    // If not, move to the next permutation
-    if (!rule_stack.top()->compare(self_stack.top(), star_vec, isHead))
-    {
-      if (permutations.size() != 0)
-      {
-        self_stack = permutations.top()->_self_stack;
-        rule_stack = permutations.top()->_rule_stack;
-        star_vec = permutations.top()->_stars;
-        current_index = permutations.top()->_current_index;
-        permutations.pop();
-        continue;
-      }
-      else
-        return false;
-    }
-    EBStarNode * checkLeaf = dynamic_cast<EBStarNode *>(rule_stack.top());
-    // We have 3 possibilities here.
-    // The first is if we are at a leaf node.
-    // The second is if we are at a commutative node.
-    // The third is everything else.
-    if (checkLeaf != NULL)
-    {
-      rule_stack.pop();
-      self_stack.pop();
-    }
-    else if (self_stack.top()->isCommutative())
-    {
-      EBNNaryOpTermNode * checkNNary = dynamic_cast<EBNNaryOpTermNode *>(self_stack.top());
-      std::vector<std::vector<EBTermNode *>> permutedChildren(0);
-      std::vector<std::vector<int>> nnary_index(0);
-      if (checkNNary != NULL && isHead)
-        permutedChildren = permuteChildren(self_children, nnary_index);
-      else
-        permutedChildren = permuteChildren(self_children);
-      rule_stack.pop();
-      for (unsigned int i = 0; i < rule_children.size(); ++i)
-        rule_stack.emplace(rule_children[i]);
-
-      for (unsigned int i = 0; i < permutedChildren.size(); ++i)
-      {
-        std::stack<EBTermNode *> temp_self_stack(self_stack);
-        temp_self_stack.pop();
[…]
