+++ modules/phase_field/include/utils/EBSimpleTrees.h
@@ -1,23 +1,24 @@
 
 #pragma once
 
-#include "ExpressionBuilder.h"
+#include "ExpressionBuilderToo.h"
 
 struct SimplificationRules
 {
-  SimplificationRules(std::vector<std::vector<ExpressionBuilder::EBTerm>> rules)
+  SimplificationRules(std::vector<std::vector<ExpressionBuilderToo::EBTerm>> rules)
   {
     for (auto rule : rules)
     {
-      _simplification_trees.emplace_back(std::make_pair(rule[0].cloneRoot(), rule[1].cloneRoot()));
+      _simplification_trees.emplace_back(
+          std::make_pair(rule[0].cloneRoot()->toNNary(), rule[1].cloneRoot()->toNNary()));
     }
   }
   operator std::vector<
-      std::pair<ExpressionBuilder::EBTermNode *, ExpressionBuilder::EBTermNode *>>()
+      std::pair<ExpressionBuilderToo::EBTermNode *, ExpressionBuilderToo::EBTermNode *>>()
   {
     return _simplification_trees;
   }
-  std::vector<std::pair<ExpressionBuilder::EBTermNode *, ExpressionBuilder::EBTermNode *>>
+  std::vector<std::pair<ExpressionBuilderToo::EBTermNode *, ExpressionBuilderToo::EBTermNode *>>
       _simplification_trees;
 };
 
@@ -27,12 +28,12 @@ public:
   PrepEBSimpleTrees();
 
   operator std::vector<
-      std::pair<ExpressionBuilder::EBTermNode *, ExpressionBuilder::EBTermNode *>>()
+      std::pair<ExpressionBuilderToo::EBTermNode *, ExpressionBuilderToo::EBTermNode *>>()
   {
     return _prep_simplification_trees;
   }
 
-  std::vector<std::pair<ExpressionBuilder::EBTermNode *, ExpressionBuilder::EBTermNode *>>
+  std::vector<std::pair<ExpressionBuilderToo::EBTermNode *, ExpressionBuilderToo::EBTermNode *>>
       _prep_simplification_trees;
 };
 
@@ -42,11 +43,11 @@ public:
   EBSimpleTrees();
 
   operator std::vector<
-      std::pair<ExpressionBuilder::EBTermNode *, ExpressionBuilder::EBTermNode *>>()
+      std::pair<ExpressionBuilderToo::EBTermNode *, ExpressionBuilderToo::EBTermNode *>>()
   {
     return _simplification_trees;
   }
 
-  std::vector<std::pair<ExpressionBuilder::EBTermNode *, ExpressionBuilder::EBTermNode *>>
+  std::vector<std::pair<ExpressionBuilderToo::EBTermNode *, ExpressionBuilderToo::EBTermNode *>>
       _simplification_trees;
 };
