+++ modules/phase_field/include/utils/ExpressionBuilderToo.h
@@ -1,0 +1,1056 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#pragma once
+
+#include <vector>
+#include <ostream>
+#include <sstream>
+#include <iomanip>
+
+#include "MooseError.h"
+#include "MooseUtils.h"
+#include "TensorTempl.h"
+#include "libmesh/libmesh_common.h"
+
+using namespace libMesh;
+
+class ExpressionBuilderToo
+{
+public:
+  ExpressionBuilderToo(){};
+
+  // forward delcarations
+  class EBTerm;
+  class EBTermNode;
+  struct Comparer;
+  typedef std::vector<EBTerm> EBTermList;
+  typedef std::vector<EBTermNode *> EBTermNodeList;
+
+  /// Base class for nodes in the expression tree
+  class EBTermNode
+  {
+  public:
+    virtual ~EBTermNode(){};
+    virtual EBTermNode * clone() const = 0;
+
+    virtual std::string stringify() const = 0;
+    virtual int precedence() const = 0;
+    friend std::ostream & operator<<(std::ostream & os, const EBTermNode & node)
+    {
+      return os << node.stringify();
+    }
+    virtual std::vector<EBTermNode *> getChildren() { return std::vector<EBTermNode *>(0); }
+    virtual bool compare(EBTermNode *, std::vector<EBTermNode *> &, bool, int &) = 0;
+
+    virtual bool
+    compareRule(EBTermNode * rule,
+                std::vector<EBTermNode *> & stars,
+                std::vector<unsigned int> & current_index,
+                std::vector<unsigned int> changed_indeces = std::vector<unsigned int>(0),
+                bool checkNNaryLeaf = false);
+
+    virtual EBTermNode * simplify(std::pair<EBTermNode *, EBTermNode *> rule, bool & changed);
+
+    virtual EBTermNode * replaceRule(std::vector<EBTermNode *>) { return this; }
+
+    virtual bool isCommutative() const { return false; }
+
+    std::vector<std::vector<EBTermNode *>>
+    permuteChildren(std::vector<EBTermNode *> children, std::vector<unsigned int> changed_indeces);
+    std::vector<std::vector<EBTermNode *>>
+    permuteChildren(std::vector<EBTermNode *> children,
+                    std::vector<std::vector<unsigned int>> & nnary_index,
+                    int max_depth,
+                    std::vector<unsigned int> changed_indeces,
+                    int depth = 0,
+                    std::vector<unsigned int> current_index = std::vector<unsigned int>(0));
+
+    virtual EBTermNode * toNNary() { return this; }
+
+    virtual EBTermNode * substitute(std::vector<std::pair<EBTermNode *, EBTermNode *>> subs);
+
+    virtual bool isConstant() = 0;
+    virtual EBTermNode * constantFolding() { return this; };
+
+    virtual Real getRealValue() { mooseError("This has no real value right now"); };
+
+    virtual EBTermNode * derivative(const std::string & derivative_comp) = 0;
+  };
+
+  /// Template class for leaf nodes holding numbers in the expression tree
+  template <typename T>
+  class EBNumberNode : public EBTermNode
+  {
+    T _value;
+
+  public:
+    EBNumberNode(T value) : _value(value){};
[…]

[…]
