+++ modules/doc/content/newsletter/2020_03.md
@@ -18,3 +18,26 @@ gained some powerful features, with many more to come in the near future.
    see [vectorpostprocessors/Statistics.md].
 
 4. An [XML](XMLOutput.md) output object was created to allow for single file VPP output.
+
+## Improved Mechanical and Termal Contact Jacobians
+
+Jacobians for kinematic-formulation base mechanical contact were improved so that their accuracy is
+not impacted by variable scaling. This change results in significantly improved linear and nonlinear
+convergence for contact problems used in conjunction with scaling (either manual or
+automatic). Thermal contact Jacobians were also improved in the month of February. We now supply
+dependence of the thermal contact residuals on master-side degrees of freedom, including both
+(block) on-diagonal temperature and off-diagonal displacements degrees of freedom.
+
+## libMesh level changes
+
+### Adaptivity with LAGRANGE_VEC
+
+The libMesh update in the month of February enabled adaptivity with the LAGRANGE_VEC finite element
+family. The update also enables used of periodic boundary conditions with LAGRANGE_VEC.
+
+### Much faster periodic boundaries
+
+This was actually in a previous libMesh update, but the improved performance make it worth
+mentioniong here. Periodic boundaries now use a tree-search (O(log N) complexity) instead of a
+linear (O(N)) search. Consequently, MOOSE users of the periodic BC capability should expect
+signicant speed-up of their simulations (especially for large problems).
