### 2020-03-03 5dda1ad
Removed unused garbage
Refs #19747
