### 2020-03-03 dd6195a
Added checks for heat structure parameters
Make sure that users provide the same number of widths, n_part_elems and materials
parameters.
