+++ modules/contact/doc/content/source/constraints/TangentialMortarMechanicalContact.md
@@ -15,5 +15,3 @@ be `TangentialMortarMechanicalContact` instances for both x and y displacement c
 !syntax inputs /Constraints/TangentialMortarMechanicalContact
 
 !syntax children /Constraints/TangentialMortarMechanicalContact
-
-!bibtex bibliography
