+++ modules/doc/content/index.md
@@ -129,6 +129,7 @@ A 3D soil desiccation simulation using phase-field for cohesive fracture model, 
 - [Physics Modules:](modules/index.md)
 
   - [Heat Conduction](HeatConduction.md)
+  - [Geochemistry](geochemistry/index.md)
   - [Navier Stokes](NavierStokes/index.md)
   - [Solid Mechanics](syntax/Modules/TensorMechanics/index.md)
   - [Contact](contact/index.md)
