+++ modules/stochastic_tools/src/userobject/PolynomialChaos.C
@@ -10,6 +10,7 @@
 #include "PolynomialChaos.h"
 
 #include "Sampler.h"
+#include "CartesianProduct.h"
 
 registerMooseObject("StochasticToolsApp", PolynomialChaos);
 
@@ -24,6 +25,16 @@ PolynomialChaos::validParams()
   params.addRequiredParam<unsigned int>("order", "Maximum polynomial order.");
   params.addRequiredParam<std::vector<DistributionName>>(
       "distributions", "Names of the distributions samples were taken from.");
+  params.addParam<bool>(
+      "output_mean", false, "True to output mean from computed coefficients to postprocessor.");
+  params.addParam<bool>(
+      "output_stdv",
+      false,
+      "True to output standard deviation from computed coefficients to postprocessor.");
+  params.addParam<bool>(
+      "output_skew", false, "True to output skewness from computed coefficients to postprocessor.");
+  params.addParam<bool>(
+      "output_kurt", false, "True to output kurtosis from computed coefficients to postprocessor.");
 
   return params;
 }
@@ -34,8 +45,33 @@ PolynomialChaos::PolynomialChaos(const InputParameters & parameters)
     _quad_sampler(nullptr),
     _tuple(generateTuple(getParam<std::vector<DistributionName>>("distributions").size(), _order)),
     _ncoeff(_tuple.size()),
-    _coeff(_ncoeff)
+    _coeff(_ncoeff),
+    _mean_pp(nullptr),
+    _stdv_pp(nullptr),
+    _skew_pp(nullptr),
+    _kurt_pp(nullptr)
 {
+  InputParameters params = _app.getFactory().getValidParams("Receiver");
+  if (getParam<bool>("output_mean"))
+  {
+    _fe_problem.addPostprocessor("Receiver", name() + "_mean", params);
+    _mean_pp = &_fe_problem.getPostprocessorValue(name() + "_mean");
+  }
+  if (getParam<bool>("output_stdv"))
+  {
+    _fe_problem.addPostprocessor("Receiver", name() + "_stdv", params);
+    _stdv_pp = &_fe_problem.getPostprocessorValue(name() + "_stdv");
+  }
+  if (getParam<bool>("output_skew"))
+  {
+    _fe_problem.addPostprocessor("Receiver", name() + "_skew", params);
+    _skew_pp = &_fe_problem.getPostprocessorValue(name() + "_skew");
+  }
+  if (getParam<bool>("output_kurt"))
+  {
+    _fe_problem.addPostprocessor("Receiver", name() + "_kurt", params);
+    _kurt_pp = &_fe_problem.getPostprocessorValue(name() + "_kurt");
+  }
 }
 
 void
@@ -97,6 +133,15 @@ PolynomialChaos::finalize()
   if (!_quad_sampler)
     for (unsigned int i = 0; i < _ncoeff; ++i)
       _coeff[i] /= _sampler->getNumberOfRows();
+
+  if (_mean_pp)
+    *_mean_pp = computeMean();
+  if (_stdv_pp)
+    *_stdv_pp = computeSTD();
+  if (_skew_pp)
+    *_skew_pp = computeSkewness();
+  if (_kurt_pp)
+    *_kurt_pp = computeKurtosis();
 }
 
 void
@@ -126,6 +171,75 @@ PolynomialChaos::evaluate(const std::vector<Real> & x) const
     Real tmp = _coeff[i];
     for (unsigned int d = 0; d < _ndim; ++d)
       tmp *= poly_val(d, _tuple[i][d]);
+    val += tmp;
+  }
+
[…]
