+++ modules/stochastic_tools/include/samplers/SobolSampler.h
@@ -15,8 +15,15 @@ class SobolSampler;
 
 template <>
 InputParameters validParams<SobolSampler>();
+
 /**
- * A class used to perform Monte Carlo Sampling
+ * A class used to perform Monte Carlo sampling for performing Sobol sensitivity analysis.
+ *
+ * The created matrices are stacked in the following order, following the nomenclature from
+ * Saltelli (2002), "Making best use of model evaluations to compute sensitivity indices"
+ *
+ * with re-sampling: [M2, N_1, ..., N_n, N_-1, ..., N-n, M1]
+ * without re-sampling: [M2, N_1, ..., N_n, M1]
  */
 class SobolSampler : public Sampler
 {
@@ -32,16 +39,25 @@ protected:
 
   ///@{
   /// Sobol Monte Carlo matrices, these are sized and cleared to avoid keeping large matrices around
-  DenseMatrix<Real> _a_matrix;
-  DenseMatrix<Real> _b_matrix;
+  DenseMatrix<Real> _m1_matrix;
+  DenseMatrix<Real> _m2_matrix;
   ///@}
-
-  /// Storage for distribution objects to be utilized
-  std::vector<Distribution const *> _distributions;
 
   /// Distribution names
   const std::vector<DistributionName> & _distribution_names;
 
+  /// Flag for building the re-sampling matrix for computing second order sensitivity indices
+  const bool & _resample;
+
+  /// Number of inputs
+  const std::size_t _num_inputs;
+
+  /// Number of matrices
+  const dof_id_type _num_matrices;
+
   /// The number of rows per matrix
   const dof_id_type _num_rows_per_matrix;
+
+  /// Storage for distribution objects to be utilized
+  std::vector<Distribution const *> _distributions;
 };
