+++ modules/phase_field/src/vectorpostprocessors/FeatureVolumeVectorPostprocessor.C
@@ -274,7 +274,10 @@ FeatureVolumeVectorPostprocessor::accumulateBoundaryFaces(
 
   // Accumulate the boundary area/length into the dominant feature. Do not use the integral value
   if (_single_feature_per_elem && dominant_feature_id != FeatureFloodCount::invalid_id)
-    _feature_volumes[dominant_feature_id] += _assembly.elementVolume(elem->side_ptr(side).get());
+  {
+    std::unique_ptr<const Elem> side_elem = elem->build_side_ptr(side);
+    _feature_volumes[dominant_feature_id] += _assembly.elementVolume(side_elem.get());
+  }
 }
 
 Real
