+++ framework/src/vectorpostprocessors/WorkBalance.C
@@ -193,8 +193,8 @@ public:
 
       // Build the side so we can compute its volume
       auto side_elem = elem->build_side_ptr(side);
-      // FIXME: I don't know how to get the volume here, there is hardly anything to
-      // work with in this threaded loop
+      // NOTE: we do not want to account for different coordinate systems here, so
+      // using volume from libmesh elem is fine here
       auto volume = side_elem->volume();
       _local_partition_surface_area += volume;
 
