+++ modules/phase_field/src/postprocessors/FauxGrainTracker.C
@@ -12,6 +12,7 @@
 // MOOSE includes
 #include "MooseMesh.h"
 #include "MooseVariable.h"
+#include "Assembly.h"
 
 registerMooseObject("PhaseFieldApp", FauxGrainTracker);
 
