+++ framework/include/base/Assembly.h
@@ -1260,7 +1260,7 @@ public:
   void addCachedJacobianContributions();
 
   /// On-demand computation of volume element accounting for RZ/RSpherical
-  Real elementVolume(const Elem * elem);
+  Real elementVolume(const Elem * elem) const;
 
   /**
    * Set the pointer to the XFEM controller object
