+++ modules/heat_conduction/doc/content/bib/heat_conduction.bib
@@ -19,3 +19,12 @@
   year={2013},
   publisher={Elsevier Science}
 }
+
+@book{nafems_std_bench,
+  title={{The Standard NAFEMS Benchmarks, Rev. 3}},
+  author={National Agency for Finite Element Methods and Standards},
+  publisher = {NAFEMS},
+  month={October},
+  year={1990}
+}
+
