### 2020-03-04 7fda0ae
config.yml modified to ensure geochemistry doc is build.  Splitting chemical_reactions and geochmistry bibs.  Refs #14789
