+++ modules/geochemistry/doc/content/bib/geochemistry.bib
@@ -1,21 +1,3 @@
-@article{guo2013,
-  title={{A parallel, fully coupled, fully implicit solution to reactive transport in porous media using the preconditioned Jacobian-Free Newton-Krylov Method}},
-  author={Guo, Luanjing and Huang, Hai and Gaston, Derek R and Permann, Cody J and Andrs, David and Redden, George D and Lu, Chuan and Fox, Don T and Fujita, Yoshiko},
-  journal={Advances in Water Resources},
-  volume={53},
-  pages={101--108},
-  year={2013}
-}
-
-@article{lichtner1996,
-  title={Continuum formulation of multicomponent-multiphase reactive transport},
-  author={Lichtner, Peter C},
-  journal={Reviews in Mineralogy and Geochemistry},
-  volume={34},
-  pages={1--81},
-  year={1996}
-}
-
 @book{bethke_2007, place={Cambridge}, edition={2}, title={Geochemical and Biogeochemical Reaction Modeling}, DOI={10.1017/CBO9780511619670}, publisher={Cambridge University Press}, author={Bethke, Craig M.}, year={2007}}
 
 @book{gwb_essentials,
