+++ modules/chemical_reactions/doc/content/bib/chemical_reactions.bib
@@ -15,51 +15,3 @@
   pages={1--81},
   year={1996}
 }
-
-@book{bethke_2007, place={Cambridge}, edition={2}, title={Geochemical and Biogeochemical Reaction Modeling}, DOI={10.1017/CBO9780511619670}, publisher={Cambridge University Press}, author={Bethke, Craig M.}, year={2007}}
-
-@book{gwb_essentials,
-author = {Bethke, C. M and Farrell, B. and Yeakel, S.},
-year = {2020},
-title = {{The Geochemists's Workbench, Release 14: GWB Essentials Guide}},
-publisher = {Aqueous Solutions LLC, Champaign, UL, USA}
-}
-
-@book{gwb_reference,
-author = {Bethke, C. M and Farrell, B. and Yeakel, S.},
-year = {2020},
-title = {{The Geochemists's Workbench, Release 14: GWB Reference Manual}},
-publisher = {Aqueous Solutions LLC, Champaign, UL, USA},
-url = {https://www.gwb.com/pdf/GWB12/GWBreference.pdf}
-}
-
-@book{prausnitz,
-author = {Prausnitz, John M. and Lichtenthaler, Rudiger N. and de Azevedo, Edmundo Gomes},
-year = {1998},
-title = {{Molecular Thermodynamics of Fluid-Phase Equilibria}},
-edition = {3},
-publisher = {Prentice Hall}
-}
-
-@techreport{toughreact,
-title={{TOUGHREACT V3.0-OMP Reference Manual: A Paralell Simulation Program for Non-Isothermal Multiphase Geochemical Reactive Transport}},
-author={Xu, Tianfu and Sonnenthal, Eric and Spycher, Nicolas and Zhen, Liang and Millar, Norman and Pruess, Karsten},
-institution={LBNL},
-year={2014}
-}
-
-@article{spycher1988,
-  title={{Fugacity coefficients of H$_{2}$, CO$_{2}$, CH$_{4}$, H$_{2}$ and of H$_{2}$--CO$_{2}$--CH$_{4}$ mixutres: A virial equation treatment for moderate pressures and temperatures applicable to calculations of hydrothermal boiling}},
-  author={Spycher, Nicolas F. and Reed, Mark H.},
-  journal={Geochemica et Cosmochimica Acta},
-  volume={52},
-  pages={739--749},
-  year={1988}
-}
-
-@techreport{palandri,
-title={{A compilation of rate parameters of water-mineral interaction kinetics for application to geochemical modelling}},
-author={Palandri and Kharaka},
-institution={USGS Open File Report 2004-1068},
-year={2004}
-}
