+++ modules/doc/config.yml
@@ -3,6 +3,7 @@ Content:
     - framework/doc/content
     - modules/chemical_reactions/doc/content
     - modules/contact/doc/content
+    - modules/geochemistry/doc/content
     - modules/fluid_properties/doc/content
     - modules/functional_expansion_tools/doc/content
     - modules/heat_conduction/doc/content
@@ -48,6 +49,7 @@ Extensions:
             chemical_reactions: !include modules/chemical_reactions/doc/hidden.yml
             #combined: !include modules/combined/doc/hidden.yml
             contact: !include modules/contact/doc/hidden.yml
+            geochemistry: !include modules/geochemistry/doc/hidden.yml
             fluid_properties: !include modules/fluid_properties/doc/hidden.yml
             #functional_expansion_tools: !include modules/functional_expansion_tools/doc/hidden.yml
             heat_conduction: !include modules/heat_conduction/doc/hidden.yml
