### 2020-03-04 5fa9b58
Augment sparsity pattern based on coupling information
Components now have to hint the spasity pattern, so that no matrix
reallocations are performed.  For example, if 2 flow channels are
coupled with a junction, the junction has to inform the simulation
about this connection.

Closes #19805
