+++ modules/thm/src/components/JunctionOneToOne.C
@@ -3,6 +3,7 @@
 #include "FlowModelSinglePhase.h"
 #include "FlowModelTwoPhase.h"
 #include "FlowModelTwoPhaseNCG.h"
+#include "THMMesh.h"
 
 registerMooseObject("THMApp", JunctionOneToOne);
 
@@ -18,6 +19,28 @@ validParams<JunctionOneToOne>()
 }
 
 JunctionOneToOne::JunctionOneToOne(const InputParameters & params) : FlowJunction(params) {}
+
+void
+JunctionOneToOne::setupMesh()
+{
+  FlowJunction::setupMesh();
+
+  std::vector<dof_id_type> connected_elems;
+  const std::map<dof_id_type, std::vector<dof_id_type>> & node_to_elem = _mesh.nodeToElemMap();
+  for (auto & nid : _nodes)
+  {
+    const auto & it = node_to_elem.find(nid);
+    if (it == node_to_elem.end())
+      mooseError(name(), ": failed to find node ", nid, "in the mesh!");
+
+    const std::vector<dof_id_type> & elems = it->second;
+    for (const auto & e : elems)
+      connected_elems.push_back(e);
+  }
+
+  if (connected_elems.size() == 2)
+    _sim.augmentSparsity(connected_elems[0], connected_elems[1]);
+}
 
 void
 JunctionOneToOne::check() const
