### 2020-03-04 d1fa2b5
Consolidate contact actions Refs. #14813
Contact pressure, penetration, and contact area are now a part of the general contact action.
