+++ modules/contact/include/actions/ContactAction.h
@@ -60,8 +60,10 @@ protected:
   const MooseEnum _model;
   const MooseEnum _formulation;
   const MooseEnum _system;
+  const MooseEnum _order;
   const MeshGeneratorName _mesh_gen_name;
   const bool _ping_pong_protection;
+
 
 private:
   void addMortarContact();
