### 2020-03-04 279277f
Apply clang format
