### 2020-03-04 0987390
Apply clang format
