+++ modules/tensor_mechanics/src/kernels/InertialForceShell.C
@@ -1,0 +1,653 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "InertialForceShell.h"
+#include "SubProblem.h"
+#include "libmesh/utility.h"
+#include "MooseVariable.h"
+#include "Assembly.h"
+#include "NonlinearSystem.h"
+#include "AuxiliarySystem.h"
+#include "MooseMesh.h"
+#include "DenseMatrix.h"
+
+#include "libmesh/utility.h"
+#include "libmesh/quadrature.h"
+#include "libmesh/enum_quadrature_type.h"
+#include "libmesh/fe_type.h"
+#include "libmesh/string_to_enum.h"
+#include "libmesh/quadrature_gauss.h"
+
+registerMooseObject("TensorMechanicsApp", InertialForceShell);
+
+defineLegacyParams(InertialForceShell);
+
+InputParameters
+InertialForceShell::validParams()
+{
+  InputParameters params = TimeKernel::validParams();
+  params.addClassDescription("Calculates the residual for the inertial force/moment and the "
+                             "contribution of mass dependent Rayleigh damping and HHT time "
+                             "integration scheme.");
+  params.set<bool>("use_displaced_mesh") = true;
+  params.addRequiredCoupledVar(
+      "rotations",
+      "The rotational variables appropriate for the simulation geometry and coordinate system");
+  params.addRequiredCoupledVar(
+      "displacements",
+      "The displacement variables appropriate for the simulation geometry and coordinate system");
+  params.addCoupledVar("velocities", "Translational velocity variables");
+  params.addCoupledVar("accelerations", "Translational acceleration variables");
+  params.addCoupledVar("rotational_velocities", "Rotational velocity variables");
+  params.addCoupledVar("rotational_accelerations", "Rotational acceleration variables");
+  params.addRangeCheckedParam<Real>(
+      "beta", "beta>0.0", "beta parameter for Newmark Time integration");
+  params.addRangeCheckedParam<Real>(
+      "gamma", "gamma>0.0", "gamma parameter for Newmark Time integration");
+  params.addParam<MaterialPropertyName>("eta",
+                                        0.0,
+                                        "Name of material property or a constant real "
+                                        "number defining the eta parameter for the "
+                                        "Rayleigh damping.");
+  params.addRangeCheckedParam<Real>("alpha",
+                                    0.0,
+                                    "alpha >= -0.3333 & alpha <= 0.0",
+                                    "Alpha parameter for mass dependent numerical damping induced "
+                                    "by HHT time integration scheme");
+  params.addParam<MaterialPropertyName>(
+      "density",
+      "density",
[…]
