### 2020-03-04 14383af
Add template method for getting Sampler objects
(refs #14784)
