### 2020-03-04 8a487bc
Create unit test for Sobol indices
(refs #14784)
