+++ modules/stochastic_tools/unit/src/TestStochasticToolsSobolCalculators.C
@@ -1,0 +1,265 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+#include <vector>
+
+#include "gtest/gtest.h"
+#include "SobolCalculators.h"
+#include "MooseRandom.h"
+#include "libmesh/communicator.h"
+#include "libmesh/parallel_object.h"
+
+using namespace StochasticTools;
+
+Real
+g_function(const std::vector<Real> & q_vector, const DenseMatrix<Real> & data, std::size_t row)
+{
+  Real y = 1;
+  for (std::size_t c = 0; c < data.n(); ++c)
+    y *= (std::abs(4 * data(row, c) - 2) + q_vector[c]) / (1 + q_vector[c]);
+  return y;
+}
+
+void
+evaluate(const std::vector<Real> & q_vector,
+         const DenseMatrix<Real> & data,
+         std::vector<Real> & out)
+{
+  for (std::size_t i = 0; i < data.m(); ++i)
+    out.push_back(g_function(q_vector, data, i));
+}
+
+Real
+model_association_index(const std::vector<Real> & data, std::size_t count)
+{
+  return 1 - std::accumulate(data.begin(), data.begin() + count, 0.);
+}
+
+std::vector<Real>
+sobolidx(const std::vector<Real> & q_vector, std::size_t K, bool resample = true)
+{
+  MooseRandom generator;
+  generator.seed(0, 1980);
+  generator.seed(1, 1949);
+
+  const std::size_t n = q_vector.size();
+
+  DenseMatrix<Real> M1(K, n); // "sample" matrix
+  DenseMatrix<Real> M2(K, n); // "re-sample" matrix
+
+  for (std::size_t r = 0; r < K; ++r)
+    for (std::size_t c = 0; c < n; ++c)
+    {
+      M1(r, c) = generator.rand(0);
+      M2(r, c) = generator.rand(1);
+    }
+
+  std::vector<DenseMatrix<Real>> Ni(n, DenseMatrix<Real>(K, n));
+  std::vector<DenseMatrix<Real>> Nni(n, DenseMatrix<Real>(K, n));
+  for (std::size_t j = 0; j < n; ++j)
+    for (std::size_t r = 0; r < K; ++r)
+      for (std::size_t c = 0; c < n; ++c)
+      {
+        Ni[j](r, c) = j == c ? M1(r, c) : M2(r, c);
+        Nni[j](r, c) = j == c ? M2(r, c) : M1(r, c);
+      }
+
+  // Output data
+  std::vector<Real> data;
+
+  // a0
+  evaluate(q_vector, M2, data);
+
+  // ai
+  for (std::size_t i = 0; i < n; ++i)
+    evaluate(q_vector, Ni[i], data);
+
+  // ani
+  if (resample)
+    for (std::size_t i = 0; i < n; ++i)
[…]
