+++ modules/stochastic_tools/include/samplers/CartesianProductSampler.h
@@ -33,7 +33,7 @@ protected:
   /// Helper object for computing the CartesianProcduct values
   // This is a pointer because it cannot be created until the grid vectors are assembled from
   // the input parameters.
-  std::unique_ptr<const StochasticTools::CartesianProduct> _cp_ptr = nullptr;
+  std::unique_ptr<const StochasticTools::CartesianProduct<Real>> _cp_ptr = nullptr;
 
 private:
   /// PerfGraph timer
