### 2020-03-04 5663110
Preliminary (wrong) residual for rotational degrees of freedom.
