+++ modules/tensor_mechanics/src/kernels/InertialForceShell.C
@@ -98,14 +98,6 @@ InertialForceShell::InertialForceShell(const InputParameters & parameters)
     _accel_num(_ndisp),
     _rot_vel_num(_nrot),
     _rot_accel_num(_nrot),
-    _area(coupledValue("area")),
-    _Ay(coupledValue("Ay")),
-    _Az(coupledValue("Az")),
-    _Ix(_has_Ix ? coupledValue("Ix") : _zero),
-    _Iy(coupledValue("Iy")),
-    _Iz(coupledValue("Iz")),
-    _beta(_has_beta ? getParam<Real>("beta") : 0.1),
-    _gamma(_has_gamma ? getParam<Real>("gamma") : 0.1),
     _eta(getMaterialProperty<Real>("eta")),
     _alpha(getParam<Real>("alpha")),
     _original_local_config(getMaterialPropertyByName<RankTwoTensor>("initial_rotation")),
@@ -329,12 +321,12 @@ InertialForceShell::computeResidual()
       _0g1_vector.push_back(0.5 * _thickness[qp_xy] * _v2[0]);
 
       DenseMatrix<Real> G1(3, 2);
-      G1(0, 0) = _0g1_vector[0][0];
-      G1(1, 0) = _0g1_vector[0][1];
-      G1(2, 0) = _0g1_vector[0][2];
-      G1(0, 1) = _0g1_vector[1][0];
-      G1(1, 1) = _0g1_vector[1][1];
-      G1(2, 1) = _0g1_vector[1][2];
+      G1(0, 0) = _0g1_vector[0](0);
+      G1(1, 0) = _0g1_vector[0](1);
+      G1(2, 0) = _0g1_vector[0](2);
+      G1(0, 1) = _0g1_vector[1](0);
+      G1(1, 1) = _0g1_vector[1](1);
+      G1(2, 1) = _0g1_vector[1](2);
       DenseMatrix<Real> G1T(2, 3);
       G1.get_transpose(G1T);
 
@@ -343,12 +335,12 @@ InertialForceShell::computeResidual()
       _0g2_vector.push_back(0.5 * _thickness[qp_xy] * _v2[1]);
 
       DenseMatrix<Real> G2(3, 2);
-      G2(0, 0) = _0g2_vector[0][0];
-      G2(1, 0) = _0g2_vector[0][1];
-      G2(2, 0) = _0g2_vector[0][2];
-      G2(0, 1) = _0g2_vector[1][0];
-      G2(1, 1) = _0g2_vector[1][1];
-      G2(2, 1) = _0g2_vector[1][2];
+      G2(0, 0) = _0g2_vector[0](0);
+      G2(1, 0) = _0g2_vector[0](1);
+      G2(2, 0) = _0g2_vector[0](2);
+      G2(0, 1) = _0g2_vector[1](0);
+      G2(1, 1) = _0g2_vector[1](1);
+      G2(2, 1) = _0g2_vector[1](2);
 
       DenseMatrix<Real> G2T(2, 3);
       G2.get_transpose(G2T);
@@ -358,12 +350,12 @@ InertialForceShell::computeResidual()
       _0g3_vector.push_back(0.5 * _thickness[qp_xy] * _v2[2]);
 
       DenseMatrix<Real> G3(3, 2);
-      G3(0, 0) = _0g3_vector[0][0];
-      G3(1, 0) = _0g3_vector[0][1];
-      G3(2, 0) = _0g3_vector[0][2];
-      G3(0, 1) = _0g3_vector[1][0];
-      G3(1, 1) = _0g3_vector[1][1];
-      G3(2, 1) = _0g3_vector[1][2];
[…]
