### 2020-03-04 3d4207f
Remove documentation associated with obsolete contact actions.
