+++ modules/contact/doc/content/source/actions/NodalAreaAction.md
@@ -1,1 +1,0 @@
-!template load file=stubs/moose_action.md.template name=NodalAreaAction syntax=/Contact/NodalAreaAction
