+++ modules/contact/doc/content/source/actions/NodalAreaVarAction.md
@@ -1,1 +1,0 @@
-!template load file=stubs/moose_action.md.template name=NodalAreaVarAction syntax=/Contact/NodalAreaVarAction
