### 2020-03-04 6ce74a5
Remove documentation associated with obsolete contact actions.
