+++ modules/contact/doc/content/source/actions/ContactPressureAuxAction.md
@@ -1,1 +1,0 @@
-!template load file=stubs/moose_action.md.template name=ContactPressureAuxAction syntax=/Contact/ContactPressureAuxAction
