+++ modules/stochastic_tools/include/samplers/CartesianProductSampler.h
@@ -34,4 +34,8 @@ protected:
   // This is a pointer because it cannot be created until the grid vectors are assembled from
   // the input parameters.
   std::unique_ptr<const StochasticTools::CartesianProduct> _cp_ptr = nullptr;
+
+private:
+  /// PerfGraph timer
+  const PerfID _perf_compute_sample;
 };
