+++ modules/stochastic_tools/src/utils/SobolCalculators.C
@@ -7,6 +7,7 @@
 //* Licensed under LGPL 2.1, please see LICENSE for details
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 #include "SobolCalculators.h"
+#include "MooseError.h"
 
 namespace StochasticTools
 {
