+++ modules/doc/content/newsletter/2020_03.md
@@ -41,3 +41,36 @@ This was actually in a previous libMesh update, but the improved performance mak
 mentioniong here. Periodic boundaries now use a tree-search (O(log N) complexity) instead of a
 linear (O(N)) search. Consequently, MOOSE users of the periodic BC capability should expect
 signicant speed-up of their simulations (especially for large problems).
+
+## Simplified TestHarness Scheduling
+
+The parallel scheduling capability of the TestHarness has been changed to reduce potential pitfalls when
+creating several related tests. Previously, the TestHarness would make all tests in a single directory
+elegible to schedule at the same time unless "prereq" parameters were added to force serialization. Now
+the TestHarness will only schedule a single test from a single directory by default. High concurrency is
+still achievable by having several different directories in a single application. The `prereq` parameter should
+still be used to declare "true" dependencies since execution by input file order is not guaranteed.
+
+It is possible to force parallel scheduling within a directory by using a new parameter `parallel_scheduling = True`
+at the top level of the test specification file. See the following example:
+
+```
+[Tests]
+  parallel_scheduling = True
+
+  [test1]
+     ...
+  []
+  [test2]
+     ...
+  []
+  [test3]
+     ...
+  []
+[]
+```
+
+Extensive use of this parameter is not recommended since it could result in unintended I/O race conditions. It can
+be difficult to catch all file related outputs. For example, consider outputs due to recover testing, which many
+developers do not routinely perform on their own workstations. However, it can be convenient to use this parameter if
+a developer has several highly related tests that they do not wish to split up into separate directories.
