+++ modules/doc/content/newsletter/2020_03.md
@@ -19,7 +19,7 @@ gained some powerful features, with many more to come in the near future.
 
 4. An [XML](XMLOutput.md) output object was created to allow for single file VPP output.
 
-## Improved Mechanical and Termal Contact Jacobians
+## Improved Mechanical and Thermal Contact Jacobians
 
 Jacobians for kinematic-formulation base mechanical contact were improved so that their accuracy is
 not impacted by variable scaling. This change results in significantly improved linear and nonlinear
@@ -40,7 +40,7 @@ family. The update also enables used of periodic boundary conditions with LAGRAN
 This was actually in a previous libMesh update, but the improved performance make it worth
 mentioniong here. Periodic boundaries now use a tree-search (O(log N) complexity) instead of a
 linear (O(N)) search. Consequently, MOOSE users of the periodic BC capability should expect
-signicant speed-up of their simulations (especially for large problems).
+significant speed-up of their simulations (especially for large problems).
 
 ## Simplified TestHarness Scheduling
 
