### 2020-03-04 525cebb
Anisotropic crack propagation along cleavage planes in crystalline materials Enhancement - Considers crack propagation along weak cleavage planes in a material Issue # 14775 closes #14775
