+++ modules/phase_field/src/kernels/ACInterfaceBetaPenalty.C
@@ -1,0 +1,183 @@
+// Considers cleavage plane anisotropy in the crack propagation
+
+#include "ACInterfaceBetaPenalty.h"
+
+registerMooseObject("PhaseFieldApp", ACInterfaceBetaPenalty);
+
+template <>
+InputParameters
+validParams<ACInterfaceBetaPenalty>()
+{
+  InputParameters params = validParams<Kernel>();
+  params.addClassDescription("Gradient energy Allen-Cahn Kernel");
+  params.addParam<MaterialPropertyName>("mob_name", "L", "The mobility used with the kernel");
+  params.addParam<MaterialPropertyName>("kappa_name", "kappa_op", "The kappa used with the kernel");
+  params.addCoupledVar("args", "Vector of nonlinear variable arguments this object depends on");
+  params.addParam<bool>("variable_L",
+                        true,
+                        "The mobility is a function of any MOOSE variable (if "
+                        "this is set to false L must be constant over the "
+                        "entire domain!)");
+  params.addRequiredParam<Real>("beta_penalty",
+                                "penalty for damage on planes not normal to z (Nguyen, 2017)");
+  params.addRequiredParam<std::vector<Real>>("cleavage_plane_normal",
+                                             "Normal to the favoured cleavage plane");
+  return params;
+}
+
+ACInterfaceBetaPenalty::ACInterfaceBetaPenalty(const InputParameters & parameters)
+  : DerivativeMaterialInterface<JvarMapKernelInterface<Kernel>>(parameters),
+    _L(getMaterialProperty<Real>("mob_name")),
+    _kappa(getMaterialProperty<Real>("kappa_name")),
+    _variable_L(getParam<bool>("variable_L")),
+    _dLdop(getMaterialPropertyDerivative<Real>("mob_name", _var.name())),
+    _d2Ldop2(getMaterialPropertyDerivative<Real>("mob_name", _var.name(), _var.name())),
+    _dkappadop(getMaterialPropertyDerivative<Real>("kappa_name", _var.name())),
+    _nvar(_coupled_moose_vars.size()),
+    _dLdarg(_nvar),
+    _d2Ldargdop(_nvar),
+    _d2Ldarg2(_nvar),
+    _dkappadarg(_nvar),
+    _gradarg(_nvar),
+    _beta_penalty(getParam<Real>("beta_penalty")),
+    _cleavage_plane_normal(getParam<std::vector<Real>>("cleavage_plane_normal"))
+{
+  // Get mobility and kappa derivatives and coupled variable gradients
+  for (unsigned int i = 0; i < _nvar; ++i)
+  {
+    MooseVariable * ivar = _coupled_standard_moose_vars[i];
+    const VariableName iname = ivar->name();
+    if (iname == _var.name())
+      paramError("args",
+                 "The kernel variable should not be specified in the "
+                 "coupled `args` parameter.");
+
+    _dLdarg[i] = &getMaterialPropertyDerivative<Real>("mob_name", iname);
+    _dkappadarg[i] = &getMaterialPropertyDerivative<Real>("kappa_name", iname);
+
[…]
