### 2020-03-05 919a03f
Fix problem vector size in unit test
(refs #14784)
