### 2020-03-05 d97ee2c
Re-write GFunction as VPP to improve performance
(refs #14784)
