+++ modules/stochastic_tools/test/include/functions/GFunction.h
@@ -8,17 +8,21 @@
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 #pragma once
 
-#include "Function.h"
+#include "GeneralVectorPostprocessor.h"
+#include "SamplerInterface.h"
 
 /* SOBOL test function, see Slaughter, Eq. 5.51 */
-class GFunction : public Function
+class GFunction : public GeneralVectorPostprocessor, public SamplerInterface
 {
 public:
   static InputParameters validParams();
   GFunction(const InputParameters & parameters);
-  virtual Real value(Real t, const Point & p) const override;
+  virtual void execute() override;
+  virtual void initialize() override {}
+  virtual void finalize() override;
 
 protected:
+  Sampler & _sampler;
   const std::vector<Real> & _q_vector;
-  const std::vector<Real> & _x_vector;
+  VectorPostprocessorValue & _values;
 };
