+++ modules/stochastic_tools/include/utils/SobolCalculators.h
@@ -7,9 +7,9 @@
 //* Licensed under LGPL 2.1, please see LICENSE for details
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 #pragma once
-
-#include "MooseTypes.h"
 #include <vector>
+#include "MooseTypes.h"
+#include "libmesh/dense_matrix.h"
 
 namespace StochasticTools
 {
