+++ modules/stochastic_tools/doc/content/source/userobject/PolynomialChaos.md
@@ -1,0 +1,175 @@
+# PolynomialChaos
+
+!syntax description /UserObjects/PolynomialChaos
+
+## Overview
+
+Polynomial chaos is a surrogate modeling technique where a quantity of interest (QoI) that is dependent on input parameters is expanded as a sum of orthogonal polynomials. Given a QoI $Q$ dependent on a set of parameters $\vec{\xi}$, the polynomial chaos expansion (PCE) is:
+
+\begin{equation}
+Q(\vec{\xi}) = \sum_{i=1}^{P}q_i\Phi_i(\vec{\xi}) ,
+\end{equation}
+
+where $P$ is the multidimensional polynomial order and $q_i$ are coefficients that are to be computed. These coefficients can be found using intrusive and non intrusive techniques. The intrusive technique is quite difficult to generalize and very computationally demanding. Since the polynomial basis is orthogonal, a non intrusive technique is developed where the coefficients are found by performing a Galerkin projection and integrating:
+
+\begin{equation}
+q_i = \frac{\left\langle Q(\vec{\xi})\Phi_i(\vec{\xi})\right\rangle}{\left\langle\Phi_i(\vec{\xi}),\Phi_i(\vec{\xi})\right\rangle},
+\label{eq:coeff}
+\end{equation}
+
+where,
+
+\begin{equation}
+\left\langle a(\vec{\xi}) b(\vec{\xi}) \right\rangle = \int_{-\infty}^{\infty}a(\vec{\xi}) b(\vec{\xi}) f(\vec{\xi}) d\vec{\xi} .
+\end{equation}
+
+The weight function ($f(\vec{\xi})$) and bases ($\Phi_i(\vec{\xi})$) are typically products of one-dimensional functions:
+
+\begin{equation}
+f(\vec{\xi}) = \prod_{d=0}^{D}f_d(\xi_d) ,
+\end{equation}
+
+\begin{equation}
+Q(\vec{\xi}) = \sum_{i=1}^{P}q_i\prod_{d=0}^{D}\phi^d_{k_{d,i}}(\xi_d) .
+\end{equation}
+
+The weighting functions are defined by the probability density function of the parameter and the polynomials are based on these distributions, [PolynomialTable] is a list of commonly used distributions and their corresponding orthogonal polynomials.
+
+!table id=PolynomialTable caption=Common probability density functions and their corresponding orthogonal polynomials
+| Distribution | Density Function ($f_d(\xi)$) | Polynomial ($\phi_i(\xi)$) | Support|
+| - | - | - | - |
+| Normal | $\frac{1}{2\pi}e^{-\xi^2/2}$ | Hermite | $[-\infty, \infty]$ |
+| Uniform | $\frac{1}{2}$ | Legendre | $[-1, 1]$ |
+| Beta | $\frac{(1-\xi)^{\alpha}(1+\xi)^{\beta}}{2^{\alpha+\beta+1}B(\alpha+1,\beta+1)}$ | Jacobi | $[-1,1]$ |
+| Exponential | $e^{-\xi}$ | Laguerre | $[0,\infty]$ |
+| Gamma | $\frac{\xi^{\alpha}e^{-\xi}}{\Gamma(\alpha+1)}$ | Generalized Laguerre | $[0,\infty]$ |
+
+The expression in [eq:coeff] can be integrated using many different techniques. One is performing a Monte Carlo integration,
+
+\begin{equation}
[…]
