### 2020-03-05 idaholab/moose#14796 @WilkAndy
Is this actually going to happen, @permcody ?    If not, then we can give JOSS the repo https://github.com/idaholab/moose   .   I think the `paper.md` documentation makes it clear we're just talking about PorousFlow, not the entirety of MOOSE and i don't think anyone else is publishing in JOSS right now.

a
