+++ modules/tensor_mechanics/src/kernels/InertialForceShell.C
@@ -64,13 +64,7 @@ InertialForceShell::validParams()
       "density",
       "density",
       "Name of Material Property  or a constant real number defining the density of the beam.");
-  params.addRequiredCoupledVar("area", "Variable containing cross-section area");
-  params.addCoupledVar("Ay", 0.0, "Variable containing first moment of area about y axis");
-  params.addCoupledVar("Az", 0.0, "Variable containing first moment of area about z axis");
-  params.addCoupledVar("Ix",
-                       "Variable containing second moment of area about x axis. Defaults to Iy+Iz");
-  params.addRequiredCoupledVar("Iy", "Variable containing second moment of area about y axis");
-  params.addRequiredCoupledVar("Iz", "Variable containing second moment of area about z axis");
+
   params.addRequiredRangeCheckedParam<unsigned int>(
       "component",
       "component<6",
@@ -277,18 +271,6 @@ InertialForceShell::computeResidual()
 
     // AMR
 
-    std::vector<RealVectorValue> local_acc;
-    local_acc.push_back(_local_accel_0);
-    local_acc.push_back(_local_accel_1);
-    local_acc.push_back(_local_accel_2);
-    local_acc.push_back(_local_accel_3);
-
-    std::vector<RealVectorValue> local_rot_acc;
-    local_rot_acc.push_back(_local_rot_accel_0);
-    local_rot_acc.push_back(_local_rot_accel_1);
-    local_rot_acc.push_back(_local_rot_accel_2);
-    local_rot_acc.push_back(_local_rot_accel_3);
-
     unsigned int dim = _current_elem->dim();
 
     FEType fe_type(Utility::string_to_enum<Order>("First"),
@@ -375,123 +357,182 @@ InertialForceShell::computeResidual()
       DenseMatrix<Real> G4T(2, 3);
       G4.get_transpose(G4T);
 
+      // Conversions to DenseVector from RealVectorValue: Make a method out of this.
+      DenseVector<Real> local_accel_dv_0;
+      DenseVector<Real> local_accel_dv_1;
+      DenseVector<Real> local_accel_dv_2;
+      DenseVector<Real> local_accel_dv_3;
+
+      DenseVector<Real> local_rot_accel_dv_0;
+      DenseVector<Real> local_rot_accel_dv_1;
+      DenseVector<Real> local_rot_accel_dv_2;
+      DenseVector<Real> local_rot_accel_dv_3;
+
+      for (unsigned int i = 0; i < 3; i++)
+      {
+        local_accel_dv_0(i) = _local_accel_0(i);
+        local_accel_dv_1(i) = _local_accel_1(i);
+        local_accel_dv_2(i) = _local_accel_2(i);
+        local_accel_dv_3(i) = _local_accel_3(i);
+
+        local_rot_accel_dv_0(i) = _local_rot_accel_0(i);
+        local_rot_accel_dv_1(i) = _local_rot_accel_1(i);
+        local_rot_accel_dv_2(i) = _local_rot_accel_2(i);
+        local_rot_accel_dv_3(i) = _local_rot_accel_3(i);
+      }
+
+      std::vector<DenseVector<Real>> local_acc;
+      local_acc.push_back(local_accel_dv_0);
+      local_acc.push_back(local_accel_dv_1);
+      local_acc.push_back(local_accel_dv_2);
+      local_acc.push_back(local_accel_dv_3);
+
+      std::vector<DenseVector<Real>> local_rot_acc;
[…]
