### 2020-03-05 3d576ff
Use getParam for distribution parameters
(refs #14729)
