+++ modules/stochastic_tools/include/distributions/BoostNormalDistribution.h
@@ -25,7 +25,4 @@ public:
   static InputParameters validParams();
 
   BoostNormalDistribution(const InputParameters & parameters);
-
-  Real getMean() const { return _distribution_unique_ptr->mean(); }
-  Real getStdDev() const { return _distribution_unique_ptr->standard_deviation(); }
 };
