### 2020-03-05 1924c62
Update equation syntax in markdown
(refs #14729)
