+++ framework/src/utils/Units.C
@@ -1,0 +1,536 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "Units.h"
+#include <stack>
+#include <algorithm>
+
+const std::map<std::string, Real> MooseUnits::_si_prefix = {{"E", 1e18},
+                                                            {"P", 1e15},
+                                                            {"T", 1e12},
+                                                            {"G", 1e9},
+                                                            {"M", 1e6},
+                                                            {"k", 1e3},
+                                                            {"h", 1e2},
+                                                            {"da", 1e1},
+                                                            {"d", 1e-1},
+                                                            {"c", 1e-2},
+                                                            {"m", 1e-3},
+                                                            {"mu", 1e-6},
+                                                            {"μ", 1e-6},
+                                                            {"n", 1e-9},
+                                                            {"p", 1e-12},
+                                                            {"f", 1e-15},
+                                                            {"a", 1e-18},
+                                                            {"z", 1e-21},
+                                                            {"y", 1e-24}};
+
+const std::vector<std::pair<std::string, MooseUnits>> MooseUnits::_unit_table = {
+    // avoid matching meters
+    {"Ohm",
+     {
+         1,
+         {{MooseUnits::BaseUnit::KILOGRAM, 1},
+          {MooseUnits::BaseUnit::METER, 2},
+          {MooseUnits::BaseUnit::SECOND, -3},
+          {MooseUnits::BaseUnit::AMPERE, -2}},
+     }},
+    {"atm",
+     {101.325e3,
+      {{MooseUnits::BaseUnit::KILOGRAM, 1},
+       {MooseUnits::BaseUnit::METER, -1},
+       {MooseUnits::BaseUnit::SECOND, -2}}}}, // Standard atmosphere
+    // avoid matching volts
+    {"eV",
+     {
+         1.602176634e-19,
+         {{MooseUnits::BaseUnit::KILOGRAM, 1},
+          {MooseUnits::BaseUnit::METER, 2},
+          {MooseUnits::BaseUnit::SECOND, -2}},
+     }}, // electronvolt
+    // avoid matching grams
+    {"erg",
+     {
+         1e-7,
+         {{MooseUnits::BaseUnit::KILOGRAM, 1},
+          {MooseUnits::BaseUnit::METER, 2},
+          {MooseUnits::BaseUnit::SECOND, -2}},
+     }},                                                  // erg
+    {"Ang", {1e-10, {{MooseUnits::BaseUnit::METER, 1}}}}, // Angstrom
+
+    // Base units
+    {"m", {1, {{MooseUnits::BaseUnit::METER, 1}}}},
+    {"g", {0.001, {{MooseUnits::BaseUnit::KILOGRAM, 1}}}},
[…]
