### 2020-03-05 f920615
Add [Surrogates] block
(refs #14729)
