### 2020-03-06 aa852b8
Removing tensorGrid function in QuadratureSampler #14710
