+++ modules/stochastic_tools/src/samplers/QuadratureSampler.C
@@ -65,42 +65,3 @@ QuadratureSampler::getQuadratureWeight(dof_id_type row_index) const
 {
   return _grid->quadratureWeight(row_index);
 }
-
-void
-QuadratureSampler::tensorGrid(const std::vector<std::vector<Real>> & points_1D,
-                              const std::vector<std::vector<Real>> & weights_1D,
-                              DenseMatrix<Real> & points,
-                              DenseVector<Real> & weights)
-{
-  unsigned int ndim = points_1D.size();
-  if (ndim == 1)
-  {
-    points.resize(points_1D[0].size(), 1);
-    weights.resize(points_1D[0].size());
-    for (unsigned int p = 0; p < points_1D[0].size(); ++p)
-    {
-      points(p, 0) = points_1D[0][p];
-      weights(p) = weights_1D[0][p];
-    }
-  }
-  else
-  {
-    std::vector<std::vector<Real>> tmpp(points_1D.begin(), points_1D.end() - 1);
-    std::vector<std::vector<Real>> tmpw(weights_1D.begin(), weights_1D.end() - 1);
-    DenseMatrix<Real> points_old;
-    DenseVector<Real> weights_old;
-    tensorGrid(tmpp, tmpw, points_old, weights_old);
-
-    points.resize(points_old.m() * points_1D[ndim - 1].size(), ndim);
-    weights.resize(points.m());
-    unsigned int k = 0;
-    for (unsigned int p = 0; p < points_1D[ndim - 1].size(); ++p)
-      for (unsigned int i = 0; i < points_old.m(); ++i)
-      {
-        weights(k) = weights_1D[ndim - 1][p] * weights_old(i);
-        for (unsigned int d = 0; d < ndim - 1; ++d)
-          points(k, d) = points_old(i, d);
-        points(k++, ndim - 1) = points_1D[ndim - 1][p];
-      }
-  }
-}
