### 2020-03-06 b29da2c
Switching to a vectorprocessor for computing PCE statistics #14710
