### 2020-03-06 idaholab/moose#14796 @WilkAndy
Thanks @permcody .   I think just a branch, pointing to the current master, is fine.  Tagging is probably unnecessary - after the JOSS paper we will simply rename `paper.md` to `joss_porous_flow_paper.md`, merge this branch with "next", and delete this branch.
