### 2020-03-06 c22a0a8
FSP: get rid of hard-coded subdomain name
Users should just use the input  parameter "blokcs" to specify what they want to include
