### 2020-03-06 38293b7
Migrated transition Functions to use new transition classes
Refs #19765
