+++ modules/thm/src/functions/PolynomialTransitionFunction.C
@@ -7,7 +7,6 @@ InputParameters
 validParams<PolynomialTransitionFunction>()
 {
   InputParameters params = validParams<SmoothTransitionFunction>();
-  params += validParams<PolynomialTransitionInterface>();
 
   params.addRequiredParam<Real>("function1_derivative_end_point", "First function");
   params.addRequiredParam<Real>("function2_derivative_end_point", "Second function");
@@ -19,25 +18,26 @@ validParams<PolynomialTransitionFunction>()
 
 PolynomialTransitionFunction::PolynomialTransitionFunction(const InputParameters & parameters)
   : SmoothTransitionFunction(parameters),
-    PolynomialTransitionInterface(this),
 
     _df1dx_end_point(getParam<Real>("function1_derivative_end_point")),
-    _df2dx_end_point(getParam<Real>("function2_derivative_end_point"))
+    _df2dx_end_point(getParam<Real>("function2_derivative_end_point")),
+
+    _transition(_x_center, _transition_width)
 {
   Point p1, p2;
   Real t1 = 0.0, t2 = 0.0;
   if (_use_time)
   {
-    t1 = _x1;
-    t2 = _x2;
+    t1 = _transition.leftEnd();
+    t2 = _transition.rightEnd();
   }
   else
   {
-    p1(_component) = _x1;
-    p2(_component) = _x2;
+    p1(_component) = _transition.leftEnd();
+    p2(_component) = _transition.rightEnd();
   }
 
-  initializeTransitionData(
+  _transition.initialize(
       _function1.value(t1, p1), _function2.value(t2, p2), _df1dx_end_point, _df2dx_end_point);
 }
 
@@ -46,7 +46,7 @@ PolynomialTransitionFunction::value(Real t, const Point & p) const
 {
   const Real x = _use_time ? t : p(_component);
 
-  return computeTransitionValue(x, _function1.value(t, p), _function2.value(t, p));
+  return _transition.value(x, _function1.value(t, p), _function2.value(t, p));
 }
 
 RealVectorValue
