+++ modules/thm/doc/content/source/functions/PolynomialTransitionFunction.md
@@ -2,7 +2,7 @@
 
 !syntax description /Functions/PolynomialTransitionFunction
 
-This function uses [PolynomialTransitionInterface.md] to smoothly transition
+This function uses [CubicTransition.md] to smoothly transition
 between two functions in time or along a spatial axis.
 
 !syntax parameters /Functions/PolynomialTransitionFunction
