### 2020-03-06 666a3bd
Removed smooth transition interface classes
Refs #19765
