### 2020-03-06 4463470
Renamed PolynomialTransitionFunction to CubicTransitionFunction
"Cubic" is a better descriptor than "Polynomial".

Refs #19765
