+++ modules/thm/doc/content/source/functions/PolynomialTransitionFunction.md
@@ -1,12 +1,0 @@
-# PolynomialTransitionFunction
-
-!syntax description /Functions/PolynomialTransitionFunction
-
-This function uses [CubicTransition.md] to smoothly transition
-between two functions in time or along a spatial axis.
-
-!syntax parameters /Functions/PolynomialTransitionFunction
-
-!syntax inputs /Functions/PolynomialTransitionFunction
-
-!syntax children /Functions/PolynomialTransitionFunction
