### 2020-03-06 6f63687
Add tests for latin hypercube sampler
(refs #14830)
