### 2020-03-06 bcfc436
Removed previous ACInterfaceBetaPenalty files
