+++ modules/phase_field/doc/content/bib/phase_field.bib
@@ -1,131 +1,0 @@
-@article {Miehe10,
-  author = {Miehe, C. and Welschinger, F. and Hofacker, M.},
-  title = {Thermodynamically consistent phase-field models of fracture: Variational principles and multi-field FE implementations},
-  journal = {International Journal for Numerical Methods in Engineering},
-  volume = {83},
-  number = {10},
-  publisher = {John Wiley & Sons, Ltd.},
-  issn = {1097-0207},
-  url = {http://dx.doi.org/10.1002/nme.2861},
-  doi = {10.1002/nme.2861},
-  pages = {1273--1311},
-  keywords = {fracture, crack propagation, phase-fields, gradient-type damage, incremental variational principles, finite elements, coupled multi-field problems},
-  year = {2010},
-}
-
-@article{kim_phase-field_1999,
-  title = {Phase-field model for binary alloys},
-  volume = {60},
-  url = {http://link.aps.org/doi/10.1103/PhysRevE.60.7186},
-  doi = {10.1103/PhysRevE.60.7186},
-  abstract = {We present a phase-field model ({PFM}) for solidification in binary alloys, which is found from the phase-field model for a pure material by direct comparison of the variables for a pure material solidification and alloy solidification. The model appears to be equivalent with the Wheeler-Boettinger-{McFadden} ({WBM}) model [A.A. Wheeler, W. J. Boettinger, and G. B. {McFadden}, Phys. Rev. A 45, 7424 (1992)], but has a different definition of the free energy density for interfacial region. An extra potential originated from the free energy density definition in the {WBM} model disappears in this model. At a dilute solution limit, the model is reduced to the Tiaden et al. model [Physica D 115, 73 (1998)] for a binary alloy. A relationship between the phase-field mobility and the interface kinetics coefficient is derived at a thin-interface limit condition under an assumption of negligible diffusivity in the solid phase. For a dilute alloy, a steady-state solution of the concentration profile across the diffuse interface is obtained as a function of the interface velocity and the resultant partition coefficient is compared with the previous solute trapping model. For one dimensional steady-state solidification, where the classical sharp-interface model is exactly soluble, we perform numerical simulations of the phase-field model: At low interface velocity, the simulated results from the thin-interface {PFM} are in excellent agreement with the exact solutions. As the partition coefficient becomes close to unit at high interface velocities, whereas, the sharp-interface {PFM} yields the correct answer.},
-  number = {6},
-  urldate = {2014-03-31},
-  journal = {Physical Review E},
-  author = {Kim, Seong Gyoon and Kim, Won Tae and Suzuki, Toshio},
-  month = dec,
-  year = {1999},
-  pages = {7186--7197},
-  annote = {This is the {KKS} model paper},
[…]

[…]
