### 2020-03-06 aa3ca76
Small changes for computing statistics with poly chaos #14710
