### 2020-03-06 b8bf0e1
ACInterfaceBetaPenalty files added -- ACInterfaceBetaPenalty documentation file updated -- ACInterfaceBetaPenalty kernel is derived from ACInterface -- phase_field.bib file updated to include reference Clayton Knap 2015
