### 2020-03-06 5f1f3d7
Document Latin hypercube sampler
(refs #14830)
