### 2020-03-06 4bf5578
Remove 'Sampler' suffix from Sampler objects
This just makes a nicer input file; the suffix is being kept on the object name to avoid name conflicts that could arise

(refs #14830)
