### 2020-03-06 c9f7ed4
Remove Distribution suffix from Distribution objects
(refs #14830)
