+++ modules/stochastic_tools/src/distributions/UniformDistribution.C
@@ -9,7 +9,8 @@
 
 #include "UniformDistribution.h"
 
-registerMooseObject("StochasticToolsApp", UniformDistribution);
+registerMooseObjectAliased("StochasticToolsApp", UniformDistribution, "Uniform");
+registerMooseObjectReplaced("StochasticToolsApp", UniformDistribution, "07/01/2020 00:00", Uniform);
 
 defineLegacyParams(UniformDistribution);
 
