### 2020-03-06 1e7edbc
Apply suggestions from code review
Co-Authored-By: Casey Icenhour <cticenho@ncsu.edu>
