+++ modules/doc/content/newsletter/2020_03.md
@@ -37,8 +37,8 @@ family. The update also enables used of periodic boundary conditions with LAGRAN
 
 ### Much faster periodic boundaries
 
-This was actually in a previous libMesh update, but the improved performance make it worth
-mentioniong here. Periodic boundaries now use a tree-search (O(log N) complexity) instead of a
+This was actually in a previous libMesh update, but the improved performance makes it worth
+mentioning here. Periodic boundaries now use a tree-search (O(log N) complexity) instead of a
 linear (O(N)) search. Consequently, MOOSE users of the periodic BC capability should expect
 significant speed-up of their simulations (especially for large problems).
 
@@ -46,7 +46,7 @@ significant speed-up of their simulations (especially for large problems).
 
 The parallel scheduling capability of the TestHarness has been changed to reduce potential pitfalls when
 creating several related tests. Previously, the TestHarness would make all tests in a single directory
-elegible to schedule at the same time unless "prereq" parameters were added to force serialization. Now
+eligible to schedule at the same time unless "prereq" parameters were added to force serialization. Now
 the TestHarness will only schedule a single test from a single directory by default. High concurrency is
 still achievable by having several different directories in a single application. The `prereq` parameter should
 still be used to declare "true" dependencies since execution by input file order is not guaranteed.
