### 2020-03-06 62203f5
Removing race condition with compilation
Make sure that we don't try to compile the combined executable
more than once.

closes #14843
