+++ modules/stochastic_tools/doc/content/source/vectorpostprocessors/PolyChaosData.md
@@ -1,0 +1,1 @@
+!template load file=stubs/moose_object.md.template name=PolyChaosData syntax=/VectorPostprocessors/PolyChaosData
