### 2020-03-07 c1bece6
Update upcoming_training.md
Fix typos

refs #14837
