### 2020-03-08 0d65130
Updating Citing reference for final version
refs #13507
