### 2020-03-08 643fcdd
Realpath can fail and should error out immediately (#14827)
