### 2020-03-08 64ddf93
comment-out the hidden.yml as suggested by @cticenhour.  Refs #14789
