+++ modules/geochemistry/doc/config.yml
@@ -11,7 +11,9 @@ Extensions:
         home: /modules/geochemistry
     MooseDocs.extensions.appsyntax:
         executable: ${MOOSE_DIR}/modules/geochemistry
-        hide: !include framework/doc/hidden.yml
+        hide:
+	    framework: !include framework/doc/hidden.yml
+	    #geochemistry: !include modules/geochemistry/doc/hidden.yml
         remove: !include framework/doc/remove.yml
         includes:
             - framework/include
