+++ modules/geochemistry/doc/content/modules/geochemistry/equilibrium_reactions.md
@@ -1,7 +1,5 @@
 # Equilibrium reactions: a primer
 
-Author: Andy Wilkins
-
 Notation and definitions are described in [geochemistry_nomenclature.md].
 
 Consider the hypothetical reactions
@@ -19,7 +17,7 @@ Now consider the change of free energy, $G$, corresponding to an increase in the
 = & \frac{\partial G}{\partial n_{B}} - \frac{\partial G}{\partial n_{C}}\frac{c}{b} - \frac{\partial G}{\partial n_{D}}\frac{d}{b} \ .
 \end{aligned}
 \end{equation}
-The first line expresses that when the number of moles of B is changed, the number of moles of C and D must change too, by the equilibrium reaction (the partial derivatives indicated "keep everything else fixed").  The second line introduces the correct stoichiometry.  That is: an increase in $n_{B}$ must be accompanied by a corresponding decrease in $n_{C}$ and $n_{D}$.
+The first line expresses that when the number of moles of B is changed, the number of moles of C and D must change too, by the equilibrium reaction (the partial derivatives indicate "keep everything else fixed").  The second line introduces the correct stoichiometry.  That is: an increase in $n_{B}$ must be accompanied by a corresponding decrease in $n_{C}$ and $n_{D}$.
 
 "Equilibrium" means a minimum in the free energy, so the above derivative must be zero.  Also, the total derivatives with respect to $n_{C}$ and $n_{D}$ must be zero, but they give the same result.  The partial derivatives, e.g. $\partial G/\partial n_{B}$, themselves shouldn't be zero, as the system cannot ever evolve along the direction of changing $n_{B}$ while keeping $n_{C}$ and $n_{D}$ fixed.
 
@@ -39,7 +37,7 @@ Given the functional forms of the $\mu$, and, say $n_{C}$ and $n_{D}$, we can fi
 
 The "theory of ideal solutions" states that for each constituent
 \begin{equation}
-\mu = mu^{0} + RT \log X \ ,
+\mu = \mu^{0} + RT \log X \ ,
 \end{equation}
 where
 
@@ -48,7 +46,7 @@ where
 - $R = 8.314\ldots\,$J.K$^{-1}$.mol$^{-1}$ is the gas constant
 - $T$ \[K\] is temperature
 - $\log$ is the natural logarithm
-- $X$ is the [mole fraction](geochemistry_nomenclature.md) of the constituent
+- $X$ \[dimensionless\] is the [mole fraction](geochemistry_nomenclature.md) of the constituent
 
 ## Mass action for ideal solutions
 
@@ -72,16 +70,11 @@ This is called "mass action".
 
 The chemical potential in the non-ideal case is written as
 \begin{equation}
-\mu = mu^{0} + RT \log a \ ,
+\mu = \mu^{0} + RT \log a \ ,
 \end{equation}
-where
+where most parameters have been defined above, but
 
-- $\mu$ \[J.mol$^{-1}$\] is the chemical potential
-- $\mu^{0}$ \[J.mol$^{-1}$\] is the constant number, independent of temperature, pressure, other constituents, etc, but which depends on the reaction in question
-- $R = 8.314\ldots\,$J.K$^{-1}$.mol$^{-1}$ is the gas constant
-- $T$ \[K\] is temperature
-- $\log$ is the natural logarithm
-- $a$ is the constituent's [activity](activity_coefficients.md), which depends on temperature and potentially other things.
+- $a$ \[dimensionless\] is the constituent's [activity](activity_coefficients.md), which depends on temperature and potentially other things.
 
[…]
