+++ modules/geochemistry/doc/content/modules/geochemistry/kinetics.md
@@ -1,7 +1,5 @@
 # Kinetics of mineralisation, redox reactions and complex species
 
-Author: Andy Wilkins
-
 Notation and definitions are described in [geochemistry_nomenclature.md].
 
 The material below follows Chapters 16 and 17 of [!cite](bethke_2007).
