+++ modules/geochemistry/doc/content/modules/geochemistry/fugacity.md
@@ -1,7 +1,5 @@
 # Fugacity
 
-Author: Andy Wilkins
-
 Notation and definitions are described in [geochemistry_nomenclature.md].
 
 The core material below follows Chapter 2 of [!cite](prausnitz).  Fugacity coefficients follow [!cite](spycher1988) as well as Appendix B of [!cite](toughreact).  Section 3.1.3.3 of [!cite](bethke_2007) offers some further equations, but I find them difficult to undrestand.
@@ -14,7 +12,7 @@ Here
 
 - $\mu$ \[J.mol$^{-1}$\] is the chemical potential
 - $\mu^{0}$ \[J.mol$^{-1}$\] is a constant
-- $f^{0}$ \[Pa\]is the gas fugacity at 1$\,$atm (or it might be $1\,\mathrm{bar}=10^{5}\,\mathrm{Pa}$ but I think the fugacity is not so precisely known that this small difference matters) and the temperature of interest
+- $f^{0}$ \[Pa\] is the gas fugacity at 1$\,$atm (or it might be $1\,\mathrm{bar}=10^{5}\,\mathrm{Pa}$ but I think the fugacity is not so precisely known that this small difference matters) and the temperature of interest
 - $R = 8.314\ldots\,$J.K$^{-1}$.mol$^{-1}$ is the gas constant
 - $T$ \[K\] is temperature
 - $\log$ is the natural logarithm

[…]
