+++ modules/porous_flow/paper.md
@@ -29,11 +29,11 @@ bibliography: paper.bib
 
 - Groundwater studies [@herron; @crosbie].  This often involves assessing the impacts of human interventions on groundwater resources, and the consequences on the biosphere.  It is frequently necessary to use large and complex models that include the effect of rainfall patterns, spatially and temporally varying evaporation and transpiration, seasonal river flows, realistic topography and hydrogeology, as well as human factors such as mines, water bores, dams, etc.  In certain situations, such as assessing the impact of mining on groundwater, it is vital that the model includes the deformation of the porous material (the rock) since this greatly impacts groundwater flow.  In other situations, it is useful to be able to track the transport of tracers and pollutants through the system, or to model the unsaturated (vadose) zone, where air and groundwater coexist in the subsurface.
 
-- Geothermal modelling: the simulation of natural and man-made geothermal systems, typically in order to quantify heat energies that may be extracted [@birdsall; @guglielmetti; @lima; @niederau].  Models in this field involve the coupled transport of heat and fluid.  Many models are often large-scale, but some focus on small regions close to wellbores to investigate wellbore integrity.  In this field, it often is useful to assess the impacts of subsurface heterogeneity, realistic pumping regimes, geochemical effects, fracture flow, and the creation of new fractures through deliberate or unintentional hydrofracturing, as well as potential microseismicity.
+- Geothermal modelling: the simulation of natural and man-made geothermal systems, typically in order to quantify heat energies that may be extracted [@birdsall; @guglielmetti; @lima; @niederau; @jin].  Models in this field involve the coupled transport of heat and fluid.  Many models are often large-scale, but some focus on small regions close to wellbores to investigate wellbore integrity.  In this field, it often is useful to assess the impacts of subsurface heterogeneity, realistic pumping regimes, geochemical effects, fracture flow, and the creation of new fractures through deliberate or unintentional hydrofracturing, as well as potential microseismicity.
 
 - Multiphase modelling, such as the simulation of CO$_{2}$ sequestration [@green2018; @lima], H$_{2}$ storage, or methane flows around coal mines [@c26050; @c25065].  These simulations involve the interaction and flow of at least two fluid phases (e.g. gas and water).  Changes in temperature often impact these flows, and realistic equations of state for the fluids involved are necessary.  For instance, CO$_{2}$ sequestration typically involves pumping high-pressure, cold CO$_{2}$ into a warmer aquifer filled with brine: the CO$_{2}$ dissolves into the brine, cooling it and changing its density and flow characteristics.  In many of these models, the altered pressure and temperature can potentially cause mechanical failure of the subsurface, and hence the creation of new flow paths and microseismicity [@lima]
 
-- Understanding mineralization [@heather] in the subsurface often involves fluid flow, heat flow and coupled 3D deformation, simulating over millennia of real time and using realistic, and therefore complicated, basin-scale geology.
+- Understanding mineralization [@heather1; @heather2] in the subsurface often involves fluid flow, heat flow and coupled 3D deformation, simulating over millennia of real time and using realistic, and therefore complicated, basin-scale geology.
 
 Other situations where ``PorousFlow`` could be used include oil and gas production, nuclear waste repositories and chemical leaching.
 
@@ -50,6 +50,6 @@ For convenience, the source code for ``PorousFlow``, https://github.com/idaholab
 
 # Acknowledgements
 
[…]
