+++ modules/porous_flow/paper.bib
@@ -80,7 +80,7 @@ url={https://www.acarp.com.au/abstracts.aspx?repId=C25065}
 @inproceedings{green2018,
 title={Community Code for Simulating {CO}$_2$ Storage: Modelling Multiphase Flow with Coupled Geomechanics and Geochemistry Using the Open-Source Multiphysics Framework {MOOSE}},
 author={Green, Christopher P. and Ennis-King, Jonathan and La Force, Tara and Wilkins, Andy},
-booktitle={14th Greenhouse Gas Control Technologies Conference Melbourne},
+booktitle={{14th Greenhouse Gas Control Technologies Conference Melbourne}},
 pages={21--26},
 year={2018}
 }
@@ -88,7 +88,7 @@ year={2018}
 @inproceedings{birdsall,
 title={{Modeling Ground Surface Deformation at the Swiss HEATSTORE Underground Thermal Energy Storage Sites}},
 author={Birdsell, Daniel T. and Saar, Martin O.},
-booktitle={Proceedings World Geothermal Congress 2020, Reykjavik, Iceland April 26 -- May 2, 2020},
+booktitle={{Proceedings World Geothermal Congress 2020, Reykjavik, Iceland April 26 -- May 2, 2020}},
 pages={1--6},
 year={2020}
 }
@@ -96,7 +96,7 @@ year={2020}
 @inproceedings{guglielmetti,
 title={{HEATSTORE SWITZERLAND: New Opportunities of Geothermal District Heating Network Sustainable Growth by High Temperature Aquifer Thermal Energy Storage Development}},
 author={Guglielmetti, L. and Alt-Epping,  P. and Birdsell, D. and de Oliveira,  F. and Diamond,  L. and Driesner, T.  and Eruteya, O. and Hollmuller, P. et al. and  Saar, M.O.},
-booktitle={Proceedings World Geothermal Congress 2020, Reykjavik, Iceland April 26 -- May 2, 2020},
+booktitle={{Proceedings World Geothermal Congress 2020, Reykjavik, Iceland April 26 -- May 2, 2020}},
 pages={1--13},
 year={2020}
 }
@@ -104,7 +104,7 @@ year={2020}
 @inproceedings{niederau,
 title={Characterization of Subsurface Heat-Transport Processes in the Canton of Aargau Using an Integrative Workflow},
 author={Niederau, Jan and Ebigbo, Anozie and Saar, Martin O.},
-booktitle={Proceedings World Geothermal Congress 2020, Reykjavik, Iceland April 26 -- May 2, 2020},
+booktitle={{Proceedings World Geothermal Congress 2020, Reykjavik, Iceland April 26 -- May 2, 2020}},
 pages={1--6},
 year={2020}
 }
@@ -118,16 +118,28 @@ title = {Impact of Effective Normal Stress on Capillary Pressure in a Single Nat
 booktitle={European Geothermal Congress, 11--14 June, 2019, The Hague, The Netherlands},
 }
 
-@article{heather,
-author = {Peter Schaubs and Heather Sheldon and Teagan Blaikie and Marcus Kunzmann and Susanne Schmid and Sam Spinks},
-title = {{Basin-scale fluid-flow models of the McArthur River mineral system: constraints from geochemistry, geophysics and sequence stratigraphy}},
-journal = {ASEG Extended Abstracts},
-volume = {2019},
-number = {1},
-pages = {1-5},
-year  = {2019},
-publisher = {Taylor & Francis},
-doi = {10.1080/22020586.2019.12073198}
+@inproceedings{heather1,
+author = {Peter M Schaubs and Heather A Sheldon},
[…]
