+++ modules/tensor_mechanics/src/kernels/InertialForceShell.C
@@ -149,12 +149,12 @@ ADInertialForceShell<compute_stage>::ADInertialForceShell(const InputParameters 
   }
 }
 
+
+
 template <ComputeStage compute_stage>
 ADReal
 ADInertialForceShell<compute_stage>::computeQpResidual()
 {
-  prepareVectorTag(_assembly, _var.number());
-
   if (_dt != 0.0)
   {
     // fetch the four nodes for _current_elem
@@ -387,120 +387,130 @@ ADInertialForceShell<compute_stage>::computeQpResidual()
       local_rot_acc.push_back(local_rot_accel_dv_2);
       local_rot_acc.push_back(local_rot_accel_dv_3);
 
-      for (unsigned int qpz = 0; qpz < _t_points.size(); ++qpz)
+      // _local_force for each of the nodes. Try vector form first (containing three displacement
+      // components)
+
+      ADReal factor_qxy = _2d_weights[qp_xy] *_ad_JxW[qp_xy] * _ad_coord[qp_xy];
+
+      for (unsigned int dim = 0; dim < 3; dim++)
       {
-        // _local_force for each of the nodes. Try vector form first (containing three displacement
-        // components)
-        _2d_weights[qp_xy] * _t_weights[qpz] * _JxW[qp_xy] * _coord[qp_xy];
+        _local_force[0](dim) +=
+            factor_qxy *
+            (_phi_map[0][qp_xy] * _phi_map[0][qp_xy] * local_acc[0](dim) +
+             _phi_map[0][qp_xy] *
+                 (G1(dim, 0) * local_rot_acc[0](0) + G1(dim, 1) * local_rot_acc[0](1)) +
+             _phi_map[0][qp_xy] * _phi_map[1][qp_xy] * local_acc[1](dim) +
+             _phi_map[0][qp_xy] *
+                 (G2(dim, 0) * local_rot_acc[1](0) + G2(dim, 1) * local_rot_acc[1](1)) +
+             _phi_map[0][qp_xy] * _phi_map[2][qp_xy] * local_acc[2](dim) +
+             _phi_map[0][qp_xy] *
+                 (G3(dim, 0) * local_rot_acc[2](0) + G3(dim, 1) * local_rot_acc[2](1)) +
+             _phi_map[0][qp_xy] * _phi_map[3][qp_xy] * local_acc[3](dim) +
+             _phi_map[0][qp_xy] *
+                 (G4(dim, 0) * local_rot_acc[3](0) + G4(dim, 1) * local_rot_acc[3](1)));
 
-        for (unsigned int dim = 0; dim < 3; dim++)
-        {
-          _local_force[0](dim) +=
-              _phi_map[0][qp_xy] * _phi_map[0][qp_xy] * local_acc[0](dim) +
-              _phi_map[0][qp_xy] *
-                  (G1(dim, 0) * local_rot_acc[0](0) + G1(dim, 1) * local_rot_acc[0](1)) +
-              _phi_map[0][qp_xy] * _phi_map[1][qp_xy] * local_acc[1](dim) +
-              _phi_map[0][qp_xy] *
-                  (G2(dim, 0) * local_rot_acc[1](0) + G2(dim, 1) * local_rot_acc[1](1)) +
-              _phi_map[0][qp_xy] * _phi_map[2][qp_xy] * local_acc[2](dim) +
-              _phi_map[0][qp_xy] *
-                  (G3(dim, 0) * local_rot_acc[2](0) + G3(dim, 1) * local_rot_acc[2](1)) +
-              _phi_map[0][qp_xy] * _phi_map[3][qp_xy] * local_acc[3](dim) +
[…]
