### 2020-03-09 45232b6
Upate documentation for Surrogate block
(refs #14729)
