+++ modules/stochastic_tools/doc/content/source/vectorpostprocessors/PolyChaosStatistics.md
@@ -1,6 +1,6 @@
 # PolyChaosStatistics
 
-This object is designed to output statistically relevant quantities from a trained polynomial chaos expansion. So far, mean, standard deviation, skewness, and kurtosis can all be computed using this postprocessor. See [PolynomialChaos][/PolynomialChaos.md] for more details on the calculation of these statistics.
+This object is designed to output statistically relevant quantities from a trained polynomial chaos expansion. So far, mean, standard deviation, skewness, and kurtosis can all be computed using this postprocessor. See [PolynomialChaos.md] for more details on the calculation of these statistics.
 
 ## Example Syntax
 
