+++ test/tests/functions/coarsened_piecewise_linear/generate_data.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
 import numpy as np
 
