+++ framework/scripts/compile_commands.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python3
+#!/usr/bin/env python33
 #* This file is part of the MOOSE framework
 #* https://www.mooseframework.org
 #*
