+++ python/MooseDocs/test/extensions/test_appsyntax.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 import os
 import unittest
 import logging
