### 2020-03-09 a119051
Cleaned up documentation of XFEM UserObjects. ref #10210
