### 2020-03-09 01cd3c2
Simplifying polynomial chaos tests #14710
