### 2020-03-09 47a3291
Remove PerfGraph calls in rand functions
(refs #14830)
