+++ framework/src/samplers/Sampler.C
@@ -68,12 +68,10 @@ Sampler::Sampler(const InputParameters & parameters)
     _perf_get_global_samples(registerTimedSection("getGlobalSamples", 1)),
     _perf_get_local_samples(registerTimedSection("getLocalSamples", 1)),
     _perf_get_next_local_row(registerTimedSection("getNextLocalRow", 1)),
-    _perf_advance_generator(registerTimedSection("advanceGenerators", 2)),
     _perf_sample_row(registerTimedSection("computeSampleRow", 2)),
     _perf_local_sample_matrix(registerTimedSection("computeLocalSampleMatrix", 2)),
     _perf_sample_matrix(registerTimedSection("computeSampleMatrix", 2)),
-    _perf_get_rand(registerTimedSection("getRand", 5)),
-    _perf_get_randl(registerTimedSection("getRandl", 5))
+    _perf_advance_generator(registerTimedSection("advanceGenerators", 2))
 {
 }
 
@@ -297,8 +295,6 @@ double
 Sampler::getRand(const unsigned int index)
 {
   mooseAssert(index < _generator.size(), "The seed number index does not exists.");
-  TIME_SECTION(_perf_get_rand);
-
   return _generator.rand(index);
 }
 
@@ -306,8 +302,6 @@ uint32_t
 Sampler::getRandl(unsigned int index, uint32_t lower, uint32_t upper)
 {
   mooseAssert(index < _generator.size(), "The seed number index does not exists.");
-  TIME_SECTION(_perf_get_randl);
-
   return _generator.randl(index, lower, upper);
 }
 
