+++ framework/include/samplers/Sampler.h
@@ -275,11 +275,9 @@ private:
   const PerfID _perf_get_global_samples;
   const PerfID _perf_get_local_samples;
   const PerfID _perf_get_next_local_row;
-  const PerfID _perf_advance_generator;
   const PerfID _perf_sample_row;
   const PerfID _perf_local_sample_matrix;
   const PerfID _perf_sample_matrix;
-  const PerfID _perf_get_rand;
-  const PerfID _perf_get_randl;
+  const PerfID _perf_advance_generator;
   ///@}
 };
