### 2020-03-09 2cd4380
Update Sobol sampler to operate with supplied Sampler objects
(refs #14784)
