+++ modules/stochastic_tools/src/samplers/SobolSampler.C
@@ -20,12 +20,10 @@ SobolSampler::validParams()
 {
   InputParameters params = Sampler::validParams();
   params.addClassDescription("Sobol variance-based sensitivity analysis Sampler.");
-  params.addRequiredParam<dof_id_type>("num_rows", "The number of rows per matrix to generate.");
-  params.addRequiredParam<std::vector<DistributionName>>(
-      "distributions",
-      "The distribution names to be sampled, the number of distributions provided defines the "
-      "number of columns per matrix.");
   params.addParam<bool>("resample", true, "Create the re-sample matrix for second-order indices.");
+  params.addRequiredParam<SamplerName>("sampler_a", "The 'sample' matrix.");
+  params.addRequiredParam<SamplerName>("sampler_b", "The 're-sample' matrix.");
+  params.suppressParameter<ExecFlagEnum>("execute_on");
   return params;
 }
 
@@ -33,21 +31,26 @@ SobolSampler::SobolSampler(const InputParameters & parameters)
   : Sampler(parameters),
     _m1_matrix(0, 0),
     _m2_matrix(0, 0),
-    _distribution_names(getParam<std::vector<DistributionName>>("distributions")),
+    _sampler_a(getSampler("sampler_a")),
+    _sampler_b(getSampler("sampler_b")),
     _resample(getParam<bool>("resample")),
-    _num_inputs(_distribution_names.size()),
-    _num_matrices(_resample ? 2 * _distribution_names.size() + 2 : _distribution_names.size() + 2),
-    _num_rows_per_matrix(getParam<dof_id_type>("num_rows")),
     _perf_sample_setup(registerTimedSection("sampleSetup", 3)),
     _perf_sample_teardown(registerTimedSection("computeTearDown", 3)),
     _perf_compute_sample(registerTimedSection("computeSample", 4))
 {
-  setNumberOfRandomSeeds(2);
+  if (_sampler_a.getNumberOfCols() != _sampler_b.getNumberOfCols())
+    paramError("sampler_a", "The supplied Sampler objects must have the same number of columns.");
 
-  for (const DistributionName & name : _distribution_names)
-    _distributions.push_back(&getDistributionByName(name));
+  if (_sampler_a.getNumberOfRows() != _sampler_b.getNumberOfRows())
+    paramError("sampler_a", "The supplied Sampler objects must have the same number of columns.");
 
-  setNumberOfCols(_distribution_names.size());
+  // Compute the number of matrices
+  const dof_id_type num_cols = _sampler_a.getNumberOfCols();
+  _num_matrices = _resample ? 2 * num_cols + 2 : num_cols + 2;
+  _num_rows_per_matrix = _sampler_a.getNumberOfRows();
+
+  // Initialize this object
+  setNumberOfCols(_sampler_a.getNumberOfCols());
   setNumberOfRows(_num_rows_per_matrix * _num_matrices);
 }
 
@@ -56,14 +59,8 @@ SobolSampler::sampleSetUp()
 {
   TIME_SECTION(_perf_sample_setup);
 
-  _m1_matrix.resize(_num_rows_per_matrix, getNumberOfCols());
-  _m2_matrix.resize(_num_rows_per_matrix, getNumberOfCols());
-  for (dof_id_type i = 0; i < _num_rows_per_matrix; ++i)
-    for (dof_id_type j = 0; j < getNumberOfCols(); ++j)
-    {
-      _m1_matrix(i, j) = _distributions[j]->quantile(this->getRand(0));
-      _m2_matrix(i, j) = _distributions[j]->quantile(this->getRand(1));
-    }
+  _m1_matrix = _sampler_a.getGlobalSamples();
[…]
