+++ modules/doc/content/application_usage/input_syntax.md
@@ -45,7 +45,7 @@ ${cmd [arg1] [arg2]...}
 
 where arguments are whitespace delimited and are allowed to be brace expressions themselves.
 Brace expressions can appear as field values either unquoted or quoted (i.e. inside a
-string-value).  MOOSE currently has four built-in brace-expression commands:
+string-value).  MOOSE currently has five built-in brace-expression commands:
 
 - `${replace <var-name>}`: looks up the parameter `var-name` in the input file searching from the
   current scope looking at consecutive parent scopes until one exists, substituting its value
@@ -60,6 +60,12 @@ string-value).  MOOSE currently has four built-in brace-expression commands:
   that can reference fields from the current input file and evaluates it with the function
   parser. Variables referencing input-file fields are looked up using the same procedure as with
   the `replace` command.
+
+- `${units <arg> <unit> [-> <to_unit>]}`: takes an argument `arg` in physical units
+  of `unit` and converts it to the unit `to_unit`. For example the expression
+  `${units 1 J/mol -> eV/at}` would evaluate to the value `1.0364269656262e-05`.
+  The `to_unit` argument is optional resulting in a no-op that serves only for documentation
+  purposes (`${units 1 J/mol}` would evaluate to the value `1.0`).
 
 Examples:
 
