### 2020-03-09 f7a8cbd
Add error tests for new errors in Sobol sampler
(refs #14784)
