### 2020-03-09 2cd8ee8
ExternalPetscSolverApp: add routines to mimic backup and restore
Add PETSc routines to mimic application backup and restore

These routines serve as demo for other developers

Closes #12595
