+++ modules/external_petsc_solver/include/petscsolver/PETScDiffusionFDM.h
@@ -24,6 +24,4 @@ PETSC_EXTERN PetscErrorCode PETScExternalSolverDestroy(TS);
 PETSC_EXTERN PetscErrorCode FormIFunction(TS, PetscReal, Vec, Vec, Vec, void *);
 PETSC_EXTERN PetscErrorCode FormIJacobian(TS, PetscReal, Vec, Vec, PetscReal, Mat, Mat, void *);
 PETSC_EXTERN PetscErrorCode FormInitialSolution(TS, Vec, void *);
-
 #endif
-
