+++ modules/external_petsc_solver/src/base/ExternalPetscSolverApp.C
@@ -12,12 +12,15 @@
 #include "AppFactory.h"
 #include "MooseSyntax.h"
 #include "PETScDiffusionFDM.h"
+#include "libmesh/petsc_vector.h"
 
-template <>
 InputParameters
-validParams<ExternalPetscSolverApp>()
+ExternalPetscSolverApp::validParams()
 {
-  InputParameters params = validParams<MooseApp>();
+  InputParameters params = MooseApp::validParams();
+
+  // By default, use preset BCs
+  params.set<bool>("use_legacy_dirichlet_bc") = false;
   return params;
 }
 
@@ -27,6 +30,14 @@ ExternalPetscSolverApp::ExternalPetscSolverApp(InputParameters parameters) : Moo
 
 #if LIBMESH_HAVE_PETSC
   PETScExternalSolverCreate(_comm->get(), &_ts);
+  DM da;
+  TSGetDM(_ts, &da);
+
+  DMCreateGlobalVector(da, &_petsc_sol);
+
+  // This is required because libMesh incorrectly treats the PETSc parallel vector as a ghost vector
+  // We should be able to remove this line of code once libMesh is updated
+  VecMPISetGhost(_petsc_sol, 0, nullptr);
 #else
   mooseError("You need to have PETSc installed to use ExternalPETScApp");
 #endif
@@ -46,6 +57,28 @@ ExternalPetscSolverApp::registerAll(Factory & f, ActionFactory & af, Syntax & /*
   Registry::registerActionsTo(af, {"ExternalPetscSolverApp"});
 
   /* register custom execute flags, action syntax, etc. here */
+}
+
+std::shared_ptr<Backup>
+ExternalPetscSolverApp::backup()
+{
+  auto backup = MooseApp::backup();
+
+  // Backup current solution
+  PetscVector<Number> petsc_sol(_petsc_sol, comm());
+  dataStore(backup->_system_data, static_cast<NumericVector<Number> &>(petsc_sol), nullptr);
+
+  return backup;
+}
+
+void
+ExternalPetscSolverApp::restore(std::shared_ptr<Backup> backup, bool for_restart)
+{
+  MooseApp::restore(backup, for_restart);
+
+  // Restore previous solution
+  PetscVector<Number> petsc_sol(_petsc_sol, comm());
+  dataLoad(backup->_system_data, static_cast<NumericVector<Number> &>(petsc_sol), nullptr);
 }
 
 void
