+++ framework/doc/content/source/utils/Units.md
@@ -1,0 +1,96 @@
+# MooseUnits
+
+`MooseUnits` is a physical units parsing and conversion helper class.
+A unit object can be constructed form a string (e.g. `N*m`, `eV/at`, `kg*(m/s)^2`).
+The unit parser supports the `*`,`/`, and `^` operators as well as parenthesis `(`, `)`.
+The argument of the `^` operator is expected to be a positive or negative integer.
+
+Upon parsing all units are resolved to a combination of the
+[seven base SI units](https://en.wikipedia.org/wiki/SI_base_unit) (`m`, `g`,
+`s`, `A`, `K`, `mol`, `cd`) with their respective exponents and a prefactor. A
+canonical form is generated that permits unit comparisons to determine
+conformity (i.e. whether two units can be converted between).
+
+`Mooseunits` supports prefixing all units with a
+[metric prefix](https://en.wikipedia.org/wiki/Metric_prefix) ranging from `Y` (yotta)
+to `y` (yocto).
+
+## Supported derived units
+
+The following units are currently supported
+
+| Symbol | Base units | Name | Description |
+| - | - | - | - |
+| `Ohm` | $\text{kg}\cdot \text{m}^{2}\cdot \text{s}^{-3}\cdot \text{A}^{-2}$ | Ohm | resistance, impedance, reactance|
+| `atm` | $101325\, \text{kg}\cdot \text{m}^{-1}\cdot \text{s}^{-2}$ | Standard atmosphere | pressure |
+| `eV` | $1.60218\cdot10^{-19}\, \text{kg}\cdot \text{m}^{2}\cdot \text{s}^{-2}$ | electron Volt | energy |
+| `erg` | $10^{-7}\, \text{kg}\cdot \text{m}^{2}\cdot \text{s}^{-2}$ | | |
+| `Ang` | $10^{-10}\, \text{m}$ | Angstrom | length |
+| `m` | $\text{m}$ | meter | length |
+| `g` | $0.001\, \text{kg}$ | gram | mass |
+| `s` | $\text{s}$ | second | time |
+| `A` | $\text{A}$ | Ampere | electric current |
+| `K` | $\text{K}$ | Kelvin | temperature |
+| `mol` | $6.02214\cdot10^{23}\, \text{at}$ | mole | amount of substance|
+| `cd` | $\text{cd}$ | candela | luminous intensity |
+| `N` | $\text{kg}\cdot \text{m}\cdot \text{s}^{-2}$ | Newton | force, weight |
+| `Pa` | $\text{kg}\cdot \text{m}^{-1}\cdot \text{s}^{-2}$ | Pascal | pressure, stress |
+| `J` | $\text{kg}\cdot \text{m}^{2}\cdot \text{s}^{-2}$ | Joule | energy, work, heat |
+| `W` | $\text{kg}\cdot \text{m}^{2}\cdot \text{s}^{-3}$ | Watt | power, radiant flux |
+| `C` | $\text{A}\cdot \text{s}$ | Coulomb | electric charge |
+| `V` | $\text{kg}\cdot \text{m}^{2}\cdot \text{s}^{-3}\cdot \text{A}^{-1}$ | Volt | voltage (electrical potential), emf |
+| `F` | $\text{kg}^{-1}\cdot \text{m}^{-2}\cdot \text{s}^{4}\cdot \text{A}^{2}$ | Farad | capacitance |
+| `S` | $\text{kg}^{-1}\cdot \text{m}^{-2}\cdot \text{s}^{3}\cdot \text{A}^{2}$ | Siemens | electrical conductance |
[…]
