### 2020-03-10 a512b7a
Temporary backwards and forward compatibility for DualNumber
Refs #000
