+++ modules/geochemistry/doc/content/modules/geochemistry/index.md
@@ -2,7 +2,7 @@
 
 The `geochemistry` module is designed to solve reactive transport in geochemical systems.  The `geochemistry` module is currently in development.  The documentation describes the module's functionality, but does not currently include any description of the MOOSE implementation.
 
-The `geochemistry` module's functionality is a subset of that described in [!cite](bethke_2007).  Various aspects have not yet been included, such as the virial Pitzer/HMW models for activity.  This is because the functionality of the `geochemistry` module is currently focussed on simulation of "GeoTES" systems, which operate between around $25^{\circ}$C and $300^{\circ}$C.
+The `geochemistry` module's functionality is a subset of that described in [!cite](bethke_2007).  Various aspects have not yet been included, such as the virial Pitzer/HMW models for activity.  This is because the functionality of the `geochemistry` module is currently focussed on simulation of "GeoTES" systems, which operate between between $25^{\circ}$C and $300^{\circ}$C.
 
 The `geochemistry` module consists of two broad parts: the reactions and the transport.
 
@@ -20,6 +20,11 @@ The `geochemistry` module consists of two broad parts: the reactions and the tra
 ## Transport
 
 - [Transport in the `geochemistry` module](transport.md)
+
+## Tests and Examples
+
+Although the MOOSE code is incomplete, some [documentation of future `geochemistry` tests](tests_and_examples/geochemistry_tests_and_examples.md) has been written.
+
 
 
 

[…]
