+++ modules/geochemistry/doc/content/modules/geochemistry/tests_and_examples/surface_complexation.md
@@ -1,24 +1,94 @@
-# Sorption of selenate in loamy soil
+# Sorption onto ferric hydroxide
 
-This example closely follows Section 9.6 of [!cite](bethke_2007).
+This example closely follows Section 10.4 of [!cite](bethke_2007).
 
-We explore the sorption of selenate (SeO$_{4}^{2-}$) as predicted by the Langmuir approach.  It is assumed
+We explore the sorption of mercury, lead and sulfate onto hydrous ferric oxide at pH 4 and 8.
 
-- the sample consists of 1$\,$kg of water and 500$\,$g of dry soil (corresponding to 189$\,$cm$^{3}$ of dry soil);
-- the number of moles of sorbing sites per gram of dry soil is $0.62\times 10^{-9}\,$mol.g$^{-1}$.  This sets the total mole number of sorbing sites, $M_{p}$;
-- the equilibrium constant for Langmuir sorption is $5.4\times 10^{-6}$;
-- the pH is 7.5;
-- charge balance is enforced on Na$^{+}$.
+It is assumed that there are two sorbing sites:
+\begin{equation}
+A_{p}=\left(\ \mathrm{>(s)FeOH},\  \mathrm{>(w)FeOH}\,\right) \ .
+\end{equation}
+[!cite](bethke_2007) uses the convention that a ">" indicates something to do with sorption.  The sorbing reactions, written in the [documentation on equilibrium reactions](equilibrium.md) as
+\begin{equation}
+A_{q} \rightleftharpoons \nu_{wq}A_{w} + \sum_{i}\nu_{iq}A_{i} + \sum_{k}\nu_{kq}A_{k} + \sum_{m}\nu_{mq}A_{m} + \nu_{pq}A_{p} \ ,
+\end{equation}
+are, specifically:
+\begin{equation}
+\begin{aligned}
+\mathrm{>(w)FeOH}_{2}^{+} & \rightleftharpoons \mathrm{>(w)FeOH} + \mathrm{H}^{+} & \log_{10}K=-7.29\ , \\
+\mathrm{>(w)FeO}^{-} & \rightleftharpoons \mathrm{>(w)FeOH} - \mathrm{H}^{+} & \log_{10}K=8.93\ , \\
+\mathrm{>(w)FeOHg}^{+} & \rightleftharpoons \mathrm{>(w)FeOH} - \mathrm{H}^{+} + \mathrm{Hg}^{2+} & \log_{10}K=-6.45\ , \\
+\mathrm{>(w)FeOPb}^{+} & \rightleftharpoons \mathrm{>(w)FeOH} - \mathrm{H}^{+} + \mathrm{Pb}^{2+} & \log_{10}K=-0.3\ , \\
+\mathrm{>(w)FeSO}_{4}^{-} & \rightleftharpoons \mathrm{>(w)FeOH} + \mathrm{H}^{+} + \mathrm{SO}_{4}^{2-}  - \mathrm{H}_{2}\mathrm{O} & \log_{10}K=-7.78\ , \\
+\mathrm{>(w)FeOHSO}_{4}^{2-} & \rightleftharpoons \mathrm{>(w)FeOH} + \mathrm{SO}_{4}^{2-} & \log_{10}K=-0.79\ , \\
+\mathrm{>(s)FeOH}_{2}^{+} & \rightleftharpoons \mathrm{>(s)FeOH} + \mathrm{H}^{+} & \log_{10}K=-7.29\ , \\
+\mathrm{>(s)FeO}^{-} & \rightleftharpoons \mathrm{>(s)FeOH} - \mathrm{H}^{+} & \log_{10}K=8.93\ , \\
[…]
