### 2020-03-10 370ce8e
Fix broken automatic reference section in Moosedocs
(refs #14830)
