### 2020-03-10 4b93ac4
Update Sampler objects to use Real instead of double
(refs #14830)
