### 2020-03-10 eab9881
Handle multiple source file matches for source auto link feature
(refs #13657)
