### 2020-03-10 7359ce8
Migrate Sampler objects to TheWarehouse
(refs #14784)
