+++ framework/src/problems/FEProblemBase.C
@@ -218,7 +218,6 @@ FEProblemBase::FEProblemBase(const InputParameters & parameters)
     _aux(nullptr),
     _coupling(Moose::COUPLING_DIAG),
     _distributions(/*threaded=*/false),
-    _samplers(_app.getExecuteOnEnum()),
     _material_props(
         declareRestartableDataWithContext<MaterialPropertyStorage>("material_props", &_mesh)),
     _bnd_material_props(
@@ -1961,17 +1960,26 @@ FEProblemBase::addSampler(std::string type, const std::string & name, InputParam
   {
     std::shared_ptr<Sampler> obj = _factory.create<Sampler>(type, name, parameters, tid);
     obj->init();
-    _samplers.addObject(obj, tid);
+    theWarehouse().add(obj);
   }
 }
 
 Sampler &
 FEProblemBase::getSampler(const std::string & name, THREAD_ID tid)
 {
-  if (!_samplers.hasActiveObject(name, tid))
-    mooseError("Unable to find Sampler " + name);
-
-  return *(_samplers.getActiveObject(name, tid));
+  std::vector<Sampler *> objs;
+  theWarehouse()
+      .query()
+      .condition<AttribSystem>("Sampler")
+      .condition<AttribThread>(tid)
+      .condition<AttribName>(name)
+      .queryInto(objs);
+  if (objs.empty())
+    mooseError(
+        "Unable to find Sampler with name '" + name +
+        "', if you are attempting to access this object in the constructor of another object then "
+        "the object being retrieved must occur prior to the caller within the input file.");
+  return *(objs[0]);
 }
 
 bool
@@ -3736,14 +3744,19 @@ FEProblemBase::executeSamplers(const ExecFlagType & exec_type)
   // do a serial loop.
   for (THREAD_ID tid = 0; tid < libMesh::n_threads(); ++tid)
   {
-    const auto & objects = _samplers[exec_type].getActiveObjects(tid);
+    std::vector<Sampler *> objects;
+    theWarehouse()
+        .query()
+        .condition<AttribSystem>("Sampler")
+        .condition<AttribThread>(tid)
+        .condition<AttribExecOns>(exec_type)
+        .queryInto(objects);
+
     if (!objects.empty())
     {
       TIME_SECTION(_execute_samplers_timer);
-
-      _samplers.setup(exec_type);
-      for (auto & sampler : objects)
-        sampler->execute();
+      FEProblemBase::objectSetupHelper<Sampler>(objects, exec_type);
+      FEProblemBase::objectExecuteHelper<Sampler>(objects);
     }
   }
 }
@@ -3764,7 +3777,6 @@ FEProblemBase::updateActiveObjects()
     _residual_materials.updateActive(tid);
     _jacobian_materials.updateActive(tid);
     _discrete_materials.updateActive(tid);
-    _samplers.updateActive(tid);
   }
 
   _control_warehouse.updateActive();
