+++ framework/src/distributions/Distribution.C
@@ -18,6 +18,7 @@ Distribution::validParams()
   InputParameters params = MooseObject::validParams();
   params.addRequiredParam<std::string>("type", "class/type name identifying the distribution");
   params.registerBase("Distribution");
+  params.registerSystemAttributeName("Distribution");
   return params;
 }
 
