### 2020-03-10 10128e1
Add template argument to Sampler friend
(refs #14784)
