+++ modules/phase_field/src/userobjects/DiscreteNucleationMap.C
@@ -23,7 +23,7 @@ DiscreteNucleationMap::validParams()
   params.addClassDescription("Generates a spatial smoothed map of all nucleation sites with the "
                              "data of the DiscreteNucleationInserter for use by the "
                              "DiscreteNucleation material.");
-  params.addParam<Real>("radius", 0.0, "Radius for the inserted nuclei");
+  // params.addParam<Real>("radius", 0.0, "Radius for the inserted nuclei");
   params.addParam<Real>("int_width", 0.0, "Nucleus interface width for smooth nuclei");
   params.addRequiredParam<UserObjectName>("inserter", "DiscreteNucleationInserter user object");
   params.addCoupledVar("periodic",
@@ -39,7 +39,7 @@ DiscreteNucleationMap::DiscreteNucleationMap(const InputParameters & parameters)
     _mesh_changed(false),
     _inserter(getUserObject<DiscreteNucleationInserterBase>("inserter")),
     _periodic(isCoupled("periodic") ? coupled("periodic") : -1),
-    _radius(getParam<Real>("radius")),
+    //_radius(getParam<Real>("radius")),
     _int_width(getParam<Real>("int_width")),
     _nucleus_list(_inserter.getNucleusList())
 {
@@ -75,26 +75,38 @@ DiscreteNucleationMap::execute()
       Real r, rmin = std::numeric_limits<Real>::max();
 
       // find the distance to the closest nucleus
+      Real local_radius = 0.0;
       for (unsigned i = 0; i < _nucleus_list.size(); ++i)
       {
         // use a non-periodic or periodic distance
+        //r = _periodic < 0
+        //        ? (_q_point[qp] - _nucleus_list[i].second).norm()
+        //        : _mesh.minPeriodicDistance(_periodic, _q_point[qp], _nucleus_list[i].second);
+
+        // modified this to the struct format to get the center of the nucleus
         r = _periodic < 0
-                ? (_q_point[qp] - _nucleus_list[i].second).norm()
-                : _mesh.minPeriodicDistance(_periodic, _q_point[qp], _nucleus_list[i].second);
+            ? (_q_point[qp] - _nucleus_list[i].center).norm()
+            : _mesh.minPeriodicDistance(_periodic, _q_point[qp], _nucleus_list[i].center);
         if (r < rmin)
+        {
           rmin = r;
+          //grab the radius of the nucleus that this qp is closest to
+          local_radius = _nucleus_list[i].radius;
+  //        _console<<"nucleus_list["<<i<<"].radius = "<<_nucleus_list[i].radius<<std::endl;
+  //        _console<<"nucleus_list["<<i<<"].center = "<<_nucleus_list[i].center<<std::endl;
+        }
       }
 
       // compute intensity value with smooth interface
       Real value = 0.0;
-      if (rmin <= _radius - _int_width / 2.0) // Inside circle
+      if (rmin <= local_radius - _int_width / 2.0) // Inside circle
       {
         active_nuclei++;
         value = 1.0;
       }
-      else if (rmin < _radius + _int_width / 2.0) // Smooth interface
+      else if (rmin < local_radius + _int_width / 2.0) // Smooth interface
       {
-        Real int_pos = (rmin - _radius + _int_width / 2.0) / _int_width;
+        Real int_pos = (rmin - local_radius + _int_width / 2.0) / _int_width;
         active_nuclei++;
         value = (1.0 + std::cos(int_pos * libMesh::pi)) / 2.0;
       }
