+++ modules/phase_field/src/userobjects/DiscreteNucleationInserterBase.C
@@ -27,3 +27,21 @@ DiscreteNucleationInserterBase::DiscreteNucleationInserterBase(const InputParame
 {
   setRandomResetFrequency(EXEC_TIMESTEP_END);
 }
+
+template<>
+void
+dataStore(std::ostream & stream, DiscreteNucleationInserterBase::NucleusLocation & nl, void * context)
+{
+   storeHelper(stream, nl.time, context);
+   storeHelper(stream, nl.center, context);
+   storeHelper(stream, nl.radius, context);
+}
+
+template<>
+void
+dataLoad(std::istream & stream, DiscreteNucleationInserterBase::NucleusLocation & nl, void * context)
+{
+  loadHelper(stream, nl.time, context);
+  loadHelper(stream, nl.center, context);
+  loadHelper(stream, nl.radius, context);
+}
