+++ modules/phase_field/src/userobjects/DiscreteNucleationInserter.C
@@ -25,6 +25,12 @@ DiscreteNucleationInserter::validParams()
   params.addRequiredParam<MaterialPropertyName>(
       "probability", "Probability density for inserting a discrete nucleus");
   params.addRequiredParam<Real>("hold_time", "Time to keep each nucleus active");
+
+  params.addParam<bool>("fixed_radius", true, "flag if the radius is fixed or not");
+  params.addParam<Real>("radius", 0.0, "fixed radius (if using)");
+  params.addParam<MaterialPropertyName>("radius_material", "r_crit", "variable radius material property name");
+  params.addParam<bool>("Poisson_statistics", true, "flag if time-dependent or time-independent statistics are used");
+
   return params;
 }
 
@@ -32,8 +38,15 @@ DiscreteNucleationInserter::DiscreteNucleationInserter(const InputParameters & p
   : DiscreteNucleationInserterBase(parameters),
     _probability(getMaterialProperty<Real>("probability")),
     _hold_time(getParam<Real>("hold_time")),
-    _local_nucleus_list(declareRestartableData("local_nucleus_list", NucleusList(0)))
+    _local_nucleus_list(declareRestartableData("local_nucleus_list", NucleusList(0))),
+    _fixed_radius(getParam<bool>("fixed_radius")),
+    _radius(getParam<Real>("radius")),
+    _local_radius( !_fixed_radius ? getMaterialProperty<Real>("radius_material")
+                   : getZeroMaterialProperty<Real>("dummy_zero_material")),
+    _poisson_stats(getParam<bool>("Poisson_statistics"))
 {
+  if (_fixed_radius && MooseUtils::absoluteFuzzyLessEqual(_radius, 0, libMesh::TOLERANCE * libMesh::TOLERANCE))
+    mooseError("fixed nucleus radius size must be greater than zero");
 }
 
 void
@@ -48,7 +61,8 @@ DiscreteNucleationInserter::initialize()
     unsigned int i = 0;
     while (i < _local_nucleus_list.size())
     {
-      if (_local_nucleus_list[i].first <= _fe_problem.time())
+      //if (_local_nucleus_list[i].first <= _fe_problem.time())
+      if (_local_nucleus_list[i].time <= _fe_problem.time())
       {
         // remove entry (by replacing with last element and shrinking size by one)
         _local_nucleus_list[i] = _local_nucleus_list.back();
@@ -79,15 +93,22 @@ DiscreteNucleationInserter::execute()
 
     const Real random = getRandomReal();
 
+    // branch the operation for using Poisson statistics (with time) or
+    // time-independent probability (e.g., recrystallization fraction)
+    if (!_poisson_stats)
+    {
+      if (getRandomReal() < rate)
+        addNucleus(qp);
+    }
+    else
+    {
     // We check the random number against the inverse of the zero probability.
     // for performance reasons we do a quick check against the linearized form of
     // that probability, which is always strictly larger than the actual probability.
     // The expression below should short circuit and the expensive exponential
     // should rarely get evaluated
-    if (random < rate * _fe_problem.dt() && random < (1.0 - std::exp(-rate * _fe_problem.dt())))
-    {
-      _local_nucleus_list.push_back(NucleusLocation(_fe_problem.time() + _hold_time, _q_point[qp]));
[…]
