### 2020-03-10 a38b870
Discrete nucleation variable size insertion #14544
added variable-radius discrete nucleation and non-poisson statistic nucleation rate (fraction)

commented comments

removed commented code from previous version

updated documentation

removed commented inputs from tests and examples

added error message for using radius in DiscreteNucleationMap

style changes

committed file needed for a test

added correct issue number to new test

addressed Daniel's PR comments

style changes

cleanup
