+++ modules/phase_field/doc/content/source/userobjects/DiscreteNucleationMap.md
@@ -3,12 +3,11 @@
 !syntax description /UserObjects/DiscreteNucleationMap
 
 The map objects builds a smooth map of nucleation sites in the simulation
-domain, expanding the point description from the
+domain, expanding the (time, x, y, z, radius) description from the
 [`DiscreteNucleationInserter`](/DiscreteNucleationInserter.md) to a description
 in which nucleation sites have a finite size and a smooth interface. This object
-takes three parameters
+takes two parameters,
 
-- ```radius``` - the radius of the area in which each nucleation site modifies the local free energy. This should be chosen sufficiently large to generate a stable precipitate, but small enough to minimize the influence on the phase field evolution
 - ```periodic``` - a variable to take the periodicity information of the simulation domain
 - ```inserter``` - the ```DiscreteNucleationInserter``` object that maintains the nucleus list
 
