+++ framework/src/constraints/MortarConstraintBase.C
@@ -18,12 +18,9 @@ InputParameters
 MortarConstraintBase::validParams()
 {
   InputParameters params = Constraint::validParams();
-  params.addRequiredParam<BoundaryName>("master_boundary",
-                                        "The name of the master boundary sideset.");
-  params.addRequiredParam<BoundaryName>("slave_boundary",
-                                        "The name of the slave boundary sideset.");
-  params.addRequiredParam<SubdomainName>("master_subdomain", "The name of the master subdomain.");
-  params.addRequiredParam<SubdomainName>("slave_subdomain", "The name of the slave subdomain.");
+  params += MortarInterface::validParams();
+  params += TwoMaterialPropertyInterface::validParams();
+
   params.addRelationshipManager(
       "AugmentSparsityOnInterface",
       Moose::RelationshipManagerType::GEOMETRIC | Moose::RelationshipManagerType::ALGEBRAIC,
@@ -52,27 +49,20 @@ MortarConstraintBase::validParams()
       "compute_primal_residuals", true, "Whether to compute residuals for the primal variable.");
   params.addParam<bool>(
       "compute_lm_residuals", true, "Whether to compute Lagrange Multiplier residuals");
-  params.addParam<bool>(
-      "periodic",
-      false,
-      "Whether this constraint is going to be used to enforce a periodic condition. This has the "
-      "effect of changing the normals vector for projection from outward to inward facing");
   return params;
 }
 
 MortarConstraintBase::MortarConstraintBase(const InputParameters & parameters)
   : Constraint(parameters),
-    CoupleableMooseVariableDependencyIntermediateInterface(this, true),
+    NeighborCoupleableMooseVariableDependencyIntermediateInterface(this, false, false),
+    MortarInterface(this),
+    TwoMaterialPropertyInterface(this, Moose::EMPTY_BLOCK_IDS, getBoundaryIDs()),
     MooseVariableInterface<Real>(this,
                                  true,
                                  "variable",
                                  Moose::VarKindType::VAR_NONLINEAR,
                                  Moose::VarFieldType::VAR_FIELD_STANDARD),
     _fe_problem(*getCheckedPointerParam<FEProblemBase *>("_fe_problem_base")),
-    _slave_id(_mesh.getBoundaryID(getParam<BoundaryName>("slave_boundary"))),
-    _master_id(_mesh.getBoundaryID(getParam<BoundaryName>("master_boundary"))),
-    _slave_subdomain_id(_mesh.getSubdomainID(getParam<SubdomainName>("slave_subdomain"))),
-    _master_subdomain_id(_mesh.getSubdomainID(getParam<SubdomainName>("master_subdomain"))),
     _var(isParamValid("variable")
              ? &_subproblem.getStandardVariable(_tid, parameters.getMooseType("variable"))
              : nullptr),
@@ -101,13 +91,6 @@ MortarConstraintBase::MortarConstraintBase(const InputParameters & parameters)
     _phys_points_slave(_assembly.qPointsFace()),
     _phys_points_master(_assembly.qPointsFaceNeighbor())
 {
-  if (_mesh.dimension() == 3)
-    mooseError("Mortar cannot currently be run in three dimensions. It's on the to-do list!");
-
-  _fe_problem.createMortarInterface(std::make_pair(_master_id, _slave_id),
-                                    std::make_pair(_master_subdomain_id, _slave_subdomain_id),
-                                    getParam<bool>("use_displaced_mesh"),
-                                    getParam<bool>("periodic"));
 }
 
 void
