### 2020-03-11 9213d4d
Build map from lower-d to higher-d subdomains for mortar
