### 2020-03-11 3c0ef0c
Allow material reinit and swap from all SubProblem derived classes
