### 2020-03-11 9f92bfb
Allow material reinit and swap from all SubProblem derived classes
