+++ framework/src/problems/DisplacedProblem.C
@@ -987,3 +987,55 @@ DisplacedProblem::couplingMatrix() const
 {
   return _mproblem.couplingMatrix();
 }
+
+void
+DisplacedProblem::swapBackMaterials(THREAD_ID tid)
+{
+  _mproblem.swapBackMaterials(tid);
+}
+
+void
+DisplacedProblem::swapBackMaterialsFace(THREAD_ID tid)
+{
+  auto && elem = _assembly[tid]->elem();
+  unsigned int side = _assembly[tid]->side();
+
+  _mproblem.swapBackMaterialsFace(_ref_mesh.elemPtr(elem->id()), side, tid);
+}
+
+void
+DisplacedProblem::swapBackMaterialsNeighbor(THREAD_ID tid)
+{
+  // NOTE: this will not work with h-adaptivity
+  auto && neighbor = _assembly[tid]->neighbor();
+  unsigned int neighbor_side = neighbor->which_neighbor_am_i(_assembly[tid]->elem());
+
+  _mproblem.swapBackMaterialsNeighbor(_ref_mesh.elemPtr(neighbor->id()), neighbor_side, tid);
+}
+
+void
+DisplacedProblem::reinitMaterialsFace(SubdomainID blk_id, THREAD_ID tid, bool swap_stateful)
+{
+  if (hasActiveMaterialProperties(tid))
+    _mproblem.reinitMaterialsFace(_ref_mesh.elemPtr(_assembly[tid]->elem()->id()),
+                                  _assembly[tid]->side(),
+                                  _assembly[tid]->qRuleFace()->n_points(),
+                                  blk_id,
+                                  tid,
+                                  swap_stateful);
+}
+
+void
+DisplacedProblem::reinitMaterialsNeighbor(SubdomainID blk_id, THREAD_ID tid, bool swap_stateful)
+{
+  if (hasActiveMaterialProperties(tid))
+  {
+    // NOTE: this will not work with h-adaptivity
+    auto && neighbor = _assembly[tid]->neighbor();
+    unsigned int neighbor_side = neighbor->which_neighbor_am_i(_assembly[tid]->elem());
+    unsigned int n_points = _assembly[tid]->qRuleFace()->n_points();
+
+    _mproblem.reinitMaterialsNeighbor(
+        _ref_mesh.elemPtr(neighbor->id()), neighbor_side, n_points, blk_id, tid, swap_stateful);
+  }
+}
