### 2020-03-11 74f6fbd
Reinit materials for mortar
Leverages new abstract virtuals in SubProblem for reiniting
and swapping (via sentinels) materials.

Closes #14821
