+++ test/include/constraints/GapHeatConductanceTest.h
@@ -22,7 +22,8 @@ public:
 protected:
   ADReal computeQpResidual(Moose::MortarType mortar_type) final;
 
-  const Real _gap_conductance_constant;
+  const ADMaterialProperty(Real) & _slave_gap_conductance;
+  const ADMaterialProperty(Real) & _master_gap_conductance;
 
   usingMortarConstraintMembers;
 };
