+++ modules/stochastic_tools/src/vectorpostprocessors/PolyChaosLocalSensitivity.C
@@ -1,0 +1,170 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+// Stocastic Tools Includes
+#include "PolyChaosLocalSensitivity.h"
+
+#include "Sampler.h"
+
+registerMooseObject("StochasticToolsApp", PolyChaosLocalSensitivity);
+
+defineLegacyParams(PolyChaosLocalSensitivity);
+
+InputParameters
+PolyChaosLocalSensitivity::validParams()
+{
+  InputParameters params = GeneralVectorPostprocessor::validParams();
+  params.addClassDescription(
+      "Tool for calculating local sensitivity with polynomial chaos expansion.");
+  params += SamplerInterface::validParams();
+
+  params.addRequiredParam<UserObjectName>("pc_name", "Name of PolynomialChaos.");
+
+  params.addParam<std::vector<unsigned int>>(
+      "sensitivity_parameters",
+      std::vector<unsigned int>(0),
+      "Parameters to take derivative with respect to, matchs columns in training sampler. "
+      "Default is to find local sensitivity of all parameters.");
+  params.addParam<std::vector<Real>>(
+      "local_points",
+      std::vector<Real>(0),
+      "Points specifying desired location of sensitivity measurement.");
+  params.addParam<SamplerName>("local_points_sampler",
+                               "Sampler specifying desired location of sensitivity measurement.");
+  params.addParam<bool>("output_points", false, "True to include local points in output.");
+  return params;
+}
+
+PolyChaosLocalSensitivity::PolyChaosLocalSensitivity(const InputParameters & parameters)
+  : GeneralVectorPostprocessor(parameters),
+    SamplerInterface(this),
+    _pc_uo(getUserObject<PolynomialChaos>("pc_name")),
+    _sampler(isParamValid("local_points_sampler") ? &getSampler("local_points_sampler") : nullptr),
+    _points(getParam<std::vector<Real>>("local_points")),
+    _sdim(getParam<std::vector<unsigned int>>("sensitivity_parameters")),
+    _output_points(getParam<bool>("output_points")),
+    _initialized(false)
+{
+  if (!_sampler && _points.size() == 0)
+    paramError(
+        "local_points",
+        "Must specify either local_points or local_points_sampler to obtain local sensitivities.");
+}
+
+void
+PolyChaosLocalSensitivity::initialize()
+{
+  // This needs to be put in initialize since the user object doesn't know the
+  // number of parameters until after initialSetup.
+  if (!_initialized)
+  {
+    if (_sdim.size() == 0)
+    {
+      _sdim.resize(_pc_uo.getNumberOfParameters());
+      for (unsigned int d = 0; d < _pc_uo.getNumberOfParameters(); ++d)
+        _sdim[d] = d;
+    }
+
+    if (_points.size() % _pc_uo.getNumberOfParameters() != 0)
+      mooseError("Number of values in local_points is not divisible by number of parameters in "
+                 "Polynomial Chaos model.");
+    else if (_sampler)
+    {
+      if (_sampler->getNumberOfCols() != _pc_uo.getNumberOfParameters())
[…]
