### 2020-03-11 6bb3146
Allow for number literals outside of exponents (#14827)
