+++ modules/geochemistry/doc/content/modules/geochemistry/geochemistry_nomenclature.md
@@ -6,15 +6,31 @@
 
 Species are real things that exist in solution, for instance HCO$_{3}^{-}$, CO$_{2}$(aq), CaHCO$_{3}^{+}$.  They never have negative mass.
 
+### Phase
+
+A phase is a region of space that is physically distinct, mechanically separable, and homogeneous in its composition and properties.  In geochemical systems, the phases are:
+
+- a fluid phase composed of water and its dissolved constituents;
+- potentially, a separate phase corresponding to each of the minerals in contact with the fluid phase;
+- potentially, separate gas phases.  If different gases (CO$_{2}$, O$_{2}$, etc) are buffered separately with known fugacity, as is assumed in `geochemistry`, each gas counts as a distinct phase.
+
 ### Components
 
 Components are useful abstractions.  They might or might not be real things, so can even have negative mass.  For instance, consider an alkaline solution (more OH$^{-}$ than H$^{+}$).  Choose H$_{2}$O and H$^{+}$ to be the component basis set.  Each hydroxyl ion, $\mathrm{H}_{2}\mathrm{O} - \mathrm{H}^{+} \rightarrow \mathrm{OH}^{-}$, is made up of a water molecule less a hydrogen ion.  Since this is an alkaline solution, the overall composition has a positive amount of water and a negative amount of H$^{+}$.  The molality of the the H$^{+}$ speices is positive, however.
 
-A complete set of components forms a basis for the reaction system.
+A complete set of components forms a *basis* for the reaction system.
 
 ### Basis
 
-A linearly-independent complete set of components: every other chemical in the system is composed of these components, and none of these components can be composed from the others.
+The set of components used in a geochemical model is the calculation's basis.  There is no unique basis that describes a system, but the basis must satisfy these rules:
+
+1. each species and phase considered in the model must be a combination of the components in the basis;
+2. the number of components is minimal to satisfy (1)
+3. the components must be linearly independent of one another, that is, no reactions exist to form one component from the others.
+
+That is, the basis is a linearly-independent complete set of components: every other chemical in the system is composed of these components, and none of these components can be composed from the others.
+
+The number of components in the basis is $1 + N_{i} + N_{k} + N_{m} + N_{p}$ (the "1" comes from the water component), while the number of phases is $1 + N_{k} + N_{m} + N_{p}$.  The Gibbs phase rule, with fixed temperature and pressure, states that the number of degrees of freedom is the difference of these, or just $N_{i}$.  Hence, given a constraint on the concentration or activity of each basis species, the system's equilibrium state can be determined.  However, calculating the extent of the system --- the amounts of fluid, minerals, gases and sorption sites --- that are present, requires the extra $1 + N_{k} + N_{m} + N_{p}$ pieces of information.
 
 ### Moles
 
@@ -54,7 +70,7 @@ Defined in terms of the chemical potential,
 where
 
 - $\mu$ \[J.mol$^{-1}$\] is the chemical potential
-- $\mu^{0}$ \[J.mol$^{-1}$\] is the constant number, independent of temperature, pressure, other constituents, etc, but which depends on the reaction in question
[…]
