### 2020-03-11 f1e4b98
Adding the frictional and glued beam constraint source code and a set of regression tests for frictionless, glued and frictional contact in a simple beam example. Issue #14873.
