+++ modules/tensor_mechanics/src/constraints/NodalFrictionalConstraint.C
@@ -9,7 +9,10 @@
 
 // MOOSE includes
 #include "NodalFrictionalConstraint.h"
+#include "NodalConstraintUtils.h"
 #include "MooseMesh.h"
+#include "Assembly.h"
+#include "SystemBase.h"
 
 #include "libmesh/mesh_inserter_iterator.h"
 #include "libmesh/parallel.h"
@@ -28,16 +31,20 @@ validParams<NodalFrictionalConstraint>()
   InputParameters params = validParams<NodalConstraint>();
   params.addRequiredParam<BoundaryName>("boundary", "The master boundary");
   params.addRequiredParam<BoundaryName>("slave", "The slave boundary");
-  params.addRequiredParam<Real>("friction_coefficient", "The friction coefficient");
-  params.addRequiredParam<Real>("normal_force", "The force in the normal direction.");
-  params.addRequiredParam<Real>("tangential_penalty", "Stiffness of the spring in the tangential direction.");
+  params.addRequiredParam<Real>("friction_coefficient",
+                                "Friction coefficient for slippage in the normal direction");
+  params.addRequiredParam<Real>("normal_force",
+                                "Normal force used together with friction_coefficient to compute "
+                                "the normal frictional capacity.");
+  params.addRequiredParam<Real>("tangential_penalty",
+                                "Stiffness of the spring in the tangential direction.");
   return params;
 }
 
 NodalFrictionalConstraint::NodalFrictionalConstraint(const InputParameters & parameters)
   : NodalConstraint(parameters),
-    _master_node_set_id(getParam<BoundaryName>("boundary")),
-    _slave_node_set_id(getParam<BoundaryName>("slave")),
+    _master_boundary_id(getParam<BoundaryName>("boundary")),
+    _slave_boundary_id(getParam<BoundaryName>("slave")),
     _normal_force(getParam<Real>("normal_force")),
     _tangential_penalty(getParam<Real>("tangential_penalty")),
     _friction_coefficient(getParam<Real>("friction_coefficient")),
@@ -50,7 +57,8 @@ NodalFrictionalConstraint::NodalFrictionalConstraint(const InputParameters & par
   if (temp_formulation == "penalty")
     _formulation = Moose::Penalty;
   else if (temp_formulation == "kinematic")
-    mooseError("NodalFrictionalConstraint: Kinematic formulation is currently not supported for this constraint.");
+    mooseError("NodalFrictionalConstraint: Kinematic formulation is currently not supported for "
+               "this constraint.");
   else
     mooseError("Formulation must be set to Penalty.");
 }
@@ -68,21 +76,21 @@ NodalFrictionalConstraint::updateConstrainedNodes()
   _connected_nodes.clear();
   _master_conn.clear();
 
-  std::vector<dof_id_type> slave_nodelist = _mesh.getNodeList(_mesh.getBoundaryID(_slave_node_set_id));
-  std::vector<dof_id_type> master_nodelist = _mesh.getNodeList(_mesh.getBoundaryID(_master_node_set_id));
-
-  std::vector<dof_id_type>::iterator in;
+  std::vector<dof_id_type> slave_nodelist =
+      _mesh.getNodeList(_mesh.getBoundaryID(_slave_boundary_id));
+  std::vector<dof_id_type> master_nodelist =
+      _mesh.getNodeList(_mesh.getBoundaryID(_master_boundary_id));
 
   // Fill in _connected_nodes, which defines slave nodes in the base class
-  for (in = slave_nodelist.begin(); in != slave_nodelist.end(); ++in)
+  for (auto in : slave_nodelist)
   {
[…]
