+++ modules/tensor_mechanics/include/constraints/NodalStickConstraint.h
@@ -7,8 +7,7 @@
 //* Licensed under LGPL 2.1, please see LICENSE for details
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
-#ifndef NODALSTICKCONSTRAINT_H
-#define NODALSTICKCONSTRAINT_H
+#pragma once
 
 #include "NodalConstraint.h"
 
@@ -22,19 +21,10 @@ class NodalStickConstraint : public NodalConstraint
 public:
   NodalStickConstraint(const InputParameters & parameters);
 
-  /**
-   * Called on this object when the mesh changes
-   */
   virtual void meshChanged() override;
 
-  /**
-   * Computes the nodal residual.
-   */
   virtual void computeResidual(NumericVector<Number> & residual) override;
 
-  /**
-   * Computes the jacobian for the current element.
-   */
   virtual void computeJacobian(SparseMatrix<Number> & jacobian) override;
 
 protected:
@@ -43,27 +33,19 @@ protected:
    */
   void updateConstrainedNodes();
 
-  /**
-   * Computes the residual for the current slave node
-   */
   virtual Real computeQpResidual(Moose::ConstraintType type) override;
 
-  /**
-   * Computes the jacobian for the constraint
-   */
   virtual Real computeQpJacobian(Moose::ConstraintJacobianType type) override;
 
   /// Holds the slave node set or side set
-  BoundaryName _master_node_set_id;
+  BoundaryName _master_boundary_id;
 
   /// Holds the slave node set or side set
-  BoundaryName _slave_node_set_id;
+  BoundaryName _slave_boundary_id;
 
-  /// Tangential stiffness of spring
+  /// Tangential stiffness of spring in all directions
   const Real & _penalty;
 
   /// master node id connected to each slave node in _connected_nodes
   std::vector<dof_id_type> _master_conn;
 };
-
-#endif /* NodalStickConstraint_H */
