### 2020-03-11 49566a7
DistributedRectilinearMeshGenerator: Declare the specializations in the header
Closes #11485
