+++ framework/src/meshgenerators/DistributedRectilinearMeshGenerator.C
@@ -18,13 +18,8 @@
 #include "libmesh/periodic_boundaries.h"
 #include "libmesh/periodic_boundary_base.h"
 #include "libmesh/unstructured_mesh.h"
-#include "libmesh/edge_edge2.h"
-#include "libmesh/edge_edge3.h"
-#include "libmesh/edge_edge4.h"
 #include "libmesh/mesh_communication.h"
 #include "libmesh/remote_elem.h"
-#include "libmesh/face_quad4.h"
-#include "libmesh/cell_hex8.h"
 #include "libmesh/partitioner.h"
 
 // TIMPI includes
@@ -114,13 +109,13 @@ DistributedRectilinearMeshGenerator::DistributedRectilinearMeshGenerator(
 }
 
 template <>
-inline dof_id_type
-DistributedRectilinearMeshGenerator::num_neighbors<Edge2>(const dof_id_type nx,
-                                                          const dof_id_type /*ny*/,
-                                                          const dof_id_type /*nz*/,
-                                                          const dof_id_type i,
-                                                          const dof_id_type /*j*/,
-                                                          const dof_id_type /*k*/)
+dof_id_type
+DistributedRectilinearMeshGenerator::numNeighbors<Edge2>(const dof_id_type nx,
+                                                         const dof_id_type /*ny*/,
+                                                         const dof_id_type /*nz*/,
+                                                         const dof_id_type i,
+                                                         const dof_id_type /*j*/,
+                                                         const dof_id_type /*k*/)
 {
   // The ends only have one neighbor
   if (i == 0 || i == nx - 1)
@@ -130,15 +125,15 @@ DistributedRectilinearMeshGenerator::num_neighbors<Edge2>(const dof_id_type nx,
 }
 
 template <>
-inline void
-DistributedRectilinearMeshGenerator::get_neighbors<Edge2>(const dof_id_type nx,
-                                                          const dof_id_type /*ny*/,
-                                                          const dof_id_type /*nz*/,
-                                                          const dof_id_type i,
-                                                          const dof_id_type /*j*/,
-                                                          const dof_id_type /*k*/,
-                                                          std::vector<dof_id_type> & neighbors,
-                                                          const bool /*corner*/)
+void
+DistributedRectilinearMeshGenerator::getNeighbors<Edge2>(const dof_id_type nx,
+                                                         const dof_id_type /*ny*/,
+                                                         const dof_id_type /*nz*/,
+                                                         const dof_id_type i,
+                                                         const dof_id_type /*j*/,
+                                                         const dof_id_type /*k*/,
+                                                         std::vector<dof_id_type> & neighbors,
+                                                         const bool /*corner*/)
 
 {
   neighbors[0] = i - 1;
@@ -154,24 +149,24 @@ DistributedRectilinearMeshGenerator::get_neighbors<Edge2>(const dof_id_type nx,
 }
 
 template <>
-inline void
-DistributedRectilinearMeshGenerator::get_indices<Edge2>(const dof_id_type /*nx*/,
-                                                        const dof_id_type /*ny*/,
-                                                        const dof_id_type elem_id,
-                                                        dof_id_type & i,
-                                                        dof_id_type & /*j*/,
-                                                        dof_id_type & /*k*/)
+void
+DistributedRectilinearMeshGenerator::getIndices<Edge2>(const dof_id_type /*nx*/,
+                                                       const dof_id_type /*ny*/,
+                                                       const dof_id_type elem_id,
+                                                       dof_id_type & i,
+                                                       dof_id_type & /*j*/,
+                                                       dof_id_type & /*k*/)
 {
[…]
