+++ framework/include/meshgenerators/DistributedRectilinearMeshGenerator.h
@@ -11,8 +11,17 @@
 
 #include "MeshGenerator.h"
 
+#include "libmesh/face_quad4.h"
+#include "libmesh/cell_hex8.h"
+#include "libmesh/edge_edge2.h"
+
 /**
- * Mesh generated from parameters
+ * This class works by first creating a "distributed ual graph" of the element connectivity based
+ * on a linear partition of mesh before ever building an elements. It then uses
+ * PetscExternalPartitioner to partition that graph - assigning elements to processors. Then, each
+ * processor can read the partition map and build only the elements that need to be on that
+ * processor. Final steps include adding in "ghosted" elements and making sure that boundary
+ * conditions are right.
  */
 class DistributedRectilinearMeshGenerator : public MeshGenerator
 {
@@ -23,19 +32,34 @@ public:
 
   std::unique_ptr<MeshBase> generate() override;
 
+  /**
+   * Build a cube mesh
+   *
+   * @param mesh Distributed UnstructuredMesh
+   * @param nx The number of elements in the x direction
+   * @param ny The number of elements in the y direction
+   * @param nz The number of elements in the z direction
+   * @param xmin Lower X Coordinate of the generated mesh
+   * @param xmax Upper X Coordinate of the generated mesh
+   * @param ymin Lower Y Coordinate of the generated mesh
+   * @param ymax Upper Y Coordinate of the generated mesh
+   * @param zmin Lower Z Coordinate of the generated mesh
+   * @param zmax Upper Z Coordinate of the generated mesh
+   * @param type The element type
+   */
   template <typename T>
-  void buildCube(UnstructuredMesh & mesh,
-                  const unsigned int nx,
-                  unsigned int ny,
-                  unsigned int nz,
-                  const Real xmin,
-                  const Real xmax,
-                  const Real ymin,
-                  const Real ymax,
-                  const Real zmin,
-                  const Real zmax,
-                  const ElemType type,
-                  bool verbose);
+  void buildCube(UnstructuredMesh & /*mesh*/,
+                 const unsigned int /*nx*/,
+                 unsigned int /*ny*/,
+                 unsigned int /*nz*/,
+                 const Real /*xmin*/,
+                 const Real /*xmax*/,
+                 const Real /*ymin*/,
+                 const Real /*ymax*/,
+                 const Real /*zmin*/,
+                 const Real /*zmax*/,
+                 const ElemType /*type*/,
+                 bool /*verbose*/);
 
   /**
    * Get the element ID for a given hex
@@ -47,17 +71,37 @@ public:
    * @param k The z index of this element
    * @return The ID of the i,j element
    */
-
   template <typename T>
-  inline dof_id_type elemId(const dof_id_type /*nx*/,
-                             const dof_id_type /*ny*/,
-                             const dof_id_type /*i*/,
-                             const dof_id_type /*j*/,
-                             const dof_id_type /*k*/)
+  dof_id_type elemId(const dof_id_type /*nx*/,
+                     const dof_id_type /*ny*/,
+                     const dof_id_type /*i*/,
+                     const dof_id_type /*j*/,
+                     const dof_id_type /*k*/)
   {
     mooseError(
         "elemId not implemented for this element type in DistributedRectilinearMeshGenerator");
   }
+
+  template <>
+  dof_id_type elemId<Edge2>(const dof_id_type nx,
+                            const dof_id_type ny,
+                            const dof_id_type i,
+                            const dof_id_type j,
+                            const dof_id_type k);
+
+  template <>
+  dof_id_type elemId<Quad4>(const dof_id_type nx,
+                            const dof_id_type ny,
+                            const dof_id_type i,
+                            const dof_id_type j,
+                            const dof_id_type k);
+
[…]
