+++ framework/src/meshgenerators/DistributedRectilinearMeshGenerator.C
@@ -37,8 +37,6 @@ DistributedRectilinearMeshGenerator::validParams()
   InputParameters params = PetscExternalPartitioner::validParams();
   params += MeshGenerator::validParams();
 
-  params.addParam<bool>("verbose", false, "Turn on verbose printing for the mesh generation");
-
   MooseEnum dims("1=1 2 3");
   params.addRequiredParam<MooseEnum>(
       "dim", dims, "The dimension of the mesh to be generated"); // Make this parameter required
@@ -89,7 +87,6 @@ DistributedRectilinearMeshGenerator::validParams()
 DistributedRectilinearMeshGenerator::DistributedRectilinearMeshGenerator(
     const InputParameters & parameters)
   : MeshGenerator(parameters),
-    _verbose(getParam<bool>("verbose")),
     _dim(getParam<MooseEnum>("dim")),
     _nx(declareMeshProperty("num_elements_x", getParam<dof_id_type>("nx"))),
     _ny(declareMeshProperty("num_elements_y", getParam<dof_id_type>("ny"))),
@@ -213,13 +210,9 @@ DistributedRectilinearMeshGenerator::addElement<Edge2>(const dof_id_type nx,
                                                        const dof_id_type elem_id,
                                                        const processor_id_type pid,
                                                        const ElemType /*type*/,
-                                                       MeshBase & mesh,
-                                                       bool verbose)
+                                                       MeshBase & mesh)
 {
   BoundaryInfo & boundary_info = mesh.get_boundary_info();
-
-  if (verbose)
-    Moose::out << "Adding elem: " << elem_id << " pid: " << pid << std::endl;
 
   auto node_offset = elem_id;
 
@@ -231,9 +224,6 @@ DistributedRectilinearMeshGenerator::addElement<Edge2>(const dof_id_type nx,
       mesh.add_point(Point(static_cast<Real>(node_offset + 1) / nx, 0, 0), node_offset + 1);
   node1_ptr->set_unique_id() = node_offset + 1;
   node1_ptr->processor_id() = pid;
-
-  if (verbose)
-    Moose::out << "Adding elem: " << elem_id << std::endl;
 
   Elem * elem = new Edge2;
   elem->set_id(elem_id);
@@ -416,13 +406,9 @@ DistributedRectilinearMeshGenerator::addElement<Quad4>(const dof_id_type nx,
                                                        const dof_id_type elem_id,
                                                        const processor_id_type pid,
                                                        const ElemType type,
-                                                       MeshBase & mesh,
-                                                       bool verbose)
+                                                       MeshBase & mesh)
 {
   BoundaryInfo & boundary_info = mesh.get_boundary_info();
-
-  if (verbose)
-    Moose::out << "Adding elem: " << elem_id << " pid: " << pid << std::endl;
 
   // Bottom Left
   auto node0_ptr = mesh.add_point(Point(static_cast<Real>(i) / nx, static_cast<Real>(j) / ny, 0),
@@ -644,16 +630,9 @@ DistributedRectilinearMeshGenerator::addElement<Hex8>(const dof_id_type nx,
                                                       const dof_id_type elem_id,
                                                       const processor_id_type pid,
                                                       const ElemType type,
[…]
