### 2020-03-11 84a3a61
Re-factor mesh meta data as generic 'extra' data that apps could create
(refs #14875)
