### 2020-03-11 d8acca5
Re-factor mesh meta data as generic 'extra' data that apps could create
(refs #14875)
