### 2020-03-11 06413e7
Re-factor mesh meta data as generic 'extra' data that apps could create
(refs #14785)
