+++ framework/src/interfaces/MeshMetaDataInterface.C
@@ -23,5 +23,6 @@ RestartableDataValue &
 MeshMetaDataInterface::registerMetaDataOnApp(const std::string & name,
                                              std::unique_ptr<RestartableDataValue> data)
 {
-  return _meta_data_app.registerRestartableData(name, std::move(data), 0, true, true);
+  return _meta_data_app.registerRestartableData(
+      name, std::move(data), 0, true, MooseApp::MESH_META_DATA);
 }
