### 2020-03-11 73a05fb
Re-factor mesh meta data as generic 'extra' data that apps could create
(refs #14875)
