### 2020-03-11 08cd885
Moved explicit specialization out of class
