+++ framework/include/meshgenerators/DistributedRectilinearMeshGenerator.h
@@ -81,27 +81,6 @@ public:
         "elemId not implemented for this element type in DistributedRectilinearMeshGenerator");
   }
 
-  template <>
-  dof_id_type elemId<Edge2>(const dof_id_type nx,
-                            const dof_id_type ny,
-                            const dof_id_type i,
-                            const dof_id_type j,
-                            const dof_id_type k);
-
-  template <>
-  dof_id_type elemId<Quad4>(const dof_id_type nx,
-                            const dof_id_type ny,
-                            const dof_id_type i,
-                            const dof_id_type j,
-                            const dof_id_type k);
-
-  template <>
-  dof_id_type elemId<Hex8>(const dof_id_type nx,
-                           const dof_id_type ny,
-                           const dof_id_type i,
-                           const dof_id_type j,
-                           const dof_id_type k);
-
   /**
    * Get the number of neighbors this element will have
    *
@@ -124,30 +103,6 @@ public:
     mooseError("numNeighbors not implemented for this element type in "
                "DistributedRectilinearMeshGenerator");
   }
-
-  template <>
-  dof_id_type numNeighbors<Edge2>(const dof_id_type nx,
-                                  const dof_id_type ny,
-                                  const dof_id_type nz,
-                                  const dof_id_type i,
-                                  const dof_id_type j,
-                                  const dof_id_type k);
-
-  template <>
-  dof_id_type numNeighbors<Quad4>(const dof_id_type nx,
-                                  const dof_id_type ny,
-                                  const dof_id_type nz,
-                                  const dof_id_type i,
-                                  const dof_id_type j,
-                                  const dof_id_type k);
-
-  template <>
-  dof_id_type numNeighbors<Hex8>(const dof_id_type nx,
-                                 const dof_id_type ny,
-                                 const dof_id_type nz,
-                                 const dof_id_type i,
-                                 const dof_id_type j,
-                                 const dof_id_type k);
 
   /**
    * Get the IDs of the neighbors of a given element
@@ -175,36 +130,6 @@ public:
                "DistributedRectilinearMeshGenerator");
   }
 
-  template <>
-  void getNeighbors<Edge2>(const dof_id_type nx,
-                           const dof_id_type ny,
-                           const dof_id_type nz,
-                           const dof_id_type i,
-                           const dof_id_type j,
-                           const dof_id_type k,
-                           std::vector<dof_id_type> & neighbors,
-                           const bool corner);
-
-  template <>
-  void getNeighbors<Quad4>(const dof_id_type nx,
-                           const dof_id_type ny,
-                           const dof_id_type nz,
-                           const dof_id_type i,
-                           const dof_id_type j,
-                           const dof_id_type k,
-                           std::vector<dof_id_type> & neighbors,
-                           const bool corner);
-
-  template <>
-  void getNeighbors<Hex8>(const dof_id_type nx,
-                          const dof_id_type ny,
-                          const dof_id_type nz,
-                          const dof_id_type i,
-                          const dof_id_type j,
-                          const dof_id_type k,
-                          std::vector<dof_id_type> & neighbors,
-                          const bool corner);
-
   /**
    * The ID of the i,j,k node
    *
@@ -227,30 +152,6 @@ public:
     mooseError(
         "nodeId not implemented for this element type in DistributedRectilinearMeshGenerator");
   }
-
-  template <>
-  dof_id_type nodeId<Edge2>(const ElemType type,
-                            const dof_id_type nx,
-                            const dof_id_type ny,
-                            const dof_id_type i,
-                            const dof_id_type j,
-                            const dof_id_type k);
-
-  template <>
-  dof_id_type nodeId<Quad4>(const ElemType type,
-                            const dof_id_type nx,
[…]
