### 2020-03-11 b089194
Adding control object that can set a bool value in a component
Refs #19815
