### 2020-03-11 1ea6a7d
Update unit test for change to the TheWarehouse
(refs #14784)
