### 2020-03-11 558bcdc
Add interface kernel test for displaced
Refs #14876
