### 2020-03-11 12cb920
Be more specific about contact inputs
