+++ test/src/base/MooseTestApp.C
@@ -22,6 +22,13 @@ InputParameters
 MooseTestApp::validParams()
 {
   InputParameters params = MooseApp::validParams();
+
+  // Flag for testing MooseApp::getRestartableDataMap error message
+  params.addCommandLineParam<bool>("test_getRestartableDataMap_error",
+                                   "--test_getRestartableDataMap_error",
+                                   false,
+                                   "Call getRestartableDataMap with a bad name.");
+
   /* MooseTestApp is special because it will have its own
    * binary and we want the default to allow test objects.
    */
@@ -42,6 +49,9 @@ MooseTestApp::MooseTestApp(const InputParameters & parameters) : MooseApp(parame
 {
   MooseTestApp::registerAll(
       _factory, _action_factory, _syntax, !getParam<bool>("disallow_test_objects"));
+
+  if (getParam<bool>("test_getRestartableDataMap_error"))
+    getRestartableDataMap("slaughter");
 }
 
 MooseTestApp::~MooseTestApp() {}
