### 2020-03-11 ec83d23
Improve assert so that an empty if doesn't appear
(refs #14875)
