### 2020-03-11 b87bebe
Update validParams in chemical_reactions
(refs #14135)
