+++ modules/chemical_reactions/src/bcs/ChemicalOutFlowBC.C
@@ -11,11 +11,12 @@
 
 registerMooseObject("ChemicalReactionsApp", ChemicalOutFlowBC);
 
-template <>
+defineLegacyParams(ChemicalOutFlowBC);
+
 InputParameters
-validParams<ChemicalOutFlowBC>()
+ChemicalOutFlowBC::validParams()
 {
-  InputParameters params = validParams<IntegratedBC>();
+  InputParameters params = IntegratedBC::validParams();
   params.addClassDescription("Chemical flux boundary condition");
   return params;
 }
