+++ modules/chemical_reactions/src/auxkernels/KineticDisPreRateAux.C
@@ -11,11 +11,12 @@
 
 registerMooseObject("ChemicalReactionsApp", KineticDisPreRateAux);
 
-template <>
+defineLegacyParams(KineticDisPreRateAux);
+
 InputParameters
-validParams<KineticDisPreRateAux>()
+KineticDisPreRateAux::validParams()
 {
-  InputParameters params = validParams<AuxKernel>();
+  InputParameters params = AuxKernel::validParams();
   params.addCoupledVar("log_k", 0.0, "The equilibrium constant of the dissolution reaction");
   params.addRequiredParam<std::vector<Real>>("sto_v",
                                              "The stoichiometric coefficients of reactant species");
