+++ modules/chemical_reactions/src/kernels/PrimaryConvection.C
@@ -11,11 +11,12 @@
 
 registerMooseObject("ChemicalReactionsApp", PrimaryConvection);
 
-template <>
+defineLegacyParams(PrimaryConvection);
+
 InputParameters
-validParams<PrimaryConvection>()
+PrimaryConvection::validParams()
 {
-  InputParameters params = validParams<Kernel>();
+  InputParameters params = Kernel::validParams();
   params.addRequiredCoupledVar("p", "Pressure");
   RealVectorValue g(0, 0, 0);
   params.addParam<RealVectorValue>("gravity", g, "Gravity vector (default is (0, 0, 0))");
