### 2020-03-11 bc6ced0
Update validParams in external_petsc_solver
(refs #14135)
