### 2020-03-11 e6d7818
Update validParams in contact
(refs #14135)
