+++ modules/contact/src/constraints/TangentialMortarLMMechanicalContact.C
@@ -12,30 +12,35 @@
 
 registerADMooseObject("MooseApp", TangentialMortarLMMechanicalContact);
 
-defineADValidParams(
-    TangentialMortarLMMechanicalContact,
-    ADMortarConstraint,
-    params.addRequiredParam<NonlinearVariableName>("slave_disp_y",
-                                                   "The y displacement variable on the slave face");
-    params.addParam<NonlinearVariableName>("master_disp_y",
-                                           "The y displacement variable on the master face");
-    params.addRequiredParam<NonlinearVariableName>(
-        "contact_pressure",
-        "The normal contact pressure; oftentimes this may be a separate lagrange multiplier "
-        "variable");
-    params.addRequiredParam<Real>("friction_coefficient", "The friction coefficient");
+defineADLegacyParams(TangentialMortarLMMechanicalContact);
 
-    MooseEnum ncp_type("min fb", "fb");
-    params.addParam<MooseEnum>("ncp_function_type",
-                               ncp_type,
-                               "The type of the nonlinear complimentarity function; options are "
-                               "min or fb where fb stands for Fischer-Burmeister");
-    params.addParam<Real>("c",
-                          1,
-                          "Parameter for balancing the size of the velocity and the pressures");
-    params.set<bool>("compute_primal_residuals") = false;
-    params.addClassDescription("Ensures that the Karush-Kuhn-Tucker conditions of Coulomb "
-                               "frictional contact are satisfied"););
+template <ComputeStage compute_stage>
+InputParameters
+TangentialMortarLMMechanicalContact<compute_stage>::validParams()
+{
+  auto params = MortarConstraintBase::validParams();
+  params.addRequiredParam<NonlinearVariableName>("slave_disp_y",
+                                                 "The y displacement variable on the slave face");
+  params.addParam<NonlinearVariableName>("master_disp_y",
+                                         "The y displacement variable on the master face");
+  params.addRequiredParam<NonlinearVariableName>(
+      "contact_pressure",
+      "The normal contact pressure; oftentimes this may be a separate lagrange multiplier "
+      "variable");
+  params.addRequiredParam<Real>("friction_coefficient", "The friction coefficient");
+
+  MooseEnum ncp_type("min fb", "fb");
+  params.addParam<MooseEnum>("ncp_function_type",
+                             ncp_type,
+                             "The type of the nonlinear complimentarity function; options are "
+                             "min or fb where fb stands for Fischer-Burmeister");
+  params.addParam<Real>(
+      "c", 1, "Parameter for balancing the size of the velocity and the pressures");
+  params.set<bool>("compute_primal_residuals") = false;
+  params.addClassDescription("Ensures that the Karush-Kuhn-Tucker conditions of Coulomb "
+                             "frictional contact are satisfied");
+  return params;
+}
 
 template <ComputeStage compute_stage>
 TangentialMortarLMMechanicalContact<compute_stage>::TangentialMortarLMMechanicalContact(
