+++ modules/fluid_properties/test/src/materials/MultiComponentFluidPropertiesMaterialPT.C
@@ -11,11 +11,12 @@
 
 registerMooseObject("FluidPropertiesTestApp", MultiComponentFluidPropertiesMaterialPT);
 
-template <>
+defineLegacyParams(MultiComponentFluidPropertiesMaterialPT);
+
 InputParameters
-validParams<MultiComponentFluidPropertiesMaterialPT>()
+MultiComponentFluidPropertiesMaterialPT::validParams()
 {
-  InputParameters params = validParams<Material>();
+  InputParameters params = Material::validParams();
   params.addRequiredCoupledVar("pressure", "pressure (Pa)");
   params.addRequiredCoupledVar("temperature", "temperature (K)");
   params.addRequiredCoupledVar("xmass", "Solute mass fraction (-)");
