### 2020-03-11 1a63af3
Update validParams in level_set
(refs #14135)
