### 2020-03-11 1f9e94e
Update validParams in misc
(refs #14135)
