+++ modules/misc/include/kernels/CoefDiffusion.h
@@ -12,15 +12,11 @@
 #include "Kernel.h"
 #include "Function.h"
 
-// Forward Declarations
-class CoefDiffusion;
-
-template <>
-InputParameters validParams<CoefDiffusion>();
-
 class CoefDiffusion : public Kernel
 {
 public:
+  static InputParameters validParams();
+
   CoefDiffusion(const InputParameters & parameters);
 
 protected:
