### 2020-03-11 d312174
Update validParams for navier_stokes
(refs #14135)
