+++ modules/navier_stokes/src/postprocessors/INSExplicitTimestepSelector.C
@@ -13,11 +13,12 @@
 
 registerMooseObject("NavierStokesApp", INSExplicitTimestepSelector);
 
-template <>
+defineLegacyParams(INSExplicitTimestepSelector);
+
 InputParameters
-validParams<INSExplicitTimestepSelector>()
+INSExplicitTimestepSelector::validParams()
 {
-  InputParameters params = validParams<ElementPostprocessor>();
+  InputParameters params = ElementPostprocessor::validParams();
 
   params.addClassDescription("Postprocessor that computes the minimum value of h_min/|u|, where "
                              "|u| is coupled in as an aux variable.");
