+++ modules/navier_stokes/include/bcs/MomentumFreeSlipBC.h
@@ -22,6 +22,8 @@ InputParameters validParams<MomentumFreeSlipBC>();
 class MomentumFreeSlipBC : public NodalNormalBC
 {
 public:
+  static InputParameters validParams();
+
   MomentumFreeSlipBC(const InputParameters & parameters);
   virtual ~MomentumFreeSlipBC();
 
