### 2020-03-11 ff0a486
Update validParams for peridynamics
(refs #14135)
