+++ modules/phase_field/include/auxkernels/FeatureFloodCountAux.h
@@ -29,6 +29,8 @@ public:
    * Factory constructor, takes parameters so that all derived classes can be built using the same
    * constructor.
    */
+  static InputParameters validParams();
+
   FeatureFloodCountAux(const InputParameters & parameters);
 
   virtual ~FeatureFloodCountAux() {}
