+++ modules/porous_flow/include/materials/PorousFlowRelativePermeabilityBase.h
@@ -24,6 +24,8 @@ InputParameters validParams<PorousFlowRelativePermeabilityBase>();
 class PorousFlowRelativePermeabilityBase : public PorousFlowMaterialBase
 {
 public:
+  static InputParameters validParams();
+
   PorousFlowRelativePermeabilityBase(const InputParameters & parameters);
 
 protected:
