### 2020-03-11 352d6a7
Update validParams in rdg
(refs #14135)
