+++ modules/rdg/src/userobjects/BoundaryFluxBase.C
@@ -9,11 +9,12 @@
 
 #include "BoundaryFluxBase.h"
 
-template <>
+defineLegacyParams(BoundaryFluxBase);
+
 InputParameters
-validParams<BoundaryFluxBase>()
+BoundaryFluxBase::validParams()
 {
-  InputParameters params = validParams<GeneralUserObject>();
+  InputParameters params = GeneralUserObject::validParams();
   return params;
 }
 
