+++ modules/richards/src/userobjects/RichardsSeff2waterVG.C
@@ -14,11 +14,12 @@
 
 registerMooseObject("RichardsApp", RichardsSeff2waterVG);
 
-template <>
+defineLegacyParams(RichardsSeff2waterVG);
+
 InputParameters
-validParams<RichardsSeff2waterVG>()
+RichardsSeff2waterVG::validParams()
 {
-  InputParameters params = validParams<RichardsSeff>();
+  InputParameters params = RichardsSeff::validParams();
   params.addRequiredRangeCheckedParam<Real>("al",
                                             "al > 0",
                                             "van-Genuchten alpha parameter.  Must "
