### 2020-03-11 36ab4b9
Update validParams in solid_mechanics
(refs #14135)
