+++ modules/solid_mechanics/include/materials/IsotropicTempDepHardening.h
@@ -22,6 +22,8 @@ InputParameters validParams<IsotropicTempDepHardening>();
 class IsotropicTempDepHardening : public IsotropicPlasticity
 {
 public:
+  static InputParameters validParams();
+
   IsotropicTempDepHardening(const InputParameters & parameters);
 
 protected:
