+++ modules/solid_mechanics/include/materials/PowerLawCreepModel.h
@@ -19,6 +19,8 @@ InputParameters validParams<PowerLawCreepModel>();
 class PowerLawCreepModel : public ReturnMappingModel
 {
 public:
+  static InputParameters validParams();
+
   PowerLawCreepModel(const InputParameters & parameters);
 
 protected:
