+++ modules/solid_mechanics/src/postprocessors/HomogenizedElasticConstants.C
@@ -14,11 +14,12 @@
 
 registerMooseObject("SolidMechanicsApp", HomogenizedElasticConstants);
 
-template <>
+defineLegacyParams(HomogenizedElasticConstants);
+
 InputParameters
-validParams<HomogenizedElasticConstants>()
+HomogenizedElasticConstants::validParams()
 {
-  InputParameters params = validParams<ElementAverageValue>();
+  InputParameters params = ElementAverageValue::validParams();
   params.addRequiredCoupledVar("dx_xx", "solution in xx");
   params.addRequiredCoupledVar("dy_xx", "solution in xx");
   params.addCoupledVar("dz_xx", "solution in xx");
