+++ modules/solid_mechanics/src/materials/CLSHPlasticMaterial.C
@@ -12,12 +12,13 @@
 
 registerMooseObject("SolidMechanicsApp", CLSHPlasticMaterial);
 
-template <>
+defineLegacyParams(CLSHPlasticMaterial);
+
 InputParameters
-validParams<CLSHPlasticMaterial>()
+CLSHPlasticMaterial::validParams()
 {
-  InputParameters params = validParams<SolidModel>();
-  params += validParams<CLSHPlasticModel>();
+  InputParameters params = SolidModel::validParams();
+  params += CLSHPlasticModel::validParams();
   return params;
 }
 
