### 2020-03-11 fecf4a5
Update validParams for xfem
(refs #14135)
