+++ modules/xfem/src/base/XFEMApp.C
@@ -15,11 +15,12 @@
 #include "AppFactory.h"
 #include "MooseSyntax.h"
 
-template <>
+defineLegacyParams(XFEMApp);
+
 InputParameters
-validParams<XFEMApp>()
+XFEMApp::validParams()
 {
-  InputParameters params = validParams<MooseApp>();
+  InputParameters params = MooseApp::validParams();
 
   params.set<bool>("automatic_automatic_scaling") = false;
 
