### 2020-03-12 28d76db
Set required mat props for mortar constraints
Do this by looping over MaterialInterface::getMatPropDepedencies
