+++ modules/stochastic_tools/doc/content/source/vectorpostprocessors/PolyChaosSobolStatistics.md
@@ -1,0 +1,31 @@
+# PolyChaosLocalSensitivity
+
+!syntax description /VectorPostprocessors/SobolStatistics
+
+## Overview
+
+This object computes Sobol sensitivity indices based on a trained [PolynomialChaos.md] surrogate model. The alogrithm is based on computations described in [!cite](Sudret_2008). The sensitivities to be computed are defined by the input parameter `sensitivity_order`: `first` computes first-order sensitivities, `second` computes second-order sensitivities, `all` computes all sensitivities (replacing `first` and `second`), and `total` comutes the total sensitivities.
+
+## Example Syntax
+
+The following input snippet uses the GFunction test object, which is defined as:
+
+!equation
+g(x_1, x_2, \ldots, x_n) = \prod_{i=1}^n\frac{|4x_i-2|+q_i}{1+q_i},
+
+where $\vec{q}$ is defined in the input file, $x$ is the vector of model data, and $n$ is the number
+of input variables. The $x$ data is generated by QuadratureSampler object.
+
+!listing poly_chaos/sobol.i block=VectorPostprocessors
+
+## Example Output
+
+The resulting output contains a single column of all the sensitivities and columnes containing the indicies. The column `i_T` is the total sensitivity index, `i_1` is the first index, `i_2` second, and so on to `i_n`. For sensitivity orders less than $n$, the last index is repeated and can be ignored.
+
+!listing poly_chaos/gold/sobol_out_sobol_0002.csv
+
+!syntax parameters /VectorPostprocessors/PolyChaosSobolStatistics
+
+!syntax inputs /VectorPostprocessors/PolyChaosSobolStatistics
+
+!syntax children /VectorPostprocessors/PolyChaosSobolStatistics
