+++ modules/stochastic_tools/src/vectorpostprocessors/PolyChaosSobolStatistics.C
@@ -1,0 +1,195 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+// Stocastic Tools Includes
+#include "PolyChaosSobolStatistics.h"
+
+#include "libmesh/utility.h"
+
+registerMooseObject("StochasticToolsApp", PolyChaosSobolStatistics);
+
+defineLegacyParams(PolyChaosSobolStatistics);
+
+InputParameters
+PolyChaosSobolStatistics::validParams()
+{
+  InputParameters params = GeneralVectorPostprocessor::validParams();
+  params.addClassDescription(
+      "Compute SOBOL statistics values of a trained PolynomialChaos surrogate.");
+
+  params.addRequiredParam<UserObjectName>("pc_name", "Name of PolynomialChaos.");
+
+  MultiMooseEnum stats("first second all total");
+  params.addRequiredParam<MultiMooseEnum>(
+      "sensitivity_order",
+      stats,
+      "Order of sensitivities to compute. The all option computes all possible orders.");
+  return params;
+}
+
+PolyChaosSobolStatistics::PolyChaosSobolStatistics(const InputParameters & parameters)
+  : GeneralVectorPostprocessor(parameters),
+    _pc_uo(getUserObject<PolynomialChaos>("pc_name")),
+    _order(getParam<MultiMooseEnum>("sensitivity_order")),
+    _initialized(false),
+    _stats(declareVector("value")),
+    _total_ind(_order.contains("total") ? &declareVector("i_T") : nullptr)
+{
+}
+
+void
+PolyChaosSobolStatistics::initialize()
+{
+  // This needs to be put in initialize since the user object doesn't know the
+  // number of parameters until after initialSetup.
+  if (!_initialized)
+  {
+    _nind = 0;
+    _nval = 0;
+
+    if (_order.contains("all"))
+    {
+      _nind = _pc_uo.getNumberOfParameters();
+      _nval = 1;
+      for (unsigned int j = 0; j < _pc_uo.getNumberOfParameters(); ++j)
+        _nval *= 2;
+      _nval -= 1;
+
+      for (unsigned int i = 1; i <= _nind; ++i)
+      {
+        std::set<std::set<unsigned int>> indi =
+            buildSobolIndices(i, _pc_uo.getNumberOfParameters(), 0);
+        _ind.insert(indi.begin(), indi.end());
+      }
+    }
+    else
+    {
+      if (_order.contains("first"))
+      {
+        _nind = 1;
+        _nval = _pc_uo.getNumberOfParameters();
+        std::set<std::set<unsigned int>> indi =
+            buildSobolIndices(_nind, _pc_uo.getNumberOfParameters(), 0);
+        _ind.insert(indi.begin(), indi.end());
+      }
+      if (_order.contains("second"))
+      {
+        _nind = 2;
+        _nval += Utility::binomial(_pc_uo.getNumberOfParameters(), (unsigned int)2);
+        std::set<std::set<unsigned int>> indi =
+            buildSobolIndices(_nind, _pc_uo.getNumberOfParameters(), 0);
+        _ind.insert(indi.begin(), indi.end());
+      }
+    }
+
+    _ind_vector.reserve(_nind);
+    for (unsigned int i = 0; i < _nind; ++i)
[…]
