+++ modules/stochastic_tools/doc/content/bib/stochastic_tools.bib
@@ -36,3 +36,15 @@
   year={1979},
   publisher={Taylor \& Francis}
 }
+
+@article{Sudret_2008,
+title = "Global sensitivity analysis using polynomial chaos expansions",
+journal = "Reliability Engineering & System Safety",
+volume = "93",
+number = "7",
+pages = "964 - 979",
+year = "2008",
+note = "Bayesian Networks in Dependability",
+doi = "10.1016/j.ress.2007.04.002",
+author = "Bruno Sudret",
+}
