### 2020-03-12 55c5582
Allowing vector name as input for results of training of surrogate model #14710
