+++ modules/geochemistry/doc/content/modules/geochemistry/tests_and_examples/non_unique.md
@@ -1,0 +1,7 @@
+# Examples of models that exhibit nonunique solutions
+
+This page follows Chapter 12 of [!cite](bethke_2007).
+
+TODO
+
+!bibtex bibliography
