+++ modules/phase_field/src/kernels/ADAllenCahn.C
@@ -17,7 +17,7 @@ template <ComputeStage compute_stage>
 InputParameters
 ADAllenCahn<compute_stage>::validParams()
 {
-  InputParameters params = ADAllenCahnBase<compute_stage>::validParams();
+  InputParameters params = ADAllenCahnBase<compute_stage, Real>::validParams();
   params.addClassDescription("Allen-Cahn Kernel that uses a DerivativeMaterial Free Energy");
   params.addRequiredParam<MaterialPropertyName>(
       "f_name", "Base name of the free energy function F defined in a DerivativeParsedMaterial");
