+++ modules/phase_field/src/kernels/ADAllenCahnBase.C
@@ -8,14 +8,3 @@
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
 #include "ADAllenCahnBase.h"
-
-template <ComputeStage compute_stage, typename T>
-InputParameters
-ADAllenCahnBase<compute_stage, T>::validParams()
-{
-  InputParameters params = ADKernelValue<compute_stage>::validParams();
-  params.addClassDescription(
-      "Allen-Cahn bulk contribution Kernel with forward mode automatic differentiation");
-  params.addParam<MaterialPropertyName>("mob_name", "L", "The mobility used with the kernel");
-  return params;
-}
