+++ modules/phase_field/src/kernels/ADSplitCHWResBase.C
@@ -8,14 +8,3 @@
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
 #include "ADSplitCHWResBase.h"
-
-template <ComputeStage compute_stage, typename T>
-InputParameters
-ADSplitCHWResBase<compute_stage, T>::validParams()
-{
-  InputParameters params = ADKernelGrad<compute_stage>::validParams();
-  params.addClassDescription(
-      "Split formulation Cahn-Hilliard Kernel for the chemical potential variable");
-  params.addParam<MaterialPropertyName>("mob_name", "mobtemp", "The mobility used with the kernel");
-  return params;
-}
