### 2020-03-12 5a54409
Update validParams in functional_expansion_tools
(refs #14135)
