+++ modules/functional_expansion_tools/include/auxkernels/FunctionSeriesToAux.h
@@ -23,6 +23,7 @@ InputParameters validParams<FunctionSeriesToAux>();
 class FunctionSeriesToAux : public FunctionAux
 {
 public:
+  static InputParameters validParams();
+
   FunctionSeriesToAux(const InputParameters & parameters);
 };
-
