### 2020-03-12 1a89587
Fixup extraneous newlines at end of headers in functional_expansion_tools
