+++ modules/functional_expansion_tools/include/utils/Hashing.h
@@ -91,4 +91,3 @@ hashCombine(Real time, const libMesh::Point & point)
   return seed;
 }
 }
-
