### 2020-03-12 fbedd04
Fixing parallel bug in SumPostprocessor
Closes #19806
