### 2020-03-12 2ef0027
SumPostprocessor can handle more then 2 values
Refs #19747
