### 2020-03-12 64335fa
Revert "Merge pull request #14856 from dschwen/rapid_dev_paper_inputs"
This reverts commit d2dbdbaad1035cecc5015b2b071ead5b5f637c60, reversing
changes made to 40f90ce91bf51610248c7eb14506bafedb448f06.
