+++ framework/include/kernels/MaterialPropertyValue.h
@@ -1,27 +1,0 @@
-//* This file is part of the MOOSE framework
-//* https://www.mooseframework.org
-//*
-//* All rights reserved, see COPYRIGHT for full restrictions
-//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
-//*
-//* Licensed under LGPL 2.1, please see LICENSE for details
-//* https://www.gnu.org/licenses/lgpl-2.1.html
-
-#pragma once
-
-#include "KernelValue.h"
-
-class MaterialPropertyValue : public KernelValue
-{
-public:
-  static InputParameters validParams();
-
-  MaterialPropertyValue(const InputParameters & parameters);
-
-protected:
-  virtual Real precomputeQpResidual() override;
-  virtual Real precomputeQpJacobian() override;
-
-  const Real _kernel_sign;
-  const MaterialProperty<Real> & _prop;
-};
