### 2020-03-12 d850169
Use inefficient helper function to construct the Jacobian.
