+++ modules/tensor_mechanics/src/kernels/InertialForceShell.C
@@ -149,11 +149,9 @@ ADInertialForceShell<compute_stage>::ADInertialForceShell(const InputParameters 
   }
 }
 
-
-
-template <ComputeStage compute_stage>
-ADReal
-ADInertialForceShell<compute_stage>::computeQpResidual()
+template <>
+void
+ADInertialForceShell<RESIDUAL>::computeResidual()
 {
   if (_dt != 0.0)
   {
@@ -390,7 +388,7 @@ ADInertialForceShell<compute_stage>::computeQpResidual()
       // _local_force for each of the nodes. Try vector form first (containing three displacement
       // components)
 
-      ADReal factor_qxy = _2d_weights[qp_xy] *_ad_JxW[qp_xy] * _ad_coord[qp_xy];
+      Real factor_qxy = _2d_weights[qp_xy] * _ad_JxW[qp_xy] * _ad_coord[qp_xy];
 
       for (unsigned int dim = 0; dim < 3; dim++)
       {
@@ -520,16 +518,468 @@ ADInertialForceShell<compute_stage>::computeQpResidual()
       _global_force_1 = _original_local_config[0] * _local_force[1];
       _global_force_2 = _original_local_config[0] * _local_force[2];
       _global_force_3 = _original_local_config[0] * _local_force[3];
+      _local_re(0) = _global_force_0(_component);
+      _local_re(1) = _global_force_1(_component);
+      _local_re(2) = _global_force_2(_component);
+      _local_re(3) = _global_force_3(_component);
     }
     else
-     // Only two rotational components: \alpha and \beta.
+    // Only two rotational components: \alpha and \beta.
     {
       _global_moment_0 = _original_local_config[0] * _local_moment[0];
       _global_moment_1 = _original_local_config[0] * _local_moment[1];
       _global_moment_2 = _original_local_config[0] * _local_moment[2];
       _global_moment_3 = _original_local_config[0] * _local_moment[3];
+
+      _local_re(0) = _global_moment_0(_component - 3);
+      _local_re(1) = _global_moment_1(_component - 3);
+      _local_re(2) = _global_moment_2(_component - 3);
+      _local_re(3) = _global_moment_3(_component - 3);
     }
 
+    accumulateTaggedLocalResidual();
+
+    if (_has_save_in)
+    {
+      Threads::spin_mutex::scoped_lock lock(Threads::spin_mtx);
+      for (unsigned int i = 0; i < _save_in.size(); i++)
+        _save_in[i]->sys().solution().add_vector(_local_re, _save_in[i]->dofIndices());
+    }
+  }
+}
+
+template <ComputeStage compute_stage>
+ADReal
+ADInertialForceShell<compute_stage>::computeADResidualNode()
+{
+    ADReal returnValue = 0.0;
+
+  // Use _i to return a nodal variable of _i with component _component.
+  if (_dt != 0.0)
+  {
+    // fetch the four nodes for _current_elem
+    std::vector<const Node *> node;
+    for (unsigned int i = 0; i < 4; ++i)
+      node.push_back(_current_elem->node_ptr(i));
+
+    // Fetch the solution for the two end nodes at time t
+    NonlinearSystemBase & nonlinear_sys = _fe_problem.getNonlinearSystemBase();
+
[…]
