### 2020-03-12 49b3231
Adding auxkernel to compute the grain boundary velocity #14881
Added auxkernel that calculates grain boundary velocity #14881

Added tests for grain boundary velocity

Updated style

Adding documentation

re-upload with the changes requested. Changed name, description, and other items of style nature.

added new static member scheme for validParams.

Final Change to manage chenges in github. Fixed header error.

Fixed styling issues

Fixed Documentation and added some information

Fixed the style.

Fixing current errors: No test file, decreasing complexity of test, added documentation, and som
[…]
