### 2020-03-12 63b6bb2
Update syntax after merge
(refs #14784)
