+++ modules/stochastic_tools/doc/content/source/samplers/MonteCarloSampler.md
@@ -1,6 +1,6 @@
-# MonteCarloSampler
+# MonteCarlo
 
-MonteCarloSampler samples the provided distribution using a traditional
+The MonteCarlo object samples the provided distribution using a traditional
 [Monte Carlo](https://en.wikipedia.org/wiki/Monte_Carlo_method) sampling.
 
 ## Example Input Syntax
