+++ modules/stochastic_tools/doc/content/source/vectorpostprocessors/SobolStatistics.md
@@ -17,7 +17,7 @@ The following input snippet uses the GFunction test object, which is defined as:
 g(x_1, x_2, \ldots, x_n) = \prod_{i=1}^n\frac{|4x_i-2|+q_i}{1+q_i},
 
 where $\vec{q}$ is defined in the input file, $x$ is the vector of model data, and $n$ is the number
-of input variables. The $x$ data is generated by SobolSampler object.
+of input variables. The $x$ data is generated by the Sobol sampler object.
 
 !listing sobol_statistics/sobol.i block=Samplers VectorPostprocessors
 
