### 2020-03-13 523520a
adding descriptions of examples that exhibit nonunique solutions, and examples of various reaction paths.  Refs #14832
