### 2020-03-13 05cf67b
add descriptions of kinetic examples.  Refs #14832
