+++ modules/geochemistry/doc/content/modules/geochemistry/tests_and_examples/kinetic_quartz_arrhenius.md
@@ -1,0 +1,7 @@
+# Quartz deposition in a fracture
+
+Section 26.2 of [!cite](bethke_2007) describes quartz deposition in a hydrothermal fracture.
+
+TODO
+
+!bibtex bibliography
