+++ modules/geochemistry/doc/content/modules/geochemistry/tests_and_examples/pickup.md
@@ -1,0 +1,5 @@
+# Adding fluids with different temperatures
+
+TODO: Section 22.2 of [!cite](bethke_2007)
+
+!bibtex bibliography
