+++ modules/geochemistry/doc/content/modules/geochemistry/tests_and_examples/flush.md
@@ -1,6 +1,7 @@
 # Flushing minerals
 
 Chapter 13.3 of [!cite](bethke_2007) describes a "flush" process, in which pure water is added to the system, and at the same time an equal volume of the equilibrium solution, consisting of solvent water, primary aqueous species, disolved minerals and disolved gases, is removed.  In the code, this is achieved by altering $M_{w}$, $M_{i}$, $M_{k}$ and $M_{m}$ appropriately.
-TODO
+
+TODO: 30.2 of Bethke
 
 !bibtex bibliography
