+++ modules/geochemistry/doc/content/modules/geochemistry/tests_and_examples/flow_through.md
@@ -6,6 +6,6 @@ Chapter 13.3 of [!cite](bethke_2007) describes a "flow-through" process, whereby
 - then setting $n_{k}$ to a tiny number (but not zero, otherwise it might be swapped out of the basis);
 - setting the mole numbers of any surface components to zero, $M_{p}=0$, as well as the molalities of unoccupied surface sites, $m_{p}=0$ and adsorbed species, $m_{q}=0$.
 
-TODO
+TODO: Section 24.3 of Bethke
 
 !bibtex bibliography
