### 2020-03-13 9050825
add descriptions of flush, flow-through, pickup and arrhenius kinetic examples.  Refs #14832
