+++ modules/geochemistry/doc/content/modules/geochemistry/tests_and_examples/kinetic_quartz_arrhenius.md
@@ -1,7 +1,23 @@
 # Quartz deposition in a fracture
 
-Section 26.2 of [!cite](bethke_2007) describes quartz deposition in a hydrothermal fracture.
+Section 26.2 of [!cite](bethke_2007) describes quartz deposition in a hydrothermal fracture.  The setup is similar to [quartz dissolution](kinetic_quartz.md) but with a different form of the reaction rate.  The reaction is
+\begin{equation}
+\mathrm{Quartz} \rightarrow \mathrm{SiO}_{2}\mathrm{(aq)} \ ,
+\end{equation}
+with rate
+\begin{equation}
+r = A_{s} e^{-E/(RT)}\left( 1 - \frac{Q}{K} \right) \ .
+\end{equation}
+It is assumed that:
 
-TODO
+- there is 1$\,$kg of water;
+- the initial temperature is 300$^{\circ}$C and this steadily reduces to 25$^{\circ}$C over the course of 1 year;
+- the mineral quartz is used instead of SiO$_{2}$(aq) in the basis;
+- initially 400$\,$g of quartz is added to the water;
+- the specific surface area is $A_{s} = 2.35\times 10^{-5}\,$cm$^{2}$/g(quartz);
+- the activation energy is $E=72800\,$J.mol$^{-1}$;
+- all other silica-containing minerals are prevented from precipitating.
+
+[!cite](bethke_2007) presents results in Figures 26.3 and 26.4.
 
 !bibtex bibliography
