### 2020-03-13 7eee7ab
Added derivatives to Reynolds number material
Closures such as friction factor depend on Reynolds number,
so these changes are necessary if one wants the Jacobians
for such a closure.

Refs idaholab/sockeye#19719
