### 2020-03-14 474db7d
Revert "Allow material reinit and swap from all SubProblem derived classes"
This reverts commit 95a5cb5ed5f740630eb75444fb13cde22afb3487.
