+++ framework/src/loops/ComputeIndicatorThread.C
@@ -75,7 +75,7 @@ ComputeIndicatorThread::onElement(const Elem * elem)
 
   // Set up Sentinel class so that, even if reinitMaterials() throws, we
   // still remember to swap back during stack unwinding.
-  SwapBackSentinel sentinel(_fe_problem, &SubProblem::swapBackMaterials, _tid);
+  SwapBackSentinel sentinel(_fe_problem, &FEProblemBase::swapBackMaterials, _tid);
 
   _fe_problem.reinitMaterials(_subdomain, _tid);
 
@@ -151,10 +151,11 @@ ComputeIndicatorThread::onInternalSide(const Elem * elem, unsigned int side)
 
       // Set up Sentinels so that, even if one of the reinitMaterialsXXX() calls throws, we
       // still remember to swap back during stack unwinding.
-      SwapBackSentinel face_sentinel(_fe_problem, &SubProblem::swapBackMaterialsFace, _tid);
+      SwapBackSentinel face_sentinel(_fe_problem, &FEProblemBase::swapBackMaterialsFace, _tid);
       _fe_problem.reinitMaterialsFace(block_id, _tid);
 
-      SwapBackSentinel neighbor_sentinel(_fe_problem, &SubProblem::swapBackMaterialsNeighbor, _tid);
+      SwapBackSentinel neighbor_sentinel(
+          _fe_problem, &FEProblemBase::swapBackMaterialsNeighbor, _tid);
       _fe_problem.reinitMaterialsNeighbor(neighbor->subdomain_id(), _tid);
 
       const std::vector<std::shared_ptr<InternalSideIndicator>> & indicators =
