### 2020-03-14 678ce92
Revert "Allow material reinit and swap from all SubProblem derived classes"
This reverts commit 95a5cb5ed5f740630eb75444fb13cde22afb3487.
