+++ framework/src/problems/FEProblemBase.C
@@ -3063,73 +3063,33 @@ FEProblemBase::reinitMaterials(SubdomainID blk_id, THREAD_ID tid, bool swap_stat
 }
 
 void
-FEProblemBase::reinitMaterialsFace(const Elem * elem,
-                                   unsigned int side,
-                                   unsigned int n_points,
-                                   SubdomainID blk_id,
-                                   THREAD_ID tid,
-                                   bool swap_stateful)
-{
-  _bnd_material_data[tid]->resize(n_points);
-
-  if (swap_stateful && !_bnd_material_data[tid]->isSwapped())
-    _bnd_material_data[tid]->swap(*elem, side);
-
-  if (_discrete_materials[Moose::FACE_MATERIAL_DATA].hasActiveBlockObjects(blk_id, tid))
-    _bnd_material_data[tid]->reset(
-        _discrete_materials[Moose::FACE_MATERIAL_DATA].getActiveBlockObjects(blk_id, tid));
-
-  if (_jacobian_materials[Moose::FACE_MATERIAL_DATA].hasActiveBlockObjects(blk_id, tid) &&
-      _currently_computing_jacobian)
-    _bnd_material_data[tid]->reinit(
-        _jacobian_materials[Moose::FACE_MATERIAL_DATA].getActiveBlockObjects(blk_id, tid));
-
-  if (_residual_materials[Moose::FACE_MATERIAL_DATA].hasActiveBlockObjects(blk_id, tid) &&
-      !_currently_computing_jacobian)
-    _bnd_material_data[tid]->reinit(
-        _residual_materials[Moose::FACE_MATERIAL_DATA].getActiveBlockObjects(blk_id, tid));
-}
-
-void
 FEProblemBase::reinitMaterialsFace(SubdomainID blk_id, THREAD_ID tid, bool swap_stateful)
 {
   if (hasActiveMaterialProperties(tid))
-    reinitMaterialsFace(_assembly[tid]->elem(),
-                        _assembly[tid]->side(),
-                        _assembly[tid]->qRuleFace()->n_points(),
-                        blk_id,
-                        tid,
-                        swap_stateful);
-}
+  {
+    auto && elem = _assembly[tid]->elem();
+    unsigned int side = _assembly[tid]->side();
+    unsigned int n_points = _assembly[tid]->qRuleFace()->n_points();
 
-void
-FEProblemBase::reinitMaterialsNeighbor(const Elem * neighbor,
-                                       unsigned int neighbor_side,
-                                       unsigned int n_points,
-                                       SubdomainID blk_id,
-                                       THREAD_ID tid,
-                                       bool swap_stateful)
-{
-  // NOTE: this will not work with h-adaptivity
-  _neighbor_material_data[tid]->resize(n_points);
+    _bnd_material_data[tid]->resize(n_points);
 
-  // Only swap if requested
-  if (swap_stateful)
-    _neighbor_material_data[tid]->swap(*neighbor, neighbor_side);
+    if (swap_stateful && !_bnd_material_data[tid]->isSwapped())
+      _bnd_material_data[tid]->swap(*elem, side);
 
-  if (_discrete_materials[Moose::NEIGHBOR_MATERIAL_DATA].hasActiveBlockObjects(blk_id, tid))
-    _neighbor_material_data[tid]->reset(
-        _discrete_materials[Moose::NEIGHBOR_MATERIAL_DATA].getActiveBlockObjects(blk_id, tid));
[…]
