### 2020-03-14 60c0841
Use _fe_problem to reinit higher dimensional elements
I choose to do this for a couple of reasons:

1. If using _fe_problem then I don't have to change the API
of SwapBackSentinel.
2. This is much safer. This allows the MortarConstraints to use
material properties  coming from both displaced and undisplaced
materials. It's hard for me to imagine many scenarios where a user
would want to use an undisplaced material with a displaced constraint
or visa versa, but one example would be if using a
GenericConstantMaterial whose property  values are totally unrelated to
whether the mesh is displaced. 
[…]
