+++ framework/src/loops/ComputeMortarFunctor.C
@@ -8,6 +8,7 @@
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
 #include "ComputeMortarFunctor.h"
+#include "FEProblemBase.h"
 #include "SubProblem.h"
 #include "Assembly.h"
 #include "ADMortarConstraint.h"
@@ -26,9 +27,13 @@ template <ComputeStage compute_stage>
 ComputeMortarFunctor<compute_stage>::ComputeMortarFunctor(
     const std::vector<std::shared_ptr<MortarConstraintBase>> & mortar_constraints,
     const AutomaticMortarGeneration & amg,
-    SubProblem & subproblem)
+    SubProblem & subproblem,
+    FEProblemBase & fe_problem,
+    bool displaced)
   : _amg(amg),
     _subproblem(subproblem),
+    _fe_problem(fe_problem),
+    _displaced(displaced),
     _assembly(_subproblem.assembly(0)),
     _qrule_msm(_assembly.qRuleMortar())
 {
@@ -63,7 +68,7 @@ ComputeMortarFunctor<compute_stage>::operator()()
     const auto & mp_deps = mc->getMatPropDependencies();
     needed_mat_props.insert(mp_deps.begin(), mp_deps.end());
   }
-  _subproblem.setActiveMaterialProperties(needed_mat_props, /*tid=*/0);
+  _fe_problem.setActiveMaterialProperties(needed_mat_props, /*tid=*/0);
 
   // Array to hold custom quadrature point locations on the slave and master sides
   std::vector<Point> custom_xi1_pts, custom_xi2_pts;
@@ -138,16 +143,23 @@ ComputeMortarFunctor<compute_stage>::operator()()
       custom_xi1_pts[qp] = xi1_eta;
     }
 
+    const Elem * reinit_slave_elem = slave_ip;
+
+    // If we're on the displaced mesh, we need to get the corresponding undisplaced elem before
+    // calling _fe_problem.reinitElemFaceRef
+    if (_displaced)
+      reinit_slave_elem = _fe_problem.mesh().elemPtr(reinit_slave_elem->id());
+
     // reinit the variables/residuals/jacobians on the slave interior
-    _subproblem.reinitElemFaceRef(
-        slave_ip, slave_side_id, _slave_boundary_id, TOLERANCE, &custom_xi1_pts);
+    _fe_problem.reinitElemFaceRef(
+        reinit_slave_elem, slave_side_id, _slave_boundary_id, TOLERANCE, &custom_xi1_pts);
 
     // reinit higher-dimensional element face materials
     {
       // Set up Sentinels so that, even if one of the reinitMaterialsXXX() calls throws, we
       // still remember to swap back during stack unwinding.
-      SwapBackSentinel face_sentinel(_subproblem, &SubProblem::swapBackMaterialsFace, /*tid=*/0);
-      _subproblem.reinitMaterialsFace(slave_ip->subdomain_id(), /*tid=*/0);
+      SwapBackSentinel face_sentinel(_fe_problem, &FEProblemBase::swapBackMaterialsFace, /*tid=*/0);
+      _fe_problem.reinitMaterialsFace(slave_ip->subdomain_id(), /*tid=*/0);
     }
 
     if (_has_master)
@@ -161,15 +173,22 @@ ComputeMortarFunctor<compute_stage>::operator()()
         custom_xi2_pts[qp] = xi2_eta;
       }
 
+      const Elem * reinit_master_elem = master_ip;
+
+      // If we're on the displaced mesh, we need to get the corresponding undisplaced elem before
+      // calling _fe_problem.reinitElemFaceRef
+      if (_displaced)
[…]
