### 2020-03-14 f4aa301
PassKey pattern to protect Assembly::JxWNeighbor
Use the PassKey pattern to protect access of JxWNeighbor. This API
was developed for NodeFaceConstraints but users may be tempted to
incorrectly use this for things like DGKernels or InterfaceKernrels.
So using PassKey we can limit this to be accessible to
NodeFaceConstraint.

Closes #14890
