+++ framework/include/functions/VectorPostprocessorFunction.h
@@ -29,11 +29,18 @@ public:
   static InputParameters validParams();
 
   VectorPostprocessorFunction(const InputParameters & parameters);
-  virtual Real value(Real /*t*/, const Point & pt) const override;
+  virtual Real value(Real t, const Point & p) const override;
 
 protected:
   std::unique_ptr<LinearInterpolation> _linear_interp;
-  const unsigned int _component;
   const VectorPostprocessorValue & _argument_column;
   const VectorPostprocessorValue & _value_column;
+
+  /// if the "component" parameter is specified, its value is assigned here and
+  /// function values are interpolated W.R.T. spatial coordinates in that direction,
+  /// otherwise, they are interpolated W.R.T time
+  const unsigned int & _component;
+
+  /// used to indicate that the "component" parameter has not been specified
+  static const unsigned int NO_COMPONENT;
 };
