+++ modules/tensor_mechanics/src/kernels/InertialForceShell.C
@@ -287,264 +287,345 @@ ADInertialForceShell<RESIDUAL>::computeResidual()
     // Loosely following notation in: "On finite element nonlinear analysis of general shell
     // structures", PhD thesis by Said Bolourchi (1975).
 
-    for (unsigned int qp_xy = 0; qp_xy < _2d_points.size(); ++qp_xy)
-    {
-      // Update 0g vectors at plane quadrature points.
-      _0g1_vector.clear();
-      _0g1_vector.push_back(-0.5 * _thickness[qp_xy] * _v1[0]);
-      _0g1_vector.push_back(0.5 * _thickness[qp_xy] * _v2[0]);
-
-      DenseMatrix<Real> G1(3, 2);
-      G1(0, 0) = _0g1_vector[0](0);
-      G1(1, 0) = _0g1_vector[0](1);
-      G1(2, 0) = _0g1_vector[0](2);
-      G1(0, 1) = _0g1_vector[1](0);
-      G1(1, 1) = _0g1_vector[1](1);
-      G1(2, 1) = _0g1_vector[1](2);
-      DenseMatrix<Real> G1T(2, 3);
-      G1.get_transpose(G1T);
-
-      _0g2_vector.clear();
-      _0g2_vector.push_back(-0.5 * _thickness[qp_xy] * _v1[1]);
-      _0g2_vector.push_back(0.5 * _thickness[qp_xy] * _v2[1]);
-
-      DenseMatrix<Real> G2(3, 2);
-      G2(0, 0) = _0g2_vector[0](0);
-      G2(1, 0) = _0g2_vector[0](1);
-      G2(2, 0) = _0g2_vector[0](2);
-      G2(0, 1) = _0g2_vector[1](0);
-      G2(1, 1) = _0g2_vector[1](1);
-      G2(2, 1) = _0g2_vector[1](2);
-
-      DenseMatrix<Real> G2T(2, 3);
-      G2.get_transpose(G2T);
-
-      _0g3_vector.clear();
-      _0g3_vector.push_back(-0.5 * _thickness[qp_xy] * _v1[2]);
-      _0g3_vector.push_back(0.5 * _thickness[qp_xy] * _v2[2]);
-
-      DenseMatrix<Real> G3(3, 2);
-      G3(0, 0) = _0g3_vector[0](0);
-      G3(1, 0) = _0g3_vector[0](1);
-      G3(2, 0) = _0g3_vector[0](2);
-      G3(0, 1) = _0g3_vector[1](0);
-      G3(1, 1) = _0g3_vector[1](1);
-      G3(2, 1) = _0g3_vector[1](2);
-
-      DenseMatrix<Real> G3T(2, 3);
-      G3.get_transpose(G3T);
-
-      _0g4_vector.clear();
-      _0g4_vector.push_back(-0.5 * _thickness[qp_xy] * _v1[3]);
-      _0g4_vector.push_back(0.5 * _thickness[qp_xy] * _v2[3]);
-
-      DenseMatrix<Real> G4(3, 2);
-      G4(0, 0) = _0g4_vector[0](0);
-      G4(1, 0) = _0g4_vector[0](1);
-      G4(2, 0) = _0g4_vector[0](2);
-      G4(0, 1) = _0g4_vector[1](0);
-      G4(1, 1) = _0g4_vector[1](1);
-      G4(2, 1) = _0g4_vector[1](2);
-
-      DenseMatrix<Real> G4T(2, 3);
-      G4.get_transpose(G4T);
-
-      // Conversions to DenseVector from RealVectorValue: Make a method out of this.
-      DenseVector<Real> local_accel_dv_0;
-      DenseVector<Real> local_accel_dv_1;
[…]
