### 2020-03-16 f00427e
Documented XFEMAction and XFEMSingleVariableConstraint.
