### 2020-03-16 97114c3
Error unless realpath() succeeds
If it fails then the return buffer contents are undefined so it's not
safe to continue anyway.

This fixes compilation with -Werror for me.

Refs #1542
