+++ modules/stochastic_tools/include/actions/LoadSurrogateDataAction.h
@@ -1,0 +1,24 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#pragma once
+
+#include "Action.h"
+class SurrogateModel;
+
+class LoadSurrogateDataAction : public Action
+{
+public:
+  static InputParameters validParams();
+  LoadSurrogateDataAction(InputParameters params);
+  virtual void act() override;
+
+private:
+  void load(SurrogateModel & model);
+};
