+++ modules/stochastic_tools/src/surrogates/PolynomialChaos.C
@@ -20,20 +20,28 @@ PolynomialChaos::validParams()
   InputParameters params = SurrogateModel::validParams();
   params.addClassDescription("Computes and evaluates polynomial chaos surrogate model.");
 
-  params.addRequiredParam<unsigned int>("order", "Maximum polynomial order.");
-  params.addRequiredParam<std::vector<DistributionName>>(
+  // Training parameters
+  params.addParam<SamplerName>("training_sampler", "Training set defined by a sampler object.");
+  params.addParam<VectorPostprocessorName>(
+      "stochastic_results", "Vectorpostprocessor with results of samples created by trainer");
+
+  params.addParam<unsigned int>("order", "Maximum polynomial order.");
+  params.addParam<std::vector<DistributionName>>(
       "distributions", "Names of the distributions samples were taken from.");
 
+  params.addParamNamesToGroup("training_sampler stochastic_results order distributions",
+                              "training");
   return params;
 }
 
 PolynomialChaos::PolynomialChaos(const InputParameters & parameters)
   : SurrogateModel(parameters),
-    _order(getParam<unsigned int>("order")),
-    _quad_sampler(nullptr),
-    _tuple(generateTuple(getParam<std::vector<DistributionName>>("distributions").size(), _order)),
-    _ncoeff(_tuple.size()),
-    _coeff(_ncoeff)
+    _ndim(declareModelData<unsigned int>("_ndim")),
+    _tuple(declareModelData<std::vector<std::vector<unsigned int>>>("_tuple")),
+    _ncoeff(declareModelData<unsigned int>("_ncoeff")),
+    _coeff(declareModelData<std::vector<Real>>("_coeff")),
+    _poly(declareModelData<std::vector<std::unique_ptr<const PolynomialQuadrature::Polynomial>>>(
+        "_poly"))
 {
   MooseUtils::linearPartitionItems(
       _ncoeff, n_processors(), processor_id(), _n_local_coeff, _st_local_coeff, _end_local_coeff);
@@ -42,24 +50,37 @@ PolynomialChaos::PolynomialChaos(const InputParameters & parameters)
 void
 PolynomialChaos::initialSetup()
 {
-  SurrogateModel::initialSetup();
+  if (isTraining())
+  {
+    // Setup data needed for training
+    _order = getParam<unsigned int>("order");
+    _tuple = generateTuple(getParam<std::vector<DistributionName>>("distributions").size(), _order);
+    _ncoeff = _tuple.size();
+    _coeff.resize(_ncoeff, 0);
+    _values_ptr = &getVectorPostprocessorValue("stochastic_results",
+                                               getParam<SamplerName>("training_sampler"));
 
-  // Special circumstances if sampler is quadrature
-  _quad_sampler = dynamic_cast<QuadratureSampler *>(_sampler);
+    // Sampler
+    _sampler = &getSamplerByName(getParam<SamplerName>("training_sampler"));
+    _ndim = _sampler->getNumberOfCols();
 
-  std::vector<DistributionName> dname = getParam<std::vector<DistributionName>>("distributions");
-  // Check if sampler dimensionality matches number of distributions
-  if (dname.size() != _ndim)
-    mooseError("Sampler number of columns does not match number of inputted distributions.");
-  for (auto nm : dname)
-    _poly.push_back(PolynomialQuadrature::makePolynomial(&getDistributionByName(nm)));
+    // Special circumstances if sampler is quadrature
+    _quad_sampler = dynamic_cast<QuadratureSampler *>(_sampler);
+
[…]
