+++ modules/tensor_mechanics/src/interfacekernels/CZMInterfaceKernel.C
@@ -21,6 +21,7 @@ CZMInterfaceKernel::validParams()
                                         " component == 0, ==> X"
                                         " component == 1, ==> Y"
                                         " component == 2, ==> Z");
+  params.set<bool>("_use_undisplaced_reference_points") = true;
 
   params.addRequiredCoupledVar("displacements", "the string containing displacement variables");
 
