+++ framework/src/problems/FEProblemBase.C
@@ -328,7 +328,8 @@ FEProblemBase::FEProblemBase(const InputParameters & parameters)
     _u_dot_old_requested(false),
     _u_dotdot_old_requested(false),
     _has_mortar(false),
-    _num_grid_steps(0)
+    _num_grid_steps(0),
+    _displaced_neighbor_ref_pts("invert_elem_phys use_undisplaced_ref unset", "unset")
 {
   _time = 0.0;
   _time_old = 0.0;
@@ -1751,7 +1752,18 @@ FEProblemBase::reinitNeighbor(const Elem * elem, unsigned int side, THREAD_ID ti
   _aux->reinitNeighborFace(neighbor, neighbor_side, bnd_id, tid);
 
   if (_displaced_problem && _reinit_displaced_face)
-    _displaced_problem->reinitNeighbor(_displaced_mesh->elemPtr(elem->id()), side, tid);
+  {
+    // There are cases like for cohesive zone modeling without significant sliding where we cannot
+    // use FEInterface::inverse_map in Assembly::reinitElemAndNeighbor in the displaced problem
+    // because the physical points coming from the element don't actually lie on the neighbor. In
+    // that case we instead pass over the reference points from the undisplaced calculation
+    const std::vector<Point> * displaced_ref_pts = nullptr;
+    if (_displaced_neighbor_ref_pts == "use_undisplaced_ref")
+      displaced_ref_pts = &_assembly[tid]->qRuleNeighbor()->get_points();
+
+    _displaced_problem->reinitNeighbor(
+        _displaced_mesh->elemPtr(elem->id()), side, tid, displaced_ref_pts);
+  }
 }
 
 void
@@ -2549,11 +2561,38 @@ FEProblemBase::addDGKernel(const std::string & dg_kernel_name,
                            const std::string & name,
                            InputParameters & parameters)
 {
+  bool use_undisplaced_reference_points = parameters.get<bool>("_use_undisplaced_reference_points");
+
   if (_displaced_problem && parameters.get<bool>("use_displaced_mesh"))
   {
     parameters.set<SubProblem *>("_subproblem") = _displaced_problem.get();
     parameters.set<SystemBase *>("_sys") = &_displaced_problem->nlSys();
     _reinit_displaced_face = true;
+
+    if (use_undisplaced_reference_points)
+    {
+      if (_displaced_neighbor_ref_pts == "invert_elem_phys")
+        mooseError(
+            "Cannot use elem-neighbor objects which rely on 1) undisplaced reference points and 2) "
+            "inversion of master elem physical points in the same simulation");
+      else if (_displaced_neighbor_ref_pts == "unset")
+        _displaced_neighbor_ref_pts = "use_undisplaced_ref";
+      else if (_displaced_neighbor_ref_pts != "use_undisplaced_ref")
+        mooseError("_displaced_neighbor_ref_pts has an invalid state ",
+                   std::string(_displaced_neighbor_ref_pts));
+    }
+    else
+    {
+      if (_displaced_neighbor_ref_pts == "use_undisplaced_ref")
+        mooseError(
+            "Cannot use elem-neighbor objects which rely on 1) undisplaced reference points and 2) "
+            "inversion of master elem physical points in the same simulation");
+      else if (_displaced_neighbor_ref_pts == "unset")
+        _displaced_neighbor_ref_pts = "invert_elem_phys";
+      else if (_displaced_neighbor_ref_pts != "invert_elem_phys")
+        mooseError("_displaced_neighbor_ref_pts has an invalid state ",
+                   std::string(_displaced_neighbor_ref_pts));
+    }
   }
   else
   {
@@ -2583,11 +2622,38 @@ FEProblemBase::addInterfaceKernel(const std::string & interface_kernel_name,
[…]
