+++ framework/include/problems/FEProblemBase.h
@@ -2122,6 +2122,10 @@ private:
 
   /// Number of steps in a grid sequence
   unsigned int _num_grid_steps;
+
+  /// MooseEnum describing how to obtain reference points for displaced mesh dgkernels and/or
+  /// interface kernels. Options are invert_elem_phys, use_undisplaced_ref, and the default unset.
+  MooseEnum _displaced_neighbor_ref_pts;
 };
 
 template <typename T>
