### 2020-03-16 42b26ed
Allow people to be stupid again
