+++ framework/src/meshgenerators/SideSetsFromPointsGenerator.C
@@ -89,8 +89,8 @@ SideSetsFromPointsGenerator::generate()
       {
         // This is the side that we want to paint our sideset with
         // First get the normal
-        _fe_face->reinit(elem, side);
         const std::vector<Point> & normals = _fe_face->get_normals();
+        _fe_face->reinit(elem, side);
 
         flood(elem, normals[0], boundary_ids[i], *mesh);
       }
