### 2020-03-16 9c4f182
Copy JxW neighbor after reiniting neighbor FE objects
Closes #14870
