### 2020-03-17 c3edbf5
Add different JxW test
Refs #14870
