+++ modules/geochemistry/doc/content/modules/geochemistry/geochemistry_database_reader.md
@@ -1,0 +1,153 @@
+# The `geochemistry` database reader
+
+A description of the GWB database format has been given [elsewhere](database.md).  This page describes the database reader of the `geochemistry` module.
+
+## Database format
+
+The database contains info on the following
+
+- The "standard" temperatures \[$^{\circ}$C\] at which the equilibrium coefficients, activity coefficients, etc, are provided.
+- The type of interpolation used to compute equilibrium coefficients, etc, at other temperatures other than standard temperatures.
+- The steam saturation curve: pressures \[bar = $10^{5}\,$Pa\] at the standard temperatures.
+- The [Debye-Huckel coefficients](activity_coefficients.md) evaluated at the standard temperatures
+- The neutral-species [activity coefficients](activity_coefficients.md) evaluated at the standard temperatures
+- The coefficients required to compute the [activity of water](activity_coefficients.md) evaluated at the standard temperatures
+- Definitions of the names, chemical formulae and mole weight \[g\] of elements.
+- The basis species, including their
+
+  - name
+  - charge
+  - ion size (for activity computation)
+  - mole weight \[g\]
+  - elemental decomposition
+
+- Optionally, information regarding sorbing sites (which are simply considered as another basis species by the code), including their
+
+  - name
+  - charge
+  - mole weight \[g\]
+  - elemental decomposition
+
+- Optionally, the redox pairs, including their
+
+  - name
+  - charge
+  - ion size (for activity computation)
+  - mole weight \[g\]
+  - equilibrium reaction stoichiometry, which may involve all basis species as well as redox pairs that have already been defined (the file is read in order)
+  - the equilibrium constant, evaluated at each of the standard temperatures, for the equilibrium reaction
+
+- Optionally, secondary aqueous species, including their
+
+  - name
+  - charge
+  - ion size (for activity computation)
+  - mole weight \[g\]
+  - equilibrium reaction stoichiometry, which may involve all basis species as well as redox pairs
+  - the equilibrium constant, evaluated at each of the standard temperatures, for the equilibrium reaction
+
+- Optionally, information concerning the free electron, including its
+
+  - name
+  - charge
+  - ion size (for activity computation)
+  - mole weight \[g\]
+  - equilibrium reaction stoichiometry, which may involve all basis species as well as redox pairs
+  - the equilibrium constant, evaluated at each of the standard temperatures, for the equilibrium reaction
+
+- Optionally, information concerning minerals, including their
+
+  - name
+  - mole volume \[cm$^{3}$\]
+  - mole weight \[g\]
+  - equilibrium reaction stoichiometry, which may involve all basis species as well as redox pairs
+  - the equilibrium constant, evaluated at each of the standard temperatures, for the equilibrium reaction
+
+- Optionally, information concerning gases, including their
+
+  - name
+  - coefficients for computing fugacity
+  - equilibrium reaction stoichiometry, which may involve all basis species as well as redox pairs
+  - the equilibrium constant, evaluated at each of the standard temperatures, for the equilibrium reaction
+
+- Optionally, information regarding sorbing minerals that are involved in surface complexation reactions, including their
+
+  - name, which must be one of the minerals already defined
+  - their specific surface area \[m$^{2}$.g$^{-1}$\]
+  - information regarding the density \[mol/mol(mineral)\] of each sorbing sites on their surface.  The sorbing sites must have already been defined, as above
+
+- Optionally, information concerning surface species involved in sorbing, including their
+
+  - name
+  - charge
+  - mole weight \[g\]
+  - equilibrium reaction stoichiometry, which may involve all basis species, redox pairs and will involve at least one sorbing site (which are considered as just another basis species by the code)
+  - $\log_{10}K$ for the reaction, as well as its derivative with respect to temperature
+
+
+## Reader workflow
+
+### Initial parsing
+
[…]
