+++ modules/geochemistry/python/tests/test_eq36reader.py
@@ -1,0 +1,166 @@
+#!/usr/bin/env python3
+#* This file is part of the MOOSE framework
+#* https://www.mooseframework.org
+#*
+#* All rights reserved, see COPYRIGHT for full restrictions
+#* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+#*
+#* Licensed under LGPL 2.1, please see LICENSE for details
+#* https://www.gnu.org/licenses/lgpl-2.1.html
+
+import unittest
+from dbclass import ThermoDB
+from readers import eq36_reader
+
+class TestGWBReader(unittest.TestCase):
+    """
+    Test that the Geochemist WorkBench database reader functions correctly
+    """
+
+    def readDatabase(self):
+        """
+        Read the database
+        """
+        with open('testdata/eq36testdata.dat', 'r') as dbfile:
+            dblist = dbfile.readlines()
+
+        # Parse the database
+        self.db = eq36_reader.readDatabase(dblist)
+
+    def testTemperature(self):
+        """
+        Test that the temperatures are correctly parsed
+        """
+        self.readDatabase()
+        gold = ['0.0100', '25.0000', '60.0000', '100.0000',
+                '150.0000', '200.0000', '250.0000', '300.0000']
+
+        self.assertEqual(self.db.temperatures, gold)
+
+    def testPressure(self):
+        """
+        Test that the pressures are correctly parsed
+        """
+        self.readDatabase()
+        gold = ['1.0132', '1.0132', '1.0132', '1.0132',
+                '4.7572', '15.5365', '39.7365', '85.8378']
+
+        self.assertEqual(self.db.pressures, gold)
+
+    def testActivityModel(self):
+        """
+        Test that the activity model is set correctly
+        """
+        self.readDatabase()
+        gold = 'debye-huckel'
+        adhgold = ['0.4939', '0.5114', '0.5465', '0.5995',
+                   '0.6855', '0.7994', '0.9593', '1.2180']
+        bdhgold = ['0.3253', '0.3288', '0.3346', '0.3421',
+                   '0.3525', '0.3639', '0.3766', '0.3925']
+        bdotgold = ['0.0394', '0.0410', '0.0438', '0.0460',
+                    '0.0470', '0.0470', '0.0340', '0.0000']
+
+        self.assertEqual(self.db.activity_model, gold)
+        self.assertEqual(self.db.adh, adhgold)
+        self.assertEqual(self.db.bdh, bdhgold)
+        self.assertEqual(self.db.bdot, bdotgold)
+
+
+    def testFugacityModel(self):
+        """
+        Test that the fugacity model is set correctly
+        """
+        self.readDatabase()
+        gold = None
+
+        self.assertEqual(self.db.fugacity_model, gold)
+
+    def testNeutralSpecies(self):
+        """
+        Test that the neutral species coefficients are correctly parsed
+        """
+        self.readDatabase()
+        gold = {'co2': {'(coefficients': ['-1.0312', '0.0012806', '255.9', '0.4445', '-0.001606']}}
+
+        self.assertDictEqual(self.db.neutral_species, gold)
+
+    def testElements(self):
+        """
+        Test that the elements are correctly parsed
+        """
+        self.readDatabase()
+        gold = {'O': {'molecular weight': '15.99940'},
+                'Ag': {'molecular weight': '107.86820'}}
+
+        self.assertDictEqual(self.db.elements, gold)
+
+    def testBasisSpecies(self):
+        """
+        Test that the basis species are correctly parsed
+        """
+        self.readDatabase()
+        gold = {'H2O': {'charge': '0.0', 'radius': '3.0', 'molecular weight': '18.015',
+                         'elements': {'H': '2.0000', 'O': '1.0000'}},
[…]
