+++ modules/geochemistry/python/readers/gwb_reader.py
@@ -1,0 +1,451 @@
+#!/usr/bin/env python3
+#* This file is part of the MOOSE framework
+#* https://www.mooseframework.org
+#*
+#* All rights reserved, see COPYRIGHT for full restrictions
+#* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+#*
+#* Licensed under LGPL 2.1, please see LICENSE for details
+#* https://www.gnu.org/licenses/lgpl-2.1.html
+
+from dbclass import ThermoDB
+
+def readDatabase(dblist):
+    """
+    Read a Geochemists Workbench database file
+    """
+
+    activity_model = None
+    fugacity_model = None
+    temperatures = []
+    pressures = []
+    adh = []
+    bdh = []
+    bdot = []
+    elements = {}
+    neutral_species = {}
+    basis_species = {}
+    redox_couples = {}
+    secondary_species = {}
+    free_electron = {}
+    mineral_species = {}
+    gas_species = {}
+    oxides = {}
+
+    line = 0
+
+    while line < len(dblist):
+
+        if 'activity model' in dblist[line]:
+            activity_model = dblist[line].split(':')[1].strip()
+
+        if 'fugacity model' in dblist[line]:
+            fugacity_model = dblist[line].split(':')[1].strip()
+
+        if 'temperatures' in dblist[line]:
+            line = line+1
+            while dblist[line].strip()[0].isnumeric():
+                temperatures.extend(dblist[line].split())
+                line = line+1
+
+        if 'pressures' in dblist[line]:
+            line = line+1
+            while len(pressures) < len(temperatures):
+                pressures.extend(dblist[line].split())
+                line = line+1
+
+        if activity_model == 'debye-huckel':
+            # Read Debye-Huckel activity coefficients
+            if 'debye huckel a' in dblist[line]:
+                line = line+1
+                while len(adh) < len(temperatures):
+                    adh.extend(dblist[line].split())
+                    line = line+1
+
+            if 'debye huckel b' in dblist[line]:
+                line = line+1
+                while len(bdh) < len(temperatures):
+                    bdh.extend(dblist[line].split())
+                    line = line+1
+
+            if 'bdot' in dblist[line]:
+                line = line+1
+                while len(bdot) < len(temperatures):
+                    bdot.extend(dblist[line].split())
+                    line = line+1
+
+        # Read the neutral species activity coefficients
+        labels = ['a', 'b', 'c', 'd']
+        if 'co2' in dblist[line]:
+            if 'co2' not in neutral_species:
+                neutral_species['co2'] = {}
+
+            key = labels[int(dblist[line].split('*')[-1].split()[2])-1]
+            line = line+1
+            vals = []
+            while len(vals) < len(temperatures):
+                vals.extend(dblist[line].split())
+                line = line+1
+
+            neutral_species['co2'][key] = vals
+            line = line-1
+
+        if 'h2o' in dblist[line]:
+            if 'h2o' not in neutral_species:
+                neutral_species['h2o'] = {}
+
+            key = labels[int(dblist[line].split('*')[-1].split()[2])-1]
+            line = line+1
+            vals = []
+            while len(vals) < len(temperatures):
+                vals.extend(dblist[line].split())
+                line = line+1
+
+            neutral_species['h2o'][key] = vals
+            line = line-1
+
+        if 'elements\n' in dblist[line]:
+            # Read the elements
+            line = line+1
+
+            reading_elements = True
+            while reading_elements:
+                if dblist[line].split():
+                    elemlist = dblist[line].replace('(','').replace(')','').split()
+                    elements[elemlist[1]] = {}
+                    elements[elemlist[1]]['name'] = elemlist[0]
+                    elements[elemlist[1]]['molecular weight'] = elemlist[4]
+                line = line+1
+
[…]

[…]
