### 2020-03-17 03d75ae
Thermodynamic database reader utility
This utility reads the JSON database of geochemical reactions
and enables easy access to species and reaction information.

Refs #14897
