+++ modules/geochemistry/include/utils/GeochemicalDatabaseReader.h
@@ -1,0 +1,384 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#pragma once
+
+#include "json/json.h"
+#include "MooseTypes.h"
+
+/**
+ * Data structure for basis (primary) species.
+ * Members are:
+ * * species name
+ * * elemental composition (elements and weights)
+ * * ionic radius (Angstrom)
+ * * charge
+ * * molecular weight (g)
+ */
+struct GeochemistryBasisSpecies
+{
+  GeochemistryBasisSpecies(){};
+
+  std::string name;
+  std::map<std::string, Real> elements;
+  Real radius;
+  Real charge;
+  Real molecular_weight;
+};
+
+/**
+ * Data structure for secondary equilibrium species.
+ * Members are:
+ * * Species name
+ * * Basis species and stoichiometric coefficients present
+ * * Equilibrium constant at given temperatures
+ * * ionic radius (Angstrom)
+ * * charge
+ * * molecular weight (g)
+ */
+struct GeochemistryEquilibriumSpecies
+{
+  GeochemistryEquilibriumSpecies(){};
+
+  std::string name;
+  std::map<std::string, Real> basis_species;
+  std::vector<Real> equilibrium_const;
+  Real radius;
+  Real charge;
+  Real molecular_weight;
+};
+
+/**
+ * Data structure for mineral species.
+ * Members are:
+ * * Species name
+ * * Molecular volume (units?)
+ * * Basis species and stoichiometric coefficients present
+ * * Equilibrium constant at given temperatures ()
+ * * molecular weight (g)
+ */
+struct GeochemistryMineralSpecies
+{
+  GeochemistryMineralSpecies(){};
+
+  std::string name;
+  Real molecular_volume;
+  std::map<std::string, Real> basis_species;
+  std::vector<Real> equilibrium_const;
+  Real molecular_weight;
+};
+
+/**
+ * Data structure for mineral species.
+ * Members are:
+ * * Species name
+ * * Basis species and stoichiometric coefficients present
+ * * Equilibrium constant at given temperatures ()
+ * * molecular weight (g)
+ * * Spycher-Reed fugacity coefficients
+ * * Pcrit, Tcrit and omega - Tsonopoulos fugacity coefficients
+ */
+struct GeochemistryGasSpecies
+{
+  GeochemistryGasSpecies(){};
+
+  std::string name;
+  std::map<std::string, Real> basis_species;
+  std::vector<Real> equilibrium_const;
+  Real molecular_weight;
+  std::vector<Real> chi;
+  Real Pcrit;
+  Real Tcrit;
+  Real omega;
+};
+
+/**
+ * Data structure for redox species.
+ * Members are:
+ * * Species name
+ * * Basis species and stoichiometric coefficients present
+ * * Equilibrium constant at given temperatures
+ * * ionic radius (Angstrom)
+ * * charge
+ * * molecular weight (g)
+ */
+struct GeochemistryRedoxSpecies
+{
+  GeochemistryRedoxSpecies(){};
+
+  std::string name;
+  std::map<std::string, Real> basis_species;
+  std::vector<Real> equilibrium_const;
+  Real radius;
+  Real charge;
+  Real molecular_weight;
+};
+
+/**
+ * Data structure for oxide species.
+ * Members are:
+ * * Species name
[…]
