+++ modules/geochemistry/src/utils/GeochemicalDatabaseReader.C
@@ -1,0 +1,561 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "GeochemicalDatabaseReader.h"
+
+#include "MooseUtils.h"
+#include "Conversion.h"
+#include "string"
+#include <fstream>
+
+GeochemicalDatabaseReader::GeochemicalDatabaseReader(const std::string filename)
+  : _filename(filename)
+{
+  read(_filename);
+}
+
+void
+GeochemicalDatabaseReader::read(FileName filename)
+{
+  MooseUtils::checkFileReadable(filename);
+
+  // Read the JSON database
+  std::ifstream jsondata(filename);
+  jsondata >> _root;
+}
+
+std::string
+GeochemicalDatabaseReader::getActivityModel() const
+{
+  return _root["Header"]["activity model"].asString();
+}
+
+std::string
+GeochemicalDatabaseReader::getFugacityModel() const
+{
+  return _root["Header"]["fugacity model"].asString();
+}
+
+std::vector<Real>
+GeochemicalDatabaseReader::getTemperatures()
+{
+  // Read temperature points
+  if (_root["Header"].isMember("temperatures"))
+  {
+    auto temperatures = _root["Header"]["temperatures"];
+    _temperature_points.resize(temperatures.size());
+    for (unsigned int i = 0; i < temperatures.size(); ++i)
+      _temperature_points[i] = std::stod(temperatures[i].asString());
+  }
+
+  return _temperature_points;
+}
+
+std::vector<Real>
+GeochemicalDatabaseReader::getPressures()
+{
+  // Read pressure points
+  if (_root["Header"].isMember("pressures"))
+  {
+    auto pressures = _root["Header"]["pressures"];
+    _pressure_points.resize(pressures.size());
+    for (unsigned int i = 0; i < pressures.size(); ++i)
+      _pressure_points[i] = std::stod(pressures[i].asString());
+  }
+
+  return _pressure_points;
+}
+
+GeochemistryDebyeHuckel
+GeochemicalDatabaseReader::getDebyeHuckel()
+{
+  if (getActivityModel() == "debye-huckel")
+  {
+    if (_root["Header"].isMember("adh"))
+    {
+      std::vector<Real> adhvals(_root["Header"]["adh"].size());
+      for (unsigned int i = 0; i < _root["Header"]["adh"].size(); ++i)
+        adhvals[i] = std::stod(_root["Header"]["adh"][i].asString());
+      _debye_huckel.adh = adhvals;
+    }
+
+    if (_root["Header"].isMember("bdh"))
+    {
+      std::vector<Real> bdhvals(_root["Header"]["bdh"].size());
+      for (unsigned int i = 0; i < _root["Header"]["bdh"].size(); ++i)
+        bdhvals[i] = std::stod(_root["Header"]["bdh"][i].asString());
+      _debye_huckel.bdh = bdhvals;
+    }
+
+    if (_root["Header"].isMember("bdot"))
+    {
+      std::vector<Real> bdotvals(_root["Header"]["bdot"].size());
+      for (unsigned int i = 0; i < _root["Header"]["bdot"].size(); ++i)
+        bdotvals[i] = std::stod(_root["Header"]["bdot"][i].asString());
+      _debye_huckel.bdot = bdotvals;
+    }
+  }
+  else
[…]
