+++ modules/geochemistry/unit/src/GeochemicalDatabaseReaderTest.C
@@ -1,0 +1,421 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "gtest/gtest.h"
+
+#include "GeochemicalDatabaseReader.h"
+
+TEST(GeochemicalDatabaseReaderTest, getActivityModel)
+{
+  GeochemicalDatabaseReader database("data/moose_testdb.json");
+
+  EXPECT_EQ(database.getActivityModel(), "debye-huckel");
+}
+
+TEST(GeochemicalDatabaseReaderTest, getFugacityModel)
+{
+  GeochemicalDatabaseReader database("data/moose_testdb.json");
+
+  EXPECT_EQ(database.getFugacityModel(), "tsonopoulos");
+}
+
+TEST(GeochemicalDatabaseReaderTest, getTemperatures)
+{
+  GeochemicalDatabaseReader database("data/moose_testdb.json");
+
+  // Get the temperature points from the database and compare with the expected
+  // valules
+  auto temperature_points = database.getTemperatures();
+
+  std::vector<Real> temperature_points_gold{0.0, 25.0, 60.0, 100.0, 150.0, 200.0, 250.0, 300.0};
+  EXPECT_EQ(temperature_points, temperature_points_gold);
+}
+
+TEST(GeochemicalDatabaseReaderTest, getPressures)
+{
+  GeochemicalDatabaseReader database("data/moose_testdb.json");
+
+  // Get the pressure points from the database and compare with the expected
+  // valules
+  auto pressure_points = database.getPressures();
+
+  std::vector<Real> pressure_points_gold{
+      1.0134, 1.0134, 1.0134, 1.0134, 4.7600, 15.5490, 39.7760, 85.9270};
+  EXPECT_EQ(pressure_points, pressure_points_gold);
+}
+
+TEST(GeochemicalDatabaseReaderTest, getDebyeHuckel)
+{
+  GeochemicalDatabaseReader database("data/moose_testdb.json");
+
+  // Get the Debye-Huckel from the database and compare with the expected
+  // valules
+  auto dh = database.getDebyeHuckel();
+
+  std::vector<Real> adh_gold{.4913, .5092, .5450, .5998, .6898, .8099, .9785, 1.2555};
+  std::vector<Real> bdh_gold{.3247, .3283, .3343, .3422, .3533, .3655, .3792, .3965};
+  std::vector<Real> bdot_gold{.0174, .0410, .0440, .0460, .0470, .0470, .0340, 0.0000};
+
+  EXPECT_EQ(dh.adh, adh_gold);
+  EXPECT_EQ(dh.bdh, bdh_gold);
+  EXPECT_EQ(dh.bdot, bdot_gold);
+}
+
+TEST(GeochemicalDatabaseReaderTest, getNeutralSpeciesActivity)
+{
+  GeochemicalDatabaseReader database("data/moose_testdb.json");
+
+  // Get the neutral species activity coefficients from the database
+  // and compare with the expected valules
+  auto nsa = database.getNeutralSpeciesActivity();
+
[…]
