### 2020-03-17 b9884ea
Base node proc_id on owning element proc_id
This makes my test work in replicated mode in dbg mode in parallel
but not in distributed mode. I think I'm going to have to reivive
PR #14456 for that
