### 2020-03-17 0e2ea68
python database converters now convert sorption stuff for the GWB format.  Small fixes, including the path to find included libraries.  Refs #14897
