+++ modules/geochemistry/python/tests/test_eq36reader.py
@@ -8,7 +8,12 @@
 #* Licensed under LGPL 2.1, please see LICENSE for details
 #* https://www.gnu.org/licenses/lgpl-2.1.html
 
+import os
+import sys
 import unittest
+# so we can find our libraries, no matter how we're called
+findbin = os.path.dirname(os.path.realpath(sys.argv[0]))
+sys.path.append(os.path.join(findbin, "../"))
 from dbclass import ThermoDB
 from readers import eq36_reader
 
