+++ modules/geochemistry/python/readers/gwb_reader.py
@@ -31,6 +31,8 @@ def readDatabase(dblist):
     mineral_species = {}
     gas_species = {}
     oxides = {}
+    sorbing_minerals = {}
+    surface_species = {}
 
     line = 0
 
@@ -135,11 +137,24 @@ def readDatabase(dblist):
 
                     # Charge, ionic radius and molecular weight
                     if 'charge' in dblist[line]:
-                        data = dblist[line].replace('=', ' ').split()
-                        basis_species[species]['charge'] = data[1]
-                        basis_species[species]['radius'] = data[4]
-                        basis_species[species]['molecular weight'] = data[8]
-                        line = line+1
+                        data = dblist[line][dblist[line].index("charge") + 6:].split("=")[1]
+                        basis_species[species]['charge'] = data.split()[0]
+                        if len(data.split()) <= 1:
+                            line = line+1
+                    if 'ion size' in dblist[line]:
+                        data = dblist[line][dblist[line].index("ion size") + 8:].split("=")[1]
+                        if len(data.split()) <= 1 or data.split()[1] != "A":
+                            raise ValueError("ion size must be measured in Angstroms")
+                        basis_species[species]['radius'] = data.split()[0]
+                        if len(data.split()) <= 2:
+                            line = line + 1
+                    if 'mole wt.' in dblist[line]:
+                        data = dblist[line][dblist[line].index("mole wt.") + 8:].split("=")[1]
+                        if len(data.split()) <= 1 or data.split()[1] != "g":
+                            raise ValueError("molecular weight must be measured in grams")
+                        basis_species[species]['molecular weight'] = data.split()[0]
+                        if len(data.split()) <= 2:
+                            line = line + 1
 
                     # Elements in basis species
                     if 'elements' in dblist[line]:
@@ -171,11 +186,24 @@ def readDatabase(dblist):
 
                     # Charge, ionic radius and molecular weight
                     if 'charge' in dblist[line]:
-                        data = dblist[line].replace('=', ' ').split()
-                        redox_couples[couple]['charge'] = data[1]
-                        redox_couples[couple]['radius'] = data[4]
-                        redox_couples[couple]['molecular weight'] = data[8]
-                        line = line+1
+                        data = dblist[line][dblist[line].index("charge") + 6:].split("=")[1]
+                        redox_couples[couple]['charge'] = data.split()[0]
+                        if len(data.split()) <= 1:
+                            line = line+1
+                    if 'ion size' in dblist[line]:
+                        data = dblist[line][dblist[line].index("ion size") + 8:].split("=")[1]
+                        if len(data.split()) <= 1 or data.split()[1] != "A":
+                            raise ValueError("ion size must be measured in Angstroms")
+                        redox_couples[couple]['radius'] = data.split()[0]
+                        if len(data.split()) <= 2:
+                            line = line + 1
+                    if 'mole wt.' in dblist[line]:
+                        data = dblist[line][dblist[line].index("mole wt.") + 8:].split("=")[1]
+                        if len(data.split()) <= 1 or data.split()[1] != "g":
+                            raise ValueError("molecular weight must be measured in grams")
+                        redox_couples[couple]['molecular weight'] = data.split()[0]
+                        if len(data.split()) <= 2:
+                            line = line + 1
 
                     # Species in redox couples
                     if 'species' in dblist[line]:
@@ -289,7 +317,47 @@ def readDatabase(dblist):
                     reading_freeelectron = False
                     line = line-1
 
-        if 'minerals\n' in dblist[line]:
+        if 'sorbing minerals\n' in dblist[line]:
+            # Read sorbing minerals
+            line = line+1
+            reading_sorbing_mineral_species = True
+            while reading_sorbing_mineral_species:
+                if dblist[line].strip():
+                    # Mineral name
[…]
