+++ modules/geochemistry/python/tests/test_dbutils.py
@@ -8,10 +8,15 @@
 #* Licensed under LGPL 2.1, please see LICENSE for details
 #* https://www.gnu.org/licenses/lgpl-2.1.html
 
+import os
+import sys
 import unittest
 import json
 import io
 from contextlib import redirect_stdout
+# so we can find our libraries, no matter how we're called
+findbin = os.path.dirname(os.path.realpath(sys.argv[0]))
+sys.path.append(os.path.join(findbin, "../"))
 import dbutils
 
 class TestDBUtils(unittest.TestCase):

[…]
