### 2020-03-17 2217a42
fix suffix for mint package, closes #14912
