### 2020-03-17 759d9a8
Update to static member validParams (#334)
