### 2020-03-17 d0ac87b
Using references for vector and set inputs #14836
