### 2020-03-17 0c34a70
tests: added tests for face info data and verified results by hand (at least some of them)
