### 2020-03-17 4b0c975
Added material_dependency_check flag #14914
