### 2020-03-17 67c8a62
Added material_dependency_check flag #14914
