### 2020-03-17 b95c5f8
Added validate_coupling flag
