### 2020-03-17 ad254c1
Update validate_coupling comment
Co-Authored-By: Daniel Schwen <daniel@schwen.de>
