### 2020-03-17 3af9790
Update validate_coupling comment
Co-Authored-By: Daniel Schwen <daniel@schwen.de>
