+++ framework/include/problems/FEProblemBase.h
@@ -2018,7 +2018,7 @@ protected:
   bool _material_coverage_check;
 
   /// Determines whether a check to verify material dependencies on every subdomain
-  bool _material_dependency_check;
+  const bool _material_dependency_check;
 
   /// Maximum number of quadrature points used in the problem
   unsigned int _max_qps;
