### 2020-03-17 20cf2f9
better error checking in gwb_reader.  add sorbtion stuff to dbclass.  Refs #14897
