+++ modules/geochemistry/python/readers/gwb_reader.py
@@ -78,7 +78,7 @@ def readDatabase(dblist):
 
         # Read the neutral species activity coefficients
         labels = ['a', 'b', 'c', 'd']
-        if 'co2' in dblist[line]:
+        if 'c co2' in dblist[line]:
             if 'co2' not in neutral_species:
                 neutral_species['co2'] = {}
 
@@ -92,7 +92,7 @@ def readDatabase(dblist):
             neutral_species['co2'][key] = vals
             line = line-1
 
-        if 'h2o' in dblist[line]:
+        if 'c h2o' in dblist[line]:
             if 'h2o' not in neutral_species:
                 neutral_species['h2o'] = {}
 
@@ -106,7 +106,7 @@ def readDatabase(dblist):
             neutral_species['h2o'][key] = vals
             line = line-1
 
-        if 'elements\n' in dblist[line]:
+        if 'elements\n' in dblist[line] and len(dblist[line].split()) == 2:
             # Read the elements
             line = line+1
 
@@ -114,6 +114,8 @@ def readDatabase(dblist):
             while reading_elements:
                 if dblist[line].split():
                     elemlist = dblist[line].replace('(','').replace(')','').split()
+                    if len(elemlist) != 6 or elemlist[2] != "mole" or elemlist[3] != "wt.=" or elemlist[5] != "g":
+                        raise ValueError("Element line incorrectly specified.  Offending line is: " + dblist[line])
                     elements[elemlist[1]] = {}
                     elements[elemlist[1]]['name'] = elemlist[0]
                     elements[elemlist[1]]['molecular weight'] = elemlist[4]
@@ -123,7 +125,7 @@ def readDatabase(dblist):
                     reading_elements = False
                     line = line-1
 
-        if 'basis species\n' in dblist[line]:
+        if 'basis species\n' in dblist[line] and len(dblist[line].split()) == 3:
             # Read the basis species
             line = line+1
             reading_basis = True
@@ -144,20 +146,20 @@ def readDatabase(dblist):
                     if 'ion size' in dblist[line]:
                         data = dblist[line][dblist[line].index("ion size") + 8:].split("=")[1]
                         if len(data.split()) <= 1 or data.split()[1] != "A":
-                            raise ValueError("ion size must be measured in Angstroms")
+                            raise ValueError("ion size must be measured in Angstroms.  Offending line is: " + dblist[line])
                         basis_species[species]['radius'] = data.split()[0]
                         if len(data.split()) <= 2:
                             line = line + 1
                     if 'mole wt.' in dblist[line]:
                         data = dblist[line][dblist[line].index("mole wt.") + 8:].split("=")[1]
                         if len(data.split()) <= 1 or data.split()[1] != "g":
-                            raise ValueError("molecular weight must be measured in grams")
+                            raise ValueError("molecular weight must be measured in grams.  Offending line is: " + dblist[line])
                         basis_species[species]['molecular weight'] = data.split()[0]
                         if len(data.split()) <= 2:
                             line = line + 1
 
                     # Elements in basis species
-                    if 'elements' in dblist[line]:
+                    if 'elements in species' in dblist[line]:
                         num_elements = int(dblist[line].split()[0])
                         while len(basis_species[species]['elements']) < num_elements:
                             line = line+1
@@ -172,7 +174,7 @@ def readDatabase(dblist):
                     reading_basis = False
                     line = line-1
 
-        if 'redox couples\n' in dblist[line]:
+        if 'redox couples\n' in dblist[line] and len(dblist[line].split()) == 3:
             # Read the redox couples
             line = line+1
             reading_redox = True
@@ -193,20 +195,20 @@ def readDatabase(dblist):
[…]
