+++ modules/peridynamics/src/kernels/GeneralizedPlaneStrainOffDiagHNOSPD.C
@@ -1,0 +1,193 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "GeneralizedPlaneStrainOffDiagHNOSPD.h"
+#include "MooseVariableScalar.h"
+#include "PeridynamicsMesh.h"
+#include "RankTwoTensor.h"
+#include "RankFourTensor.h"
+
+registerMooseObject("PeridynamicsApp", GeneralizedPlaneStrainOffDiagHNOSPD);
+
+defineLegacyParams(GeneralizedPlaneStrainOffDiagHNOSPD);
+
+InputParameters
+GeneralizedPlaneStrainOffDiagHNOSPD::validParams()
+{
+  InputParameters params = MechanicsBaseNOSPD::validParams();
+  params.addClassDescription(
+      "Class for calculating off-diagonal Jacobian corresponding to "
+      "coupling between displacements (or temperature) with scalar out-of-plane strain for "
+      "generalized plane strain using HNOSPD formulation");
+
+  params.addCoupledVar("scalar_out_of_plane_strain",
+                       "Scalar variable for strain in the out-of-plane direction");
+
+  return params;
+}
+
+GeneralizedPlaneStrainOffDiagHNOSPD::GeneralizedPlaneStrainOffDiagHNOSPD(
+    const InputParameters & parameters)
+  : MechanicsBaseNOSPD(parameters),
+    _scalar_out_of_plane_strain_var_num(coupledScalar("scalar_out_of_plane_strain"))
+{
+  // Consistency check
+  if (_disp_var.size() != 2)
+    mooseError("GeneralizedPlaneStrain only works for two dimensional models!");
+}
+
+void
+GeneralizedPlaneStrainOffDiagHNOSPD::computeOffDiagJacobianScalar(unsigned int jvar_num)
+{
+  if (jvar_num == _scalar_out_of_plane_strain_var_num)
+  {
+    prepare();
+
+    if (_var.number() == _disp_var[0]->number())
+      if (_use_full_jacobian)
+        computeDispFullOffDiagJacobianScalar(0, jvar_num);
+      else
+        computeDispPartialOffDiagJacobianScalar(0, jvar_num);
+    else if (_var.number() == _disp_var[1]->number())
+      if (_use_full_jacobian)
+        computeDispFullOffDiagJacobianScalar(1, jvar_num);
+      else
+        computeDispPartialOffDiagJacobianScalar(1, jvar_num);
+    else if (_temp_coupled ? _var.number() == _temp_var->number() : 0)
+      computeTempOffDiagJacobianScalar(jvar_num);
+  }
+}
+
+void
+GeneralizedPlaneStrainOffDiagHNOSPD::computeDispPartialOffDiagJacobianScalar(unsigned int component,
+                                                                             unsigned int jvar_num)
+{
+  // off-diagonal jacobian entries on the column and row corresponding to
+  // scalar_out_of_plane_strain for coupling with displacements
+
+  DenseMatrix<Number> & ken = _assembly.jacobianBlock(_var.number(), jvar_num);
+  DenseMatrix<Number> & kne = _assembly.jacobianBlock(jvar_num, _var.number());
[…]
