+++ modules/peridynamics/src/kernels/HorizonStabilizedFiniteStrainMechanicsNOSPD.C
@@ -1,0 +1,468 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "HorizonStabilizedFiniteStrainMechanicsNOSPD.h"
+#include "PeridynamicsMesh.h"
+
+registerMooseObject("PeridynamicsApp", HorizonStabilizedFiniteStrainMechanicsNOSPD);
+
+defineLegacyParams(HorizonStabilizedFiniteStrainMechanicsNOSPD);
+
+InputParameters
+HorizonStabilizedFiniteStrainMechanicsNOSPD::validParams()
+{
+  InputParameters params = MechanicsBaseNOSPD::validParams();
+  params.addClassDescription(
+      "Class for calculating residual and Jacobian for the horizon-stabilized "
+      "non-ordinary state-based peridynamic correspondence model under finite "
+      "strain assumptions");
+
+  params.addRequiredParam<unsigned int>(
+      "component",
+      "An integer corresponding to the variable this kernel acts on (0 for x, 1 for y, 2 for z)");
+
+  return params;
+}
+
+HorizonStabilizedFiniteStrainMechanicsNOSPD::HorizonStabilizedFiniteStrainMechanicsNOSPD(
+    const InputParameters & parameters)
+  : MechanicsBaseNOSPD(parameters),
+    _dgrad_old(getMaterialPropertyOld<RankTwoTensor>("deformation_gradient")),
+    _E_inc(getMaterialProperty<RankTwoTensor>("strain_increment")),
+    _R_inc(getMaterialProperty<RankTwoTensor>("rotation_increment")),
+    _component(getParam<unsigned int>("component"))
+{
+}
+
+void
+HorizonStabilizedFiniteStrainMechanicsNOSPD::computeLocalResidual()
+{
+  // For finite strain formulation, the _stress tensor gotten from material class is the
+  // Cauchy stress (Sigma). the first Piola-Kirchhoff stress (P) is then obtained as
+  // P = J * Sigma * inv(F)^T.
+  // Nodal force states are based on the first Piola-Kirchhoff stress tensors (P).
+  // i.e., T = (J * Sigma * inv(F)^T) * inv(Shape) * xi * multi.
+  // Cauchy stress is calculated as Sigma_n+1 = Sigma_n + R * (C * dt * D) * R^T
+
+  std::vector<Real> nodal_force_comp(_nnodes);
+  for (unsigned int nd = 0; nd < _nnodes; ++nd)
+    nodal_force_comp[nd] = _multi[nd] *
+                           ((_dgrad[nd].det() * _stress[nd] * _dgrad[nd].inverse().transpose()) *
+                            _shape2[nd].inverse())
+                               .row(_component) *
+                           (nd == 0 ? 1 : -1) * _origin_vec;
+
+  _local_re(0) = -(nodal_force_comp[0] - nodal_force_comp[1]) * _bond_status;
+  _local_re(1) = -_local_re(0);
+}
+
+void
+HorizonStabilizedFiniteStrainMechanicsNOSPD::computeLocalJacobian()
+{
[…]

[…]
