### 2020-03-18 41e603a
some const additions to make interfacing easier
