### 2020-03-18 5beeef5
some const additions to make interfacing easier
