+++ modules/geochemistry/src/utils/GeochemicalDatabaseReader.C
@@ -126,10 +126,10 @@ GeochemicalDatabaseReader::getElements()
 }
 
 std::map<std::string, GeochemistryBasisSpecies>
-GeochemicalDatabaseReader::getBasisSpecies(std::vector<std::string> names)
+GeochemicalDatabaseReader::getBasisSpecies(const std::vector<std::string> & names)
 {
   // Parse the basis species specified in names
-  for (auto & species : names)
+  for (const auto & species : names)
     if (_root["basis species"].isMember(species))
     {
       GeochemistryBasisSpecies dbs;
@@ -155,10 +155,10 @@ GeochemicalDatabaseReader::getBasisSpecies(std::vector<std::string> names)
 }
 
 std::map<std::string, GeochemistryEquilibriumSpecies>
-GeochemicalDatabaseReader::getEquilibriumSpecies(std::vector<std::string> names)
+GeochemicalDatabaseReader::getEquilibriumSpecies(const std::vector<std::string> & names)
 {
   // Parse the secondary species specified in names
-  for (auto & species : names)
+  for (const auto & species : names)
     if (_root["secondary species"].isMember(species))
     {
       GeochemistryEquilibriumSpecies dbs;
@@ -190,10 +190,10 @@ GeochemicalDatabaseReader::getEquilibriumSpecies(std::vector<std::string> names)
 }
 
 std::map<std::string, GeochemistryMineralSpecies>
-GeochemicalDatabaseReader::getMineralSpecies(std::vector<std::string> names)
+GeochemicalDatabaseReader::getMineralSpecies(const std::vector<std::string> & names)
 {
   // Parse the mineral species specified in names
-  for (auto & species : names)
+  for (const auto & species : names)
     if (_root["mineral species"].isMember(species))
     {
       GeochemistryMineralSpecies dbs;
@@ -224,10 +224,10 @@ GeochemicalDatabaseReader::getMineralSpecies(std::vector<std::string> names)
 }
 
 std::map<std::string, GeochemistryGasSpecies>
-GeochemicalDatabaseReader::getGasSpecies(std::vector<std::string> names)
+GeochemicalDatabaseReader::getGasSpecies(const std::vector<std::string> & names)
 {
   // Parse the gas species specified in names
-  for (auto & species : names)
+  for (const auto & species : names)
     if (_root["gas species"].isMember(species))
     {
       GeochemistryGasSpecies dbs;
@@ -276,10 +276,10 @@ GeochemicalDatabaseReader::getGasSpecies(std::vector<std::string> names)
 }
 
 std::map<std::string, GeochemistryRedoxSpecies>
-GeochemicalDatabaseReader::getRedoxSpecies(std::vector<std::string> names)
+GeochemicalDatabaseReader::getRedoxSpecies(const std::vector<std::string> & names)
 {
   // Parse the redox species specified in names
-  for (auto & species : names)
+  for (const auto & species : names)
     if (_root["redox couples"].isMember(species))
     {
       GeochemistryRedoxSpecies dbs;
@@ -311,7 +311,7 @@ GeochemicalDatabaseReader::getRedoxSpecies(std::vector<std::string> names)
 }
 
 std::map<std::string, GeochemistryOxideSpecies>
-GeochemicalDatabaseReader::getOxideSpecies(std::vector<std::string> names)
+GeochemicalDatabaseReader::getOxideSpecies(const std::vector<std::string> & names)
 {
   // Parse the oxide species specified in names
   for (auto & species : names)
@@ -373,7 +373,7 @@ GeochemicalDatabaseReader::getNeutralSpeciesActivity()
 }
 
 std::vector<std::string>
-GeochemicalDatabaseReader::equilibriumReactions(std::vector<std::string> names) const
[…]
