### 2020-03-18 2d0e44b
add initialization to avoid compiler warning/errors.
