### 2020-03-18 7f4e834
Change std::stod to MooseUtils::convert
Refs #14897
