+++ modules/geochemistry/src/utils/GeochemicalDatabaseReader.C
@@ -14,8 +14,7 @@
 #include "string"
 #include <fstream>
 
-GeochemicalDatabaseReader::GeochemicalDatabaseReader(const std::string filename)
-  : _filename(filename)
+GeochemicalDatabaseReader::GeochemicalDatabaseReader(const FileName filename) : _filename(filename)
 {
   read(_filename);
 }
@@ -51,7 +50,7 @@ GeochemicalDatabaseReader::getTemperatures()
     auto temperatures = _root["Header"]["temperatures"];
     _temperature_points.resize(temperatures.size());
     for (unsigned int i = 0; i < temperatures.size(); ++i)
-      _temperature_points[i] = std::stod(temperatures[i].asString());
+      _temperature_points[i] = MooseUtils::convert<Real>(temperatures[i].asString());
   }
 
   return _temperature_points;
@@ -66,7 +65,7 @@ GeochemicalDatabaseReader::getPressures()
     auto pressures = _root["Header"]["pressures"];
     _pressure_points.resize(pressures.size());
     for (unsigned int i = 0; i < pressures.size(); ++i)
-      _pressure_points[i] = std::stod(pressures[i].asString());
+      _pressure_points[i] = MooseUtils::convert<Real>(pressures[i].asString());
   }
 
   return _pressure_points;
@@ -81,7 +80,7 @@ GeochemicalDatabaseReader::getDebyeHuckel()
     {
       std::vector<Real> adhvals(_root["Header"]["adh"].size());
       for (unsigned int i = 0; i < _root["Header"]["adh"].size(); ++i)
-        adhvals[i] = std::stod(_root["Header"]["adh"][i].asString());
+        adhvals[i] = MooseUtils::convert<Real>(_root["Header"]["adh"][i].asString());
       _debye_huckel.adh = adhvals;
     }
 
@@ -89,7 +88,7 @@ GeochemicalDatabaseReader::getDebyeHuckel()
     {
       std::vector<Real> bdhvals(_root["Header"]["bdh"].size());
       for (unsigned int i = 0; i < _root["Header"]["bdh"].size(); ++i)
-        bdhvals[i] = std::stod(_root["Header"]["bdh"][i].asString());
+        bdhvals[i] = MooseUtils::convert<Real>(_root["Header"]["bdh"][i].asString());
       _debye_huckel.bdh = bdhvals;
     }
 
@@ -97,7 +96,7 @@ GeochemicalDatabaseReader::getDebyeHuckel()
     {
       std::vector<Real> bdotvals(_root["Header"]["bdot"].size());
       for (unsigned int i = 0; i < _root["Header"]["bdot"].size(); ++i)
-        bdotvals[i] = std::stod(_root["Header"]["bdot"][i].asString());
+        bdotvals[i] = MooseUtils::convert<Real>(_root["Header"]["bdot"][i].asString());
       _debye_huckel.bdot = bdotvals;
     }
   }
@@ -118,7 +117,7 @@ GeochemicalDatabaseReader::getElements()
     {
       _elements[el].name = _root["elements"][el]["name"].asString();
       _elements[el].molecular_weight =
-          std::stod(_root["elements"][el]["molecular weight"].asString());
+          MooseUtils::convert<Real>(_root["elements"][el]["molecular weight"].asString());
     }
   }
 
@@ -136,13 +135,14 @@ GeochemicalDatabaseReader::getBasisSpecies(const std::vector<std::string> & name
 
       auto basis_species = _root["basis species"][species];
       dbs.name = species;
-      dbs.radius = std::stod(basis_species["radius"].asString());
-      dbs.charge = std::stod(basis_species["charge"].asString());
[…]
