### 2020-03-18 2fd0669
methods added to GeochemicalDatabaseReader that allow querying of internal datastructures.  stuff added to data/moose_testdb.json to enable more varied error checking
