+++ modules/geochemistry/python/tests/test_gwbreader.py
@@ -166,6 +166,8 @@ class TestGWBReader(unittest.TestCase):
                            'Pcrit': '46.0',
                            'Tcrit': '190.4',
                            'omega': '.011',
+                           'a' : '0.0',
+                           'b' : '0.0',
                            'logk': ['-2.6487', '-2.8202', '-2.9329', '-2.9446',
                                     '-2.9163', '-2.7253', '-2.4643', '-2.1569']},
                 'N2(g)': {'species': {'N2(aq)': '1.000'},
@@ -173,6 +175,8 @@ class TestGWBReader(unittest.TestCase):
                           'Pcrit': '33.9',
                           'Tcrit': '126.2',
                           'omega': '.039',
+                           'a' : '0.0',
+                           'b' : '0.0',
                           'logk': ['-2.9620', '-3.1848', '-3.3320', '-3.2902',
                                    '-3.1631', '-2.9499', '-2.7827', '-2.3699']}}
 
