### 2020-03-18 862e2e9
New clang format
refs #14260
