### 2020-03-18 c0ad81a
Add copy constructor from another Query
refs #14260
