### 2020-03-18 d80dc44
Removing legacy param stuff from poly chaos objects #14710
