+++ modules/stochastic_tools/include/vectorpostprocessors/PolyChaosStatistics.h
@@ -15,9 +15,6 @@
 
 class PolyChaosStatistics;
 
-template <>
-InputParameters validParams<PolyChaosStatistics>();
-
 class PolyChaosStatistics : public GeneralVectorPostprocessor
 {
 public:
