### 2020-03-18 750bf3f
Removing legacy param stuff from poly chaos objects #14710
