### 2020-03-18 cfce2ac
Explicitly cast bool to Real (#14918)
