+++ modules/stochastic_tools/src/vectorpostprocessors/PolynomialChaosData.C
@@ -8,12 +8,12 @@
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
 // Stocastic Tools Includes
-#include "PolyChaosData.h"
+#include "PolynomialChaosData.h"
 
-registerMooseObject("StochasticToolsApp", PolyChaosData);
+registerMooseObject("StochasticToolsApp", PolynomialChaosData);
 
 InputParameters
-PolyChaosData::validParams()
+PolynomialChaosData::validParams()
 {
   InputParameters params = GeneralVectorPostprocessor::validParams();
   params.addClassDescription("Tool for extracting data from polynomial chaos user object and "
@@ -22,7 +22,7 @@ PolyChaosData::validParams()
   return params;
 }
 
-PolyChaosData::PolyChaosData(const InputParameters & parameters)
+PolynomialChaosData::PolynomialChaosData(const InputParameters & parameters)
   : GeneralVectorPostprocessor(parameters),
     _pc_uo(getUserObject<PolynomialChaos>("pc_name")),
     _coeff(_pc_uo.getCoefficients()),
@@ -32,7 +32,7 @@ PolyChaosData::PolyChaosData(const InputParameters & parameters)
 }
 
 void
-PolyChaosData::initialize()
+PolynomialChaosData::initialize()
 {
   // This needs to be put in initialize since the user object doesn't know the
   // number of parameters until after initialSetup.
@@ -44,7 +44,7 @@ PolyChaosData::initialize()
 }
 
 void
-PolyChaosData::execute()
+PolynomialChaosData::execute()
 {
   _coeff_vector.clear();
   _coeff_vector = _coeff;
