+++ modules/stochastic_tools/doc/content/source/vectorpostprocessors/PolynomialChaosData.md
@@ -1,0 +1,1 @@
+!template load file=stubs/moose_object.md.template name=PolynomialChaosData syntax=/VectorPostprocessors/PolynomialChaosData

[…]
