### 2020-03-18 cbe23bd
Implement nodeToElemPtrMap() to replace nodeToElemMap()
refs #14920
