+++ conda/petsc/build.sh
@@ -1,0 +1,130 @@
+#!/bin/bash
+set -eu
+export PETSC_DIR=$SRC_DIR
+export PETSC_ARCH=arch-conda-c-opt
+
+if [[ $mpi == "openmpi" ]]; then
+  export OMPI_MCA_plm=isolated
+  export OMPI_MCA_rmaps_base_oversubscribe=yes
+  export OMPI_MCA_btl_vader_single_copy_mechanism=none
+elif [[ $mpi == "moose-mpich" ]]; then
+  export HYDRA_LAUNCHER=fork
+fi
+
+unset CFLAGS CPPFLAGS CXXFLAGS FFLAGS LIBS
+if [[ $(uname) == Darwin ]]; then
+    export LDFLAGS="${LDFLAGS:-} -Wl,-headerpad_max_install_names"
+    ADDITIONAL_ARGS="--with-blas-lib=libblas${SHLIB_EXT} --with-lapack-lib=liblapack${SHLIB_EXT}"
+else
+    ADDITIONAL_ARGS="--download-fblaslapack=1"
+fi
+
+if [[ $(uname) == Darwin ]]; then
+    TUNING="-march=core2 -mtune=haswell"
+else
+    TUNING="-march=nocona -mtune=haswell"
+fi
+
+# for MPI discovery
+export C_INCLUDE_PATH=$PREFIX/include
+export CPLUS_INCLUDE_PATH=$PREFIX/include
+export FPATH_INCLUDE_PATH=$PREFIX/include
+
+BUILD_CONFIG=`cat <<"EOF"
+  --COPTFLAGS=-O3 \
+  --CXXOPTFLAGS=-O3 \
+  --FOPTFLAGS=-O3 \
+  --with-x=0 \
+  --with-mpi=1 \
+  --with-ssl=0 \
+  --with-openmp=1 \
+  --with-debugging=0 \
+  --with-cxx-dialect=C++11 \
+  --with-shared-libraries=1 \
+  --download-mumps=1 \
+  --download-hypre=1 \
+  --download-metis=1 \
+  --download-slepc=1 \
+  --download-ptscotch=1 \
+  --download-parmetis=1 \
+  --download-scalapack=1 \
+  --download-superlu_dist=1
+EOF
+`
+
+python ./configure ${BUILD_CONFIG} ${ADDITIONAL_ARGS:-} \
+       AR="${AR:-ar}" \
+       CC="mpicc" \
+       CXX="mpicxx" \
+       FC="mpifort" \
+       F90="mpifort" \
+       F77="mpifort" \
+       CFLAGS="${TUNING}" \
+       CXXFLAGS="${TUNING}" \
+       LDFLAGS="${LDFLAGS:-}" \
+       --prefix=$PREFIX || (cat configure.log && exit 1)
+
+# Verify that gcc_ext isn't linked
+for f in $PETSC_ARCH/lib/petsc/conf/petscvariables $PETSC_ARCH/lib/pkgconfig/PETSc.pc; do
+  if grep gcc_ext $f; then
+    echo "gcc_ext found in $f"
+    exit 1
+  fi
+done
+
+sedinplace() {
+  if [[ $(uname) == Darwin ]]; then
+    sed -i "" "$@"
+  else
+    sed -i"" "$@"
+  fi
+}
+
+# Remove abspath of ${BUILD_PREFIX}/bin/python
+sedinplace "s%${BUILD_PREFIX}/bin/python%python%g" $PETSC_ARCH/include/petscconf.h
+sedinplace "s%${BUILD_PREFIX}/bin/python%python%g" $PETSC_ARCH/lib/petsc/conf/petscvariables
[…]
