+++ conda/petsc/run_test.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+set -e
+
+export PETSC_DIR=${PREFIX}
+
+pkg-config --cflags PETSc | grep -v isystem
