+++ conda/petsc/conda_build_config.yaml
@@ -1,0 +1,35 @@
+CONDA_BUILD_SYSROOT:           # [osx]
+  - /opt/MacOSX10.9.sdk        # [osx]
+
+macos_min_version:             # [osx]
+  - 10.9                       # [osx]
+
+macos_machine:                 # [osx]
+  - x86_64-apple-darwin13.4.0  # [osx]
+
+MACOSX_DEPLOYMENT_TARGET:      # [osx]
+  - 10.9                       # [osx]
+
+mpi:
+  - moose-mpich
+
+moose_mpich:
+  - 3.3.2
+
+libblas:
+  - 3.8.0 15_openblas # [osx]
+
+libcblas:
+  - 3.8.0 15_openblas # [osx]
+
+liblapack:
+  - 3.8.0 15_openblas # [osx]
+
+libopenblas:
+  - 0.3.8 h3d69b6c_0  # [osx]
+
+pin_run_as_build:
+  moose_mpich: x.x.x
+  libblas: x.x.x
+  libcblas: x.x.x
+  liblapack: x.x.x

[…]
