### 2020-03-18 69e81e2
prepare_for_use in BreakMeshByBlockGenerator
There's no guarantee that we're going to prepare later. During
construction, MooseMesh._needs_prepare_for_use is initialized to
`false`; I think this is a fair approach...if at any point during the
MeshGeneration phase someone does something that needs mesh preparation,
they should call `prepare_for_use` on the spot. In that vein, we call
`prepare_for_use` here because we've done an operation that requires
updating the parallel counts.
