### 2020-03-18 6ec8105
Fix documentation: equation for the moment about local Z axis in the beam element. Refs. #14772
