### 2020-03-18 c940d50
Adding Smolyak sparse grid #14833
