### 2020-03-18 b01b36d
Change a couple of signs in the beam element internal force calculation for readability.
