### 2020-03-18 da274e8
Documenting sparse grids #14833
