### 2020-03-18 f84b62b
Apply clang format.
