### 2020-03-19 56ae579
more methods added to GeochemicalDatabaseReader that i am using in other code.  More tests added to units tests for these methods and others i added yesterday.  moose_testdb.json extended to allow further testing, etc
