+++ modules/geochemistry/src/utils/GeochemicalDatabaseReader.C
@@ -371,6 +371,38 @@ GeochemicalDatabaseReader::getSurfaceSpecies(const std::vector<std::string> & na
   return _surface_species;
 }
 
+std::map<std::string, GeochemistrySorbingMinerals>
+GeochemicalDatabaseReader::getSorbingMinerals(const std::vector<std::string> & names)
+{
+  // Parse the sorbing minerals specified in names
+  for (const auto & species : names)
+    if (_root["sorbing minerals"].isMember(species))
+    {
+      GeochemistrySorbingMinerals sm;
+
+      auto sorbing_mineral = _root["sorbing minerals"][species];
+      sm.name = species;
+      sm.surface_area = MooseUtils::convert<Real>(sorbing_mineral["surface area"].asString());
+
+      std::map<std::string, Real> basis_species;
+      unsigned ind = 0;
+      for (auto & site : sorbing_mineral["sorbing site"])
+      {
+        basis_species[site.asString()] =
+            MooseUtils::convert<Real>(sorbing_mineral["site density"][ind].asString());
+        ind += 1;
+      }
+
+      sm.basis_species = basis_species;
+
+      _sorbing_minerals[species] = sm;
+    }
+    else
+      mooseError(species + " does not exist in database " + _filename);
+
+  return _sorbing_minerals;
+}
+
 std::map<std::string, GeochemistryNeutralSpeciesActivity>
 GeochemicalDatabaseReader::getNeutralSpeciesActivity()
 {
@@ -608,4 +640,24 @@ bool
 GeochemicalDatabaseReader::isRedoxSpecies(const std::string & name) const
 {
   return _root["redox couples"].isMember(name);
+}
+
+bool
+GeochemicalDatabaseReader::isSorbingMineral(const std::string & name) const
+{
+  return _root["sorbing minerals"].isMember(name);
+}
+
+std::vector<std::string>
+GeochemicalDatabaseReader::secondarySpeciesNames() const
+{
+  std::vector<std::string> names(_root["secondary species"].getMemberNames());
+  return names;
+}
+
+std::vector<std::string>
+GeochemicalDatabaseReader::redoxCoupleNames() const
+{
+  std::vector<std::string> names(_root["redox couples"].getMemberNames());
+  return names;
 }
