### 2020-03-19 bc6292b
Method to print out species information
