### 2020-03-19 f1c2e48
Method to print out species information
