### 2020-03-19 55642c3
Unit test more methods in GeochemicalDatabaseReader
Refs #14897
