+++ modules/geochemistry/unit/src/GeochemicalDatabaseReaderTest.C
@@ -480,6 +480,42 @@ TEST(GeochemicalDatabaseReaderTest, isBasisSpecies)
   EXPECT_FALSE(database.isBasisSpecies(">(s)FeO-"));
 }
 
+TEST(GeochemicalDatabaseReaderTest, isSecondarySpecies)
+{
+  GeochemicalDatabaseReader database("data/moose_testdb.json");
+
+  EXPECT_FALSE(database.isSecondarySpecies("Ca++"));
+  EXPECT_FALSE(database.isSecondarySpecies("H2O"));
+  EXPECT_FALSE(database.isSecondarySpecies("Ag"));
+  EXPECT_TRUE(database.isSecondarySpecies("CO2(aq)"));
+  EXPECT_TRUE(database.isSecondarySpecies("CO3--"));
+  EXPECT_FALSE(database.isSecondarySpecies("Calcite"));
+  EXPECT_FALSE(database.isSecondarySpecies("Fe(OH)3(ppd)"));
+  EXPECT_FALSE(database.isSecondarySpecies("CH4(g)"));
+  EXPECT_FALSE(database.isSecondarySpecies("(O-phth)--"));
+  EXPECT_FALSE(database.isSecondarySpecies("Fe+++"));
+  EXPECT_FALSE(database.isSecondarySpecies("Cu2O"));
+  EXPECT_FALSE(database.isSecondarySpecies(">(s)FeO-"));
+}
+
+TEST(GeochemicalDatabaseReaderTest, isMineralSpecies)
+{
+  GeochemicalDatabaseReader database("data/moose_testdb.json");
+
+  EXPECT_FALSE(database.isMineralSpecies("Ca++"));
+  EXPECT_FALSE(database.isMineralSpecies("H2O"));
+  EXPECT_FALSE(database.isMineralSpecies("Ag"));
+  EXPECT_FALSE(database.isMineralSpecies("CO2(aq)"));
+  EXPECT_FALSE(database.isMineralSpecies("CO3--"));
+  EXPECT_TRUE(database.isMineralSpecies("Calcite"));
+  EXPECT_TRUE(database.isMineralSpecies("Fe(OH)3(ppd)"));
+  EXPECT_FALSE(database.isMineralSpecies("CH4(g)"));
+  EXPECT_FALSE(database.isMineralSpecies("(O-phth)--"));
+  EXPECT_FALSE(database.isMineralSpecies("Fe+++"));
+  EXPECT_FALSE(database.isMineralSpecies("Cu2O"));
+  EXPECT_FALSE(database.isMineralSpecies(">(s)FeO-"));
+}
+
 TEST(GeochemicalDatabaseReaderTest, isRedoxSpecies)
 {
   GeochemicalDatabaseReader database("data/moose_testdb.json");
@@ -498,6 +534,24 @@ TEST(GeochemicalDatabaseReaderTest, isRedoxSpecies)
   EXPECT_FALSE(database.isRedoxSpecies(">(s)FeO-"));
 }
 
+TEST(GeochemicalDatabaseReaderTest, isGasSpecies)
+{
+  GeochemicalDatabaseReader database("data/moose_testdb.json");
+
+  EXPECT_FALSE(database.isGasSpecies("Ca++"));
+  EXPECT_FALSE(database.isGasSpecies("H2O"));
+  EXPECT_FALSE(database.isGasSpecies("Ag"));
+  EXPECT_FALSE(database.isGasSpecies("CO2(aq)"));
+  EXPECT_FALSE(database.isGasSpecies("CO3--"));
+  EXPECT_FALSE(database.isGasSpecies("Calcite"));
[…]
