### 2020-03-19 adba20b
Update documentation to describe database format
Refs #14897

Refs #14897
