+++ modules/geochemistry/doc/content/modules/geochemistry/database.md
@@ -2,68 +2,128 @@
 
 Notation and definitions are described in [geochemistry_nomenclature.md].
 
-The `geochemistry` module works with chemical databases in a special MOOSE format, and the module contains a python file that translates between MOOSE format and other formats such as EQ3/6.  This page describes the GWB format, using a database file downloaded from the geochemist workbench site.  See the "Thermo Datasets" chapter of [!cite](gwb_reference) and [https://academy.gwb.com](https://academy.gwb.com)
+The `geochemistry` module works with chemical databases in a simple [JSON](https://www.json.org) format,
+making it easy to parse in a variety of applications or languages.
 
-This page goes through the database file section-by-section.
+## Python database converter
 
-## Descriptive header
-
-The database file begins with a descriptive header:
-
-```
-dataset of thermodynamic data for gwb programs
-dataset format: jan19
-activity model: debye-huckel
-fugacity model: tsonopoulos
-```
-
-!alert note 
-Only the [Debye-Huckel activity model](activity_coefficients.md) model is currently supported.
-Only the [Spycher-Reed fugacity model](fugacity.md) model is currently supported.
-
-## Temperature definition
-
-All equilibrium coefficients, activity coefficients, etc, are evaluated at a set of temperatures.  These are assumed to be in $^{\circ}$C, and defined by the line(s)
+A python utility The module contains a python utility that translates between the MOOSE format and other formats such as EQ3/6 or [GWB](gwb_database.md). A database in either of these formats can be translated to a MOOSE JSON database using
 
 ```
-* temperatures
-         0.0000     25.0000     60.0000    100.0000
-       150.0000    200.0000    250.0000    300.0000
+geochemistry/python/database_converter.py -i gwb_database --format gwb -o moose_database.json
 ```
 
-If a numerical simulation is performed at a temperature that is not one of these 8 values, a fourth-order polynomial fit to the data is used.
+for example. Using the supplied python conversion utility ensures that the database produced adheres to the required
+format.
 
-## Steam saturation curve
-
-The pressures --- in bars (1 bar $= 10^{5}\,$Pa) --- along the steam saturation curve, at the temperatures defined above are given in the form
-
-```
-* pressures
-         1.0134      1.0134      1.0134      1.0134
-         4.7600     15.5490     39.7760     85.9270
-```
-
-## Debye-Huckel coefficients
-
-The [Debye-Huckel coefficients](activity_coefficients.md) are given next.  These are evaluated at the temperatures defined above
+A full list of the functionality can be obtained using the `--help` flag
 
 ```
-* debye huckel a (adh)
-          .4913       .5092       .5450       .5998
-          .6898       .8099       .9785      1.2555
-* debye huckel b (bdh)
-          .3247       .3283       .3343       .3422
-          .3533       .3655       .3792       .3965
-* bdot
-          .0174       .0410       .0440       .0460
-          .0470       .0470       .0340      0.0000
+geochemistry/python/database_converter.py --help
+usage: database_converter.py [-h] -i INPUT --format {gwb,eq36} [-o OUTPUT]
+
+Utility to read data from thermodynamic database and write to MOOSE
+thermodynamic database (in JSON format)
+
+optional arguments:
+  -h, --help            show this help message and exit
+  -i INPUT, --input INPUT
[…]
