+++ modules/phase_field/unit/Makefile
@@ -2,15 +2,15 @@
 ################### MOOSE Application Standard Makefile #######################
 ###############################################################################
 #
-# Required Environment variables (one of the following)
-# PACKAGES_DIR  - Location of the MOOSE redistributable package
-#
 # Optional Environment variables
-# MOOSE_DIR     - Root directory of the MOOSE project
-# FRAMEWORK_DIR - Location of the MOOSE framework
+# MOOSE_DIR        - Root directory of the MOOSE project
+# MODULE_DIR       - Location of the MOOSE modules directory
+# FRAMEWORK_DIR    - Location of the MOOSE framework
 #
 ###############################################################################
-MOOSE_DIR          ?= $(shell dirname `pwd`)
+THIS_MODULE_DIR    ?= $(shell dirname `pwd`)
+MODULE_DIR         ?= $(shell dirname $(THIS_MODULE_DIR))
+MOOSE_DIR          ?= $(shell dirname $(MODULE_DIR))
 FRAMEWORK_DIR      ?= $(MOOSE_DIR)/framework
 ADDITIONAL_CPPFLAGS += -Wall -Wextra
 ###############################################################################
@@ -24,23 +24,21 @@ include $(FRAMEWORK_DIR)/build.mk
 include $(FRAMEWORK_DIR)/moose.mk
 
 ################################## MODULES ####################################
-XFEM               := yes
-CHEMICAL_REACTIONS := yes
-RDG                := yes
+PHASE_FIELD := yes
 include           $(MOOSE_DIR)/modules/modules.mk
 ###############################################################################
 
-APPLICATION_DIR  := $(MOOSE_DIR)/unit
-APPLICATION_NAME := moose-unit
+APPLICATION_DIR  := $(shell pwd)
+APPLICATION_NAME := phase_field-unit
 BUILD_EXEC       := yes
-app_BASE_DIR     :=      # Intentionally blank
+app_BASE_DIR     :=
 DEP_APPS    ?= $(shell $(FRAMEWORK_DIR)/scripts/find_dep_apps.py $(APPLICATION_NAME))
 include $(FRAMEWORK_DIR)/app.mk
 
 # Find all the MOOSE unit test source files and include their dependencies.
-moose_unit_srcfiles := $(shell find $(MOOSE_DIR)/unit/src -name "*.C")
-moose_unit_deps := $(patsubst %.C, %.$(obj-suffix).d, $(moose_unit_srcfiles))
--include $(moose_unit_deps)
+stochastic_tools_unit_srcfiles := $(shell find $(APPLICATION_DIR)/src -name "*.C")
+stochastic_tools_unit_deps := $(patsubst %.C, %.$(obj-suffix).d, $(stochastic_tools_unit_srcfiles))
+-include $(stochastic_tools_unit_deps)
 
 ###############################################################################
 # Additional special case targets should be added here
