### 2020-03-19 2a2765a
Move XFEM unit tests
(refs #14927)
