### 2020-03-19 50f384b
Move tensor mechanics unit tests and remove solid mechanics unit tests
(refs #14927)
