+++ modules/tensor_mechanics/unit/include/gtest_include.h
@@ -1,0 +1,14 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#pragma once
+
+#include "libmesh/ignore_warnings.h"
+#include "gtest/gtest.h"
+#include "libmesh/restore_warnings.h"
