### 2020-03-19 9fbf40a
Move chemical reactions unit tests
(refs #14927)
