### 2020-03-19 fd54c64
Complete unit test update and optional running from run_tests
(refs #14927)
