+++ unit/Makefile
@@ -24,7 +24,6 @@ include $(FRAMEWORK_DIR)/build.mk
 include $(FRAMEWORK_DIR)/moose.mk
 
 ################################## MODULES ####################################
-RDG                := yes
 include           $(MOOSE_DIR)/modules/modules.mk
 ###############################################################################
 
