### 2020-03-19 e9847d0
Explicitly set `skip_find_neighbors=false`
libmesh/libmesh#2473 will hopefully add an allow_find_neighbors
API that has the same structure as allow_renumbering. However,
in for it to work existing users of `prepare_for_use` need to
explicitly state whether they want to `find_neighbors` or not
