+++ framework/src/mesh/RinglebMesh.C
@@ -210,7 +210,7 @@ RinglebMesh::buildMesh()
   }
 
   /// Find neighbors, etc.
-  mesh.prepare_for_use();
+  mesh.prepare_for_use(false, false);
 
   /// Create the triangular elements if required by the user
   if (_triangles)
