+++ framework/src/mesh/MooseMesh.C
@@ -365,7 +365,7 @@ MooseMesh::prepare(bool force)
     {
       CONSOLE_TIMED_PRINT("Preparing for use");
 
-      getMesh().prepare_for_use();
+      getMesh().prepare_for_use(false, false);
     }
   }
   else
@@ -375,7 +375,7 @@ MooseMesh::prepare(bool force)
     // Call prepare_for_use() and DO NOT allow renumbering
     getMesh().allow_renumbering(false);
     if (force || _needs_prepare_for_use)
-      getMesh().prepare_for_use();
+      getMesh().prepare_for_use(false, false);
   }
 
   // Collect (local) subdomain IDs
