+++ framework/src/meshgenerators/ConcentricCircleMeshGenerator.C
@@ -750,8 +750,8 @@ ConcentricCircleMeshGenerator::generate()
       MeshTools::Modification::change_boundary_id(other_mesh, 5, 2);
       MeshTools::Modification::change_boundary_id(other_mesh, 6, 3);
       MeshTools::Modification::change_boundary_id(other_mesh, 7, 4);
-      mesh->prepare_for_use(false);
-      other_mesh.prepare_for_use(false);
+      mesh->prepare_for_use(false, false);
+      other_mesh.prepare_for_use(false, false);
       mesh->stitch_meshes(other_mesh, 1, 3, TOLERANCE, true);
       mesh->get_boundary_info().sideset_name(1) = "left";
       mesh->get_boundary_info().sideset_name(2) = "bottom";
@@ -763,8 +763,8 @@ ConcentricCircleMeshGenerator::generate()
       MeshTools::Modification::change_boundary_id(other_mesh, 1, 5);
       MeshTools::Modification::change_boundary_id(other_mesh, 2, 1);
       MeshTools::Modification::change_boundary_id(other_mesh, 5, 2);
-      mesh->prepare_for_use(false);
-      other_mesh.prepare_for_use(false);
+      mesh->prepare_for_use(false, false);
+      other_mesh.prepare_for_use(false, false);
       mesh->stitch_meshes(other_mesh, 1, 1, TOLERANCE, true);
 
       MeshTools::Modification::change_boundary_id(*mesh, 2, 1);
@@ -789,8 +789,8 @@ ConcentricCircleMeshGenerator::generate()
       MeshTools::Modification::change_boundary_id(other_mesh, 5, 4);
       MeshTools::Modification::change_boundary_id(other_mesh, 6, 1);
       MeshTools::Modification::change_boundary_id(other_mesh, 7, 2);
-      mesh->prepare_for_use(false);
-      other_mesh.prepare_for_use(false);
+      mesh->prepare_for_use(false, false);
+      other_mesh.prepare_for_use(false, false);
       mesh->stitch_meshes(other_mesh, 2, 4, TOLERANCE, true);
       mesh->get_boundary_info().sideset_name(1) = "left";
       mesh->get_boundary_info().sideset_name(2) = "bottom";
@@ -802,8 +802,8 @@ ConcentricCircleMeshGenerator::generate()
       MeshTools::Modification::change_boundary_id(other_mesh, 1, 5);
       MeshTools::Modification::change_boundary_id(other_mesh, 2, 1);
       MeshTools::Modification::change_boundary_id(other_mesh, 5, 2);
-      mesh->prepare_for_use(false);
-      other_mesh.prepare_for_use(false);
+      mesh->prepare_for_use(false, false);
+      other_mesh.prepare_for_use(false, false);
       mesh->stitch_meshes(other_mesh, 2, 2, TOLERANCE, true);
 
       MeshTools::Modification::change_boundary_id(*mesh, 3, 2);
@@ -837,8 +837,8 @@ ConcentricCircleMeshGenerator::generate()
       MeshTools::Modification::change_boundary_id(*mesh, 5, 3);
       MeshTools::Modification::change_boundary_id(*mesh, 6, 4);
       MeshTools::Modification::change_boundary_id(*mesh, 7, 1);
-      mesh->prepare_for_use(false);
-      other_mesh.prepare_for_use(false);
+      mesh->prepare_for_use(false, false);
+      other_mesh.prepare_for_use(false, false);
       mesh->stitch_meshes(other_mesh, 4, 2, TOLERANCE, true);
       mesh->get_boundary_info().sideset_name(1) = "left";
       mesh->get_boundary_info().sideset_name(2) = "bottom";
[…]
