+++ modules/tensor_mechanics/src/actions/TensorMechanicsActionBase.C
@@ -42,7 +42,7 @@ TensorMechanicsActionBase::validParams()
 
   params.addRequiredParam<std::vector<VariableName>>(
       "displacements", "The nonlinear displacement variables for the problem");
-  params.addParam<VariableName>("temperature", "The temperature");
+  params.addParam<std::vector<VariableName>>("temperature", "The temperature");
 
   MooseEnum strainType("SMALL FINITE", "SMALL");
   params.addParam<MooseEnum>("strain", strainType, "Strain formulation");
