+++ modules/tensor_mechanics/src/kernels/StressDivergenceTensors.C
@@ -37,10 +37,15 @@ StressDivergenceTensors::validParams()
                        "The name of the temperature variable used in the "
                        "ComputeThermalExpansionEigenstrain.  (Not required for "
                        "simulations without temperature coupling.)");
-  params.addParam<std::string>(
+  params.addDeprecatedParam<MaterialPropertyName>(
       "thermal_eigenstrain_name",
       "thermal_eigenstrain",
-      "The eigenstrain_name used in the ComputeThermalExpansionEigenstrain.");
+      "The eigenstrain_name used in the ComputeThermalExpansionEigenstrain.",
+      "Please provide a list of all eigenstrains in 'eigenstrain_names' instead");
+  params.addParam<std::vector<MaterialPropertyName>>(
+      "eigenstrain_names",
+      "List of eigenstrains used in the strain calculation. Used for computing their derivaties "
+      "for off-diagonal Jacobian terms.");
   params.addCoupledVar("out_of_plane_strain",
                        "The name of the out_of_plane_strain variable used in the "
                        "WeakPlaneStress kernel.");
@@ -70,10 +75,6 @@ StressDivergenceTensors::StressDivergenceTensors(const InputParameters & paramet
     _disp_var(_ndisp),
     _temp_coupled(isCoupled("temperature")),
     _temp_var(_temp_coupled ? coupled("temperature") : 0),
-    _deigenstrain_dT(_temp_coupled ? &getMaterialPropertyDerivative<RankTwoTensor>(
-                                         getParam<std::string>("thermal_eigenstrain_name"),
-                                         getVar("temperature", 0)->name())
-                                   : nullptr),
     _out_of_plane_strain_coupled(isCoupled("out_of_plane_strain")),
     _out_of_plane_strain(_out_of_plane_strain_coupled ? &coupledValue("out_of_plane_strain")
                                                       : nullptr),
@@ -86,6 +87,27 @@ StressDivergenceTensors::StressDivergenceTensors(const InputParameters & paramet
 {
   for (unsigned int i = 0; i < _ndisp; ++i)
     _disp_var[i] = coupled("displacements", i);
+
+  if (_temp_coupled)
+  {
+    for (auto eigenstrain_name : getParam<std::vector<MaterialPropertyName>>("eigenstrain_names"))
+      _deigenstrain_dT.push_back(&getMaterialPropertyDerivative<RankTwoTensor>(
+          eigenstrain_name, getVar("temperature", 0)->name()));
+
+    // Handle the deprecated 'thermal_eigenstrain_name' parameter: delete this code when removed
+    if (_deigenstrain_dT.size() == 0)
+    {
+      const auto thermal_eigenstrain_name =
+          getParam<MaterialPropertyName>("thermal_eigenstrain_name");
+      _deigenstrain_dT.push_back(&getMaterialPropertyDerivative<RankTwoTensor>(
+          thermal_eigenstrain_name, getVar("temperature", 0)->name()));
+    }
+    else
+    {
+      if (parameters.isParamSetByUser("thermal_eigenstrain_name"))
+        mooseError("Cannot specify both 'thermal_eigenstrain_name' and 'eigenstrain_names'");
+    }
+  }
 
   // Checking for consistency between mesh size and length of the provided displacements vector
   if (_out_of_plane_direction != 2 && _ndisp != 3)
@@ -340,8 +362,14 @@ StressDivergenceTensors::computeQpOffDiagJacobian(unsigned int jvar)
 
   // off-diagonal Jacobian with respect to a coupled temperature variable
[…]
