+++ modules/tensor_mechanics/src/kernels/StressDivergenceRZTensors.C
@@ -94,6 +94,10 @@ StressDivergenceRZTensors::computeQpOffDiagJacobian(unsigned int jvar)
 
   if (_temp_coupled && jvar == _temp_var)
   {
+    RankTwoTensor total_deigenstrain_dT;
+    for (const auto deigenstrain_dT : _deigenstrain_dT)
+      total_deigenstrain_dT += (*deigenstrain_dT)[_qp];
+
     Real jac = 0.0;
     if (_component == 0)
     {
@@ -102,7 +106,7 @@ StressDivergenceRZTensors::computeQpOffDiagJacobian(unsigned int jvar)
           jac -= (_grad_test[_i][_qp](0) * _Jacobian_mult[_qp](0, 0, k, l) +
                   _test[_i][_qp] / _q_point[_qp](0) * _Jacobian_mult[_qp](2, 2, k, l) +
                   _grad_test[_i][_qp](1) * _Jacobian_mult[_qp](0, 1, k, l)) *
-                 (*_deigenstrain_dT)[_qp](k, l);
+                 total_deigenstrain_dT(k, l);
       return jac * _phi[_j][_qp];
     }
     else if (_component == 1)
@@ -111,7 +115,7 @@ StressDivergenceRZTensors::computeQpOffDiagJacobian(unsigned int jvar)
         for (unsigned l = 0; l < LIBMESH_DIM; ++l)
           jac -= (_grad_test[_i][_qp](1) * _Jacobian_mult[_qp](1, 1, k, l) +
                   _grad_test[_i][_qp](0) * _Jacobian_mult[_qp](1, 0, k, l)) *
-                 (*_deigenstrain_dT)[_qp](k, l);
+                 total_deigenstrain_dT(k, l);
       return jac * _phi[_j][_qp];
     }
   }
