### 2020-03-19 1454b2c
surfaceSpeciesNames method added to GeochemicalDatabaseReader
