+++ modules/xfem/doc/content/syntax/XFEM/index.md
@@ -1,1 +1,78 @@
-!template load file=stubs/moose_system.md.template name=XFEM syntax=/XFEM
+# XFEM Action System
+
+## Overview
+
+The XFEM Action is an object required when running MOOSE simulations using XFEM. It provides
+general XFEM solution parameters to the `XFEMApp` class. It has several functions:
+
+- Provide `qrule` variable to `XFEMApp` to determine which quadrature weighting rule to use.
+- Specify `output_cut_plane` variable to output cut elements to the exodus mesh.
+- Optional: Provide parameter values for prescribed, incremental crack growth.
+- Optional: Setup of near tip enrichment parameters.
+- Optional: Allow users to control the amount of debug information printed during a simulation.
+
+## Constructed MooseObjects
+
+Three of the above functions pass parameters to `XFEMApp` for use in setting up relevant options and
+problems: the `qrule` variable, incremental crack growth, and `debug_output_level`. The remaining
+functions have components that replace specific MooseObjects to make input files more concise.
+The following table details the types of classes replaced by utilizing the `output_cut_plane`
+option in the `XFEMAction`.
+
+!table id=incr_crack_xfem_action_table caption=Correspondence Among Action Functionality and MooseObjects for `XFEMAction` when `output_cut_plane` = `true`.
+| Functionality     | Replaced Classes   | Associated Parameters   |
+|-------------------|--------------------|-------------------------|
+| Add `AuxVariables` for various cut parameters and volume fraction calculations | [AuxVariables](/AuxVariables/index.md) |   |
+| Add `AuxKernels` for various cut parameters and volume fraction calculations | [XFEMVolFracAux](/XFEMVolFracAux.md) [XFEMCutPlaneAux](/XFEMCutPlaneAux.md) |   |
+
+Several objects are instantiated when using the `output_cut_plane` option. These `AuxVariables` and
+`AuxKernels` store data defining the XFEM cut: element volume fraction as well as definitions for
+two planes (their x, y, z origins and normals). This allows the cut location to be tracked
+throughout the simulation and saves the user from having to manually set up each `AuxVariable` and
+its associated `AuxKernel` (`XFEMVolFracAux` for the volume fraction and `XFEMCutPlaneAux` for cut
+plane AuxVariables).
+
+Next, the class types replaced when using crack tip enrichment are detailed:
+
+!table id=near_tip_xfem_action_table caption=Correspondence Among Action Functionality and MooseObjects for `XFEMAction` when `use_crack_tip_enrichment` = `true`.
+| Functionality     | Replaced Classes   | Associated Parameters   |
+|-------------------|--------------------|-------------------------|
+| Add enrichment displacement variables | [Variables](/syntax/Variables/index.md) | `enrichment_displacements`: vector of enrichment displacement names |
+| Add stress divergence tensor kernels for crack tip enrichment | [CrackTipEnrichmentStressDivergenceTensors](/CrackTipEnrichmentStressDivergenceTensors.md) | `crack_front_definition`, `enrichment_displacements`, `displacements` |
+| Add crack tip enrichment boundary conditions | [CrackTipEnrichmentCutOffBC](/CrackTipEnrichmentCutOffBC.md) | `cut_off_boundary`, `cut_off_radius`, `crack_front_definition` |
+
+When crack tip enrichment is used, `enrichment_displacements` are added as first-order, Lagrange
+Variables. Next, a `CrackTipEnrichmentStressDivergenceTensors` kernel is added for each enrichment
+displacement variable. Each kernel uses the `crack_front_definition`, `enrichment_displacements`,
+and `diplacements` provided in the XFEM Action. All enrichment displacement variables also have an
[…]
