+++ modules/xfem/doc/content/modules/xfem/theory/theory.md
@@ -1,0 +1,7 @@
+# XFEM Module Theory
+
+## Introduction
+
+Extended finite element method (XFEM) 
+
+An example of a strong discontinuity includes a propagating crack, while a discrete material interface would be a weak discontinuity.
