+++ modules/xfem/doc/content/source/actions/XFEMAction.md
@@ -1,4 +1,4 @@
-# XFEMAction
+# XFEM Action
 
 !syntax description /XFEM/XFEMAction
 
@@ -12,15 +12,9 @@ visualized on Exodus mesh files with the `output_cut_plane` variable.
 Parameters specific to cracking application cases are also present in this action. For cracks that
 grow according to a consistent increment, `use_crack_growth_increment` and `crack_growth_increment`
 can be defined. To employ enrichments for cracks with tips ending within an element,
-`use_crack_tip_enrichment` can be set to `true` and parameters to used in the crack tip enrichment
-must be detailed: `crack_front_definition`, `displacements`, `enrichment_displacements`,
+`use_crack_tip_enrichment` can be set to `true` and parameters used in the crack tip enrichment must
+be detailed: `crack_front_definition`, `displacements`, `enrichment_displacements`,
 `cut_off_boundary`, `cut_off_radius`.
 
-!! For a more detailed description of what objects are generated by employing the various options
-available in this action, please see the syntax page.
-
-## Example Input File Syntax
-
-!listing modules/xfem/test/tests/crack_tip_enrichment/edge_crack_2d.i block=XFEM
-
-!syntax parameters /XFEM/XFEMAction
+For a more detailed description of what objects are generated by employing the various options
+available in this action, please see the [XFEM Action syntax](syntax/XFEM/index.md) page.
