+++ modules/xfem/doc/content/modules/xfem/theory/algorithms.md
@@ -1,0 +1,4 @@
+# XFEM Module Algorithms
+
+!alert construction
+The XFEM module Algorithms page needs documentation.
