### 2020-03-20 a847a4e
Rebase fixes, reviewer comments, and improve logic and testing for PolynomialChaos for replicated and distributed training data
(refs #14875)
