+++ framework/include/vectorpostprocessors/VectorPostprocessorData.h
@@ -177,19 +177,25 @@ private:
     VectorPostprocessorVectors & operator=(VectorPostprocessorVectors &&) = default;
     ///@}
 
+    /// The state (pointers to the correct data) of the vectors for the VPP object, see
+    /// the header for the declaration of the VectorPostprocessorState class
     std::vector<std::pair<std::string, VectorPostprocessorState>> _values;
 
+    /// The following flags default to false, they can each be switched to true from within the
+    /// getVectorPostprocessorHelper. Once true they remain true. When they become true depends on
+    /// the flag. The various getter methods should be inspected to see how they are set.
+
     /// Boolean indicating whether these vectors contain complete history (append mode)
-    bool _contains_complete_history;
+    bool _contains_complete_history = false;
 
     /// Boolean indicating whether the vector will already be replicated in parallel by the VPP
-    bool _is_broadcast;
+    bool _is_broadcast = false;
 
     /// Boolean indicating whether any old vectors have been requested.
-    bool _needs_old;
+    bool _needs_old = false;
 
     /// Flag if data is distributed
-    bool _is_distributed;
+    bool _is_distributed = false;
   };
 
   /// The VPP data store in a map: VPP Name to vector storage
