### 2020-03-20 3af653c
Re-factor and Fixes after rebase
(refs #14875)
