### 2020-03-20 cab6b85
Re-factor and Fixes after rebase
(refs #14875)
