### 2020-03-20 3afecc9
Adding Clenshaw-Curtis sparse grid #14833
