+++ modules/stochastic_tools/src/utils/PolynomialQuadrature.C
@@ -61,6 +61,14 @@ Polynomial::computeDerivative(const unsigned int /*order*/,
   return 0;
 }
 
+void
+Polynomial::clenshawQuadrature(const unsigned int /*order*/,
+                               std::vector<Real> & /*points*/,
+                               std::vector<Real> & /*weights*/) const
+{
+  ::mooseError("Clenshaw quadrature has not been implemented for this polynomial type.");
+}
+
 Real
 Polynomial::productIntegral(const std::vector<unsigned int> order) const
 {
@@ -143,6 +151,18 @@ Legendre::quadrature(const unsigned int order,
                      std::vector<Real> & weights) const
 {
   gauss_legendre(order, points, weights, _lower_bound, _upper_bound);
+}
+
+void
+Legendre::clenshawQuadrature(const unsigned int order,
+                             std::vector<Real> & points,
+                             std::vector<Real> & weights) const
+{
+  clenshaw_curtis(order, points, weights);
+  Real dx = (_upper_bound - _lower_bound) / 2.0;
+  Real xav = (_upper_bound + _lower_bound) / 2.0;
+  for (unsigned int i = 0; i < points.size(); ++i)
+    points[i] = points[i] * dx + xav;
 }
 
 Real
@@ -318,6 +338,45 @@ gauss_hermite(const unsigned int order,
   }
 }
 
+void
+clenshaw_curtis(const unsigned int order, std::vector<Real> & points, std::vector<Real> & weights)
+{
+  // Number of points needed
+  unsigned int N = order + (order % 2);
+  points.resize(N + 1);
+  weights.resize(N + 1);
+
+  if (N == 0)
+  {
+    points[0] = 0;
+    weights[0] = 1;
+    return;
+  }
+
+  std::vector<Real> dk(N / 2 + 1);
+  for (unsigned int k = 0; k <= (N / 2); ++k)
+    dk[k] = ((k == 0 || k == (N / 2)) ? 1.0 : 2.0) / (1.0 - 4.0 * (Real)k * (Real)k);
+
+  for (unsigned int n = 0; n <= (N / 2); ++n)
+  {
+    Real theta = (Real)n * M_PI / ((Real)N);
+    points[n] = -std::cos(theta);
+    for (unsigned int k = 0; k <= (N / 2); ++k)
+    {
+      Real Dnk =
+          ((n == 0 || n == (N / 2)) ? 0.5 : 1.0) * std::cos((Real)k * theta * 2.0) / ((Real)N);
+      weights[n] += Dnk * dk[k];
+    }
+  }
+
+  for (unsigned int n = 0; n < (N / 2); ++n)
+  {
+    points[N - n] = -points[n];
+    weights[N - n] = weights[n];
+  }
+  weights[N / 2] *= 2.0;
+}
+
 TensorGrid::TensorGrid(const std::vector<unsigned int> & npoints,
                        std::vector<std::unique_ptr<const Polynomial>> & poly)
 {
@@ -405,6 +464,51 @@ SmolyakGrid::gridIndex(const unsigned int n) const
   ::mooseError("Point index requested is greater than number of points.");
 
   return 0;
+}
+
+ClenshawCurtisGrid::ClenshawCurtisGrid(const unsigned int max_order,
+                                       std::vector<std::unique_ptr<const Polynomial>> & poly)
+  : _ndim(poly.size())
+{
+  // Compute full tensor tuple
+  std::vector<std::vector<unsigned int>> tuple_1d(_ndim);
+  for (unsigned int d = 0; d < _ndim; ++d)
+  {
+    tuple_1d[d].resize(max_order);
[…]
