### 2020-03-20 e055403
Adding Clenshaw-Curtis sparse grid #14833
