### 2020-03-20 fa8a903
stuff added to data/moose_testdb.json to enable testing of wacky cases
