### 2020-03-20 4a7d8bb
trying to get pictures and equations to work...
