+++ modules/xfem/doc/content/syntax/XFEM/index.md
@@ -37,7 +37,7 @@ Next, the class types replaced when using crack tip enrichment are detailed:
 !table id=near_tip_xfem_action_table caption=Correspondence Among Action Functionality and MooseObjects for `XFEMAction` when `use_crack_tip_enrichment` = `true`.
 | Functionality     | Replaced Classes   | Associated Parameters   |
 |-------------------|--------------------|-------------------------|
-| Add enrichment displacement variables | [Variables](/syntax/Variables/index.md) | `enrichment_displacements`: vector of enrichment displacement names |
+| Add enrichment displacement variables | [Variables](syntax/Variables/index.md) | `enrichment_displacements`: vector of enrichment displacement names |
 | Add stress divergence tensor kernels for crack tip enrichment | [CrackTipEnrichmentStressDivergenceTensors](/CrackTipEnrichmentStressDivergenceTensors.md) | `crack_front_definition`, `enrichment_displacements`, `displacements` |
 | Add crack tip enrichment boundary conditions | [CrackTipEnrichmentCutOffBC](/CrackTipEnrichmentCutOffBC.md) | `cut_off_boundary`, `cut_off_radius`, `crack_front_definition` |
 
