### 2020-03-20 a75d311
Fixup MooseDocs for python 3.6
(refs #14255)
