### 2020-03-20 efc7802
Skip tests that rely on unicode on python 3.6 and linux
(refs #14255)
