+++ python/TestHarness/tests/test_CSVValidationTester.py
@@ -6,12 +6,14 @@
 #*
 #* Licensed under LGPL 2.1, please see LICENSE for details
 #* https://www.gnu.org/licenses/lgpl-2.1.html
-
+import sys
+import unittest
 import subprocess
 from TestHarnessTestCase import TestHarnessTestCase
 
 class TestHarnessTester(TestHarnessTestCase):
 
+    @unittest.skipIf(sys.platform == 'linux' and sys.version_info[0] == 3 and sys.version_info[1] < 7, "Python 3.6 print doesn't handle \xb1 on linux")
     def testCSVValidationTester(self):
         """
         Test for correct operation of CSV validation tester
@@ -25,6 +27,7 @@ class TestHarnessTester(TestHarnessTestCase):
         self.assertRegexpMatches(output, r'test_harness\.csv_validation_tester_01.*?OK')
         self.assertRegexpMatches(output, r'test_harness\.csv_validation_tester_02.*?FAILED \(DIFF\)')
 
+    @unittest.skipIf(sys.platform == 'linux' and sys.version_info[0] == 3 and sys.version_info[1] < 7, "Python 3.6 print doesn't handle \xb1 on linux")
     def testCSVValidationTesterVerbose(self):
         """
         Test for correct operation of CSV validation tester in verbose mode
