### 2020-03-20 608dc4f
Change VTKWindow for change that occured in VTK 8
(refs #8807)
