### 2020-03-20 fc9c23e
Skip find neighbors
