### 2020-03-23 21439a3
add vi solver option to phase field fracture ref#14946
