### 2020-03-23 df28f2b
add vi solver option to phase field fracture ref#14946
