+++ modules/xfem/doc/content/source/actions/XFEMAction.md
@@ -1,1 +1,7 @@
-!template load file=stubs/moose_action.md.template name=XFEMAction syntax=/XFEM/XFEMAction
+# XFEM Action
+
+## Overview
+
+The XFEM Action takes in parameters that generally apply to how one wants to use XFEM and provides
+these to XFEM. See the description, example use, and parameters on the
+[XFEM Action syntax](syntax/XFEM/index.md) page.
