+++ modules/xfem/doc/content/syntax/XFEM/index.md
@@ -1,1 +1,85 @@
-!template load file=stubs/moose_system.md.template name=XFEM syntax=/XFEM
+# XFEM Action System
+
+!syntax description /XFEM/XFEMAction
+
+## Overview
+
+The `XFEM` block must be supplied to run any MOOSE simulations using XFEM. It provides general
+solution parameters to XFEM. Including this block in the input file serves several functions:
+
+- Causes an algorithm to be run at defined intervals to split the mesh based on a set of evolving defined interfaces.
+- Causes a modified quadrature rule to be used to correctly integrate elements split by these interfaces.
+- Creates objects to generate output related to the XFEM algorithm.
+- Provides a syntax block where various parameters controlling the behavior of XFEM can be specified by the user, including:
+
+  - Provide `qrule` variable to XFEM to determine which quadrature weighting rule to use.
+  - Specify `output_cut_plane` variable to output cut elements to the exodus mesh.
+  - Optional: Provide parameter values for prescribed, incremental crack growth.
+  - Optional: Setup of near tip enrichment parameters.
+  - Optional: Allow users to control the amount of debugging information printed during a simulation.
+
+## Constructed MooseObjects
+
+Three of the above user-settable parameters pass parameters to the XFEM object for use in setting up
+relevant options: the `qrule` variable, incremental crack growth, and `debug_output_level`. The
+remaining parameters result in the creation of various objects, which allows for input files to be
+more concise. The following table details the types of classes created by utilizing the
+`output_cut_plane` option in the `XFEMAction`.
+
+!table id=incr_crack_xfem_action_table caption=Correspondence Among Action Functionality and MooseObjects for `XFEMAction` when `output_cut_plane` = `true`.
+| Functionality     | Replaced Classes   | Associated Parameters   |
+|-------------------|--------------------|-------------------------|
+| Add `AuxVariables` for various cut parameters and volume fraction calculations | [AuxVariables](/AuxVariables/index.md) |   |
+| Add `AuxKernels` for various cut parameters and volume fraction calculations | [XFEMVolFracAux](/XFEMVolFracAux.md) [XFEMCutPlaneAux](/XFEMCutPlaneAux.md) |   |
+
+Several objects are instantiated when using the `output_cut_plane` option. These `AuxVariables` and
+`AuxKernels` store data defining the XFEM cut: element volume fraction as well as definitions for
+two planes (their x, y, z origins and normals). This allows the cut location to be tracked
+throughout the simulation and saves the user from having to manually set up each `AuxVariable` and
+its associated `AuxKernel` (`XFEMVolFracAux` for the volume fraction and `XFEMCutPlaneAux` for cut
+plane AuxVariables).
+
+Next, the class types created when using crack tip enrichment are detailed:
+
+!table id=near_tip_xfem_action_table caption=Correspondence Among Action Functionality and MooseObjects for `XFEMAction` when `use_crack_tip_enrichment` = `true`.
+| Functionality     | Replaced Classes   | Associated Parameters   |
+|-------------------|--------------------|-------------------------|
+| Add enrichment displacement variables | [Variables](syntax/Variables/index.md) | `enrichment_displacements`: vector of enrichment displacement names |
+| Add stress divergence tensor kernels for crack tip enrichment | [CrackTipEnrichmentStressDivergenceTensors](/CrackTipEnrichmentStressDivergenceTensors.md) | `crack_front_definition`, `enrichment_displacements`, `displacements` |
[…]
