### 2020-03-23 af928cd
Update colorbar font_size in peacock not chigger
