+++ framework/src/utils/Units.C
@@ -21,7 +21,8 @@ const std::vector<std::pair<std::string, MooseUnits>> MooseUnits::_unit_table = 
     // avoid matching meters
     {"Ohm",
      {
-         1,
+         1, // conversion factor
+         0, // additive shift (currently only used for Celsius and Fahrenheit)
          {{MooseUnits::BaseUnit::KILOGRAM, 1},
           {MooseUnits::BaseUnit::METER, 2},
           {MooseUnits::BaseUnit::SECOND, -3},
@@ -29,6 +30,7 @@ const std::vector<std::pair<std::string, MooseUnits>> MooseUnits::_unit_table = 
      }},
     {"atm",
      {101.325e3,
+      0,
       {{MooseUnits::BaseUnit::KILOGRAM, 1},
        {MooseUnits::BaseUnit::METER, -1},
        {MooseUnits::BaseUnit::SECOND, -2}}}}, // Standard atmosphere
@@ -36,6 +38,7 @@ const std::vector<std::pair<std::string, MooseUnits>> MooseUnits::_unit_table = 
     {"eV",
      {
          1.602176634e-19,
+         0,
          {{MooseUnits::BaseUnit::KILOGRAM, 1},
           {MooseUnits::BaseUnit::METER, 2},
           {MooseUnits::BaseUnit::SECOND, -2}},
@@ -44,76 +47,92 @@ const std::vector<std::pair<std::string, MooseUnits>> MooseUnits::_unit_table = 
     {"erg",
      {
          1e-7,
+         0,
          {{MooseUnits::BaseUnit::KILOGRAM, 1},
           {MooseUnits::BaseUnit::METER, 2},
           {MooseUnits::BaseUnit::SECOND, -2}},
-     }},                                                  // erg
-    {"Ang", {1e-10, {{MooseUnits::BaseUnit::METER, 1}}}}, // Angstrom
+     }}, // erg
+
+    // avoid matching Farad or Coulomb
+    {"degC", {1, 273.15, {{MooseUnits::BaseUnit::KELVIN, 1}}}},         // Celsius
+    {"degF", {5.0 / 9.0, 459.67, {{MooseUnits::BaseUnit::KELVIN, 1}}}}, // Fahrenheit
+    {"degR", {5.0 / 9.0, 0, {{MooseUnits::BaseUnit::KELVIN, 1}}}},      // Rankine
+
+    {"Ang", {1e-10, 0, {{MooseUnits::BaseUnit::METER, 1}}}}, // Angstrom
 
     // Base units
-    {"m", {1, {{MooseUnits::BaseUnit::METER, 1}}}},
-    {"g", {0.001, {{MooseUnits::BaseUnit::KILOGRAM, 1}}}},
-    {"s", {1, {{MooseUnits::BaseUnit::SECOND, 1}}}},
-    {"A", {1, {{MooseUnits::BaseUnit::AMPERE, 1}}}},
-    {"K", {1, {{MooseUnits::BaseUnit::KELVIN, 1}}}},
-    {"mol", {6.02214076e23, {{MooseUnits::BaseUnit::COUNT, 1}}}},
-    {"cd", {1, {{MooseUnits::BaseUnit::CANDELA, 1}}}},
+    {"m", {1, 0, {{MooseUnits::BaseUnit::METER, 1}}}},
+    {"g", {0.001, 0, {{MooseUnits::BaseUnit::KILOGRAM, 1}}}},
+    {"s", {1, 0, {{MooseUnits::BaseUnit::SECOND, 1}}}},
+    {"A", {1, 0, {{MooseUnits::BaseUnit::AMPERE, 1}}}},
+    {"K", {1, 0, {{MooseUnits::BaseUnit::KELVIN, 1}}}},
[…]
