+++ modules/contact/doc/content/source/constraints/SparsityBasedContactConstraint.md
@@ -1,1 +1,0 @@
-!template load file=stubs/moose_object.md.template name=SparsityBasedContactConstraint syntax=/Constraints/SparsityBasedContactConstraint
