+++ modules/contact/doc/content/source/constraints/OneDContactConstraint.md
@@ -1,1 +1,0 @@
-!template load file=stubs/moose_object.md.template name=OneDContactConstraint syntax=/Constraints/OneDContactConstraint
