### 2020-03-23 387802b
Deprecate contact constraint, issue #14813:   - GluedContactConstraint   - MultiDContactConstraint   - OneDContactConstraint   - SparsityBasedContactConstraint
