+++ modules/contact/doc/content/source/constraints/GluedContactConstraint.md
@@ -1,1 +1,0 @@
-!template load file=stubs/moose_object.md.template name=GluedContactConstraint syntax=/Constraints/GluedContactConstraint

[…]
