### 2020-03-24 b8a4d28
refactor moose variable heirarchy
