+++ framework/include/variables/MooseVariableFE.h
@@ -49,47 +49,38 @@ InputParameters validParams<ArrayMooseVariable>();
  *
  */
 template <typename OutputType>
-class MooseVariableFE : public MooseVariableFEBase
+class MooseVariableFE : public MooseVariableField<OutputType>
 {
 public:
-  // type for gradient, second and divergence of template class OutputType
-  typedef typename TensorTools::IncrementRank<OutputType>::type OutputGradient;
-  typedef typename TensorTools::IncrementRank<OutputGradient>::type OutputSecond;
-  typedef typename TensorTools::DecrementRank<OutputType>::type OutputDivergence;
+  using OutputGradient = typename MooseVariableField<OutputType>::OutputGradient;
+  using OutputSecond = typename MooseVariableField<OutputType>::OutputSecond;
+  using OutputDivergence = typename MooseVariableField<OutputType>::OutputDivergence;
 
-  // shortcut for types storing values on quadrature points
-  typedef MooseArray<OutputType> FieldVariableValue;
-  typedef MooseArray<OutputGradient> FieldVariableGradient;
-  typedef MooseArray<OutputSecond> FieldVariableSecond;
-  typedef MooseArray<OutputType> FieldVariableCurl;
-  typedef MooseArray<OutputDivergence> FieldVariableDivergence;
+  using FieldVariableValue = typename MooseVariableField<OutputType>::FieldVariableValue;
+  using FieldVariableGradient = typename MooseVariableField<OutputType>::FieldVariableGradient;
+  using FieldVariableSecond = typename MooseVariableField<OutputType>::FieldVariableSecond;
+  using FieldVariableCurl = typename MooseVariableField<OutputType>::FieldVariableCurl;
+  using FieldVariableDivergence = typename MooseVariableField<OutputType>::FieldVariableDivergence;
 
-  // shape function type for the template class OutputType
-  typedef typename Moose::ShapeType<OutputType>::type OutputShape;
+  using OutputShape = typename MooseVariableField<OutputType>::OutputShape;
+  using OutputShapeGradient = typename MooseVariableField<OutputType>::OutputShapeGradient;
+  using OutputShapeSecond = typename MooseVariableField<OutputType>::OutputShapeSecond;
+  using OutputShapeDivergence = typename MooseVariableField<OutputType>::OutputShapeDivergence;
 
-  // type for gradient, second and divergence of shape functions of template class OutputType
-  typedef typename TensorTools::IncrementRank<OutputShape>::type OutputShapeGradient;
-  typedef typename TensorTools::IncrementRank<OutputShapeGradient>::type OutputShapeSecond;
-  typedef typename TensorTools::DecrementRank<OutputShape>::type OutputShapeDivergence;
+  using FieldVariablePhiValue = typename MooseVariableField<OutputType>::FieldVariablePhiValue;
+  using FieldVariablePhiGradient = typename MooseVariableField<OutputType>::FieldVariablePhiGradient;
+  using FieldVariablePhiSecond = typename MooseVariableField<OutputType>::FieldVariablePhiSecond;
+  using FieldVariablePhiCurl = typename MooseVariableField<OutputType>::FieldVariablePhiCurl;
+  using FieldVariablePhiDivergence = typename MooseVariableField<OutputType>::FieldVariablePhiDivergence;
 
-  // shortcut for types storing shape function values on quadrature points
-  typedef MooseArray<std::vector<OutputShape>> FieldVariablePhiValue;
-  typedef MooseArray<std::vector<OutputShapeGradient>> FieldVariablePhiGradient;
-  typedef MooseArray<std::vector<OutputShapeSecond>> FieldVariablePhiSecond;
[…]
