### 2020-04-14 #2118 @jedbrown
@dflanderka @jbrezmorf @Paulie14 @janhybs Would any of you from the Flow123 project be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
