### 2020-04-15 #2118 @rpodgorney
I can provide a review.

Robert Podgorney, PhD, PG
Lead, Geothermal and Fossil Energy
Energy and Environment Division
Idaho National Laboratory

Office: 208-526-1524<tel:208-526-1524>
Cell: 208-520-9361<tel:208-520-9361>
email: robert.podgorney@inl.gov<mailto:robert.podgorney@inl.gov>

-sent from my mobile device, please excuse any typos or autocorrect errors

Statement of Confidentiality
The contents of this e-mail message and any attachments are confidential and are intended solely for addressee. The information may also be legally privileged. This transmission is sent in trust, for the sole purpose of delivery to the intended recipient. If you have received this transmission in error, any use, reproduction or dissemination of this transmission is strictly prohibited. If you are not the intended recipient, please immediately notify the sender by reply e-mail or phone and delete this message and its attachments, if any.
