### 2020-05-10 #2176 @arfon
> @jedbrown Hi Jed. Trying to proceed with the review I realized that my JOSS invitation has expired and I can not fill the checklist. Sorry, I didn't notice the invitation before.

☝️ You should be re-invited now.
