### 2020-05-11 idaholab/moose#15268 @jbrezmorf
PorousFlow JOSS review improvement tips
[PorousFlow JOSS review issue](https://github.com/openjournals/joss-reviews/issues/2176)

**Installation**
Minor issue with following [Moose instructions](https://mooseframework.org/getting_started/installation/conda.html). Activation of the conda environment using `conda activate moose` didn't work, the `source activate moose` did the job. Not clear to me as the former syntax should apply for the latest conda version I have used (following the instructions).

***Build***
Both MOOSE and PorousFlow builds fine. However, it took me some time to figure out that PorousFlow build actually makes the final MOOSE application. It would be helpful to add sort of "Get started" instructions to the first tutorial. In particular the formulation "The PorousFlow module is a library of physics for fluid" was confusing for me.

***Documentation***
Overall the documentation is complete and well structured. The tutorials cover the main applications of the framework and help a lot to begin with creation of own configuration files.

***Lava tutorial***
- 1phase case works fine and the solution seems to be physically correct, in particular, the fingering phenomenon is captured.
- using 4 threads leads to speedup 2.6 (from 20min to 7.5min), quite good
- 2phase case, can not proceed over the time 270352
  the nonlinear solver fails to converge which leads to timestep reduction down to about 0.001.
  I stop the calculation after 90min of the single thread calculation.
- The porosity is a random field, however it is not clear how the seed of a pseudorandom generator is determined. According to [the general MOOSE doc](https://mooseframework.inl.gov/old/wiki/MooseTraining/random/), it seems that the seed is not the same between application runs, but it is in contradiction to the claim that the PRNG is consistent for serial and parallel run. I found this generic page about PRNG as unsufficient to really understand the "multi-level random seeding scheme" in use. Having a command-line option to set the global application seed would be helpful.
- Please add a piece of information about the raw execution times of the tutorial problems. Knowing how demanding the lava tutorial is I would do the tests on the cluster using more cores.

***fracture transient***
Very good agreement between the 2d fracture `fine_thick_fracture_transient.i` and 1d fracture `fine_transient.i`.
