### 2020-05-14 idaholab/moose#15268 @loganharbour
I know that `conda activate` is the new and preferred method of activating Conda environments. However, we have found that `source activate` is necessary on some of our testing machines.

What OS did you counter this on and with what version of conda?
