### 2020-05-14 idaholab/moose#15268 @loganharbour
One more request: is `conda activate` simply unavailable or do you end up with an error?
