### 2020-05-15 idaholab/moose#15268 @jbrezmorf
It resulted in:
CommandNotFoundError: Your shell has not been properly configured to use 'conda activate'.

While the shell configuration steps make no difference. I'm suspicious that it can be caused by some previous installations of older conda versions.
