### 2020-05-15 idaholab/moose#15268 @loganharbour
What version of Conda (`conda -V`) and what operating system are you using?

I'm trying to replicate this so we can add some additional notes to the documentation.
