### 2020-05-17 idaholab/moose#15268 @jbrezmorf
```
 ~$ ~/miniconda3/bin/conda -V
conda 4.8.2
 ~$ lsb_release -a
No LSB modules are available.
Distributor ID: Ubuntu
Description:    Ubuntu 18.04.4 LTS
Release:        18.04
Codename:       bionic
```
