### 2020-05-17 #2176 @jbrezmorf
Software:
=======
no major issues, some tips fo improvements in
[the issue](https://github.com/idaholab/moose/issues/15268).

- fully coupled THM multiphase
- use generic Newton method for solving the system of equations
- claimed implementation of all types of THM relationships
- supported by detailed documentation both for usage as for used numerical methods

Summary:
=======
well written,  a typographic error in Acknowledgements: "Philipp Sch{”a}dle."
