### 2020-05-20 idaholab/moose#15268 @cpgr
Thanks for the thoughts on some of the examples @jbrezmorf. I'll add some notes to the documentation and input files for those examples that are more computationally expensive to make the user aware before they find out for themselves.

I'll double check the two phase convective mixing example too. It did work for me when I made it but it is possible that something has changed.

Thanks again for the suggestions!
