### 2020-05-20 idaholab/moose#15268 @WilkAndy
Thanks @cpgr.  For some reason i missed this issue completely - ooops.  I just set about fixing the umlauts in Phil's name: see #15320 (changed `{"a}` to `{\"a}` in the Acknowledgements).  So you don't need me to address any of the reviewer comments?

Remember to put in a PR to the `joss_porous_flow` branch - not to `next`.
