### 2020-05-20 idaholab/moose#15268 @cpgr
Oh, I would have pushed to `next` if you hadn't reminded me!

I'll look after all them today or tonight. I did notice that the paper is missing some capitilisation in the references for conference proceedings - things like the country names being lowercase. Can you fix them when you do #15320?
