### 2020-05-20 idaholab/moose#15268 @WilkAndy
If you want to check that the paper looks OK, go to https://whedon.theoj.org/ and enter your moose github repo address and your branch name.

I couldn't get the umlauts to work on Schaedle - i remember mucking around with that at the time of submission too.  The umlauts work OK in the bib file with the `{\"a}` method.  I know that Phil spells his name Schaedle (that is `ae` instead of `{\"a}`) so I think we should use that in the Acknowledgements, unless you can get the umlauts working.
