### 2020-05-20 idaholab/moose#15268 @WilkAndy
I'm a bit snowed under with a lot of admin stuff right now.  Unless you want to combine #15320 and #15268 into one PR, let's leave the umlauts and capitalisation until @rpodgorney has also done his review.  He might notice other things that need tweaking.   Just note the stuff you haven't done here or in #15320 and i'll address them in a couple of days.
