### 2020-05-20 4382a25
Update lava lamp documentation
Refs #15268
