### 2020-05-20 #2176 @arfon
:wave: @rpodgorney - today we [reopened JOSS](https://blog.joss.theoj.org/2020/05/reopening-joss) for new submissions and are checking in on our existing reviews. Do you think you might be able to wrap up your review in the next 2-3 weeks?
