### 2020-06-21 a046e2b
Fix spelling in paper.md
Closes #15320
