### 2020-06-22 #2176 @rpodgorney
Sorry for letting this slip.  All in all no issues at all, the code, tests, and documentation are al superb.

The paper is well written and contains just the right amount of details without becoming overly verbose.
