### 2020-06-22 #2176 @rpodgorney
@arfon this is my first review for JOSS.  I've looked through the code, documentation, and tests -- and feel it is OK to publish.  Other than the checklist, is there anything else I need to do to complete/document my approval?
