### 2020-06-22 #2176 @arfon
> @arfon this is my first review for JOSS. I've looked through the code, documentation, and tests -- and feel it is OK to publish. Other than the checklist, is there anything else I need to do to complete/document my approval?

Nope, that's it. Over to @jedbrown to decide on next steps.
