### 2020-07-03 #2176 @jedbrown
Sorry about my delay. This is looking great.
1. There is a period missing after the Lime reference (end of paragraph).
2. "simply solve simplistic" -> "only solve simplistic" or some other phrasing.
3. Could you make any comments in the comparison with other software, perhaps in terms of capability, performance, and easy of use/extension?
4. You're welcome to include a figure if you'd like.
5. Would it make sense to write down the basic equations that are solved? Perhaps one or two lines for each of THMC just to get the structural gist across, or a figure representing the role and coupling. (It's sort of abrupt to launch directly into prior applications).
6. Can you double-check capitalization in your references? Our toolchain uses pandoc-citeproc, which is based on apa.csl. Its semantics are closer to biblatex than bibtex, which is required to be precisely APA-conforming. Anyway, there are extra braces needed, e.g., `{I}celand` or `{Iceland}` are needed.
