### 2020-07-07 idaholab/moose#15539 @WilkAndy
Some JOSS PorousFlow changes
## Reason

The suggestions below come from the JOSS editor ( https://github.com/openjournals/joss-reviews/issues/2176#issuecomment-653517195 ) .   They need to be done in porous_flow/paper.md and porous_flow/paper.bib

## Design

- [x] There is a period missing after the Lime reference (end of paragraph).
- [x] "simply solve simplistic" -> "only solve simplistic" or some other phrasing.
- [x] Could you make any comments in the comparison with other software, perhaps in terms of capability, performance, and easy of use/extension?
- [x] You're welcome to include a figure if you'd like.
- [x] Would it make sense to write down the basic equations that are solved? Perhaps one or two lines for each of THMC just to get the structural gist across, or a figure representing the role and coupling. (It's sort of abrupt to launch directly into prior applications).
- [x] Can you double-check capitalization in your references? Our toolchain uses pandoc-citeproc, which is based on apa.csl. Its semantics are closer to biblatex than bibtex, which is required to be precisely APA-conforming. Anyway, there are extra braces needed, e.g., `{I}celand` or `{Iceland}` are needed.

Also:

```
Reference check summary:

OK DOIs

- 10.1007/s12665-012-1546-x is OK
- 10.1007/s00366-006-0049-3 is OK
- 10.1080/22020586.2019.12073198 is OK

MISSING DOIs

- https://doi.org/10.1016/j.softx.2020.100430 may be missing for title: MOOSE: Enabling Massively Parallel Multiphysics Simulation
- https://doi.org/10.21203/rs.2.17644/v1 may be missing for title: Impact of Effective Normal Stress on Capillary Pressure in a Single Natural Fracture

INVALID DOIs

- None
```
