### 2020-07-07 bcf4ac2
PorousFlow JOSS changes.  Refs #15539
- Regarding the equations: I've included some, although we wanted this paper to be readable by a more general audience, so have included the equations towards the end of the paper.

- Regarding comparisons with other codes: I've included another paragraph that discusses some of the benefits of using MOOSE+PorousFlow.  It's really hard to write this sort of thing without appearing to denigrate other very worthy codes!

- Done the DOIs mentioned above
