+++ modules/porous_flow/paper.bib
@@ -2,7 +2,7 @@
 title={{Groundwater numerical modelling for the Hunter subregion. Product 2.6.2 for the Hunter subregion from the Northern Sydney Basin Bioregional Assessment}},
 author={Herron, N. F. and Peeters, L. and Crosbie, R. and Marvanek, S. P. and Pagendam, D. and Ramage, A. and Rachakonda, P. K. and Wilkins, A.},
 year={2018},
-institution={Department of the Environment and Energy, Bureau of Meteorology, CSIRO and Geoscience Australia, Canberra},
+institution={{Department of the Environment and Energy, Bureau of Meteorology, CSIRO and Geoscience Australia, Canberra}},
 url={http://data.bioregionalassessments.gov.au/product/NSB/HUN/2.6.2}
 }
 
@@ -20,7 +20,7 @@ doi = {10.1007/s12665-012-1546-x}
 
 @TechReport{pflotran,
 Author = "Peter C. Lichtner and Glenn E. Hammond and Chuan Lu and Satish Karra and Gautam Bisht and Benjamin Andre and Richard T. Mills and Jitendra Kumar and Jennifer M. Frederick",
-Title = "{PFLOTRAN} User Manual",
+Title = "{PFLOTRAN} {U}ser {M}anual",
 Note = "http://documentation.pflotran.org",
 institution = "Sandia National Laboratories",
 Year = "2019" }
@@ -33,13 +33,19 @@ url = {http://www.opm-project.org}
 }
 
 
-@misc{permann2019moose,
-    title={MOOSE: Enabling Massively Parallel Multiphysics Simulation},
-    author={Cody J. Permann and Derek R. Gaston and David Andrs and Robert W. Carlsen and Fande Kong and Alexander D. Lindsay and Jason M. Miller and John W. Peterson and Andrew E. Slaughter and Roy H. Stogner and Richard C. Martineau},
-    year={2019},
-    eprint={1911.04488},
-    archivePrefix={arXiv},
-    primaryClass={cs.MS}
+@article{permann2020moose,
+   title = {{MOOSE}: Enabling massively parallel multiphysics simulation},
+  author = {Cody J. Permann and Derek R. Gaston and David Andr{\v{s}} and Robert W. Carlsen and Fande
+            Kong and Alexander D. Lindsay and Jason M. Miller and John W. Peterson and Andrew
+            E. Slaughter and Roy H. Stogner and Richard C. Martineau},
+    year = {2020},
+ journal = {{SoftwareX}},
+  volume = {11},
+   pages = {100430},
+    issn = {2352-7110},
+     doi = {https://doi.org/10.1016/j.softx.2020.100430},
+     url = {http://www.sciencedirect.com/science/article/pii/S2352711019302973},
+keywords = {Framework, Finite-element, Parallel, Multiphysics, Multiscale}
 }
 
 @article{libmesh,
@@ -55,7 +61,7 @@ doi = {10.1007/s00366-006-0049-3}
 }
 
 @techreport{petsc,
-title={{PETSc} Users Manual},
+title={{PETSc Users Manual}},
 author={Balay, Satish and Abhyankar, Shrirang and Adams, Mark and Brown, Jed and Brune, Peter and Buschelman, Kris and Dalcin, LD and Eijkhout, Victor and Gropp, W and Kaushik, Dinesh and others},
 year={2019},
 institution={Argonne National Lab. (ANL), Argonne, IL (United States)}
@@ -78,9 +84,9 @@ url={https://www.acarp.com.au/abstracts.aspx?repId=C25065}
 }
 
 @inproceedings{green2018,
[…]
