+++ modules/porous_flow/paper.md
@@ -23,7 +23,7 @@ date: 13 February 2020
 bibliography: paper.bib
 ---
 
-# Summary
+# Summary and applications
 
 ``PorousFlow`` enables simulation of transport and flow in porous media.  These are important in many practical fields of research.  For instance, ``PorousFlow`` has been used in the following applications.
 
@@ -31,20 +31,64 @@ bibliography: paper.bib
 
 - Geothermal modelling: the simulation of natural and man-made geothermal systems, typically in order to quantify heat energies that may be extracted [@birdsall; @guglielmetti; @lima; @niederau].  Models in this field involve the coupled transport of heat and fluid.  Many models are often large-scale, but some focus on small regions close to wellbores to investigate wellbore integrity.  In this field, it often is useful to assess the impacts of subsurface heterogeneity, realistic pumping regimes, geochemical effects, fracture flow, and the creation of new fractures through deliberate or unintentional hydrofracturing, as well as potential microseismicity.
 
-- Multiphase modelling, such as the simulation of CO$_{2}$ sequestration [@green2018; @lima], H$_{2}$ storage, or methane flows around coal mines [@c26050; @c25065].  These simulations involve the interaction and flow of at least two fluid phases (e.g. gas and water).  Changes in temperature often impact these flows, and realistic equations of state for the fluids involved are necessary.  For instance, CO$_{2}$ sequestration typically involves pumping high-pressure, cold CO$_{2}$ into a warmer aquifer filled with brine: the CO$_{2}$ dissolves into the brine, cooling it and changing its density and flow characteristics.  In many of these models, the altered pressure and temperature can potentially cause mechanical failure of the subsurface, and hence the creation of new flow paths and microseismicity [@lima]
+- Multiphase modelling, such as the simulation of CO$_{2}$ sequestration [@green2018; @lima], H$_{2}$ storage, or methane flows around coal mines [@c26050; @c25065].  These simulations involve the interaction and flow of at least two fluid phases (e.g. gas and water).  Changes in temperature often impact these flows, and realistic equations of state for the fluids involved are necessary.  For instance, CO$_{2}$ sequestration typically involves pumping high-pressure, cold CO$_{2}$ into a warmer aquifer filled with brine: the CO$_{2}$ dissolves into the brine, cooling it and changing its density and flow characteristics.  In many of these models, the altered pressure and temperature can potentially cause mechanical failure of the subsurface, and hence the creation of new flow paths and microseismicity [@lima].
 
 - Understanding mineralization [@heather] in the subsurface often involves fluid flow, heat flow and coupled 3D deformation, simulating over millennia of real time and using realistic, and therefore complicated, basin-scale geology.
 
 Other situations where ``PorousFlow`` could be used include oil and gas production, nuclear waste repositories and chemical leaching.
 
-In many of these cases it is not enough to simply solve simplistic, traditional flow and transport physics.  Coupling with other physics is necessary: solid mechanical deformations and stresses can be important; geochemistry can alter the flow characteristics of the subsurface; high-precision equations of state are required; the evolution of fluid components (tracers, pollutants, reactants) is of interest; unsaturated physics is crucial.  The flow and transport often impacts the subsurface structure, and conversely, the effects of these structure-changes impact the flow and transport.
[…]
