### 2020-07-07 #2176 @WilkAndy
Hi @jedbrown .   I've just submitted a PullRequest to MOOSE for the changes you request above.   Perhaps you'd like to review the changes there?   https://github.com/idaholab/moose/pull/15540

- Regarding the equations: I've included some, although we wanted this paper to be readable by a more general audience, so have included the equations towards the end of the paper.

- Regarding comparisons with other codes: I've included another paragraph that discusses some of the benefits of using MOOSE+PorousFlow. It's really hard to write this sort of thing without appearing to denigrate other very worthy codes!

- Done the DOIs mentioned above
