+++ modules/porous_flow/paper.md
@@ -37,7 +37,7 @@ bibliography: paper.bib
 
 Other situations where ``PorousFlow`` could be used include oil and gas production, nuclear waste repositories and chemical leaching.
 
-In many of these cases it is not enough to only solve simplistic, traditional flow and transport physics.  Coupling with other physics is necessary: solid mechanical deformations and stresses can be important; geochemistry can alter the flow characteristics of the subsurface; high-precision equations of state are required; the evolution of fluid components (tracers, pollutants, reactants) is of interest; unsaturated physics is crucial.  The flow and transport often impacts the subsurface structure, and conversely, the effects of these structure-changes impact the flow and transport.
+In many of these cases it is not enough to only solve traditional flow and transport physics.  Coupling with other physics is necessary: solid mechanical deformations and stresses can be important; geochemistry can alter the flow characteristics of the subsurface; high-precision equations of state are required; the evolution of fluid components (tracers, pollutants, reactants) is of interest; unsaturated physics is crucial.  The flow and transport often impacts the subsurface structure, and conversely, the effects of these structure-changes impact the flow and transport.
 
 Simulating these situations can be challenging in practice.  Many widely-used simulation codes typically focus on only a subset of physical processes, and therefore accurate modelling often requires the use of several loosely-coupled software packages.  Such an approach can be useful in many problems of interest, particularly where the characteristic time scales for each physical process are sufficiently different.  In other instances, however, the ability to tightly couple different physical processes is necessary due to comparable characteristic time scales for the different physical processes.
 ``PorousFlow`` allows simulation of all the relevant thermal-hydraulic-mechanical-chemical (THMC) physical phenomenon in a tightly-coupled framework.
@@ -86,7 +86,7 @@ A comprehensive listing of equations, including fluid equations of state, the ev
 
 Other open-source codes that can also perform tightly-coupled THMC simulations to some degree include OpenGeoSys [@opengeosys], PFLOTRAN [@pflotran] and work under the Open Porous Media initiative [@opm].
 
-Harnessing the power of MOOSE means that ``PorousFlow`` simulations efficiently use multiprocessors and automatically utilize threading/OpenMP, without developers or users having to worry about the software technicalities.  Mesh adaptivity with a variety of integration schemes and unstructured meshes may be used.  Various time-stepping schemes may be employed.  A variety of I/O formats are available, such as Exodus and VTK files as well as CSV plaintext.  The flexible plug-in architecture makes it easy for novice developers to solve new physics equations or couple together existing modules to solve complex problems.  Users may even utilising the "MultiApp" system that allows simulations to be embedded inside other simulations.  A GUI helps users create, run and analyse models.  An example ``PorousFlow`` result is shown in \autoref{fig:example}.
+One aspect that makes ``PorousFlow`` different to these codes is the ease of coupling additional physics to the base flow and transport capabilities, and the relative speed and simplicity of developing new capability using the APIs provided by MOOSE.  Harnessing the power of MOOSE means that ``PorousFlow`` simulations efficiently use multiprocessors and automatically utilize threading/OpenMP, without developers or users having to worry about the software technicalities.  From the user's perspective, mesh adaptivity with a variety of integration schemes and unstructured meshes may be used.  Various time-stepping schemes may be employed.  A variety of I/O formats are available, such as Exodus and VTK files as well as CSV plaintext.  Users may utilize the "MultiApp" system that allows simulations to be embedded inside other simulations.  A GUI helps users create, run and analyse models.  An example ``PorousFlow`` result is shown in \autoref{fig:example}.
 
[…]
