### 2020-07-10 #2176 @jedbrown
@openjournals/dev I see the paper.md in the same place. Is this a Whedon issue?
