### 2020-07-10 #2176 @jedbrown
While you're at it, could you avoid repeating the repository URL (which already exists in the sidebar)? Also, it's a bit unusual to launch into a bullet list of applications after only one short sentence saying what the package does. I suppose I don't object to the present organization, but an EiC might.

I have a question outside the context of the review: is the global conservation exact (to solver tolerance) or only up to discretization error? The skeleton velocity term not being in divergence form looks suspicious to me.
