### 2020-07-11 idaholab/moose#15591 @WilkAndy
porousflow JOSS paper: minor mods
## Reason

Need to make minor mods to the PorousFlow JOSS paper

## Design

`https://doi.org/10.1016/j.softx.2020.100430 is INVALID because of 'https://doi.org/' prefix`

`While you're at it, could you avoid repeating the repository URL (which already exists in the sidebar)? Also, it's a bit unusual to launch into a bullet list of applications after only one short sentence saying what the package does. I suppose I don't object to the present organization, but an EiC might.`
