### 2020-07-23 #2176 @WilkAndy
Hi @jedbrown

- i believe i've addressed your issues raised above
- i have checked the article proof - it looks fine
- the reference check is OK.

Good to publish?

andy
