### 2020-08-19 #2176 @WilkAndy
Hi @jedbrown .  Just wondering about the status of this JOSS paper.   I think we've addressed all referee and editorial iteams and it's read to go.  If not, please tell me what needs to be done.
