### 2020-08-30 #2176 @jedbrown
Sorry about my delays. It's looking great.
* Could you note that a Lagrangian mesh is used (so the mesh moves with `v_s`; this is done at the very start of the website docs, but didn't make it here yet)?
* Is it necessary to switch from operator to index notation for eq 2 and 3? (Perhaps yes if you don't like `:` for contraction over both indices, but I want to make sure it's an affirmative choice.)
* I think https://doi.org/10.1016/j.camwa.2020.05.014 is a preferred citation for OPM.
* "THMC" is only used twice and can just be spelled out.

Once you've addressed these, please:
- [ ] Archive the reviewed software in Zenodo or a similar DOI-granting service (e.g., figshare, an institutional repository)
