### 2020-08-30 #2176 @jedbrown
FWIW, these small changes to the paper do not need to appear in the archived version, but it should be a tagged release that includes any substantive changes resulting from the review. I see [only one tag](https://github.com/idaholab/moose/tags) in the MOOSE repository; let us know if tagging a release (`v1.0.1` or some such) is problematic.
