### 2020-10-05 #2176 @kthyng
Hi all! I am the rotating editor in chief this week. I see it has been awhile since there has been activity on this submission, but it looks pretty much done. @jedbrown what steps are left before you'd like me to take over to finish publishing? @WilkAndy have you finished the items listed just above here?
