### 2020-10-05 #2176 @WilkAndy
Hi @kthyng , I just got back from holidays and am addressing a backlog of TODO items.  I have not implemented the items mentioned above by @jedbrown, but hope to do them this week.  However, because of a disastrous computer crash at Idaho National Labs, the whole MOOSE project has been severely impacted so I don't expect the changes, tagging, etc, will be reviewed and merged within the week, so I think you can safely forget about it during your editor-in-chief week.
