### 2020-10-07 idaholab/moose#15906 @WilkAndy
PorousFlow JOSS paper enhancements
## Reason

Need to address the items at

https://github.com/openjournals/joss-reviews/issues/2176#issuecomment-629848519

## Design

This is what the editor says:

> Sorry about my delays. It's looking great.
>
>     * Could you note that a Lagrangian mesh is used (so the mesh moves with `v_s`; this is done at the very start of the website docs, but didn't make it here yet)?
>
>     * Is it necessary to switch from operator to index notation for eq 2 and 3? (Perhaps yes if you don't like `:` for contraction over both indices, but I want to make sure it's an affirmative choice.)
>
>     * I think https://doi.org/10.1016/j.camwa.2020.05.014 is a preferred citation for OPM.
>
>     * "THMC" is only used twice and can just be spelled out.
>
>
> Once you've addressed these, please:
>
>     * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.
>
>     * [ ]  Archive the reviewed software in Zenodo or a similar DOI-granting service (e.g., figshare, an institutional repository)
>
>     * [ ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
>
>     * [ ]  Please list the DOI of the archived version here.
>
>
> I can then move forward with accepting the submission.
