+++ modules/porous_flow/paper.bib
@@ -25,10 +25,12 @@ Note = "http://documentation.pflotran.org",
 institution = "Sandia National Laboratories",
 Year = "2019" }
 
-@misc{opm,
-title = {{Open Porous Media: http://www.opm-project.org}},
-key = {Open Porous Media},
-year = {2019},
+@article{opm,
+title = {{The Open Porous Media flow reservoir simulator}},
+author = {Atgeirr F. Rasmussen and Tor H. Sandve and Kai Bao and Andreas Lauser and Joakim Hove and B{\aa}rd Skaflestad and Robert Kl{\"o}fkorn and Markus Blatt and Alf B. Rustad and Ove Saevareid and Knut-Andreas Lie and Andreas Thune},
+journal = {Computers & Mathematics with Applications},
+year = {2020},
+doi = {10.1016/j.camwa.2020.05.014},
 url = {http://www.opm-project.org}
 }
 
