### 2020-10-28 #2176 @WilkAndy
@jedbrown 's final suggestions have been implemented and merged into the master branch

tag = https://github.com/idaholab/moose/releases/tag/snapshot-20-10-27   (ie, tag name is snapshot-20-10-27 )

https://zenodo.org/record/4071026#.X5kq-9sRXOQ

Metadata is correct

doi = 10.5281/zenodo.4071026
