### 2020-11-26 #2176 @WilkAndy
Hi @kthyng .   I'm pinging you about this JOSS paper.  I'm pretty sure it's ready to publish, but @jedbrown isn't responding.  I think he might have accidentally unsubscribed to this issue.   Perhaps you can take it from here?
