### 2020-11-26 #2176 @jedbrown
@WilkAndy It looks like the bib entry for that should be update. (It looks like the `21--26` is actually a date, so maybe you can check the rest of the formatting when you add the DOI.)

While you're at it, could you add the URL https://documentation.pflotran.org/ for PFLOTRAN (as recommended in the bibtex on their website)?

You can just put those small changes in your branch and we'll rebuild the PDF.
