### 2020-11-26 idaholab/moose#16312 @WilkAndy
PorousFlow: JOSS paper bib changes
## Reason

JOSS editors require some minor changes: see https://github.com/openjournals/joss-reviews/issues/2176 .

## Design

Change a date and add a URL in `porous_flow/paper.bib`
