### 2020-11-26 #2176 @WilkAndy
Hi @jedbrown.   Thanks for all your work on this.  I've made the changes your requested above and the result can be found at

https://github.com/WilkAndy/moose    branch: joss_bib_16312

You may generate the PDF from there (I have checked it works) or you may wait for my PR (https://github.com/idaholab/moose/pull/16313) to make its way through the MOOSE process, which will 1day - 1week.
