### 2020-11-26 #2176 @jedbrown
Hmm, I see this branch is in a different repository (https://github.com/WilkAndy/moose/tree/joss_bib_16312) and I don't know how to ask Whedon to build from that. But this paper should be ready for acceptance with either a manual build or once it lands in a branch of the main repository.
