### 2020-11-26 #2176 @arfon
@jedbrown - is this (☝️) what you were wanting to do with the custom branch?

@WilkAndy - has the DOI suggested by Whedon above (`10.2139/ssrn.3365603`) been determined to be correct (or incorrect). If it looks correct for the reference item, please add it to your BibTeX file.
