### 2020-11-26 #2176 @WilkAndy
No, the above is still the old version.  It is not from https://github.com/WilkAndy/moose/tree/joss_bib_16312 .    I'm happy to wait till the bib changes make their way through the review process at https://github.com/idaholab/moose/pull/16313
